### 2026-05-09 645738a
fix: ASAN check to pass --no-build-vignettes into the build_arg
