### 2026-05-09 8b974cb
fix: upgraded extendr-api to 0.9.0, pass --no-build-vignettes to rcmdcheck in ASAN
