### 2026-05-09 589d769
docs: prerequisites for different bindings and platforms
