### 2026-05-10 13b7c33
docs: update the versions of the documentation dependencies
