+++ docs/requirements.txt
@@ -1,3 +1,3 @@
 # MkDocs Material documentation dependencies
-mkdocs-material>=9.5
-mkdocs-git-revision-date-localized-plugin>=1.2
+mkdocs-material>=9.7
+mkdocs-git-revision-date-localized-plugin>=1.5
