### 2026-05-10 61b5cd9
build(r): fix Windows linkage/linkage-errors, normalize Rd style, and enable full pkgdown site generation
