+++ bindings/r/tests/testthat.R
@@ -1,1 +1,4 @@
-../../../tests/r/testthat.R
+library(testthat)
+library(rfastlowess)
+
+test_check("rfastlowess")
