+++ bindings/r/src/lib.rs
@@ -9,6 +9,9 @@
 
 use extendr_api::prelude::*;
 
+// Provide the Result alias that was removed from extendr_api::prelude in 0.9.0
+type Result<T> = std::result::Result<T, extendr_api::Error>;
+
 use fastLowess::internals::api::{
     BoundaryPolicy, RobustnessMethod,
     ScalingMethod::{self, MAD, MAR, Mean},
