### 2026-05-10 fefeb37
fix: add R build to R CMD Check workflow
