+++ bindings/r/tests/testthat.R
@@ -1,3 +1,7 @@
+# extendr S3 registration requires NOT_CRAN=true.
+# Set it here so covr/pkgcheck always measure coverage correctly.
+Sys.setenv(NOT_CRAN = "true")
+
 library(testthat)
 library(rfastlowess)
 
