+++ dev/prepare_cran.sh
@@ -6,14 +6,14 @@ echo "📦 Preparing package for CRAN submission..."
 # 1. Extract vendor archive if needed
 echo "   -> Extracting vendored dependencies..."
 if [ -f src/vendor.tar.xz ] && [ ! -d src/vendor ]; then
-    (cd src && tar -xf vendor.tar.xz)
+	(cd src && tar -xf vendor.tar.xz)
 fi
 
 # 2. Ensure cargo config exists
 mkdir -p src/cargo
 if [ ! -f src/cargo/config.toml ]; then
-    echo "   -> Creating cargo config..."
-    cat > src/cargo/config.toml << 'EOF'
+	echo "   -> Creating cargo config..."
+	cat >src/cargo/config.toml <<'EOF'
 [source.crates-io]
 replace-with = "vendored-sources"
 
@@ -25,18 +25,19 @@ fi
 # 3. Copy shared R tests
 echo "   -> Copying shared R tests..."
 rm -rf tests/testthat
+rm -f tests/testthat.R
 mkdir -p tests/testthat
 if [ -d "../../tests/r/testthat" ]; then
-    cp ../../tests/r/testthat/*.R tests/testthat/
-    cp ../../tests/r/testthat.R tests/testthat.R
+	cp ../../tests/r/testthat/*.R tests/testthat/
+	cp ../../tests/r/testthat.R tests/testthat.R
 elif [ -d "tests/testthat" ]; then
-    echo "      (Tests already present or root tests missing)"
+	echo "      (Tests already present or root tests missing)"
 fi
 
 # 4. Generate AUTHORS file
 echo "   -> Generating inst/AUTHORS..."
 mkdir -p inst
-(cd src && cargo metadata --locked --format-version 1 > ../cargo_metadata_temp.json)
+(cd src && cargo metadata --locked --format-version 1 >../cargo_metadata_temp.json)
 
 python3 -c '
 import json
