### 2026-05-10 d3d7a45
fix: ASAN workflow is now explicit about loading Rust
