### 2026-05-10 bb0980d
build(cpp): adopt camelBack API, clean lint/FFI issues, add CMake package support, and document Windows usage
