+++ tests/cpp/test_fastlowess.cpp
@@ -1,333 +1,461 @@
 #include "../../bindings/cpp/include/fastlowess.hpp"
-#include <cassert>
+
+#include <array>
 #include <cmath>
+#include <cstdio>
+#include <cstdlib>
+#include <exception>
 #include <iostream>
 #include <string>
 #include <vector>
 
-using namespace fastlowess;
+namespace {
 
-// Helper to check approximate equality
-bool is_approx(double a, double b, double epsilon = 1e-10) {
-  if (std::isnan(a) && std::isnan(b))
+constexpr double k_default_epsilon = 1e-10;
+constexpr double k_basic_fraction = 0.5;
+constexpr double k_robust_fraction = 0.7;
+constexpr double k_confidence_level = 0.95;
+constexpr double k_streaming_return_all_fraction = 0.3;
+constexpr double k_streaming_basic_fraction = 0.1;
+constexpr double k_linear_slope = 2.0;
+constexpr double k_linear_intercept = 1.0;
+constexpr double k_zero_intercept = 0.0;
+constexpr double k_interval_max_x = 10.0;
+constexpr double k_streaming_return_all_max_x = 100.0;
+constexpr double k_streaming_basic_max_x = 1000.0;
+constexpr double k_streaming_basic_sine_divisor = 100.0;
+constexpr double k_streaming_accuracy_max_x = 100.0;
+constexpr std::size_t k_small_sample_size = 5U;
+constexpr std::size_t k_interval_point_count = 20U;
+constexpr std::size_t k_streaming_return_all_point_count = 100U;
+constexpr std::size_t k_streaming_return_all_chunk_size = 5000U;
+constexpr std::size_t k_streaming_basic_point_count = 2000U;
+constexpr std::size_t k_streaming_basic_chunk_size = 1000U;
+constexpr std::size_t k_streaming_accuracy_point_count = 200U;
+constexpr std::size_t k_streaming_accuracy_chunk_size = 1000U;
+constexpr std::size_t k_online_window_capacity = 10U;
+constexpr std::size_t k_online_min_points = 3U;
+constexpr int k_robust_iterations = 3;
+
+constexpr std::array<double, k_small_sample_size> k_sample_x_values = {
+    1.0, 2.0, 3.0, 4.0, 5.0};
+constexpr std::array<double, k_small_sample_size> k_sample_y_values = {
+    2.0, 4.1, 5.9, 8.2, 9.8};
+constexpr std::array<double, k_small_sample_size> k_outlier_y_values = {
+    2.0, 4.1, 100.0, 8.2, 9.8};
+constexpr std::array<double, k_small_sample_size> k_reuse_x_values = {
+    10.0, 20.0, 30.0, 40.0, 50.0};
+constexpr std::array<double, k_small_sample_size> k_reuse_y_values = {
+    20.0, 40.0, 60.0, 80.0, 100.0};
+constexpr std::array<double, k_online_window_capacity> k_online_x_values = {
+    1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0};
+constexpr std::array<double, k_online_window_capacity> k_online_y_values = {
+    2.0, 4.0, 6.0, 8.0, 10.0, 12.0, 14.0, 16.0, 18.0, 20.0};
+
+struct SeriesData {
+  std::vector<double> x_values;
+  std::vector<double> y_values;
+};
+
+struct LinearSeriesSpec {
+  std::size_t point_count;
+  double max_x_value;
+  double slope;
+  double intercept;
+};
+
+struct SineSeriesSpec {
+  std::size_t point_count;
+  double max_x_value;
+  double divisor;
+};
+
+template <std::size_t ElementCount>
[…]
