+++ examples/cpp/streaming_smoothing.cpp
@@ -5,82 +5,116 @@
  * Demonstrates chunk-based processing for large datasets.
  */
 
+#include <algorithm>
 #include <cmath>
+#include <cstddef>
+#include <cstdio>
+#include <exception>
 #include <iostream>
 #include <random>
 #include <vector>
 
 #include "../../bindings/cpp/include/fastlowess.hpp"
 
+namespace {
+
+constexpr size_t k_point_count = 10000;
+constexpr unsigned int k_random_seed = 42;
+constexpr double k_noise_std_dev = 0.5;
+constexpr double k_sine_divisor = 10.0;
+constexpr double k_scale_divisor = 50.0;
+constexpr double k_fraction = 0.1;
+constexpr int k_chunk_size = 1000;
+constexpr int k_overlap = 100;
+constexpr size_t k_progress_interval = 2000;
+
+} // namespace
+
 int main() {
-  std::cout << "=== Streaming LOWESS Smoothing Example ===" << std::endl;
-
-  // Generate large synthetic dataset
-  const size_t n = 10000;
-  std::vector<double> x(n), y(n);
-
-  std::mt19937 gen(42);
-  std::normal_distribution<> noise(0, 0.5);
-
-  for (size_t i = 0; i < n; ++i) {
-    x[i] = static_cast<double>(i) / 100.0;
-    y[i] = std::sin(x[i] / 10.0) * x[i] / 50.0 + noise(gen);
-  }
-
-  std::cout << "Generated " << n << " data points" << std::endl;
-
-  // Streaming smoothing
   try {
+    std::cout << "=== Streaming LOWESS Smoothing Example ===\n";
+
+    // Generate large synthetic dataset
+    const size_t point_count = k_point_count;
+    std::vector<double> x_values(point_count);
+    std::vector<double> y_values(point_count);
+
+    std::seed_seq generator_seed = {k_random_seed, k_random_seed, k_random_seed,
+                                    k_random_seed};
+    std::mt19937 generator(generator_seed);
+    std::normal_distribution<> noise(0.0, k_noise_std_dev);
+
+    for (size_t point_index = 0; point_index < point_count; ++point_index) {
+      x_values[point_index] = static_cast<double>(point_index) / k_overlap;
+      y_values[point_index] =
+          ((std::sin(x_values[point_index] / k_sine_divisor) *
+            x_values[point_index]) /
+           k_scale_divisor) +
+          noise(generator);
+    }
+
+    std::cout << "Generated " << point_count << " data points\n";
+
+    // Streaming smoothing
     fastlowess::StreamingOptions opts;
-    opts.fraction = 0.1;
+    opts.fraction = k_fraction;
     opts.iterations = 2;
-    opts.chunk_size = 1000;
-    opts.overlap = 100;
+    opts.chunk_size = k_chunk_size;
+    opts.overlap = k_overlap;
     opts.return_diagnostics = true;
 
     std::cout << "\nProcessing with chunk_size=" << opts.chunk_size
-              << ", overlap=" << opts.overlap << std::endl;
+              << ", overlap=" << opts.overlap << '\n';
 
     fastlowess::StreamingLowess model(opts);
 
-    std::cout << "\nProcessing data in chunks..." << std::endl;
+    std::cout << "\nProcessing data in chunks...\n";
 
-    size_t chunk_size = opts.chunk_size;
+    const size_t chunk_size = static_cast<size_t>(opts.chunk_size);
     size_t total_processed = 0;
 
-    for (size_t i = 0; i < n; i += chunk_size) {
-      size_t current_chunk_len = std::min(chunk_size, n - i);
-      std::vector<double> x_chunk(x.begin() + i,
-                                  x.begin() + i + current_chunk_len);
-      std::vector<double> y_chunk(y.begin() + i,
[…]
