+++ bindings/cpp/build.rs
@@ -20,6 +20,32 @@ fn main() {
         .expect("Unable to generate bindings")
         .write_to_file(&output_file);
 
+    normalize_generated_header(&output_file);
+
     println!("cargo:rerun-if-changed=src/lib.rs");
     println!("cargo:rerun-if-changed=cbindgen.toml");
+}
+
+fn normalize_generated_header(output_file: &PathBuf) {
+    let header = std::fs::read_to_string(output_file).expect("Unable to read generated bindings");
+
+    let normalized = header
+        .replace(
+            "#include <cstdarg>\n#include <cstdint>\n#include <cstdlib>\n#include <ostream>\n#include <new>\n\n",
+            "",
+        )
+        .replace(
+            "`ptr` must be a valid CppLowess pointer. `x` and `y` must be valid arrays of length `n`.",
+            "`ptr` must be a valid CppLowess pointer. `x_values` and `y_values` must be valid arrays of length `n`.",
+        )
+        .replace(
+            "`ptr` must be valid. `x` and `y` must be valid arrays of length `n`.",
+            "`ptr` must be valid. `x_values` and `y_values` must be valid arrays of length `n`.",
+        )
+        .replace("const double *x,", "const double *x_values,")
+        .replace("const double *y,", "const double *y_values,");
+
+    if normalized != header {
+        std::fs::write(output_file, normalized).expect("Unable to write normalized bindings");
+    }
 }
