+++ bindings/cpp/cmake/fastlowessConfig.cmake.in
@@ -1,0 +1,5 @@
+@PACKAGE_INIT@
+
+include("${CMAKE_CURRENT_LIST_DIR}/fastlowessTargets.cmake")
+
+check_required_components(fastlowess)
