### 2026-05-10 b9e615d
fix(r): hardened looks for rustup env files in ASAN
