+++ .github/workflows/R-CMD-check.yaml
@@ -118,15 +118,23 @@ jobs:
     runs-on: ubuntu-latest
     container:
       image: rocker/r-devel-san:latest
+    env:
+      CARGO_HOME: /usr/local/cargo
+      RUSTUP_HOME: /usr/local/rustup
     defaults:
       run:
         working-directory: bindings/r
     steps:
       - uses: actions/checkout@v4
       - name: Install latest Rust
+        shell: bash
         run: |
+          export CARGO_HOME=/usr/local/cargo
+          export RUSTUP_HOME=/usr/local/rustup
           curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
-          echo "$HOME/.cargo/bin" >> $GITHUB_PATH
+          echo "$CARGO_HOME/bin" >> $GITHUB_PATH
+          echo "CARGO_HOME=$CARGO_HOME" >> $GITHUB_ENV
+          echo "RUSTUP_HOME=$RUSTUP_HOME" >> $GITHUB_ENV
       - name: Install System Dependencies
         run: |
           apt-get update
@@ -138,7 +146,9 @@ jobs:
       - name: Prepare Package Sources
         shell: bash
         run: |
-          if [ -f "$HOME/.cargo/env" ]; then
+          if [ -f "$CARGO_HOME/env" ]; then
+            . "$CARGO_HOME/env"
+          elif [ -f "$HOME/.cargo/env" ]; then
             . "$HOME/.cargo/env"
           fi
           chmod +x configure cleanup ../../dev/prepare_cran.sh
@@ -147,7 +157,9 @@ jobs:
       - name: Run ASAN Check
         shell: bash
         run: |
-          if [ -f "$HOME/.cargo/env" ]; then
+          if [ -f "$CARGO_HOME/env" ]; then
+            . "$CARGO_HOME/env"
+          elif [ -f "$HOME/.cargo/env" ]; then
             . "$HOME/.cargo/env"
           fi
           RDscript -e "rcmdcheck::rcmdcheck(args = c('--no-manual', '--no-vignettes'), build_args = c('--no-build-vignettes'), error_on = 'error', check_dir = 'check')"
