+++ .github/workflows/R-CMD-check.yaml
@@ -131,6 +131,8 @@ jobs:
         run: |
           export CARGO_HOME=/usr/local/cargo
           export RUSTUP_HOME=/usr/local/rustup
+          apt-get update
+          apt-get install -y --no-install-recommends curl ca-certificates
           curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
           echo "$CARGO_HOME/bin" >> $GITHUB_PATH
           echo "CARGO_HOME=$CARGO_HOME" >> $GITHUB_ENV
