### 2026-05-10 62ec293
fix(r): removed brittle curl dependency from the ASAN Rust bootstrap
