+++ .github/workflows/R-CMD-check.yaml
@@ -131,9 +131,15 @@ jobs:
         run: |
           export CARGO_HOME=/usr/local/cargo
           export RUSTUP_HOME=/usr/local/rustup
-          apt-get update
-          apt-get install -y --no-install-recommends curl ca-certificates
-          curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
+          if command -v curl >/dev/null 2>&1; then
+            curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
+          elif command -v wget >/dev/null 2>&1; then
+            wget -qO- https://sh.rustup.rs | sh -s -- -y
+          else
+            apt-get update
+            apt-get install -y --no-install-recommends wget ca-certificates
+            wget -qO- https://sh.rustup.rs | sh -s -- -y
+          fi
           echo "$CARGO_HOME/bin" >> $GITHUB_PATH
           echo "CARGO_HOME=$CARGO_HOME" >> $GITHUB_ENV
           echo "RUSTUP_HOME=$RUSTUP_HOME" >> $GITHUB_ENV
