### 2026-05-10 e5f9816
docs: modify cpp docs with the new results structure
