+++ docs/user-guide/robustness.md
@@ -265,7 +265,7 @@ Use robustness weights to identify potential outliers:
         .return_robustness_weights = true
     });
 
-    auto weights = result.robustness_weights();
+    auto weights = result.robustnessWeights();
     for (size_t i = 0; i < weights.size(); ++i) {
         if (weights[i] < 0.5) {
             std::cout << "Potential outlier at " << i << std::endl;
