### 2026-05-10 7f757c5
fix(r): skip vignette handling in ASAN
