### 2026-05-10 b82a2bc
fix(r): added libuv1-dev to SRR check
