+++ .github/workflows/SRR-check.yaml
@@ -37,7 +37,7 @@ jobs:
       - uses: r-lib/actions/setup-tinytex@v2
 
       - name: Install system dependencies
-        run: sudo apt-get update && sudo apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev
+        run: sudo apt-get update && sudo apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev libuv1-dev
 
       - name: Install dependencies
         run: |
