### 2026-05-10 bf35a3e
fix(python): public API now actually accepts documented array-like inputs
