### 2026-05-10 d30c26f
fix(repo): add pre-commit and pre-push git hook guards
