### 2026-05-10 4a1d8d3
chore: lint python dev scripts
