+++ dev/patch_vendor_crates.py
@@ -7,28 +7,31 @@ GPU dependencies that are not needed for the R binding.
 """
 
 import argparse
-import re
+import importlib
 import sys
 from pathlib import Path
-from typing import Any, Dict, Optional
+from typing import Any, Dict
 
 try:
     import tomllib
 except ImportError:
     import tomli as tomllib  # type: ignore
 
-import tomli_w
+
+def load_tomli_w():
+    """Import `tomli_w` lazily so editor-only envs do not fail static import resolution."""
+    return importlib.import_module("tomli_w")
 
 
 def load_workspace_values(workspace_toml: Path) -> Dict[str, Any]:
     """Load workspace package values from root Cargo.toml."""
-    with open(workspace_toml, "rb") as f:
-        data = tomllib.load(f)
-    
+    with open(workspace_toml, "rb") as file_handle:
+        data = tomllib.load(file_handle)
+
     workspace = data.get("workspace", {})
     package = workspace.get("package", {})
     deps = workspace.get("dependencies", {})
-    
+
     return {
         "package": package,
         "dependencies": deps,
@@ -63,12 +66,12 @@ def resolve_workspace_dep(dep_name: str, value: Any, workspace: Dict[str, Any]) 
             result = dict(ws_dep)
         else:
             result = {}
-        
+
         # Merge with local overrides (features, optional, etc.)
         for k, v in value.items():
             if k != "workspace":
                 result[k] = v
-        
+
         return result
     return value
 
@@ -80,9 +83,9 @@ def patch_crate_toml(
     lowess_as_path: bool = False,
 ) -> None:
     """Patch a crate's Cargo.toml to remove workspace inheritance."""
-    with open(crate_toml, "rb") as f:
-        data = tomllib.load(f)
-    
+    with open(crate_toml, "rb") as file_handle:
+        data = tomllib.load(file_handle)
+
     # Patch [package] section
     package = data.get("package", {})
     keys_to_remove = []
@@ -92,21 +95,26 @@ def patch_crate_toml(
             keys_to_remove.append(key)
         else:
             package[key] = resolved
-    
+
     for key in keys_to_remove:
         del package[key]
-    
+
     # Patch [dependencies] section
     if "dependencies" in data:
         deps = data["dependencies"]
         deps_to_remove = []
-        
+
         for dep_name, value in deps.items():
             # Remove GPU deps if requested
-            if remove_gpu and dep_name in ("wgpu", "bytemuck", "pollster", "futures-intrusive"):
+            if remove_gpu and dep_name in (
+                "wgpu",
+                "bytemuck",
+                "pollster",
+                "futures-intrusive",
+            ):
                 deps_to_remove.append(dep_name)
                 continue
-            
+
             # Handle lowess -> path dependency
             if lowess_as_path and dep_name == "lowess":
                 resolved = resolve_workspace_dep(dep_name, value, workspace)
@@ -115,30 +123,32 @@ def patch_crate_toml(
                     resolved["path"] = "../lowess"
                 deps[dep_name] = resolved
                 continue
-            
+
             deps[dep_name] = resolve_workspace_dep(dep_name, value, workspace)
-        
+
         for dep_name in deps_to_remove:
             del deps[dep_name]
-    
+
     # Patch [dev-dependencies] section
     if "dev-dependencies" in data:
         dev_deps = data["dev-dependencies"]
         for dep_name, value in dev_deps.items():
             dev_deps[dep_name] = resolve_workspace_dep(dep_name, value, workspace)
-    
+
     # Patch [features] section - empty GPU feature if needed
     if remove_gpu and "features" in data:
[…]
