### 2026-05-10 f5d32f3
chore: clean python scripts
