+++ validation/plot_foundation.py
@@ -1,0 +1,459 @@
+#!/usr/bin/env python3
+"""Core validation plots for fit quality and uncertainty."""
+
+from typing import Any, cast
+
+import matplotlib.pyplot as plt
+import pandas as pd
+import numpy as np
+
+try:
+    from .plot_common import (
+        add_rasterized_colorbar,
+        check_file,
+        column_to_numpy,
+        get_input_path,
+        plot_noisy_points,
+        plot_true_signal,
+        save_figure,
+    )
+except ImportError:
+    from plot_common import (  # type: ignore
+        add_rasterized_colorbar,
+        check_file,
+        column_to_numpy,
+        get_input_path,
+        plot_noisy_points,
+        plot_true_signal,
+        save_figure,
+    )
+
+
+def plot_degree_comparison():
+    """Plot the linear-versus-quadratic smoothing comparison."""
+    if not check_file("degree_comparison.csv"):
+        return
+    print("Plotting Degree Comparison...")
+
+    df = pd.read_csv(get_input_path("degree_comparison.csv"))
+    figure, axes = plt.subplots(1, 2, figsize=(10, 4))
+
+    ax1 = axes[0]
+    ax1.plot(
+        df["x"],
+        df["y_noisy"],
+        "o",
+        markersize=3,
+        alpha=0.4,
+        color="gray",
+        markeredgewidth=0,
+        label="Noisy data",
+        rasterized=True,
+    )
+    ax1.plot(df["x"], df["y_true"], "k-", lw=1.5, label="True signal", alpha=0.7)
+    ax1.plot(df["x"], df["y_lowess"], "b-", lw=2, label="LOWESS (Linear)")
+    ax1.plot(df["x"], df["y_fastLowess"], "r-", lw=2, label="fastLowess (Quadratic)")
+    ax1.set_xlabel("x", fontsize=12)
+    ax1.set_ylabel("y", fontsize=12)
+    ax1.set_title("LOWESS vs fastLowess: Full View", fontsize=14)
+    ax1.legend(loc="upper right")
+    ax1.grid(True, alpha=0.3)
+
+    ax2 = axes[1]
+    mask = (df["x"] > -0.6) & (df["x"] < 0.6)
+    ax2.plot(
+        df.loc[mask, "x"],
+        df.loc[mask, "y_noisy"],
+        "o",
+        markersize=4,
+        alpha=0.5,
+        color="gray",
+        markeredgewidth=0,
+        label="Noisy data",
+        rasterized=True,
+    )
+    ax2.plot(df.loc[mask, "x"], df.loc[mask, "y_true"], "k-", lw=2, label="True signal")
+    ax2.plot(
+        df.loc[mask, "x"],
+        df.loc[mask, "y_lowess"],
+        "b-",
+        lw=2.5,
+        label="LOWESS (Linear)",
+    )
+    ax2.plot(
+        df.loc[mask, "x"],
+        df.loc[mask, "y_fastLowess"],
+        "r-",
+        lw=2.5,
+        label="fastLowess (Quadratic)",
+    )
+    ax2.set_xlabel("x", fontsize=12)
+    ax2.set_ylabel("y", fontsize=12)
+    ax2.set_title("Peak Region: Linear Flattens, Quadratic Captures", fontsize=14)
+    ax2.legend(loc="lower center")
+    ax2.grid(True, alpha=0.3)
+
+    peak_x = df.loc[df["y_true"].idxmax(), "x"]
+    peak_true = df["y_true"].max()
+    peak_lowess = df.loc[df["y_true"].idxmax(), "y_lowess"]
+
+    ax2.annotate(
+        f"Gap = {peak_true - peak_lowess:.3f}",
+        xy=(peak_x, (peak_lowess + peak_true) / 2),
+        xytext=(0.3, 0.85),
[…]
