+++ benchmarks/compare.py
@@ -1,12 +1,16 @@
+"""Compare exported benchmark timings across R and Rust backends."""
+
 import json
 from pathlib import Path
-from statistics import mean, median
 
-def load_json(p: Path):
-    if not p.exists():
+
+def load_json(path: Path):
+    """Load a JSON file when it exists, otherwise return ``None``."""
+    if not path.exists():
         return None
-    with p.open("r", encoding="utf-8") as f:
-        return json.load(f)
+    with path.open("r", encoding="utf-8") as file:
+        return json.load(file)
+
 
 def pick_time_value(entry: dict):
     """Robustly pick a numeric timing from an entry."""
@@ -14,34 +18,35 @@ def pick_time_value(entry: dict):
         if key in entry:
             try:
                 return float(entry[key]), entry.get("size")
-            except Exception:
+            except (TypeError, ValueError):
                 pass
-    # fallback
+
     for k, v in entry.items():
         if isinstance(v, (int, float)):
             if k in ("iterations", "size", "runs"):
                 continue
             try:
                 return float(v), entry.get("size")
-            except Exception:
+            except (TypeError, ValueError):
                 pass
     return None, entry.get("size")
 
+
 def load_all_data(output_dir: Path):
+    """Load all exported benchmark JSON files keyed by backend label."""
     files = {
         "Rust (CPU)": output_dir / "rust_benchmark_cpu.json",
         "Rust (Serial)": output_dir / "rust_benchmark_cpu_serial.json",
         "Rust (GPU)": output_dir / "rust_benchmark_gpu.json",
         "R": output_dir / "r_benchmark.json",
     }
-    
+
     data = {}
     for label, path in files.items():
         loaded = load_json(path)
         if loaded:
-            # Flatten category structure: {category: [entries]} -> {name: entry}
             flat = {}
-            for cat, entries in loaded.items():
+            for entries in loaded.values():
                 for entry in entries:
                     name = entry.get("name")
                     if name:
@@ -49,81 +54,89 @@ def load_all_data(output_dir: Path):
             data[label] = flat
     return data
 
-def main():
-    repo_root = Path(__file__).resolve().parent
-    # walk up to workspace root
-    workspace = repo_root
+
+def find_workspace_output_dir() -> Path:
+    """Find the workspace ``output`` directory by walking upward from this script."""
+    workspace = Path(__file__).resolve().parent
     for _ in range(6):
         if (workspace / "output").exists():
             break
         if workspace.parent == workspace:
             break
         workspace = workspace.parent
-    out_dir = workspace / "output"
-    
-    data = load_all_data(out_dir)
-    
-    # Check what we have
+    return workspace / "output"
+
+
+def collect_benchmark_names(data: dict) -> list[str]:
+    """Return all benchmark names seen across backends in sorted order."""
+    all_names = set()
+    for backend_data in data.values():
+        all_names.update(backend_data.keys())
+    return sorted(all_names)
+
+
+def format_baseline(entry: dict | None) -> tuple[float | None, str]:
+    """Return the numeric baseline value and its display string."""
+    if not entry:
+        return None, "-"
+
+    baseline_value, _ = pick_time_value(entry)
+    if baseline_value and baseline_value > 0:
+        return baseline_value, f"{baseline_value:.2f}ms"
+    return None, "-"
+
+
+def format_candidate_cell(entry: dict | None, baseline_value: float | None) -> str:
+    """Format one backend cell, optionally including speedup versus baseline."""
+    if not entry:
+        return "-"
[…]
