### 2026-05-10 0aeb0c2
fix(r): R SRR/pkgcheck warnings
