### 2026-05-10 ce29054
chore(julia): lint julia scripts
