### 2026-05-10 687b7d0
chore(wasm): lint wasm code, upgrade oxlint to 1.63
