### 2026-05-10 6bf37c8
chore(nodejs): lint nodejs code, upgrade oxlint to 1.63, upgrade napi-rs/cli to 3.6
