+++ bindings/r/man/Lowess.Rd
@@ -5,24 +5,24 @@
 \title{LOWESS Batch Smoothing}
 \usage{
 Lowess(
-  fraction = 0.67,
-  iterations = 3L,
-  delta = NULL,
-  weight_function = "tricube",
-  robustness_method = "bisquare",
-  scaling_method = "mad",
-  boundary_policy = "extend",
-  confidence_intervals = NULL,
-  prediction_intervals = NULL,
-  return_diagnostics = FALSE,
-  return_residuals = FALSE,
-  return_robustness_weights = FALSE,
-  zero_weight_fallback = "use_local_mean",
-  auto_converge = NULL,
-  cv_fractions = NULL,
-  cv_method = "kfold",
-  cv_k = 5L,
-  parallel = TRUE
+    fraction = 0.67,
+    iterations = 3L,
+    delta = NULL,
+    weight_function = "tricube",
+    robustness_method = "bisquare",
+    scaling_method = "mad",
+    boundary_policy = "extend",
+    confidence_intervals = NULL,
+    prediction_intervals = NULL,
+    return_diagnostics = FALSE,
+    return_residuals = FALSE,
+    return_robustness_weights = FALSE,
+    zero_weight_fallback = "use_local_mean",
+    auto_converge = NULL,
+    cv_fractions = NULL,
+    cv_method = "kfold",
+    cv_k = 5L,
+    parallel = TRUE
 )
 }
 \arguments{
