### 2026-05-10 1028610
ci: add comprehensive sanitizers and ci make all jobs
