### 2026-05-10 e31b7f9
fix: replace deprecated npx license-checker, fix julia windows build
