### 2026-05-11 2a025eb
build: bump lowess version to 1.3.0, add missing windows cpp dependencies
