### 2026-05-11 977de3a
build: bump fastLowess version to 1.3.0
