+++ crates/fastLowess/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 # Metadata
 name = "fastLowess"
-version = "1.2.0"
+version = "1.3.0"
 authors = ["Amir Valizadeh <thisisamirv@gmail.com>"]
 edition = "2024"
 rust-version = "1.89"
@@ -26,7 +26,7 @@ include = [
 ]
 
 [dependencies]
-lowess = { path = "../lowess", version = "1.2", features = ["dev"] }
+lowess = { path = "../lowess", version = "1.3", features = ["dev"] }
 num-traits = { version = "0.2", features = ["libm"] }
 ndarray = { version = "0.17", optional = true }
 rayon = { version = "1.12", optional = true }
