### 2026-05-11 f194fc9
build: bump to v1.3.0
