+++ Makefile
@@ -100,6 +100,13 @@ WASM_TEST_DIR := tests/wasm
 # C++ bindings
 CPP_PKG := fastlowess-cpp
 CPP_DIR := bindings/cpp
+CPP_CARGO_TARGET :=
+CPP_LIBRARY_DIR := target/release
+
+ifeq ($(OS),Windows_NT)
+	CPP_CARGO_TARGET := --target x86_64-pc-windows-gnu
+	CPP_LIBRARY_DIR := target/x86_64-pc-windows-gnu/release
+endif
 
 # Julia native library paths and symbol scanners
 ifeq ($(HOST_PLATFORM),windows)
@@ -116,21 +123,27 @@ endif
 JL_SHARED_LIB_ABS := $(abspath $(JL_SHARED_LIB))
 
 ifeq ($(HOST_PLATFORM),windows)
-	CPP_SHARED_LIB := target/release/fastlowess_cpp.dll
+	CPP_SHARED_LIB := $(CPP_LIBRARY_DIR)/fastlowess_cpp.dll
+	CPP_TEST_LIB := $(CPP_LIBRARY_DIR)/libfastlowess_cpp.a
 	CPP_EXPORT_SCAN := objdump -p $(CPP_SHARED_LIB)
-	CPP_TEST_BUILD := cmake --build . --config Debug
-	CPP_TEST_RUN := PATH="../../../target/release$(PATH_SEPARATOR)$$PATH" ./Debug/test_fastlowess_suite.exe
+	CPP_TEST_BUILD := cmake --build .
+	CPP_TEST_RUN := PATH="../../../$(CPP_LIBRARY_DIR)$(PATH_SEPARATOR)$$PATH" ./test_fastlowess_suite.exe
 else ifeq ($(HOST_PLATFORM),macos)
-	CPP_SHARED_LIB := target/release/libfastlowess_cpp.dylib
+	CPP_SHARED_LIB := $(CPP_LIBRARY_DIR)/libfastlowess_cpp.dylib
+	CPP_TEST_LIB := $(CPP_SHARED_LIB)
 	CPP_EXPORT_SCAN := nm -gU $(CPP_SHARED_LIB)
 	CPP_TEST_BUILD := make
 	CPP_TEST_RUN := ./test_fastlowess_suite
 else
-	CPP_SHARED_LIB := target/release/libfastlowess_cpp.so
+	CPP_SHARED_LIB := $(CPP_LIBRARY_DIR)/libfastlowess_cpp.so
+	CPP_TEST_LIB := $(CPP_SHARED_LIB)
 	CPP_EXPORT_SCAN := nm -D $(CPP_SHARED_LIB)
 	CPP_TEST_BUILD := make
 	CPP_TEST_RUN := ./test_fastlowess_suite
 endif
+
+CPP_LIBRARY_DIR_ABS := $(abspath $(CPP_LIBRARY_DIR))
+CPP_TEST_LIB_ABS := $(abspath $(CPP_TEST_LIB))
 
 ifeq ($(HOST_PLATFORM),windows)
 	CPP_EXAMPLE_RUN_ENV := PATH="target/release$(PATH_SEPARATOR)$$PATH"
@@ -392,9 +405,9 @@ _r_impl:
 	@echo "2. Installing development packages (R & Python)..."
 	@echo "=============================================================================="
 	@$(PYTHON) -m pip install -q tomli tomli_w || true
[…]
