+++ bindings/python/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fastLowess-py"
-version = "1.2.0"
+version = "1.3.0"
 authors = ["Amir Valizadeh <thisisamirv@gmail.com>"]
 edition = "2024"
 license = "MIT OR Apache-2.0"
@@ -14,7 +14,7 @@ keywords = ["lowess", "smoothing", "statistics", "regression", "bioinformatics"]
 categories = ["science", "algorithms"]
 
 [dependencies]
-fastLowess = { version = "1.2", features = ["dev", "cpu"] }
+fastLowess = { version = "1.3", features = ["dev", "cpu"] }
 pyo3 = { version = "0.28", features = ["abi3-py38"] }
 numpy = "0.28"
 
