+++ bindings/cpp/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.15)
-project(fastlowess-cpp VERSION 1.2.0 LANGUAGES CXX)
+project(fastlowess-cpp VERSION 1.3.0 LANGUAGES CXX)
 
 include(FetchContent)
 include(CMakePackageConfigHelpers)
