+++ bindings/r/src/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rfastlowess"
-version = "1.2.0"
+version = "1.3.0"
 authors = ["Amir Valizadeh <thisisamirv@gmail.com>"]
 edition = "2024"
 license = "MIT OR Apache-2.0"
