+++ bindings/wasm/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fastlowess-wasm"
-version = "1.2.0"
+version = "1.3.0"
 authors = ["Amir Valizadeh <thisisamirv@gmail.com>"]
 edition = "2024"
 license = "MIT OR Apache-2.0"
@@ -17,7 +17,7 @@ categories = ["science", "algorithms"]
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-fastLowess = { version = "1.2", features = ["dev", "cpu"] }
+fastLowess = { version = "1.3", features = ["dev", "cpu"] }
 wasm-bindgen = "0.2"
 js-sys = "0.3"
 console_error_panic_hook = "0.1"
