+++ bindings/julia/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fastlowess-jl"
-version = "1.2.0"
+version = "1.3.0"
 authors = ["Amir Valizadeh <thisisamirv@gmail.com>"]
 edition = "2024"
 license = "MIT OR Apache-2.0"
@@ -14,7 +14,7 @@ keywords = ["lowess", "smoothing", "statistics", "regression", "bioinformatics"]
 categories = ["science", "algorithms"]
 
 [dependencies]
-fastLowess = { version = "1.2", features = ["dev", "cpu"] }
+fastLowess = { version = "1.3", features = ["dev", "cpu"] }
 
 [lib]
 name = "fastlowess_jl"

[…]
