+++ .github/workflows/R-CMD-check.yaml
@@ -27,7 +27,7 @@ jobs:
 
         steps:
             - name: Checkout Repository
-              uses: actions/checkout@v4
+              uses: actions/checkout@v6
 
             - name: Install Rust Toolchain
               uses: dtolnay/rust-toolchain@stable
