+++ .github/workflows/SRR-check.yaml
@@ -24,7 +24,7 @@ jobs:
             issues: write
         if: ${{ github.event.workflow_run.conclusion == 'success' }}
         steps:
-            - uses: actions/checkout@v4
+            - uses: actions/checkout@v6
               with:
                   fetch-depth: 0
 
