+++ .github/workflows/sanitizers.yml
@@ -29,7 +29,7 @@ jobs:
 
         steps:
             - name: Checkout Repository
-              uses: actions/checkout@v4
+              uses: actions/checkout@v6
 
             - name: Install Rust Nightly
               uses: dtolnay/rust-toolchain@nightly
@@ -38,17 +38,17 @@ jobs:
                   components: rust-src
 
             - name: Install Python
-              uses: actions/setup-python@v5
+              uses: actions/setup-python@v6
               with:
                   python-version: "3.12"
 
             - name: Setup Node
-              uses: actions/setup-node@v4
+              uses: actions/setup-node@v6
               with:
                   node-version: 22
 
             - name: Install Julia
-              uses: julia-actions/setup-julia@v2
+              uses: julia-actions/setup-julia@v3
               with:
                   version: "1.11"
 
@@ -146,7 +146,7 @@ jobs:
             run:
                 working-directory: bindings/r
         steps:
-            - uses: actions/checkout@v4
+            - uses: actions/checkout@v6
             - name: Install latest Rust
               shell: bash
               run: |
