+++ .github/workflows/release-wasm.yml
@@ -13,7 +13,7 @@ jobs:
         name: Publish WASM
         runs-on: ubuntu-latest
         steps:
-            - uses: actions/checkout@v4
+            - uses: actions/checkout@v6
 
             - name: Install Rust Toolchain
               uses: dtolnay/rust-toolchain@stable
@@ -25,7 +25,7 @@ jobs:
               run: curl https://rustwasm.github.io/wasm-pack/installer/init.sh -sSf | sh
 
             - name: Setup Node
-              uses: actions/setup-node@v4
+              uses: actions/setup-node@v6
               with:
                   node-version: 22
                   registry-url: "https://registry.npmjs.org"

[…]
