### 2026-05-13 99bef65
fix(r): prepare_cran won't delete in-package tests anymore
