### 2026-05-13 cb5a2ed
build(r): remove tests from bindings/r/
