### 2026-07-08 a281b8c
chore: add lintr, minor clang-tidy modifications
