+++ tests/cpp/test_fastlowess.cpp
@@ -153,11 +153,11 @@ void testBasicSmooth() {
   auto result = lowess.fit(sample_x_values, sample_y_values).value();
 
   assertTrue(result.valid(), "Result should be valid");
-  assertTrue(result.yVector().size() == k_small_sample_size,
+  assertTrue(result.y_vector().size() == k_small_sample_size,
              "Output length mismatch");
-  assertTrue(result.xVector().size() == k_small_sample_size,
+  assertTrue(result.x_vector().size() == k_small_sample_size,
              "X length mismatch");
-  assertApprox(result.fractionUsed(), k_basic_fraction);
+  assertApprox(result.fraction_used(), k_basic_fraction);
 }
 
 void testBasicSmoothSerial() {
@@ -172,7 +172,7 @@ void testBasicSmoothSerial() {
   auto result = lowess.fit(sample_x_values, sample_y_values).value();
 
   assertTrue(result.valid(), "Serial result should be valid");
-  assertTrue(result.yVector().size() == k_small_sample_size,
+  assertTrue(result.y_vector().size() == k_small_sample_size,
              "Serial output length mismatch");
 }
 
@@ -190,7 +190,7 @@ void testLowessWithDiagnostics() {
   auto diagnostics = result.diagnostics();
   assertTrue(diagnostics.rmse() >= 0.0, "RMSE negative");
   assertTrue(diagnostics.mae() >= 0.0, "MAE negative");
-  assertTrue(diagnostics.rSquared() >= 0.0 && diagnostics.rSquared() <= 1.0,
+  assertTrue(diagnostics.r_squared() >= 0.0 && diagnostics.r_squared() <= 1.0,
              "R2 out of range");
 }
 
@@ -221,7 +221,7 @@ void testLowessWithRobustnessWeights() {
   fastlowess::Lowess lowess(options);
   auto result = lowess.fit(sample_x_values, outlier_y_values).value();
 
-  auto robustness_weights = result.robustnessWeights();
+  auto robustness_weights = result.robustness_weights();
   assertTrue(robustness_weights.size() == k_small_sample_size,
              "Robustness weight count mismatch");
   for (const double weight_value : robustness_weights) {
@@ -243,8 +243,8 @@ void testLowessWithConfidenceIntervals() {
   fastlowess::Lowess lowess(options);
   auto result = lowess.fit(linear_data.x_values, linear_data.y_values).value();
 
-  auto confidence_lower = result.confidenceLower();
-  auto confidence_upper = result.confidenceUpper();
+  auto confidence_lower = result.confidence_lower();
+  auto confidence_upper = result.confidence_upper();
   assertTrue(confidence_lower.size() == k_interval_point_count,
              "Confidence lower size mismatch");
   assertTrue(confidence_upper.size() == k_interval_point_count,
@@ -269,9 +269,9 @@ void testLowessWithPredictionIntervals() {
   fastlowess::Lowess lowess(options);
   auto result = lowess.fit(linear_data.x_values, linear_data.y_values).value();
 
-  assertTrue(result.predictionLower().size() == k_interval_point_count,
+  assertTrue(result.prediction_lower().size() == k_interval_point_count,
              "Prediction lower size mismatch");
-  assertTrue(result.predictionUpper().size() == k_interval_point_count,
+  assertTrue(result.prediction_upper().size() == k_interval_point_count,
              "Prediction upper size mismatch");
 }
 
@@ -290,9 +290,9 @@ void testLowessReuse() {
   auto first_result = lowess.fit(sample_x_values, sample_y_values).value();
   auto second_result = lowess.fit(reuse_x_values, reuse_y_values).value();
 
-  assertTrue(first_result.yVector().size() == k_small_sample_size,
+  assertTrue(first_result.y_vector().size() == k_small_sample_size,
[…]
