### 2026-07-08 e3bdce8
docs: migrate root docs to docs/
