### 2026-07-08 09371a7
refactor: change enum-typed builder methods to string-based API
