+++ crates/lowess/src/math/boundary.rs
@@ -16,7 +16,11 @@ use alloc::vec::Vec;
 use std::vec::Vec;
 
 // External dependencies
+use core::str::FromStr;
 use num_traits::Float;
+
+// Internal dependencies
+use crate::primitives::errors::LowessError;
 
 // Policy for handling boundaries at the start and end of a data stream.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
@@ -33,6 +37,24 @@ pub enum BoundaryPolicy {
 
     // No boundary padding (standard LOWESS behavior).
     NoBoundary,
+}
+
+impl FromStr for BoundaryPolicy {
+    type Err = LowessError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_str() {
+            "extend" | "pad" => Ok(Self::Extend),
+            "reflect" | "mirror" => Ok(Self::Reflect),
+            "zero" => Ok(Self::Zero),
+            "noboundary" | "none" => Ok(Self::NoBoundary),
+            _ => Err(LowessError::InvalidOption {
+                option: "boundary_policy",
+                value: s.to_string(),
+                valid: "extend, reflect, zero, noboundary",
+            }),
+        }
+    }
 }
 
 // Apply a boundary policy to pad the input data.
