### 2026-07-08 f5b5568
fix: rust syntax in tests, examples, bench, and validation
