+++ examples/lowess/streaming_smoothing.rs
@@ -397,7 +397,7 @@ fn example_5_outlier_handling() -> Result<(), LowessError> {
     println!("Dataset: {} points with 3 outliers", n);
     println!("Testing robustness methods:\n");
 
-    let methods = vec![(Bisquare, "Bisquare"), (Huber, "Huber"), (Talwar, "Talwar")];
+    let methods = vec![("bisquare", "Bisquare"), ("huber", "Huber"), ("talwar", "Talwar")];
 
     for (method, name) in methods {
         println!("Using {} robustness:", name);
