### 2026-07-08 09d79cf
refactor: add adaptor build type aliases as the primary user-facing constructors
