### 2026-07-08 9af40c1
refactor: add parse module to define IntoEnum<E> trait and its macro-generated impls for all enum-typed builder parameters
