+++ crates/fastLowess/src/adapters/batch.rs
@@ -26,6 +26,7 @@ use std::fmt::Debug;
 use std::result::Result;
 
 // Export dependencies from lowess crate
+use crate::parse::IntoEnum;
 use lowess::internals::adapters::batch::BatchLowessBuilder;
 use lowess::internals::algorithms::regression::WLSSolver;
 use lowess::internals::algorithms::regression::ZeroWeightFallback;
@@ -98,8 +99,8 @@ impl<T: Float> ParallelBatchLowessBuilder<T> {
     }
 
     // Set the kernel weight function.
-    pub fn weight_function(mut self, wf: impl AsRef<str>) -> Self {
-        match wf.as_ref().parse::<WeightFunction>() {
+    pub fn weight_function(mut self, wf: impl IntoEnum<WeightFunction>) -> Self {
+        match wf.into_enum() {
             Ok(w) => self.base.weight_function = w,
             Err(e) => {
                 if self.base.deferred_error.is_none() {
@@ -111,8 +112,8 @@ impl<T: Float> ParallelBatchLowessBuilder<T> {
     }
 
     // Set the robustness method for outlier handling.
-    pub fn robustness_method(mut self, method: impl AsRef<str>) -> Self {
-        match method.as_ref().parse::<RobustnessMethod>() {
+    pub fn robustness_method(mut self, method: impl IntoEnum<RobustnessMethod>) -> Self {
+        match method.into_enum() {
             Ok(m) => self.base.robustness_method = m,
             Err(e) => {
                 if self.base.deferred_error.is_none() {
@@ -124,8 +125,8 @@ impl<T: Float> ParallelBatchLowessBuilder<T> {
     }
 
     // Set the zero-weight fallback policy.
-    pub fn zero_weight_fallback(mut self, fallback: impl AsRef<str>) -> Self {
-        match fallback.as_ref().parse::<ZeroWeightFallback>() {
+    pub fn zero_weight_fallback(mut self, fallback: impl IntoEnum<ZeroWeightFallback>) -> Self {
+        match fallback.into_enum() {
             Ok(f) => self.base.zero_weight_fallback = f,
             Err(e) => {
                 if self.base.deferred_error.is_none() {
@@ -137,8 +138,8 @@ impl<T: Float> ParallelBatchLowessBuilder<T> {
     }
 
     // Set the boundary handling policy.
-    pub fn boundary_policy(mut self, policy: impl AsRef<str>) -> Self {
-        match policy.as_ref().parse::<BoundaryPolicy>() {
+    pub fn boundary_policy(mut self, policy: impl IntoEnum<BoundaryPolicy>) -> Self {
+        match policy.into_enum() {
             Ok(p) => self.base.boundary_policy = p,
             Err(e) => {
                 if self.base.deferred_error.is_none() {

[…]
