### 2026-07-08 15167bb
refactor: all bindings now use binding_support in fastLowess
