+++ bindings/cpp/src/lib.rs
@@ -10,15 +10,11 @@ use std::ffi::CStr;
 use std::os::raw::{c_char, c_double, c_int, c_ulong};
 use std::ptr;
 
+use fastLowess::binding_support;
 use fastLowess::internals::adapters::online::ParallelOnlineLowess;
 use fastLowess::internals::adapters::streaming::ParallelStreamingLowess;
-use fastLowess::internals::api::{
-    BoundaryPolicy, MergeStrategy, RobustnessMethod, ScalingMethod, UpdateMode, WeightFunction,
-    ZeroWeightFallback,
-};
-use fastLowess::prelude::{
-    Batch, KFold, LOOCV, Lowess as LowessBuilder, LowessResult, MAD, MAR, Mean, Online, Streaming,
-};
+use fastLowess::internals::api::{MergeStrategy, UpdateMode};
+use fastLowess::prelude::{Batch, Lowess as LowessBuilder, LowessResult, Online, Streaming};
 
 /// Result struct that can be passed across FFI boundary.
 /// All arrays are allocated by Rust and must be freed by Rust.
@@ -45,6 +41,10 @@ pub struct CppLowessResult {
     pub residuals: *mut c_double,
     /// Robustness weights (NULL if not computed)
     pub robustness_weights: *mut c_double,
+    /// Cross-validation scores (NULL if not computed, length = cv_scores_len)
+    pub cv_scores: *mut c_double,
+    /// Number of cross-validation scores
+    pub cv_scores_len: c_ulong,
 
     /// Fraction used for smoothing
     pub fraction_used: c_double,
@@ -77,6 +77,8 @@ impl Default for CppLowessResult {
             prediction_upper: ptr::null_mut(),
             residuals: ptr::null_mut(),
             robustness_weights: ptr::null_mut(),
+            cv_scores: ptr::null_mut(),
+            cv_scores_len: 0,
             fraction_used: 0.0,
             iterations_used: -1,
             rmse: f64::NAN,
@@ -124,102 +126,6 @@ unsafe fn parse_c_str(s: *const c_char, default: &str) -> &str {
     }
 }
 
-/// Parse weight function from string.
-fn parse_weight_function(name: &str) -> Result<WeightFunction, String> {
-    match name.to_lowercase().as_str() {
-        "tricube" => Ok(WeightFunction::Tricube),
-        "epanechnikov" => Ok(WeightFunction::Epanechnikov),
-        "gaussian" => Ok(WeightFunction::Gaussian),
-        "uniform" | "boxcar" => Ok(WeightFunction::Uniform),
-        "biweight" | "bisquare" => Ok(WeightFunction::Biweight),
-        "triangle" | "triangular" => Ok(WeightFunction::Triangle),
-        "cosine" => Ok(WeightFunction::Cosine),
-        _ => Err(format!(
-            "Unknown weight function: {}. Valid: tricube, epanechnikov, gaussian, uniform, biweight, triangle, cosine",
-            name
-        )),
-    }
-}
-
-/// Parse robustness method from string.
-fn parse_robustness_method(name: &str) -> Result<RobustnessMethod, String> {
-    match name.to_lowercase().as_str() {
-        "bisquare" | "biweight" => Ok(RobustnessMethod::Bisquare),
-        "huber" => Ok(RobustnessMethod::Huber),
-        "talwar" => Ok(RobustnessMethod::Talwar),
-        _ => Err(format!(
-            "Unknown robustness method: {}. Valid: bisquare, huber, talwar",
-            name
-        )),
-    }
-}
-
-/// Parse zero weight fallback from string.
-fn parse_zero_weight_fallback(name: &str) -> Result<ZeroWeightFallback, String> {
-    match name.to_lowercase().as_str() {
[…]
