+++ crates/fastLowess/src/api.rs
@@ -18,11 +18,16 @@ use lowess::internals::api::Online as BaseOnline;
 use lowess::internals::api::Streaming as BaseStreaming;
 
 // Publicly re-exported types
+pub use lowess::internals::algorithms::regression::ZeroWeightFallback;
+pub use lowess::internals::algorithms::robustness::RobustnessMethod;
 pub use lowess::internals::api::{
-    BatchMode, Lowess, LowessAdapter, LowessBuilder, OnlineLowess, OnlineMode, StreamingLowess,
-    StreamingMode,
+    BatchMode, Lowess, LowessAdapter, LowessBuilder, MergeStrategy, OnlineLowess, OnlineMode,
+    StreamingLowess, StreamingMode, UpdateMode,
 };
 pub use lowess::internals::engine::output::LowessResult;
+pub use lowess::internals::math::boundary::BoundaryPolicy;
+pub use lowess::internals::math::kernel::WeightFunction;
+pub use lowess::internals::math::scaling::ScalingMethod;
 pub use lowess::internals::primitives::backend::Backend;
 pub use lowess::internals::primitives::errors::LowessError;
 
