+++ bindings/r/R/Lowess.R
@@ -19,7 +19,7 @@
 #' @param delta Interpolation threshold. NULL = auto.
 #' @param weight_function Kernel name. Default: "tricube".
 #' @param robustness_method Method: "bisquare", "huber", "talwar".
-#' @param scaling_method Scale estimation: "mad", "mar".
+#' @param scaling_method Scale estimation: "mad", "mar", "mean".
 #' @param boundary_policy Edge handling: "extend", "reflect", "zero",
 #'   "noboundary".
 #' @param confidence_intervals Confidence level (e.g., 0.95). NULL disables.

[…]
