### 2026-07-08 e941fc9
refactor: change add_points to singular add_point in the bindings
