+++ bindings/r/R/OnlineLowess.R
@@ -16,7 +16,7 @@
 #' model <- OnlineLowess(fraction = 0.2, window_capacity = 20)
 #' x <- 1:50
 #' y <- sin(x * 0.1) + rnorm(50, 0, 0.1)
-#' result <- model$add_points(x, y)
+#' result <- model$add_point(1.0, sin(1.0 * 0.1))
 #' plot(x, y)
 #' lines(x, result$y, col = "red")
 #' @export
@@ -44,11 +44,8 @@ OnlineLowess <- function(
     structure(
         list(
             handle = handle,
-            add_points = function(x, y) {
-                if (length(x) != length(y)) {
-                    stop("x and y must have the same length")
-                }
-                handle$add_points(as.double(x), as.double(y))
+            add_point = function(x, y) {
+                handle$add_point(as.double(x), as.double(y))
             },
             params = list(
                 fraction = fraction,
