### 2026-07-08 2784521
refactor(lowess): align with loess-rs — rename auto_converge, add ParseErrors, iterations_used, parse aliases, pub(crate) IntoEnum
