+++ tests/lowess/engine_executor_tests.rs
@@ -112,7 +112,7 @@ fn test_config_defaults() {
     );
     assert!(config.cv_kind.is_none(), "Default CV kind should be None");
     assert!(
-        config.auto_convergence.is_none(),
+        config.auto_converge.is_none(),
         "Default auto-convergence should be None"
     );
     assert!(
@@ -135,7 +135,7 @@ fn test_config_custom() {
         robustness_method: RobustnessMethod::Huber,
         cv_fractions: Some(vec![0.3, 0.5, 0.7]),
         cv_kind: None,
-        auto_convergence: Some(1e-6),
+        auto_converge: Some(1e-6),
         return_variance: None,
         boundary_policy: BoundaryPolicy::default(),
         scaling_method: ScalingMethod::default(),
@@ -156,7 +156,7 @@ fn test_config_custom() {
     assert_eq!(config.zero_weight_fallback, 1);
     assert_eq!(config.robustness_method, RobustnessMethod::Huber);
     assert_eq!(config.cv_fractions, Some(vec![0.3, 0.5, 0.7]));
-    assert_eq!(config.auto_convergence, Some(1e-6));
+    assert_eq!(config.auto_converge, Some(1e-6));
 }
 
 // ============================================================================
@@ -347,7 +347,7 @@ fn test_config_f32() {
         robustness_method: RobustnessMethod::Bisquare,
         cv_fractions: None,
         cv_kind: None,
-        auto_convergence: None,
+        auto_converge: None,
         return_variance: None,
         boundary_policy: BoundaryPolicy::default(),
         scaling_method: ScalingMethod::default(),
@@ -407,7 +407,7 @@ fn test_executor_convergence_zero_tolerance() {
         robustness_method: RobustnessMethod::Bisquare,
         cv_fractions: None,
         cv_kind: None,
-        auto_convergence: Some(0.0), // Zero tolerance
+        auto_converge: Some(0.0), // Zero tolerance
         return_variance: None,
         boundary_policy: BoundaryPolicy::Extend,
         scaling_method: ScalingMethod::default(),
@@ -445,7 +445,7 @@ fn test_executor_delta_equals_range() {
         robustness_method: RobustnessMethod::Bisquare,
         cv_fractions: None,
         cv_kind: None,
-        auto_convergence: None,
+        auto_converge: None,
         return_variance: None,
         boundary_policy: BoundaryPolicy::Extend,
         scaling_method: ScalingMethod::default(),
