### 2026-07-08 e13e116
feat: add custom_weights support
