### 2026-07-08 2fced03
fix: ParseErrors accumulation in fastLowess
