### 2026-07-08 91a6cec
docs: fix readme table formats and links
