### 2026-07-08 027285a
fix: move binding_support to internals module
