### 2026-07-09 670cc81
refactor: change adapter argument API to Lowess, StreamingLowess, and OnlineLowess
