+++ crates/fastLowess/src/api.rs
@@ -5,9 +5,9 @@
 //! with adapters that utilize all available CPU cores or GPU hardware.
 
 // Internal dependencies
-use crate::adapters::batch::ParallelBatchLowessBuilder;
-use crate::adapters::online::ParallelOnlineLowessBuilder;
-use crate::adapters::streaming::ParallelStreamingLowessBuilder;
+use crate::adapters::batch::{ParallelBatchLowess, ParallelBatchLowessBuilder};
+use crate::adapters::online::{ParallelOnlineLowess, ParallelOnlineLowessBuilder};
+use crate::adapters::streaming::{ParallelStreamingLowess, ParallelStreamingLowessBuilder};
 
 // External dependencies
 use num_traits::Float;
@@ -20,10 +20,7 @@ use lowess::internals::api::Streaming as BaseStreaming;
 // Publicly re-exported types
 pub use lowess::internals::algorithms::regression::ZeroWeightFallback;
 pub use lowess::internals::algorithms::robustness::RobustnessMethod;
-pub use lowess::internals::api::{
-    BatchMode, Lowess, LowessAdapter, LowessBuilder, MergeStrategy, OnlineLowess, OnlineMode,
-    StreamingLowess, StreamingMode, UpdateMode,
-};
+pub use lowess::internals::api::{LowessAdapter, LowessBuilder, MergeStrategy, UpdateMode};
 pub use lowess::internals::engine::output::LowessResult;
 pub use lowess::internals::math::boundary::BoundaryPolicy;
 pub use lowess::internals::math::kernel::WeightFunction;
@@ -78,7 +75,10 @@ impl<T: Float> LowessAdapter<T> for Streaming {
         base.parallel = Some(parallel);
 
         // Wrap with extension fields
-        ParallelStreamingLowessBuilder { base, parse_errors: Vec::new() }
+        ParallelStreamingLowessBuilder {
+            base,
+            parse_errors: Vec::new(),
+        }
     }
 }
 
@@ -98,6 +98,176 @@ impl<T: Float> LowessAdapter<T> for Online {
         base.parallel = Some(parallel);
 
         // Wrap with extension fields
-        ParallelOnlineLowessBuilder { base, parse_errors: Vec::new() }
+        ParallelOnlineLowessBuilder {
+            base,
+            parse_errors: Vec::new(),
+        }
+    }
+}
+
+// Entry-point wrapper types: Lowess / StreamingLowess / OnlineLowess
+// Mirror the bindings API while defaulting to parallel execution.
+
+// Macro: generate method-forwarding impls common to all three entry-point types.
+macro_rules! impl_common_builder {
+    ($t:ty) => {
+        impl Default for $t {
+            fn default() -> Self {
+                Self::new()
+            }
+        }
+        impl $t {
+            pub fn new() -> Self {
+                Self(LowessBuilder::new())
+            }
+            // string enum options
+            pub fn weight_function(mut self, s: &str) -> Self {
+                self.0 = self.0.weight_function(s);
+                self
+            }
+            pub fn robustness_method(mut self, s: &str) -> Self {
+                self.0 = self.0.robustness_method(s);
+                self
+            }
+            pub fn scaling_method(mut self, s: &str) -> Self {
+                self.0 = self.0.scaling_method(s);
+                self
+            }
+            pub fn zero_weight_fallback(mut self, s: &str) -> Self {
+                self.0 = self.0.zero_weight_fallback(s);
+                self
+            }
+            pub fn boundary_policy(mut self, s: &str) -> Self {
+                self.0 = self.0.boundary_policy(s);
+                self
+            }
+            // numeric options
+            pub fn fraction(mut self, f: f64) -> Self {
[…]
