### 2026-07-09 cc50e38
feat: improve validation
