+++ crates/fastLowess/src/lib.rs
@@ -229,10 +229,10 @@ mod api;
 // Input data handling.
 mod input;
 
-// String input compatibility for builder methods.
+// String input compatibility for builder methods (only needed for dev/binding builds).
 mod parse;
 
-// Shared helpers for language bindings.
+// Shared helpers for language bindings (only needed for dev/binding builds).
 mod binding_support;
 
 // Standard fastLowess prelude.
