+++ bindings/python/src/lib.rs
@@ -7,13 +7,13 @@ use pyo3::prelude::*;
 use std::fmt::Display;
 use std::sync::Mutex;
 
-use ::fastLowess::internals::binding_support;
 use ::fastLowess::internals::adapters::online::ParallelOnlineLowess;
 use ::fastLowess::internals::adapters::streaming::ParallelStreamingLowess;
 use ::fastLowess::internals::api::{
     Batch, BoundaryPolicy, LowessBuilder, Online, RobustnessMethod, ScalingMethod, Streaming,
     WeightFunction, ZeroWeightFallback,
 };
+use ::fastLowess::internals::binding_support;
 use ::fastLowess::prelude::LowessResult;
 
 // ============================================================================
@@ -227,7 +227,8 @@ impl PyStreamingLowess {
         return_residuals=false,
         return_robustness_weights=false,
         zero_weight_fallback="use_local_mean",
-        parallel=true
+        parallel=true,
+        merge_strategy="weighted_average"
     ))]
     #[allow(clippy::too_many_arguments)]
     fn new(
@@ -246,6 +247,7 @@ impl PyStreamingLowess {
         return_robustness_weights: bool,
         zero_weight_fallback: &str,
         parallel: bool,
+        merge_strategy: &str,
     ) -> PyResult<Self> {
         let overlap_size = overlap.unwrap_or_else(|| {
             let default = chunk_size / 10;
@@ -273,7 +275,7 @@ impl PyStreamingLowess {
                 parallel: Some(parallel),
                 chunk_size: Some(chunk_size),
                 overlap: Some(overlap_size),
-                merge_strategy: None,
+                merge_strategy: Some(merge_strategy),
                 window_capacity: None,
                 min_points: None,
                 update_mode: None,
@@ -323,6 +325,34 @@ impl PyStreamingLowess {
         })?;
 
         Ok(PyLowessResult { inner: result })
+    }
+}
+
+/// Result from a single online update step.
+#[pyclass(name = "OnlineOutput", from_py_object)]
+#[derive(Clone)]
+pub struct PyOnlineOutput {
+    /// Smoothed value for the latest point
+    #[pyo3(get)]
+    pub smoothed: f64,
+    /// Standard error (if computed)
+    #[pyo3(get)]
+    pub std_error: Option<f64>,
+    /// Residual y − smoothed (if computed)
+    #[pyo3(get)]
+    pub residual: Option<f64>,
+    /// Robustness weight for the latest point (if computed)
+    #[pyo3(get)]
+    pub robustness_weight: Option<f64>,
+    /// Number of robustness iterations performed (if tracked)
+    #[pyo3(get)]
+    pub iterations_used: Option<usize>,
+}
+
+#[pymethods]
+impl PyOnlineOutput {
+    fn __repr__(&self) -> String {
+        format!("OnlineOutput(smoothed={:.4})", self.smoothed)
     }
 }
 
@@ -407,14 +437,21 @@ impl PyOnlineLowess {
         })
     }
 
-    /// Add a single point and return smoothed value if enough points are available.
-    fn add_point(&self, x: f64, y: f64) -> PyResult<Option<f64>> {
+    /// Add a single point and return its smoothed value, or None if the window
+    /// is still filling up.
+    fn add_point(&self, x: f64, y: f64) -> PyResult<Option<PyOnlineOutput>> {
         let mut inner = self
             .inner
             .lock()
             .map_err(|e| to_py_error(format!("Mutex poisoned: {}", e)))?;
         let result = inner.add_point(x, y).map_err(to_py_error)?;
-        Ok(result.map(|o| o.smoothed))
[…]
