+++ bindings/python/Cargo.toml
@@ -15,8 +15,8 @@ categories = ["science", "algorithms"]
 
 [dependencies]
 fastLowess = { version = "1.3", features = ["dev", "cpu"] }
-pyo3 = { version = "0.28", features = ["abi3-py38"] }
-numpy = "0.28"
+pyo3 = { version = "0.29", features = ["abi3-py38"] }
+numpy = "0.29"
 
 [lib]
 name = "fastlowess"
