### 2026-07-09 68dc4e5
fix: nodejs binding
