+++ bindings/nodejs/src/lib.rs
@@ -3,10 +3,10 @@
 use napi::bindgen_prelude::*;
 use napi_derive::napi;
 
-use ::fastLowess::internals::binding_support;
 use ::fastLowess::internals::adapters::online::ParallelOnlineLowess;
 use ::fastLowess::internals::adapters::streaming::ParallelStreamingLowess;
 use ::fastLowess::internals::api::{Batch, LowessBuilder, Online, Streaming};
+use ::fastLowess::internals::binding_support;
 use ::fastLowess::prelude::{LowessError, LowessResult};
 
 /// Diagnostic statistics for the LOWESS fit.
@@ -185,10 +185,12 @@ pub struct SmoothOptions {
     pub cv_method: Option<String>,
     /// Number of folds for K-Fold CV. Default: 5.
     pub cv_k: Option<u32>,
+    /// Random seed for reproducible K-Fold cross-validation. Default: None.
+    pub cv_seed: Option<i64>,
+    /// Compute standard errors. Default: false.
+    pub return_se: Option<bool>,
     /// Enable parallel execution. Default: true.
     pub parallel: Option<bool>,
-    /// Per-observation case weights. Must have the same length as input data.
-    pub custom_weights: Option<Vec<f64>>,
 }
 
 /// Batch LOWESS smoothing.
@@ -207,12 +209,36 @@ impl Lowess {
 
     /// Fit the model.
     #[napi]
-    pub fn fit(&self, x: Float64Array, y: Float64Array) -> Result<LowessResultObj> {
-        let builder = self.create_builder()?;
+    pub fn fit(
+        &self,
+        x: Float64Array,
+        y: Float64Array,
+        custom_weights: Option<Vec<f64>>,
+    ) -> Result<LowessResultObj> {
+        let mut builder = self.create_builder()?;
+        if let Some(cw) = custom_weights {
+            if cw.len() != y.as_ref().len() {
+                return Err(Error::new(
+                    Status::InvalidArg,
+                    format!(
+                        "customWeights length ({}) must match y length ({})",
+                        cw.len(),
+                        y.as_ref().len()
+                    ),
+                ));
+            }
+            if cw.iter().any(|&w| w < 0.0) {
+                return Err(Error::new(
+                    Status::InvalidArg,
+                    "customWeights values must be non-negative".to_string(),
+                ));
+            }
+            builder = builder.custom_weights(cw);
+        }
         let model = builder
             .adapter(Batch)
             .build()
-            .map_err(|e| Error::new(Status::GenericFailure, e.to_string()))?;
+            .map_err(|e| Error::new(Status::InvalidArg, e.to_string()))?;
 
         let result = model
             .fit(x.as_ref(), y.as_ref())
@@ -223,8 +249,32 @@ impl Lowess {
 
     /// Fit the model asynchronously.
     #[napi(js_name = "fit_async")]
-    pub fn fit_async(&self, x: Float64Array, y: Float64Array) -> Result<AsyncTask<LowessTask>> {
-        let builder = self.create_builder()?;
+    pub fn fit_async(
+        &self,
+        x: Float64Array,
+        y: Float64Array,
+        custom_weights: Option<Vec<f64>>,
+    ) -> Result<AsyncTask<LowessTask>> {
+        let mut builder = self.create_builder()?;
+        if let Some(cw) = custom_weights {
+            if cw.len() != y.as_ref().len() {
+                return Err(Error::new(
+                    Status::InvalidArg,
+                    format!(
+                        "customWeights length ({}) must match y length ({})",
+                        cw.len(),
+                        y.as_ref().len()
+                    ),
+                ));
+            }
+            if cw.iter().any(|&w| w < 0.0) {
+                return Err(Error::new(
[…]

[…]
