### 2026-07-09 117e916
fix: move all lowess internals re-exports to binding_support in fastLowess
