+++ crates/fastLowess/src/adapters/batch.rs
@@ -26,15 +26,10 @@ use std::fmt::Debug;
 use std::result::Result;
 
 // Export dependencies from lowess crate
-use crate::parse::IntoEnum;
 use lowess::internals::adapters::batch::BatchLowessBuilder;
-use lowess::internals::algorithms::regression::{WLSSolver, ZeroWeightFallback};
-use lowess::internals::algorithms::robustness::RobustnessMethod;
+use lowess::internals::algorithms::regression::WLSSolver;
 use lowess::internals::engine::output::LowessResult;
 use lowess::internals::evaluation::cv::CVKind;
-use lowess::internals::evaluation::intervals::IntervalMethod;
-use lowess::internals::math::boundary::BoundaryPolicy;
-use lowess::internals::math::kernel::WeightFunction;
 use lowess::internals::primitives::backend::Backend;
 use lowess::internals::primitives::errors::LowessError;
 
@@ -59,7 +54,6 @@ impl<T: Float> Default for ParallelBatchLowessBuilder<T> {
 
 #[allow(private_bounds)]
 impl<T: Float> ParallelBatchLowessBuilder<T> {
-    // Create a new batch LOWESS builder with default parameters.
     fn new() -> Self {
         let mut base = BatchLowessBuilder::default();
         base.parallel = Some(true); // Default to parallel in fastLowess
@@ -71,150 +65,6 @@ impl<T: Float> ParallelBatchLowessBuilder<T> {
         }
     }
 
-    // Set parallel execution mode.
-    pub fn parallel(mut self, parallel: bool) -> Self {
-        self.base.parallel = Some(parallel);
-        self
-    }
-
-    // Set the execution backend.
-    pub fn backend(mut self, backend: Backend) -> Self {
-        self.base.backend = Some(backend);
-        self
-    }
-
-    // Set the smoothing fraction (span).
-    pub fn fraction(mut self, fraction: T) -> Self {
-        self.base.fraction = fraction;
-        self
-    }
-
-    // Set the number of robustness iterations.
-    pub fn iterations(mut self, iterations: usize) -> Self {
-        self.base.iterations = iterations;
-        self
-    }
-
-    // Set the delta parameter for interpolation optimization.
-    pub fn delta(mut self, delta: T) -> Self {
-        self.base.delta = Some(delta);
-        self
-    }
-
-    // Set the kernel weight function.
-    pub fn weight_function(mut self, wf: impl IntoEnum<WeightFunction>) -> Self {
-        match wf.into_enum() {
-            Ok(w) => self.base.weight_function = w,
-            Err(e) => {
-                self.parse_errors.push(e);
-            }
-        }
-        self
-    }
-
-    // Set the robustness method for outlier handling.
-    pub fn robustness_method(mut self, method: impl IntoEnum<RobustnessMethod>) -> Self {
-        match method.into_enum() {
-            Ok(m) => self.base.robustness_method = m,
-            Err(e) => {
-                self.parse_errors.push(e);
-            }
-        }
-        self
-    }
-
-    // Set the zero-weight fallback policy.
-    pub fn zero_weight_fallback(mut self, fallback: impl IntoEnum<ZeroWeightFallback>) -> Self {
-        match fallback.into_enum() {
-            Ok(f) => self.base.zero_weight_fallback = f,
-            Err(e) => {
-                self.parse_errors.push(e);
-            }
-        }
-        self
-    }
-
-    // Set the boundary handling policy.
-    pub fn boundary_policy(mut self, policy: impl IntoEnum<BoundaryPolicy>) -> Self {
-        match policy.into_enum() {
-            Ok(p) => self.base.boundary_policy = p,
-            Err(e) => {
-                self.parse_errors.push(e);
-            }
-        }
-        self
-    }
[…]
