+++ crates/fastLowess/src/binding_support.rs
@@ -4,14 +4,22 @@
 //! Node.js, Python, R, and WASM bindings so option aliases and validation
 //! behavior stay consistent across all binding frontends.
 
+use crate::adapters::batch::ParallelBatchLowessBuilder;
+use crate::adapters::online::ParallelOnlineLowessBuilder;
+use crate::adapters::streaming::ParallelStreamingLowessBuilder;
 use crate::api::{LowessBuilder, LowessError};
-use lowess::internals::adapters::online::UpdateMode;
-use lowess::internals::adapters::streaming::MergeStrategy;
-use lowess::internals::algorithms::regression::ZeroWeightFallback;
-use lowess::internals::algorithms::robustness::RobustnessMethod;
-use lowess::internals::math::boundary::BoundaryPolicy;
-use lowess::internals::math::kernel::WeightFunction;
-use lowess::internals::math::scaling::ScalingMethod;
+use crate::parse::IntoEnum;
+use lowess::internals::evaluation::intervals::IntervalMethod;
+pub use lowess::internals::primitives::backend::Backend;
+use num_traits::Float;
+
+pub use lowess::internals::adapters::online::UpdateMode;
+pub use lowess::internals::adapters::streaming::MergeStrategy;
+pub use lowess::internals::algorithms::regression::ZeroWeightFallback;
+pub use lowess::internals::algorithms::robustness::RobustnessMethod;
+pub use lowess::internals::math::boundary::BoundaryPolicy;
+pub use lowess::internals::math::kernel::WeightFunction;
+pub use lowess::internals::math::scaling::ScalingMethod;
 use std::ffi::CString;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -534,4 +542,344 @@ pub fn apply_typed_builder_options(
 
 pub fn lowess_error_message(err: &LowessError) -> String {
     err.to_string()
+}
+
+// ─── Builder setter methods ───────────────────────────────────────────────────
+//
+// These impl blocks are in binding_support.rs (compiled only with the `dev`
+// feature) so the adapter files (batch.rs, online.rs, streaming.rs) stay free
+// of any `#[cfg]` attributes.
+
+#[allow(private_bounds)]
+impl<T: Float> ParallelBatchLowessBuilder<T> {
+    pub fn parallel(mut self, parallel: bool) -> Self {
+        self.base.parallel = Some(parallel);
+        self
+    }
+
+    pub fn backend(mut self, backend: Backend) -> Self {
+        self.base.backend = Some(backend);
+        self
+    }
+
+    pub fn fraction(mut self, fraction: T) -> Self {
+        self.base.fraction = fraction;
+        self
+    }
+
+    pub fn iterations(mut self, iterations: usize) -> Self {
+        self.base.iterations = iterations;
+        self
+    }
+
+    pub fn delta(mut self, delta: T) -> Self {
+        self.base.delta = Some(delta);
+        self
+    }
+
+    pub fn weight_function(mut self, wf: impl IntoEnum<WeightFunction>) -> Self {
+        match wf.into_enum() {
+            Ok(w) => self.base.weight_function = w,
+            Err(e) => {
+                self.parse_errors.push(e);
+            }
+        }
+        self
+    }
+
+    pub fn robustness_method(mut self, method: impl IntoEnum<RobustnessMethod>) -> Self {
+        match method.into_enum() {
+            Ok(m) => self.base.robustness_method = m,
+            Err(e) => {
+                self.parse_errors.push(e);
+            }
+        }
+        self
+    }
+
+    pub fn zero_weight_fallback(mut self, fallback: impl IntoEnum<ZeroWeightFallback>) -> Self {
+        match fallback.into_enum() {
[…]
