### 2026-07-10 16abaa1
feat: centralize aliases
