### 2026-07-10 a2b2800
refactor: binding logic centralization
