+++ bindings/wasm/src/lib.rs
@@ -101,17 +101,38 @@ export class StreamingLowess {
 export class OnlineLowess {
     free(): void;
     constructor(options?: SmoothOptions, onlineOpts?: OnlineOptions);
-    /** Add a single point and get the smoothed value (or undefined if not enough points yet). */
-    add_point(x: number, y: number): number | undefined;
+    /** Add a single point and get the smoothed output (or undefined if not enough points yet). */
+    add_point(x: number, y: number): OnlineOutput | undefined;
+}
+
+/** Result from a single online update step. */
+export class OnlineOutput {
+    free(): void;
+    get smoothed(): number;
+    get std_error(): number | undefined;
+    get residual(): number | undefined;
+    get robustness_weight(): number | undefined;
+    get iterations_used(): number | undefined;
 }
 "#;
 
 use ::fastLowess::internals::adapters::online::ParallelOnlineLowess;
 use ::fastLowess::internals::adapters::streaming::ParallelStreamingLowess;
-use ::fastLowess::internals::api::{Batch, LowessBuilder, Online, Streaming};
+use ::fastLowess::internals::api::LowessBuilder;
 use ::fastLowess::internals::binding_support as shared_parse;
-use ::fastLowess::internals::binding_support::{MergeStrategy, UpdateMode};
 use ::fastLowess::prelude::LowessResult as InnerLowessResult;
+
+fn to_js_error(err: shared_parse::BindingError) -> JsValue {
+    JsValue::from_str(&err.message)
+}
+
+fn map_invalid_arg<T, E: ToString>(result: Result<T, E>) -> Result<T, JsValue> {
+    shared_parse::map_invalid_arg(result).map_err(to_js_error)
+}
+
+fn map_runtime<T, E: ToString>(result: Result<T, E>) -> Result<T, JsValue> {
+    shared_parse::map_runtime(result).map_err(to_js_error)
+}
 
 #[derive(Deserialize)]
 pub struct SmoothOptions {
@@ -135,7 +156,7 @@ pub struct SmoothOptions {
     pub cv_fractions: Option<Vec<f64>>,
     pub cv_method: Option<String>,
     pub cv_k: Option<u32>,
-    pub cv_seed: Option<u32>,
+    pub cv_seed: Option<u64>,
 }
 
 #[derive(Deserialize)]
@@ -164,6 +185,44 @@ pub struct Diagnostics {
     pub effective_df: Option<f64>,
     #[wasm_bindgen(js_name = residual_sd)]
     pub residual_sd: f64,
+}
+
+// Result of a single online update step.
+#[wasm_bindgen]
+pub struct OnlineOutput {
+    smoothed: f64,
+    std_error: Option<f64>,
+    residual: Option<f64>,
+    robustness_weight: Option<f64>,
+    iterations_used: Option<usize>,
+}
+
+#[wasm_bindgen]
+impl OnlineOutput {
+    #[wasm_bindgen(getter)]
+    pub fn smoothed(&self) -> f64 {
+        self.smoothed
+    }
+
+    #[wasm_bindgen(getter, js_name = "std_error")]
+    pub fn std_error(&self) -> Option<f64> {
+        self.std_error
+    }
+
+    #[wasm_bindgen(getter)]
+    pub fn residual(&self) -> Option<f64> {
+        self.residual
+    }
+
+    #[wasm_bindgen(getter, js_name = "robustness_weight")]
+    pub fn robustness_weight(&self) -> Option<f64> {
+        self.robustness_weight
+    }
+
+    #[wasm_bindgen(getter, js_name = "iterations_used")]
+    pub fn iterations_used(&self) -> Option<u32> {
[…]
