+++ bindings/nodejs/src/lib.rs
@@ -5,9 +5,25 @@ use napi_derive::napi;
 
 use ::fastLowess::internals::adapters::online::ParallelOnlineLowess;
 use ::fastLowess::internals::adapters::streaming::ParallelStreamingLowess;
-use ::fastLowess::internals::api::{Batch, LowessBuilder, Online, Streaming};
+use ::fastLowess::internals::api::LowessBuilder;
 use ::fastLowess::internals::binding_support;
-use ::fastLowess::prelude::{LowessError, LowessResult};
+use ::fastLowess::prelude::LowessResult as InnerLowessResult;
+
+fn to_napi_error(err: binding_support::BindingError) -> Error {
+    let status = match err.category {
+        binding_support::BindingErrorCategory::InvalidArg => Status::InvalidArg,
+        binding_support::BindingErrorCategory::Runtime => Status::GenericFailure,
+    };
+    Error::new(status, err.message)
+}
+
+fn map_invalid_arg<T, E: ToString>(result: std::result::Result<T, E>) -> Result<T> {
+    binding_support::map_invalid_arg(result).map_err(to_napi_error)
+}
+
+fn map_runtime<T, E: ToString>(result: std::result::Result<T, E>) -> Result<T> {
+    binding_support::map_runtime(result).map_err(to_napi_error)
+}
 
 /// Diagnostic statistics for the LOWESS fit.
 #[napi(object)]
@@ -17,21 +33,24 @@ pub struct Diagnostics {
     /// Mean Absolute Error.
     pub mae: f64,
     /// R-squared (coefficient of determination).
+    #[napi(js_name = "r_squared")]
     pub r_squared: f64,
     /// Akaike Information Criterion (if computed).
     pub aic: Option<f64>,
     /// Corrected AIC (if computed).
     pub aicc: Option<f64>,
     /// Effective degrees of freedom (if computed).
+    #[napi(js_name = "effective_df")]
     pub effective_df: Option<f64>,
     /// Residual standard deviation.
+    #[napi(js_name = "residual_sd")]
     pub residual_sd: f64,
 }
 
 /// Result of a LOWESS fit.
 #[napi]
 pub struct LowessResultObj {
-    inner: LowessResult<f64>,
+    inner: InnerLowessResult<f64>,
 }
 
 #[napi]
@@ -158,39 +177,95 @@ pub struct SmoothOptions {
     /// Set to 0.0 to disable interpolation.
     pub delta: Option<f64>,
     /// Weight function ("tricube", "epanechnikov", "gaussian", "uniform", "biweight", "triangle", "cosine"). Default: "tricube".
+    #[napi(js_name = "weight_function")]
     pub weight_function: Option<String>,
     /// Robustness method ("bisquare", "huber", "talwar"). Default: "bisquare".
+    #[napi(js_name = "robustness_method")]
     pub robustness_method: Option<String>,
     /// Fallback strategy when weights are zero ("use_local_mean", "return_original", "return_none"). Default: "use_local_mean".
+    #[napi(js_name = "zero_weight_fallback")]
     pub zero_weight_fallback: Option<String>,
     /// Boundary handling ("extend", "reflect", "zero", "noboundary"). Default: "extend".
+    #[napi(js_name = "boundary_policy")]
     pub boundary_policy: Option<String>,
     /// Scaling method ("mad", "mar", "mean"). Default: "mad".
+    #[napi(js_name = "scaling_method")]
     pub scaling_method: Option<String>,
     /// Auto-convergence tolerance. Default: None.
+    #[napi(js_name = "auto_converge")]
     pub auto_converge: Option<f64>,
     /// Return residuals in result. Default: false.
+    #[napi(js_name = "return_residuals")]
[…]
