### 2026-07-10 6031915
build: update dependencies to the latest compatible version
