+++ crates/lowess/Cargo.toml
@@ -27,7 +27,7 @@ include = [
 
 [dependencies]
 num-traits = { version = "0.2", features = ["libm"] }
-wide = { version = "1.4", default-features = false }
+wide = { version = "1.5", default-features = false }
 
 [features]
 default = ["std"]
