+++ crates/fastLowess/Cargo.toml
@@ -30,9 +30,9 @@ lowess = { path = "../lowess", version = "1.3", features = ["dev"] }
 num-traits = { version = "0.2", features = ["libm"] }
 ndarray = { version = "0.17", optional = true }
 rayon = { version = "1.12", optional = true }
-wgpu = { version = "29.0", optional = true }
+wgpu = { version = "30.0", optional = true }
 bytemuck = { version = "1.25", optional = true }
-pollster = { version = "0.4", optional = true }
+pollster = { version = "1.0", optional = true }
 futures-intrusive = { version = "0.5", optional = true }
 
 [features]
