### 2026-07-10 cad2eab
docs: update docs for new API
