+++ docs/api/julia.md
@@ -68,10 +68,10 @@ online = OnlineLowess(; kwargs...)
 **Methods:**
 
 ```julia
-result = add_point(online, x::Float64, y::Float64) :: Union{Float64, Nothing}
+result = add_point(online, x[1], y[1])  # returns OnlineOutput or nothing
 ```
 
-* Adds a single point to the model. Returns the smoothed value as a `Float64`, or `nothing` if not enough points have been accumulated yet.
+* Adds a single point to the sliding window. Returns `nothing` while the window is still filling (fewer than `min_points` seen), and an `OnlineOutput` once smoothing begins.
 
 ## Options Structures
 
@@ -94,44 +94,60 @@ result = add_point(online, x::Float64, y::Float64) :: Union{Float64, Nothing}
 | `return_residuals` | `Bool` | `false` | Include residuals in result |
 | `return_robustness_weights` | `Bool` | `false` | Include weights in result |
 | `parallel` | `Bool` | `true` | Enable parallel execution |
-| `cv_method` | `String` | `""` | Cross-validation method ("kfold") |
-| `cv_k` | `Int` | `5` | Number of CV folds |
-| `cv_fractions` | `Vector{Float64}` | `[]` | Manual fractions for CV grid |
-| `custom_weights` | `Vector{Float64}` | `nothing` | Per-observation weights (Batch only) |
+| `cv_method` | `String` | `"kfold"` | CV method (`"kfold"` or `"loocv"`) (Batch only) |
+| `cv_k` | `Int` | `5` | Number of folds for k-fold CV (Batch only) |
+| `cv_fractions` | `Vector{Float64}` | `Float64[]` | Fractions to test for cross-validation (Batch only) |
+| `cv_seed` | `Union{Int, Nothing}` | `nothing` | Random seed for cross-validation shuffling (Batch only) |
+| `custom_weights` | `Union{Vector{Float64}, Nothing}` | `nothing` | Per-observation case weights — passed to `fit()`, not the constructor (Batch only) |
 
 ### `StreamingOptions` (inherits `LowessOptions`)
 
 | Field | Type | Default | Description |
 | --- | --- | --- | --- |
 | `chunk_size` | `Int` | `5000` | Data chunk size |
-| `overlap` | `Int` | `500` | Overlap size (-1 for auto) |
-| `merge_strategy` | `String` | `"weighted"` | Merge strategy for overlap |
+| `overlap` | `Int` | `500` | Overlap between chunks |
+| `merge_strategy` | `String` | `"weighted_average"` | Strategy for blending overlap regions |
 
 ### `OnlineOptions` (inherits `LowessOptions`)
 
 | Field | Type | Default | Description |
 | --- | --- | --- | --- |
-| `window_capacity` | `Int` | `1000` | Max window size |
-| `min_points` | `Int` | `2` | Min points before smoothing |
-| `update_mode` | `String` | `"incremental"` | Update mode ("full" or "incremental") |
+| `window_capacity` | `Int` | `1000` | Max points in sliding window |
+| `min_points` | `Int` | `3` | Min points before smoothing starts |
+| `update_mode` | `String` | `"full"` | Update mode (`"full"` or `"incremental"`) |
+| `parallel` | `Bool` | `false` | Enable parallel execution (off by default; online LOWESS fits one point at a time) |
 
 ## Result Structure
+
+### `OnlineOutput`
+
+Returned by `add_point()` once the window has enough points (`nothing` until then).
+
[…]
