+++ docs/user-guide/robustness.md
@@ -12,7 +12,9 @@ Standard LOWESS can be biased by outliers. Robustness iterations downweight poin
 4. Refit using combined distance × robustness weights
 5. Repeat steps 2–4
 
-![Robust vs Standard](../assets/diagrams/robust_vs_standard_lowess.svg)
+![Robustness Methods](../assets/diagrams/robust_method_comparison.svg)
+
+![Robustness Iterations](../assets/diagrams/robust_iter_comparison.svg)
 
 ---
 
@@ -33,38 +35,38 @@ $$w(u) = \begin{cases} (1 - u^2)^2 & |u| < 1 \\ 0 & |u| \geq 1 \end{cases}$$
 
 === "Python"
     ```python
-    result = fl.smooth(x, y, iterations=3, robustness_method="bisquare")
+    result = fl.Lowess(iterations=3, robustness_method="bisquare").fit(x, y)
     ```
 
 === "Rust"
     ```rust
     let model = Lowess::new()
         .iterations(3)
-        .robustness_method(Bisquare)
+        .robustness_method("bisquare")
         .build()?;
     ```
 
 === "Julia"
     ```julia
-    result = smooth(x, y, iterations=3, robustness_method="bisquare")
+    result = fit(Lowess(; iterations=3, robustness_method="bisquare"), x, y)
     ```
 
 === "Node.js"
     ```javascript
-    const result = smooth(x, y, { iterations: 3, robustness_method: "bisquare" });
+    const result = new Lowess({ iterations: 3, robustness_method: "bisquare" }).fit(x, y);
     ```
 
 === "WebAssembly"
     ```javascript
-    const result = smooth(x, y, { iterations: 3, robustness_method: "bisquare" });
+    const result = new Lowess({ iterations: 3, robustness_method: "bisquare" }).fit(x, y);
     ```
 
 === "C++"
     ```cpp
-    auto result = fastlowess::smooth(x, y, {
-        .iterations = 3,
+    fastlowess::Lowess model({ .iterations = 3,
         .robustness_method = "bisquare"
-    });
+     });
+    auto result = model.fit(x, y).value();
     ```
 
 ---
@@ -84,38 +86,38 @@ $$w(u) = \begin{cases} 1 & |u| \leq k \\ k/|u| & |u| > k \end{cases}$$
 
 === "Python"
     ```python
-    result = fl.smooth(x, y, iterations=3, robustness_method="huber")
+    result = fl.Lowess(iterations=3, robustness_method="huber").fit(x, y)
     ```
 
 === "Rust"
     ```rust
     let model = Lowess::new()
         .iterations(3)
-        .robustness_method(Huber)
+        .robustness_method("huber")
         .build()?;
     ```
 
 === "Julia"
     ```julia
-    result = smooth(x, y, iterations=3, robustness_method="huber")
+    result = fit(Lowess(; iterations=3, robustness_method="huber"), x, y)
     ```
 
 === "Node.js"
     ```javascript
-    const result = smooth(x, y, { iterations: 3, robustness_method: "huber" });
+    const result = new Lowess({ iterations: 3, robustness_method: "huber" }).fit(x, y);
     ```
 
 === "WebAssembly"
     ```javascript
-    const result = smooth(x, y, { iterations: 3, robustness_method: "huber" });
+    const result = new Lowess({ iterations: 3, robustness_method: "huber" }).fit(x, y);
     ```
 
 === "C++"
     ```cpp
-    auto result = fastlowess::smooth(x, y, {
-        .iterations = 3,
+    fastlowess::Lowess model({ .iterations = 3,
         .robustness_method = "huber"
-    });
+     });
+    auto result = model.fit(x, y).value();
     ```
 
 ---
[…]
