### 2026-07-10 7e06335
fix: compilation, tests, lint, and clippy
