### 2026-07-10 d75d16c
docs: fix custome_weights passage to fit() instead of model
