### 2026-07-10 5204312
docs: add the same model->result snippet pattern to all languages
