+++ docs/index.md
@@ -223,7 +223,8 @@ See the [Installation Guide](getting-started/installation.md) for more options a
     x = np.array([1.0, 2.0, 3.0, 4.0, 5.0])
     y = np.array([2.0, 4.1, 5.9, 8.2, 9.8])
 
-    result = fl.Lowess(fraction=0.5, iterations=3).fit(x, y)
+    model = fl.Lowess(fraction=0.5, iterations=3)
+    result = model.fit(x, y)
     print(result.y)
     ```
 
@@ -252,7 +253,8 @@ See the [Installation Guide](getting-started/installation.md) for more options a
     x = [1.0, 2.0, 3.0, 4.0, 5.0]
     y = [2.0, 4.1, 5.9, 8.2, 9.8]
 
-    result = fit(Lowess(; fraction=0.5, iterations=3), x, y)
+    model = Lowess(; fraction=0.5, iterations=3)
+    result = fit(model, x, y)
     println(result.y)
     ```
 
@@ -264,7 +266,8 @@ See the [Installation Guide](getting-started/installation.md) for more options a
     const x = new Float64Array([1, 2, 3, 4, 5]);
     const y = new Float64Array([2.0, 4.1, 5.9, 8.2, 9.8]);
 
-    const result = new fl.Lowess({ fraction: 0.5, iterations: 3 }).fit(x, y);
+    const model = new fl.Lowess({ fraction: 0.5, iterations: 3 });
+    const result = model.fit(x, y);
     console.log(result.y);
     ```
 
@@ -276,7 +279,8 @@ See the [Installation Guide](getting-started/installation.md) for more options a
     const x = new Float64Array([1, 2, 3, 4, 5]);
     const y = new Float64Array([2.0, 4.1, 5.9, 8.2, 9.8]);
 
-    const result = new Lowess({ fraction: 0.5, iterations: 3 }).fit(x, y);
+    const model = new Lowess({ fraction: 0.5, iterations: 3 });
+    const result = model.fit(x, y);
     console.log(result.y);
     ```
 
