+++ docs/examples/nodejs.md
@@ -62,7 +62,8 @@ const x = Array.from({ length: 100 }, (_, i) => i * 0.1);
 const y = x.map(xi => Math.sin(xi) + Math.random() * 0.2);
 
 // Basic smoothing
-const result = new Lowess({ fraction: 0.3 }).fit(x, y);
+const model = new Lowess({ fraction: 0.3 });
+const result = model.fit(x, y);
 console.log('Smoothed values:', result.y);
 
 // With options
@@ -89,5 +90,6 @@ const options = {
     confidence_intervals: 0.95
 };
 
-const result: LowessResult = new Lowess(options).fit(x, y);
+const model = new Lowess(options);
+const result: LowessResult = model.fit(x, y);
 ```
