### 2026-07-10 3f96b96
refactor: centralize and unify defaults
