### 2026-07-10 df86fd1
refactor: centralize and unify the rest of defaults
