+++ crates/lowess/src/defaults.rs
@@ -8,6 +8,14 @@
 //! (`STREAMING_`, `ONLINE_`). Shared constants (used by every
 //! adapter) have no prefix.
 
+use crate::adapters::online::UpdateMode;
+use crate::adapters::streaming::MergeStrategy;
+use crate::algorithms::regression::ZeroWeightFallback;
+use crate::algorithms::robustness::RobustnessMethod;
+use crate::math::boundary::BoundaryPolicy;
+use crate::math::kernel::WeightFunction;
+use crate::math::scaling::ScalingMethod;
+
 // ── Streaming adapter ─────────────────────────────────────────────────────────
 
 // Default number of data points per chunk.
@@ -17,6 +25,7 @@ pub const DEFAULT_STREAMING_CHUNK_SIZE: usize = 5_000;
 pub const DEFAULT_STREAMING_OVERLAP: usize = 500;
 
 // Default merge strategy for overlapping chunk regions.
+pub const DEFAULT_STREAMING_MERGE_STRATEGY_ENUM: MergeStrategy = MergeStrategy::WeightedAverage;
 pub const DEFAULT_STREAMING_MERGE_STRATEGY: &str = "weighted_average";
 
 // ── Online adapter ────────────────────────────────────────────────────────────
@@ -28,6 +37,7 @@ pub const DEFAULT_ONLINE_WINDOW_CAPACITY: usize = 1_000;
 pub const DEFAULT_ONLINE_MIN_POINTS: usize = 2;
 
 // Default update mode for the **Online** adapter.
+pub const DEFAULT_ONLINE_UPDATE_MODE_ENUM: UpdateMode = UpdateMode::Incremental;
 pub const DEFAULT_ONLINE_UPDATE_MODE: &str = "incremental";
 
 // ── Shared (all adapters) ─────────────────────────────────────────────────────
@@ -38,26 +48,49 @@ pub const DEFAULT_FRACTION: f64 = 0.67;
 // Default number of robustness iterations.
 pub const DEFAULT_ITERATIONS: usize = 3;
 
-// Default interpolation-optimisation threshold.
-//
-// `0.0` means "auto-compute as 1 % of the x-range" for the **Batch**
-// adapter; **Streaming** and **Online** adapters use `0.0` as a literal
-// (optimisation disabled).
+// Default interpolation-optimisation threshold for **Streaming** and **Online** adapters.
+// `0.0` disables the optimisation (every point is fit individually).
 pub const DEFAULT_DELTA: f64 = 0.0;
 
+// Default interpolation-optimisation threshold for the **Batch** adapter.
+// `None` means auto-compute as 1 % of the x-range.
+pub const fn default_batch_delta<T>() -> Option<T> {
+    None
+}
+
+// Default auto-convergence tolerance: `None` disables early stopping.
+pub const fn default_auto_converge<T>() -> Option<T> {
+    None
+}
+
+// Default confidence-interval level: `None` means no confidence intervals.
+pub const fn default_confidence_level<T>() -> Option<T> {
+    None
+}
+
+// Default prediction-interval level: `None` means no prediction intervals.
+pub const fn default_prediction_level<T>() -> Option<T> {
+    None
+}
+
 // Default kernel weight function.
+pub const DEFAULT_WEIGHT_FUNCTION_ENUM: WeightFunction = WeightFunction::Tricube;
 pub const DEFAULT_WEIGHT_FUNCTION: &str = "tricube";
 
 // Default robustness weighting method.
[…]
