+++ crates/lowess/src/adapters/online.rs
@@ -134,17 +134,17 @@ impl<T: Float> OnlineLowessBuilder<T> {
             window_capacity: crate::defaults::DEFAULT_ONLINE_WINDOW_CAPACITY,
             min_points: crate::defaults::DEFAULT_ONLINE_MIN_POINTS,
             fraction: T::from(crate::defaults::DEFAULT_FRACTION).unwrap(),
-            delta: T::zero(),
+            delta: T::from(crate::defaults::DEFAULT_DELTA).unwrap(),
             iterations: crate::defaults::DEFAULT_ITERATIONS,
-            weight_function: WeightFunction::default(),
-            update_mode: UpdateMode::default(),
-            robustness_method: RobustnessMethod::default(),
-            zero_weight_fallback: ZeroWeightFallback::default(),
-            boundary_policy: BoundaryPolicy::default(),
-            scaling_method: ScalingMethod::default(),
-            compute_residuals: false,
-            return_robustness_weights: false,
-            auto_converge: None,
+            weight_function: crate::defaults::DEFAULT_WEIGHT_FUNCTION_ENUM,
+            update_mode: crate::defaults::DEFAULT_ONLINE_UPDATE_MODE_ENUM,
+            robustness_method: crate::defaults::DEFAULT_ROBUSTNESS_METHOD_ENUM,
+            zero_weight_fallback: crate::defaults::DEFAULT_ZERO_WEIGHT_FALLBACK_ENUM,
+            boundary_policy: crate::defaults::DEFAULT_BOUNDARY_POLICY_ENUM,
+            scaling_method: crate::defaults::DEFAULT_SCALING_METHOD_ENUM,
+            compute_residuals: crate::defaults::DEFAULT_RETURN_RESIDUALS,
+            return_robustness_weights: crate::defaults::DEFAULT_RETURN_ROBUSTNESS_WEIGHTS,
+            auto_converge: crate::defaults::default_auto_converge(),
             deferred_error: None,
             custom_smooth_pass: None,
             custom_cv_pass: None,
