### 2026-07-10 8655785
refactor: merge alias and parse into api, break defaults to module-level
