+++ crates/lowess/src/alias.rs
@@ -1,278 +1,0 @@
-//! Centralised string-alias maps for all option enums.
-//!
-//! All `impl FromStr` blocks for option types live here so that accepted
-//! aliases and their canonical spellings stay in one place.  Every binding
-//! frontend (`fastLowess::binding_support`) delegates to these impls instead
-//! of maintaining its own duplicated match arms.
-//!
-//! ## Canonical names
-//!
-//! The first alias listed in each match arm is the canonical (round-trip)
-//! name.  The `*_str` helpers in `fastLowess::binding_support` always return
-//! the canonical name so that `parse → str → parse` round-trips correctly.
-
-// Feature-gated imports
-#[cfg(not(feature = "std"))]
-use alloc::string::ToString;
-
-// External dependencies
-use core::str::FromStr;
-
-// Internal dependencies
-use crate::adapters::online::UpdateMode;
-use crate::adapters::streaming::MergeStrategy;
-use crate::algorithms::regression::ZeroWeightFallback;
-use crate::algorithms::robustness::RobustnessMethod;
-use crate::math::boundary::BoundaryPolicy;
-use crate::math::kernel::WeightFunction;
-use crate::math::scaling::ScalingMethod;
-use crate::primitives::errors::LowessError;
-
-// WeightFunction
-
-impl FromStr for WeightFunction {
-    type Err = LowessError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "cosine" => Ok(WeightFunction::Cosine),
-            "epanechnikov" => Ok(WeightFunction::Epanechnikov),
-            "gaussian" => Ok(WeightFunction::Gaussian),
-            "biweight" | "bisquare" => Ok(WeightFunction::Biweight),
-            "triangle" | "triangular" => Ok(WeightFunction::Triangle),
-            "tricube" => Ok(WeightFunction::Tricube),
-            "uniform" | "boxcar" => Ok(WeightFunction::Uniform),
-            _ => Err(LowessError::InvalidOption {
-                option: "weight_function",
-                value: s.to_string(),
-                valid: "tricube, epanechnikov, gaussian, uniform, biweight, triangle, cosine",
-            }),
-        }
-    }
-}
-
-// BoundaryPolicy
-
-impl FromStr for BoundaryPolicy {
-    type Err = LowessError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "extend" | "pad" => Ok(BoundaryPolicy::Extend),
-            "reflect" | "mirror" => Ok(BoundaryPolicy::Reflect),
-            "zero" => Ok(BoundaryPolicy::Zero),
-            "noboundary" | "none" => Ok(BoundaryPolicy::NoBoundary),
-            _ => Err(LowessError::InvalidOption {
-                option: "boundary_policy",
-                value: s.to_string(),
-                valid: "extend, reflect, zero, noboundary",
-            }),
-        }
-    }
-}
-
-// ScalingMethod
-
-impl FromStr for ScalingMethod {
-    type Err = LowessError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "mar" | "median_absolute_residual" => Ok(ScalingMethod::MAR),
-            "mad" | "median_absolute_deviation" => Ok(ScalingMethod::MAD),
-            "mean" | "mean_absolute_residual" => Ok(ScalingMethod::Mean),
-            _ => Err(LowessError::InvalidOption {
-                option: "scaling_method",
-                value: s.to_string(),
[…]
