+++ tests/fastLowess/gpu_tests.rs
@@ -1,10 +1,16 @@
 #![cfg(feature = "dev")]
 #![cfg(feature = "gpu")]
+use Backend::{CPU, GPU};
+use RobustnessMethod::Bisquare;
+use ScalingMethod::Mean;
+use WeightFunction::{Biweight, Cosine, Epanechnikov, Gaussian, Triangle, Tricube, Uniform};
 use approx::assert_abs_diff_eq;
 use fastLowess::internals::engine::gpu::{GLOBAL_EXECUTOR, GpuConfig, GpuExecutor};
-use fastLowess::prelude::*;
-use lowess::internals::evaluation::cv::KFold;
+use lowess::internals::algorithms::robustness::RobustnessMethod;
 use lowess::internals::math::boundary::BoundaryPolicy;
+use lowess::internals::math::kernel::WeightFunction;
+use lowess::internals::math::scaling::ScalingMethod;
+use lowess::internals::primitives::backend::Backend;
 use lowess::internals::primitives::window::Window;
 use lowess::prelude::Lowess;
 
@@ -85,7 +91,9 @@ fn test_gpu_cv_reduction() {
     // GPU Build
     let model = Lowess::new()
         .backend(GPU)
-        .cv_config(KFold(5, &[0.1, 0.2, 0.5]))
+        .cv_method("kfold")
+        .cv_k(5)
+        .cv_fractions(vec![0.1, 0.2, 0.5])
         .delta(0.0) // Force exact fit
         .build()
         .unwrap();
@@ -98,7 +106,9 @@ fn test_gpu_cv_reduction() {
         // CPU Build to compare
         let cpu_model = Lowess::new()
             .backend(CPU)
-            .cv_config(KFold(5, &[0.1, 0.2, 0.5]))
+            .cv_method("kfold")
+            .cv_k(5)
+            .cv_fractions(vec![0.1, 0.2, 0.5])
             .build()
             .unwrap();
         let cpu_res = cpu_model.fit(&x, &y).unwrap();
@@ -301,69 +311,78 @@ fn test_cpu_gpu_padding_equivalence() {
         // CPU Padding
         let (cpu_px, cpu_py) = apply_boundary_policy(&x, &y, window_size, policy);
 
-        // GPU Padding
-        let mut exec_lock = match GLOBAL_EXECUTOR.lock() {
-            Ok(g) => g,
-            Err(p) => p.into_inner(),
-        };
-        if exec_lock.is_none() {
-            match block_on(GpuExecutor::new()) {
-                Ok(e) => *exec_lock = Some(e),
-                Err(_) => {
-                    println!("GPU not available, skipping test");
-                    return;
+        // GPU Padding — lock is held only for the GPU work, released before assertions
+        let gpu_result: Option<(Vec<f32>, Vec<f32>)> = {
+            let mut exec_lock = match GLOBAL_EXECUTOR.lock() {
+                Ok(g) => g,
+                Err(p) => p.into_inner(),
+            };
+            if exec_lock.is_none() {
+                match block_on(GpuExecutor::new()) {
+                    Ok(e) => *exec_lock = Some(e),
+                    Err(_) => {
+                        println!("GPU not available, skipping test");
+                        return;
+                    }
                 }
             }
-        }
-        let exec = exec_lock.as_mut().unwrap();
+            let exec = exec_lock.as_mut().unwrap();
 
-        let pad_len = window_size / 2;
-        let total_n_padded = n + 2 * pad_len;
+            let pad_len = window_size / 2;
+            let total_n_padded = n + 2 * pad_len;
 
-        let gpu_config = GpuConfig {
-            n: total_n_padded as u32,
-            window_size: window_size as u32,
-            weight_function: 0,
[…]
