### 2026-07-11 eacfd6c
fix: gpu tests, r binding return_se order
