+++ bindings/r/man/OnlineLowess.Rd
@@ -5,24 +5,24 @@
 \title{LOWESS Online Smoothing}
 \usage{
 OnlineLowess(
-  fraction = 0.67,
-  window_capacity = 1000L,
-  min_points = 3L,
-  iterations = 3L,
-  delta = NULL,
-  weight_function = "tricube",
-  robustness_method = "bisquare",
-  scaling_method = "mad",
-  boundary_policy = "extend",
-  zero_weight_fallback = "use_local_mean",
-  update_mode = "full",
-  auto_converge = NULL,
-  return_robustness_weights = FALSE,
-  return_diagnostics = FALSE,
-  return_residuals = FALSE,
-  parallel = FALSE,
-  confidence_intervals = NULL,
-  prediction_intervals = NULL
+    fraction = 0.67,
+    window_capacity = 1000L,
+    min_points = 3L,
+    iterations = 3L,
+    delta = NULL,
+    weight_function = "tricube",
+    robustness_method = "bisquare",
+    scaling_method = "mad",
+    boundary_policy = "extend",
+    zero_weight_fallback = "use_local_mean",
+    update_mode = "full",
+    auto_converge = NULL,
+    return_robustness_weights = FALSE,
+    return_diagnostics = FALSE,
+    return_residuals = FALSE,
+    parallel = FALSE,
+    confidence_intervals = NULL,
+    prediction_intervals = NULL
 )
 }
 \arguments{
