### 2026-07-11 baf13fd
fix: verify_snippets
