+++ dev/verify_snippets.py
@@ -1,0 +1,1631 @@
+"""Extract fenced code blocks from docs/ and run them to verify correctness.
+
+Each snippet is prefixed with language-specific boilerplate that provides common
+sample data and imports, so fragment-style doc examples can execute end-to-end.
+
+Usage
+-----
+    python dev/verify_snippets.py                    # run all supported languages
+    python dev/verify_snippets.py --lang python      # Python only
+    python dev/verify_snippets.py --lang nodejs      # Node.js only
+    python dev/verify_snippets.py --lang julia       # Julia only
+    python dev/verify_snippets.py --lang r           # R only
+    python dev/verify_snippets.py --lang wasm        # WebAssembly only
+    python dev/verify_snippets.py --lang rust        # Rust only
+    python dev/verify_snippets.py --lang cpp         # C++ only
+    python dev/verify_snippets.py --file docs/api/python.md
+    python dev/verify_snippets.py --dry-run          # list snippets, don't run
+    python dev/verify_snippets.py --verbose          # show snippet source on failure
+    python dev/verify_snippets.py --output out.json  # also write JSON report
+    python dev/verify_snippets.py --timeout 60       # per-snippet timeout (seconds)
+    python dev/verify_snippets.py --stop-on-fail     # exit after first failure
+"""
+
+from __future__ import annotations
+
+import argparse
+import concurrent.futures
+import json
+import os
+import re
+import subprocess
+import sys
+import tempfile
+import textwrap
+import threading
+import time
+from collections import defaultdict
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Iterator, List, Optional, Tuple
+
+# ---------------------------------------------------------------------------
+# Paths
+# ---------------------------------------------------------------------------
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+DOCS_DIR = REPO_ROOT / "docs"
+
+# ---------------------------------------------------------------------------
+# Terminal colours (disabled on non-TTY or Windows without colour support)
+# ---------------------------------------------------------------------------
+
+_USE_COLOR = sys.stdout.isatty() and os.name != "nt" or os.environ.get("FORCE_COLOR")
+
+
+def _c(code: str, text: str) -> str:
+    return f"\033[{code}m{text}\033[0m" if _USE_COLOR else text
+
+
+def green(t: str) -> str:
+    return _c("32", t)
+
+
+def red(t: str) -> str:
+    return _c("31", t)
+
+
+def yellow(t: str) -> str:
+    return _c("33", t)
+
+
+def cyan(t: str) -> str:
+    return _c("36", t)
+
+
+def bold(t: str) -> str:
+    return _c("1", t)
+
+
+# ---------------------------------------------------------------------------
+# Python executable detection (prefer venv where fastlowess is installed)
+# ---------------------------------------------------------------------------
+
+_PYTHON_BIN: str = sys.executable  # may be replaced in main()
+
+
+def _find_python_with_fastlowess() -> str:
+    """Return the best Python executable that has fastlowess installed."""
+    candidates = [
[…]
