+++ docs/user-guide/intervals.md
@@ -38,9 +38,9 @@ Estimate uncertainty in the smoothed curve itself.
     model = fl.Lowess(fraction=0.5, confidence_intervals=0.95)
     result = model.fit(x, y)
 
-    print("Smoothed:", result["y"])
-    print("CI Lower:", result["confidence_lower"])
-    print("CI Upper:", result["confidence_upper"])
+    print("Smoothed:", result.y)
+    print("CI Lower:", result.confidence_lower)
+    print("CI Upper:", result.confidence_upper)
     ```
 
 === "Rust"
@@ -133,8 +133,8 @@ Estimate where new observations might fall.
     model = fl.Lowess(fraction=0.5, prediction_intervals=0.95)
     result = model.fit(x, y)
 
-    print("PI Lower:", result["prediction_lower"])
-    print("PI Upper:", result["prediction_upper"])
+    print("PI Lower:", result.prediction_lower)
+    print("PI Upper:", result.prediction_upper)
     ```
 
 === "Rust"
@@ -326,7 +326,7 @@ Access standard errors directly (available when intervals are computed):
     ```python
     model = fl.Lowess(confidence_intervals=0.95)
     result = model.fit(x, y)
-    print("Standard errors:", result["standard_errors"])
+    print("Standard errors:", result.standard_errors)
     ```
 
 === "Rust"
