### 2026-07-11 cc05e25
fix: r docs tests
