### 2026-07-11 d06330c
fix: wasm docs tests
