### 2026-07-11 01040e1
fix: cpp compiling and docs tests
