+++ .cargo/config.toml
@@ -1,2 +1,6 @@
 [target.x86_64-pc-windows-msvc.env]
 CC = "clang-cl"
+
+[target.x86_64-pc-windows-gnu]
+linker = "c:/rtools45/x86_64-w64-mingw32.static.posix/bin/x86_64-w64-mingw32.static.posix-gcc.exe"
+ar = "c:/rtools45/x86_64-w64-mingw32.static.posix/bin/x86_64-w64-mingw32.static.posix-ar.exe"

[…]
