### 2026-07-11 a52f9c9
fix: julia docs tests
