### 2026-07-11 093e60e
fix: rust docs tests
