+++ dev/verify_snippets.py
@@ -132,14 +132,15 @@ _TAB_ALIASES: dict[str, set[str]] = {
 _LANG_TAGS: dict[str, set[str]] = {
     "python": {"python"},
     "julia": {"julia"},
-    "nodejs": {"javascript", "js", "typescript", "ts"},
+    "nodejs": {"javascript", "js"},
     "wasm": {"javascript", "js"},
     "r": {"r"},
     "cpp": {"cpp", "c++"},
     "rust": {"rust"},
 }
 
-_PYTHON_PREAMBLE = textwrap.dedent("""\
+_PYTHON_PREAMBLE = ""
+_PYTHON_PREAMBLE_UNUSED = textwrap.dedent("""\
     # --- snippet preamble: suppress display back-end -------------------------
     import sys as _sys, os as _os
     try:
@@ -218,7 +219,8 @@ _PYTHON_PREAMBLE = textwrap.dedent("""\
     # -------------------------------------------------------------------------
 """)
 
-_JULIA_PREAMBLE = textwrap.dedent("""\
+_JULIA_PREAMBLE = ""
+_JULIA_PREAMBLE_UNUSED = textwrap.dedent("""\
     # --- snippet preamble ----------------------------------------------------
     using FastLOWESS
     using Random, Printf, Statistics
@@ -267,7 +269,8 @@ _JULIA_PREAMBLE = textwrap.dedent("""\
     # -------------------------------------------------------------------------
 """)
 
-_NODEJS_PREAMBLE = textwrap.dedent("""\
+_NODEJS_PREAMBLE = ""
+_NODEJS_PREAMBLE_UNUSED = textwrap.dedent("""\
     // --- snippet preamble ----------------------------------------------------
     'use strict';
     const fl = (() => { try { return require('fastlowess'); } catch (e) { return null; } })();
@@ -307,6 +310,21 @@ _NODEJS_PREAMBLE = textwrap.dedent("""\
 """)
 
 _R_PREAMBLE = textwrap.dedent("""\
+    suppressMessages({{
+        .libPaths(c(
+            normalizePath(file.path(
+                Sys.getenv("LOWESS_REPO_ROOT", "{repo_root}"),
+                "bindings", "r", ".r-lib"
+            ), mustWork = FALSE),
+            .libPaths()
+        ))
+        library(rfastlowess)
+    }})
+    suppressWarnings(pdf(NULL))
+    plot.new()
+""").format(repo_root=str(REPO_ROOT).replace("\\", "/"))
+
+_R_PREAMBLE_UNUSED = textwrap.dedent("""\
     # --- snippet preamble ----------------------------------------------------
     suppressMessages({{
         ### NOTE: KEEP
@@ -369,7 +387,8 @@ _R_PREAMBLE = textwrap.dedent("""\
 
 _WASM_PKG_DIR = REPO_ROOT / "bindings" / "wasm" / "pkg"
 
-_WASM_PREAMBLE = textwrap.dedent("""\
+_WASM_PREAMBLE = ""
+_WASM_PREAMBLE_UNUSED = textwrap.dedent("""\
     // --- snippet preamble (WASM) ---------------------------------------------
     'use strict';
     const _wasmPkg = (() => {{
@@ -411,8 +430,8 @@ _WASM_PREAMBLE = textwrap.dedent("""\
     // -------------------------------------------------------------------------
 """).format(wasm_pkg=str(_WASM_PKG_DIR).replace("\\", "/"))
 
-# Rust snippets are wrapped: top + snippet + bottom
[…]
