+++ dev/fix_doc_snippets.py
@@ -1,0 +1,967 @@
+#!/usr/bin/env python3
+"""
+fix_doc_snippets.py — Transform doc snippets to be self-contained.
+
+Adds minimal boilerplate to each code snippet that would otherwise be skipped
+because it lacks imports, fn main, int main, require(), etc.
+
+Usage:
+    python3 dev/fix_doc_snippets.py [--dry-run] [--lang LANG]
+
+Without --dry-run, modifies the doc files in place.
+"""
+
+from __future__ import annotations
+
+import argparse
+import re
+import sys
+from pathlib import Path
+
+REPO_ROOT = Path(__file__).parent.parent
+sys.path.insert(0, str(REPO_ROOT))
+
+from dev.verify_snippets import (  # noqa: E402
+    DOCS_DIR,
+    Snippet,
+    extract_snippets,
+    should_skip,
+)
+
+# ---------------------------------------------------------------------------
+# Standard boilerplate per language
+# ---------------------------------------------------------------------------
+
+# Python: standard imports + 100-pt noisy sine wave
+_PY_STD = """\
+import fastlowess as fl
+import numpy as np
+
+rng = np.random.default_rng(42)
+x = np.linspace(0, 2 * np.pi, 100)
+y = np.sin(x) + rng.normal(0, 0.3, 100)
+
+"""
+
+# Python: when the snippet uses matplotlib (time-series tutorial)
+_PY_TIMESERIES = """\
+import fastlowess as fl
+import numpy as np
+import matplotlib.pyplot as plt
+
+np.random.seed(42)
+t = np.linspace(0, 100, 500)
+trend_true = 10 + 0.5 * t + 3 * np.sin(t / 10)
+y = trend_true + np.random.normal(0, 3, len(t))
+
+"""
+
+# Python: genomics tutorial (uses positions / coverage)
+_PY_GENOMICS = """\
+import fastlowess as fl
+import numpy as np
+
+np.random.seed(42)
+positions = np.arange(0, 10000, 10, dtype=float)
+coverage = np.random.poisson(50, len(positions)).astype(float)
+
+"""
+
+# Python: merge.md (uses x_chunk / y_chunk)
+_PY_MERGE = """\
+import fastlowess as fl
+import numpy as np
+
+x_chunk = np.linspace(0, np.pi, 50)
+y_chunk = np.sin(x_chunk) + np.random.default_rng(42).normal(0, 0.1, 50)
+
+"""
+
+# Python: custom-weights.md extra vars (calibration_indices / sigma / Lowess)
+_PY_CUSTOM_WEIGHTS_EXTRA = """\
+import fastlowess as fl
+from fastlowess import Lowess
+import numpy as np
+
+rng = np.random.default_rng(42)
+x = np.linspace(0, 2 * np.pi, 100)
+y = np.sin(x) + rng.normal(0, 0.3, 100)
+calibration_indices = [5, 20, 40, 60, 80]
+sigma = rng.uniform(0.1, 0.5, 100)
+
+"""
+
+# Node.js: standard require + data
+_JS_STD = """\
+const {{ {imports} }} = require('fastlowess');
+
+const n = 100;
+const x = Float64Array.from({{ length: n }}, (_, i) => i * 2 * Math.PI / (n - 1));
+const y = Float64Array.from(x, (xi, i) => Math.sin(xi) + (((i * 7 + 3) % 17) / 17 - 0.5) * 0.6);
+
+"""
+
+# Node.js: merge.md (uses xChunk / yChunk)
+_JS_MERGE = """\
+const { StreamingLowess } = require('fastlowess');
+
+const n = 50;
[…]
