### 2026-07-11 99c9bb1
docs: unify docs interface between bindings/crates
