+++ dev/verify_snippets.py
@@ -140,174 +140,10 @@ _LANG_TAGS: dict[str, set[str]] = {
 }
 
 _PYTHON_PREAMBLE = ""
-_PYTHON_PREAMBLE_UNUSED = textwrap.dedent("""\
-    # --- snippet preamble: suppress display back-end -------------------------
-    import sys as _sys, os as _os
-    try:
-        import matplotlib as _mpl
-        _mpl.use("Agg")
-        import matplotlib.pyplot as plt
-        plt.show = lambda *a, **kw: None
-    except ImportError:
-        import sys as _sys2, types as _types
-        _plt_stub = _types.ModuleType('matplotlib.pyplot')
-        _plt_stub.__getattr__ = lambda _n: (lambda *a, **kw: _plt_stub)  # type: ignore
-        _sys2.modules.setdefault('matplotlib', _types.ModuleType('matplotlib'))
-        _sys2.modules.setdefault('matplotlib.pyplot', _plt_stub)
-        plt = _plt_stub  # type: ignore
-
-    # --- snippet preamble: common imports ------------------------------------
-    import numpy as np
-    import fastlowess as fl
-
-    # --- snippet preamble: sample data ---------------------------------------
-    np.random.seed(42)
-    _n = 100
-    x = np.linspace(0.0, 10.0, _n)
-    y = np.sin(x) + np.random.normal(0, 0.1, _n)
-
-    # Aliases used in various tutorials
-    t = x.copy()
-    t_irregular = x.copy()
-    y_irregular = y.copy()
-    positions = x.copy()
-    observed = y.copy()
-    times = x.copy()
-    temperatures = y.copy()
-    hours = x.copy()
-    expression = y.copy()
-    coverage = np.abs(y.copy()) * 10 + 5
-
-    # Multivariate input variables (lat/lon/x1/x2/x3 for dimensions.md examples)
-    lat = x.copy(); lon = x * 0.5
-    x1 = x.copy(); x2 = x * 0.5; x3 = x * 0.25
-    z = np.sin(x) + np.cos(x * 0.5)
-    # Python binding requires flattened 1D for multi-dim input
-    x2d = np.column_stack([x, x * 0.5]).ravel()    # (200,) flat row-major
-    x3d = np.column_stack([x, x * 0.5, x * 0.25]).ravel()  # (300,) flat
-
-    # Outlier / weight examples
-    y_with_outlier = y.copy();  y_with_outlier[50] = 100.0
-    weights = np.ones(_n)
-
-    # Streaming / chunk examples
-    chunk_size, overlap = 50, 10
-    chunk1_x, chunk1_y = x[:50].copy(), y[:50].copy()
-    chunk2_x, chunk2_y = x[50:].copy(), y[50:].copy()
-    x_chunk, y_chunk = x[:50].copy(), y[:50].copy()
-
-    # Sliding-window examples
-    data_x = list(x[:30])
-    data_y = list(y[:30])
-
-    # Calibration / uncertainty examples
-    calibration_indices = np.array([2, 5, 7, 20, 50])
-    sigma = np.random.default_rng(42).uniform(0.1, 0.5, _n)
-
-    # API doc pseudocode helpers
-    fastlowess = fl   # allow docs that use bare "fastlowess.X(...)"
-    kwargs = {'fraction': 0.3}
-    model = fl.Lowess()
-    stream = fl.StreamingLowess()
-    online = fl.OnlineLowess()
-    # Bare-name aliases so snippets that import with "from fastlowess import ..."
-    # or use the class name directly without fl. prefix work out of the box.
-    Lowess = fl.Lowess
-    StreamingLowess = fl.StreamingLowess
-    OnlineLowess = fl.OnlineLowess
[…]
