### 2026-07-11 dcf79bd
fix: examples syntax errors
