+++ examples/julia/batch_smoothing.jl
@@ -19,140 +19,140 @@ using Printf
 using Pkg
 project_name = Pkg.project().name
 if project_name != "FastLOWESS"
-    # Not in the FastLOWESS project, need to develop it
-    script_dir = @__DIR__
-    julia_pkg_dir = joinpath(dirname(script_dir), "julia")
-    if !haskey(Pkg.project().dependencies, "FastLOWESS")
-        Pkg.develop(path = julia_pkg_dir)
-    end
+	# Not in the FastLOWESS project, need to develop it
+	script_dir = @__DIR__
+	julia_pkg_dir = joinpath(dirname(script_dir), "julia")
+	if !haskey(Pkg.project().dependencies, "FastLOWESS")
+		Pkg.develop(path = julia_pkg_dir)
+	end
 end
 
 using FastLOWESS
 
 function generate_sample_data(n_points = 1000)
-    """
-    	Generate complex sample data with a trend, seasonality, and outliers.
-    	"""
-    Random.seed!(42)
-    x = collect(range(0, 50, length = n_points))
+	"""
+		Generate complex sample data with a trend, seasonality, and outliers.
+		"""
+	Random.seed!(42)
+	x = collect(range(0, 50, length = n_points))
 
-    # Trend + Seasonality
-    y_true = 0.5 .* x .+ 5 .* sin.(x .* 0.5)
+	# Trend + Seasonality
+	y_true = 0.5 .* x .+ 5 .* sin.(x .* 0.5)
 
-    # Gaussian noise
-    y = y_true .+ randn(n_points) .* 1.5
+	# Gaussian noise
+	y = y_true .+ randn(n_points) .* 1.5
 
-    # Add significant outliers (10% of data)
-    n_outliers = Int(round(n_points * 0.1))
-    outlier_indices = randperm(n_points)[1:n_outliers]
-    for i ∈ outlier_indices
-        y[i] += rand(10:20) * rand([-1, 1])
-    end
+	# Add significant outliers (10% of data)
+	n_outliers = Int(round(n_points * 0.1))
+	outlier_indices = randperm(n_points)[1:n_outliers]
+	for i ∈ outlier_indices
+		y[i] += rand(10:20) * rand([-1, 1])
+	end
 
-    return x, y, y_true
+	return x, y, y_true
 end
 
 function main()
-    println("=== FastLOWESS Batch Smoothing Example ===")
+	println("=== FastLOWESS Batch Smoothing Example ===")
 
-    # 1. Generate Data
-    x, y, y_true = generate_sample_data(1000)
-    println("Generated $(length(x)) data points with outliers.")
+	# 1. Generate Data
+	x, y, y_true = generate_sample_data(1000)
+	println("Generated $(length(x)) data points with outliers.")
 
-    # 2. Basic Smoothing (Default parameters)
-    println("Running basic smoothing...")
-    # 2. Basic Smoothing (Default parameters)
-    println("Running basic smoothing...")
-    # Use a smaller fraction (0.05) to capture the sine wave seasonality
-    l_basic = Lowess(iterations = 0, fraction = 0.05)
-    res_basic = fit(l_basic, x, y)
+	# 2. Basic Smoothing (Default parameters)
+	println("Running basic smoothing...")
+	# 2. Basic Smoothing (Default parameters)
+	println("Running basic smoothing...")
+	# Use a smaller fraction (0.05) to capture the sine wave seasonality
[…]
