### 2026-07-11 5f41ab3
ci: split julia release workflow to a two-stage flow
