+++ .github/workflows/release-julia-jll.yml
@@ -1,0 +1,268 @@
+name: Release Julia JLL
+
+on:
+  release:
+    types: [published]
+  workflow_dispatch:
+    inputs:
+      version:
+        description: "Version to update to (without v prefix, e.g., 0.1.2)"
+        required: true
+        type: string
+
+env:
+  FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true
+
+permissions:
+  contents: write
+  pull-requests: write
+
+jobs:
+  update-yggdrasil:
+    name: Update Yggdrasil
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Check required secrets
+        env:
+          YGGDRASIL_TOKEN: ${{ secrets.YGGDRASIL_TOKEN }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          if [ -z "$YGGDRASIL_TOKEN" ]; then
+            echo "⚠️ Warning: YGGDRASIL_TOKEN secret is not set."
+            echo "Please add a PAT with 'repo' permissions as a secret named YGGDRASIL_TOKEN."
+            
+            if [ -n "$GITHUB_TOKEN" ]; then
+              echo "✅ Falling back to GITHUB_TOKEN (might fail if no fork permissions)..."
+            else
+              echo "❌ Error: No tokens available."
+              exit 1
+            fi
+          else
+            echo "✅ YGGDRASIL_TOKEN is configured."
+          fi
+
+      - name: Checkout Source (Lowess Project)
+        uses: actions/checkout@v6
+        with:
+          path: lowess-project
+
+      - name: Determine Version and Commit
+        id: meta
+        run: |
+          cd lowess-project
+          if [ "${{ github.event_name }}" == "release" ]; then
+            VERSION="${{ github.event.release.tag_name }}"
+            VERSION="${VERSION#v}"
+            COMMIT="${{ github.sha }}"
+            echo "Triggered by release: $VERSION ($COMMIT)"
+          else
+            VERSION="${{ github.event.inputs.version }}"
+            COMMIT=$(git rev-parse HEAD)
+            echo "Triggered manually: $VERSION ($COMMIT)"
+          fi
+
+          # Validate version format
+          if ! echo "$VERSION" | grep -qE '^[0-9]+\.[0-9]+\.[0-9]+$'; then
+            echo "❌ Invalid version format: $VERSION (expected x.y.z)"
+            exit 1
+          fi
+
+          # Validate commit hash
+          if ! echo "$COMMIT" | grep -qE '^[0-9a-f]{40}$'; then
+            echo "❌ Invalid commit hash: $COMMIT (expected 40 hex chars)"
+            exit 1
+          fi
+
+          echo "version=$VERSION" >> $GITHUB_OUTPUT
+          echo "commit=$COMMIT" >> $GITHUB_OUTPUT
+
+      - name: Prepare Build Script
+        id: script
+        env:
+          VERSION: ${{ steps.meta.outputs.version }}
+          COMMIT: ${{ steps.meta.outputs.commit }}
+        run: |
+          # Read the template from the source repo
+          CONTENT=$(cat lowess-project/dev/build_tarballs_julia.jl)
+
+          # Replace the top-level package version only.
+          CONTENT=$(echo "$CONTENT" | sed -E "s/^version = v\"[^\"]*\"$/version = v\"$VERSION\"/")
+
+          # Replace only the GitSource commit hash line.
+          CONTENT=$(echo "$CONTENT" | sed -E "s/^([[:space:]]*)\"[0-9a-f]{40}\"\),$/\1\"$COMMIT\"),/")
+
+          # Verify replacements
[…]
