### 2026-07-11 f67a34f
ci: bootstrap pak package on windows runner
