+++ .github/workflows/R-CMD-check.yaml
@@ -53,6 +53,18 @@ jobs:
       - name: Setup Pandoc
         uses: r-lib/actions/setup-pandoc@v2
 
+      - name: Bootstrap pak (Windows)
+        if: runner.os == 'Windows'
+        shell: Rscript {0}
+        run: |
+          install.packages(
+            "pak",
+            repos = sprintf(
+              "https://r-lib.github.io/p/pak/stable/%s/%s/%s",
+              .Platform$pkgType, R.Version()$os, R.Version()$arch
+            )
+          )
+
       - name: "R: Setup Dependencies"
         uses: r-lib/actions/setup-r-dependencies@v2
         with:
