### 2026-07-11 2aef750
fix: allow useless_conversion for julia cv_seed
