### 2026-07-11 faaa338
ci: suppress compilation warnings in SRR check
