### 2026-07-11 4f04e70
chore: pass --no-check-urls to R cmd check
