+++ .github/workflows/R-CMD-check.yaml
@@ -92,7 +92,7 @@ jobs:
         with:
           upload-snapshots: true
           build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
-          args: 'c("--no-manual", "--as-cran")'
+          args: 'c("--no-manual", "--as-cran", "--no-check-urls")'
           check-dir: '"check"'
           error-on: '"error"'
           upload-results: false
