### 2026-07-11 0dd9b60
chore: clean project root
