+++ .readthedocs.yaml
@@ -7,7 +7,7 @@ build:
     python: "3.12"
 
 mkdocs:
-  configuration: mkdocs.yml
+  configuration: docs/mkdocs.yml
 
 python:
   install:
