### 2026-07-11 958eafd
build: bump rust crate versions to 2.0.0
