+++ crates/lowess/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 # Metadata
 name = "lowess"
-version = "1.3.0"
+version = "2.0.0"
 authors = ["Amir Valizadeh <thisisamirv@gmail.com>"]
 edition = "2024"
 rust-version = "1.89"
