### 2026-07-11 6db4515
build: bump bindings to v2.0.0
