+++ bindings/nodejs/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fastlowess-node"
-version = "1.3.0"
+version = "2.0.0"
 authors = ["Amir Valizadeh <thisisamirv@gmail.com>"]
 edition = "2024"
 license = "MIT OR Apache-2.0"
@@ -17,7 +17,7 @@ categories = ["science", "algorithms"]
 crate-type = ["cdylib"]
 
 [dependencies]
-fastLowess = { version = "1.3", features = ["dev", "cpu"] }
+fastLowess = { version = "2.0", features = ["dev", "cpu"] }
 napi = { version = "3.8", features = ["napi4"] }
 napi-derive = "3.5"
 
