+++ bindings/julia/julia/Project.toml
@@ -1,6 +1,6 @@
 name = "FastLOWESS"
 uuid = "5d760c97-4851-4f57-8eaf-1b3e829bbf83"
-version = "1.3.0"
+version = "2.0.0"
 authors = ["Amir Valizadeh <thisisamirv@gmail.com>"]
 edition = "2024"
 license = "MIT OR Apache-2.0"
@@ -18,7 +18,7 @@ fastlowess_jll = "7381d214-3001-5784-8d9a-8eb8e8688076"
 
 [compat]
 julia = "1.6"
-fastlowess_jll = "1.3.0"
+fastlowess_jll = "2.0.0"
 Test = "1"
 Random = "1"
 

[…]
