### 2026-07-12 023a486
ci: fix julia release
