+++ .github/workflows/release-julia-jll.yml
@@ -90,7 +90,7 @@ jobs:
           CONTENT=$(echo "$CONTENT" | sed -E "s/^version = v\"[^\"]*\"$/version = v\"$VERSION\"/")
 
           # Replace only the GitSource commit hash line.
-          CONTENT=$(echo "$CONTENT" | sed -E "s/^([[:space:]]*)\"[0-9a-f]{40}\"\),$/\1\"$COMMIT\"),/")
+          CONTENT=$(echo "$CONTENT" | sed -E "s/(GitSource\(\"[^\"]*\", \")([0-9a-f]{40})(\"\),)/\1$COMMIT\3/")
 
           # Verify replacements
           echo "=== Updated Script (first 30 lines) ==="
@@ -104,7 +104,7 @@ jobs:
           fi
 
           # Verify commit was replaced
-          if ! echo "$CONTENT" | grep -qE "^[[:space:]]*\"$COMMIT\"\),$"; then
+          if ! echo "$CONTENT" | grep -qE "GitSource\(\"[^\"]*\", \"$COMMIT\"\),"; then
             echo "❌ Failed to update commit hash in script"
             exit 1
           fi
