### 2026-07-12 5230802
build: bump nodejs modules to v2.0.0
