### 2026-07-12 46b3bf6
docs: update benchmarks
