+++ benchmarks/fastLowess/src/bench.rs
@@ -1,424 +1,0 @@
-//! Industry-level LOWESS benchmarks using Criterion.
-//!
-//! Benchmarks cover:
-//! - Scalability (1K to 50K points)
-//! - Algorithm parameters (fraction, iterations)
-//! - Real-world scenarios (financial, scientific, genomic)
-//! - Pathological cases (outliers, clustered data, high noise)
-//!
-//! For CPU backend and serial execution in bash, use
-//! For CPU backend and serial execution, use `FASTLOWESS_BACKEND=serial cargo bench`.
-//! For CPU backend and parallel execution, use `FASTLOWESS_BACKEND=parallel cargo bench`.
-//! For GPU backend, use `FASTLOWESS_BACKEND=gpu cargo bench`.
-//!
-//! For CPU backend and serial execution in PowerShell, use
-//! `$env:FASTLOWESS_BACKEND='serial'; cargo bench`.
-//! For CPU backend and parallel execution in PowerShell, use
-//! `$env:FASTLOWESS_BACKEND='parallel'; cargo bench`.
-//! For GPU backend in PowerShell, use
-//! `$env:FASTLOWESS_BACKEND='gpu'; cargo bench --features gpu`.
-
-use criterion::{BenchmarkId, Criterion, Throughput, criterion_group, criterion_main};
-use fastLowess::prelude::*;
-use lowess::internals::primitives::backend::Backend;
-use rand::prelude::*;
-use rand_distr::{Normal, Uniform};
-use std::env;
-use std::f64::consts::PI;
-use std::hint::black_box;
-
-// ============================================================================
-// Helper Functions
-// ============================================================================
-
-fn get_config() -> (bool, Backend, &'static str) {
-    match env::var("FASTLOWESS_BACKEND").ok().as_deref() {
-        Some("cpu_serial") | Some("serial") => (false, Backend::CPU, "serial"),
-        Some("gpu") => (true, Backend::GPU, "gpu"),
-        Some("cpu") | Some("parallel") => (true, Backend::CPU, "parallel"),
-        _ => (true, Backend::CPU, "parallel"),
-    }
-}
-
-// ============================================================================
-// Data Generation with Reproducible RNG
-// ============================================================================
-
-/// Generate smooth sinusoidal data with Gaussian noise.
-fn generate_sine_data(size: usize, seed: u64) -> (Vec<f64>, Vec<f64>) {
-    let mut rng = StdRng::seed_from_u64(seed);
-    let noise_dist = Normal::new(0.0, 0.2).unwrap();
-
-    let x: Vec<f64> = (0..size).map(|i| i as f64 * 10.0 / size as f64).collect();
-    let y: Vec<f64> = x
-        .iter()
-        .map(|&xi| xi.sin() + noise_dist.sample(&mut rng))
-        .collect();
-    (x, y)
-}
-
-/// Generate data with outliers (5% of points are extreme).
-fn generate_outlier_data(size: usize, seed: u64) -> (Vec<f64>, Vec<f64>) {
-    let mut rng = StdRng::seed_from_u64(seed);
-    let noise_dist = Normal::new(0.0, 0.2).unwrap();
-    let outlier_dist = Uniform::new(-5.0, 5.0).unwrap();
-
-    let x: Vec<f64> = (0..size).map(|i| i as f64 * 10.0 / size as f64).collect();
-    let mut y: Vec<f64> = x
-        .iter()
-        .map(|&xi| xi.sin() + noise_dist.sample(&mut rng))
-        .collect();
-
-    // Add 5% outliers
-    let n_outliers = size / 20;
-    for _ in 0..n_outliers {
-        let idx = rng.random_range(0..size);
[…]
