+++ benchmarks/fastLowess/src/lib.rs
@@ -1,2 +1,0 @@
-// Dummy library file to satisfy Cargo package requirements for benchmarks
-pub fn benchmark_helper() {}
