+++ benchmarks/convert_criterion.py
@@ -1,234 +1,0 @@
-#!/usr/bin/env python3
-"""
-Convert Criterion benchmark results to JSON format for comparison.
-
-Criterion stores results in:
-- target/criterion/<group>/<bench_id>/<param>/new/estimates.json (parameterized)
-- target/criterion/<group>/<bench_name>/new/estimates.json (non-parameterized)
-
-This script extracts timing data and writes it to:
-- benchmarks/output/rust_benchmark_cpu.json (parallel)
-- benchmarks/output/rust_benchmark_cpu_serial.json (serial)
-
-Usage: python3 convert_criterion.py
-"""
-
-import json
-import sys
-from pathlib import Path
-from typing import Dict, List
-
-IGNORED_CRITERION_DIRS = {"report", "new", "base", "change"}
-GROUP_SUFFIXES = ("_parallel", "_serial", "_gpu")
-
-
-def strip_backend_suffix(category: str) -> str:
-    """Return a category name without its backend suffix."""
-    for suffix in GROUP_SUFFIXES:
-        if category.endswith(suffix):
-            return category[: -len(suffix)]
-    return category
-
-
-def build_benchmark_name(clean_category: str, bench_id: str, param: str) -> str:
-    """Return the exported benchmark name for one Criterion result."""
-    if clean_category == "scalability":
-        return f"scale_{param}"
-    if clean_category in {
-        "financial",
-        "scientific",
-        "genomic",
-        "fraction",
-        "iterations",
-    }:
-        return f"{clean_category}_{param}"
-    return f"{bench_id}_{param}"
-
-
-def build_result_record(name: str, size: int, timing: dict) -> dict:
-    """Build one normalized exported benchmark result record."""
-    return {
-        "name": name,
-        "size": size,
-        "iterations": 10,
-        **timing,
-    }
-
-
-def parse_size(param: str) -> int:
-    """Convert a parameter string to an integer size when possible."""
-    try:
-        return int(param)
-    except ValueError:
-        return 0
-
-
-def parse_estimates(estimates_file: Path) -> dict | None:
-    """Parse a criterion estimates.json file and return timing in ms."""
-    try:
-        with estimates_file.open("r", encoding="utf-8") as file:
-            estimates = json.load(file)
-
-        # Criterion stores in nanoseconds
-        mean_ns = estimates.get("mean", {}).get("point_estimate", 0)
-        std_ns = estimates.get("std_dev", {}).get("point_estimate", 0)
-        median_ns = estimates.get("median", {}).get("point_estimate", 0)
-
-        # Try to get min/max from sample data
-        sample_file = estimates_file.parent / "sample.json"
-        min_ns = mean_ns
-        max_ns = mean_ns
-        if sample_file.exists():
-            try:
-                with sample_file.open("r", encoding="utf-8") as file:
-                    sample = json.load(file)
-                times = sample.get("times", [])
-                if times:
-                    min_ns = min(times)
-                    max_ns = max(times)
-            except (OSError, json.JSONDecodeError, TypeError, ValueError):
-                pass
-
-        return {
-            "mean_time_ms": mean_ns / 1_000_000,
-            "std_time_ms": std_ns / 1_000_000,
-            "median_time_ms": median_ns / 1_000_000,
-            "min_time_ms": min_ns / 1_000_000,
-            "max_time_ms": max_ns / 1_000_000,
-        }
-    except (OSError, json.JSONDecodeError, TypeError, ValueError) as error:
-        print(f"Error parsing {estimates_file}: {error}")
-        return None
-
-
-def find_criterion_results(criterion_dir: Path) -> Dict[str, List[dict]]:
-    """Find all criterion benchmark results and organize by category."""
[…]

[…]
