### 2026-07-13 fac7a23
ci: expand python support
