### 2026-08-15 52a5591
fix: make r python finding, tomli installation on ubuntu
