+++ bindings/r/DESCRIPTION
@@ -18,4 +18,4 @@ VignetteBuilder: knitr
 BugReports: https://github.com/thisisamirv/lowess-project/issues
 URL: https://github.com/thisisamirv/lowess-project, https://lowess.readthedocs.io/
 Config/rextendr/version: 0.5.0
-Config/roxygen2/version: 8.0.0
+Config/roxygen2/version: 8.1.0
