### 2026-08-15 036952e
fix: remove isolate_cargo.py, fix python lint issues
