+++ tests/python/test_gil.py
@@ -1,5 +1,6 @@
 """Regression test that verifies heavy LOWESS fitting releases the Python GIL."""
 
+import sys
 import threading
 import time
 
@@ -64,7 +65,7 @@ def test_gil_release():
         print(
             f"FAIL: Main thread was blocked! Got {ticks} ticks, expected > {expected_ticks:.2f}"
         )
-        exit(1)
+        sys.exit(1)
     else:
         print("PASS: Main thread remained responsive.")
 
