+++ Makefile
@@ -211,15 +211,8 @@ endif
 # ==============================================================================
 # lowess crate
 # ==============================================================================
-ISOLATE ?= true
-all: ISOLATE := false
-
 lowess:
-	@if [ "$(ISOLATE)" = "true" ]; then \
-		$(PYTHON) dev/isolate_cargo.py crates/lowess -- "$(MAKE)" _lowess_impl; \
-	else \
-		"$(MAKE)" _lowess_impl; \
-	fi
+	@"$(MAKE)" _lowess_impl
 
 _lowess_impl:
 	@echo "Running $(LOWESS_PKG) crate checks..."
@@ -285,11 +278,7 @@ lowess-clean:
 # fastLowess crate
 # ==============================================================================
 fastLowess:
-	@if [ "$(ISOLATE)" = "true" ]; then \
-		$(PYTHON) dev/isolate_cargo.py crates/fastLowess -- "$(MAKE)" _fastLowess_impl; \
-	else \
-		"$(MAKE)" _fastLowess_impl; \
-	fi
+	@"$(MAKE)" _fastLowess_impl
 
 _fastLowess_impl:
 	@echo "Running $(FASTLOWESS_PKG) crate checks..."
@@ -351,11 +340,7 @@ fastLowess-clean:
 # Python bindings
 # ==============================================================================
 python:
-	@if [ "$(ISOLATE)" = "true" ]; then \
-		$(PYTHON) dev/isolate_cargo.py bindings/python -- "$(MAKE)" _python_impl; \
-	else \
-		"$(MAKE)" _python_impl; \
-	fi
+	@"$(MAKE)" _python_impl
 
 _python_impl:
 	@echo "Running $(PY_PKG) checks..."
@@ -417,11 +402,7 @@ python-clean:
 # R bindings
 # ==============================================================================
 r:
-	@if [ "$(ISOLATE)" = "true" ]; then \
-		$(PYTHON) dev/isolate_cargo.py bindings/r -- "$(MAKE)" _r_impl; \
-	else \
-		"$(MAKE)" _r_impl; \
-	fi
+	@"$(MAKE)" _r_impl
 
 _r_impl:
 	@echo "Running $(R_PKG_NAME) checks..."
@@ -596,11 +577,7 @@ r-clean:
 # Julia bindings
 # ==============================================================================
 julia:
-	@if [ "$(ISOLATE)" = "true" ]; then \
-		$(PYTHON) dev/isolate_cargo.py bindings/julia -- "$(MAKE)" _julia_impl; \
-	else \
-		"$(MAKE)" _julia_impl; \
-	fi
+	@"$(MAKE)" _julia_impl
 
 _julia_impl:
 	@echo "Running $(JL_PKG) checks..."
@@ -706,11 +683,7 @@ julia-clean:
 # Node.js bindings
 # ==============================================================================
 nodejs:
-	@if [ "$(ISOLATE)" = "true" ]; then \
-		$(PYTHON) dev/isolate_cargo.py bindings/nodejs -- "$(MAKE)" _nodejs_impl; \
-	else \
-		"$(MAKE)" _nodejs_impl; \
-	fi
+	@"$(MAKE)" _nodejs_impl
 
 _nodejs_impl:
 	@echo "Running $(NODE_PKG) checks..."
@@ -748,11 +721,7 @@ nodejs-clean:
 # WebAssembly bindings
 # ==============================================================================
 wasm:
-	@if [ "$(ISOLATE)" = "true" ]; then \
[…]
