### 2026-08-15 cabebda
fix: remove now dead check_root_cargo.py
