+++ docs/changelog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documented in `docs/contributing.md` that `python3` with `tomli` and `tomli_w` is required to run the R binding's vendoring scripts, and why Python is needed for an R build.
 - Removed `dev/isolate_cargo.py` and the `ISOLATE` Makefile variable. All per-component targets (`lowess`, `fastLowess`, `python`, `r`, `julia`, `nodejs`, `wasm`, `cpp`) now call their `_*_impl` sub-target directly; Cargo's `-p <package>` flag already scopes each build to the relevant crate without workspace mutation.
 - Removed `dev/check_root_cargo.py`. This script guarded against `Cargo.toml` being left in an isolated state by `isolate_cargo.py`; since workspace isolation no longer happens, the guard is unnecessary.
+- Removed `dev/fix_doc_snippets.py`. All 80 documentation code snippets that previously required runtime transformation (missing imports, WASM ES-module rewrites, R/Julia data preambles, Node.js `require` injection) now carry their boilerplate directly in the Markdown source.
 
 **Python:**
 

[…]
