### 2026-08-15 c42ec3b
build: remove prepare_cargo
