### 2026-08-15 56f9b88
build: replace patch_vendor_crates with idiomatic sed commands
