+++ dev/patch_vendor_crates.py
@@ -1,219 +1,0 @@
-#!/usr/bin/env python3
-"""
-Patch vendored Cargo.toml files for R package build.
-
-This script replaces workspace inheritance with concrete values and removes
-GPU dependencies that are not needed for the R binding.
-"""
-
-import argparse
-import importlib
-import sys
-from pathlib import Path
-from typing import Any, Dict
-
-try:
-    import tomllib
-except ImportError:
-    import tomli as tomllib  # type: ignore
-
-
-def load_tomli_w():
-    """Import `tomli_w` lazily so editor-only envs do not fail static import resolution."""
-    return importlib.import_module("tomli_w")
-
-
-def load_workspace_values(workspace_toml: Path) -> Dict[str, Any]:
-    """Load workspace package values from root Cargo.toml."""
-    with open(workspace_toml, "rb") as file_handle:
-        data = tomllib.load(file_handle)
-
-    workspace = data.get("workspace", {})
-    package = workspace.get("package", {})
-    deps = workspace.get("dependencies", {})
-
-    return {
-        "package": package,
-        "dependencies": deps,
-    }
-
-
-def resolve_workspace_value(value: Any, key: str, workspace: Dict[str, Any]) -> Any:
-    """Resolve a workspace = true value to its concrete value."""
-    if isinstance(value, dict) and value.get("workspace") is True:
-        # Get from workspace.package
-        pkg_value = workspace["package"].get(key)
-        if pkg_value is not None:
-            # Merge with any other fields in the original value
-            result = {k: v for k, v in value.items() if k != "workspace"}
-            if isinstance(pkg_value, dict):
-                result.update(pkg_value)
-            else:
-                return pkg_value  # Simple value replacement
-            return result if result else pkg_value
-        return None  # Will be removed
-    return value
-
-
-def resolve_workspace_dep(dep_name: str, value: Any, workspace: Dict[str, Any]) -> Any:
-    """Resolve a workspace dependency to its concrete value."""
-    if isinstance(value, dict) and value.get("workspace") is True:
-        ws_dep = workspace["dependencies"].get(dep_name, {})
-        if isinstance(ws_dep, str):
-            # Simple version string in workspace
-            result = {"version": ws_dep}
-        elif isinstance(ws_dep, dict):
-            result = dict(ws_dep)
-        else:
-            result = {}
-
-        # Merge with local overrides (features, optional, etc.)
-        for k, v in value.items():
-            if k != "workspace":
-                result[k] = v
-
-        return result
-    return value
-
-
-def patch_crate_toml(
-    crate_toml: Path,
-    workspace: Dict[str, Any],
-    remove_gpu: bool = False,
-    lowess_as_path: bool = False,
-) -> None:
-    """Patch a crate's Cargo.toml to remove workspace inheritance."""
-    with open(crate_toml, "rb") as file_handle:
-        data = tomllib.load(file_handle)
-
-    # Patch [package] section
-    package = data.get("package", {})
-    keys_to_remove = []
-    for key, value in package.items():
-        resolved = resolve_workspace_value(value, key, workspace)
-        if resolved is None:
-            keys_to_remove.append(key)
-        else:
-            package[key] = resolved
-
-    for key in keys_to_remove:
-        del package[key]
-
-    # Patch [dependencies] section
-    if "dependencies" in data:
-        deps = data["dependencies"]
-        deps_to_remove = []
-
-        for dep_name, value in deps.items():
-            # Remove GPU deps if requested
-            if remove_gpu and dep_name in (
-                "wgpu",
-                "bytemuck",
-                "pollster",
-                "futures-intrusive",
-            ):
[…]
