+++ docs/changelog.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed `dev/check_root_cargo.py`. This script guarded against `Cargo.toml` being left in an isolated state by `isolate_cargo.py`; since workspace isolation no longer happens, the guard is unnecessary.
 - Removed `dev/fix_doc_snippets.py`. All 80 documentation code snippets that previously required runtime transformation (missing imports, WASM ES-module rewrites, R/Julia data preambles, Node.js `require` injection) now carry their boilerplate directly in the Markdown source.
 - Removed `dev/prepare_cargo.py`. The two actions it provided — (1) stripping `[workspace]`/`[patch.crates-io]` before vendoring and (2) appending them back afterward — are now performed inline in the Makefiles with `sed` and `printf`. The `exclude`/`restore` actions it also defined were never called.
+- Removed `dev/patch_vendor_crates.py`. The crates it patched (`fastLowess`, `lowess`) do not use workspace inheritance, so the script's only real work was removing GPU optional-dependencies and stripping the `version` field from the `lowess` path dep in `fastLowess/Cargo.toml`. Both are now done with a single `sed -i.bak` call inline in the Makefiles. This also eliminates the `tomli`/`tomli_w` pip-install step from the R build.
 
 **Python:**
 
