### 2026-08-15 fd1e2a8
build: remove check_js_licenses, replace clean_checksums with commands inline in the Makefiles
