+++ docs/changelog.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed `dev/fix_doc_snippets.py`. All 80 documentation code snippets that previously required runtime transformation (missing imports, WASM ES-module rewrites, R/Julia data preambles, Node.js `require` injection) now carry their boilerplate directly in the Markdown source.
 - Removed `dev/prepare_cargo.py`. The two actions it provided — (1) stripping `[workspace]`/`[patch.crates-io]` before vendoring and (2) appending them back afterward — are now performed inline in the Makefiles with `sed` and `printf`. The `exclude`/`restore` actions it also defined were never called.
 - Removed `dev/patch_vendor_crates.py`. The crates it patched (`fastLowess`, `lowess`) do not use workspace inheritance, so the script's only real work was removing GPU optional-dependencies and stripping the `version` field from the `lowess` path dep in `fastLowess/Cargo.toml`. Both are now done with a single `sed -i.bak` call inline in the Makefiles. This also eliminates the `tomli`/`tomli_w` pip-install step from the R build.
+- Removed `dev/clean_checksums.py`. The two things it did — strip `tests`/`benches`/`examples`/`doc` directories from the vendor tree and reset `.cargo-checksum.json` files — are now done with two `find` commands inline in the Makefiles. Cargo accepts `{"files":{}}` checksums for vendored crates so per-file verification is disabled after stripping, removing the need to recompute hashes. The R build now requires no Python scripts at all.
+- Removed `check_js_licenses.js` as it was just needed once during development.
 
 **Python:**
 
