### 2026-08-15 eaba921
build: break root Makefile into sub Makefiles at binding/crate level
