### 2026-08-16 7e2b615
build: inline makefile configs
