+++ bindings/python/Makefile
@@ -1,6 +1,35 @@
 .SHELLFLAGS := -ec
 
-include mk/config.mk
+UNAME_S := $(shell uname -s)
+
+ifeq ($(OS),Windows_NT)
+	HOST_PLATFORM  := windows
+	PATH_SEPARATOR := ;
+	STAT_SIZE_CMD  := stat -c%s
+	NPM := npm.cmd
+	NPX := npx.cmd
+else ifeq ($(UNAME_S),Darwin)
+	HOST_PLATFORM  := macos
+	PATH_SEPARATOR := :
+	STAT_SIZE_CMD  := stat -f%z
+	NPM := npm
+	NPX := npx
+else
+	HOST_PLATFORM  := linux
+	PATH_SEPARATOR := :
+	STAT_SIZE_CMD  := stat -c%s
+	NPM := npm
+	NPX := npx
+endif
+
+PYTHON      ?= python3
+PYO3_PYTHON ?= $(PYTHON)
+NODE        ?= node
+
+TEMP ?= /tmp
+ifeq ($(OS),Windows_NT)
+	TEMP := /tmp
+endif
 
 PY_PKG      := fastLowess-py
 PY_DIR      := bindings/python
