### 2026-08-16 b014e5a
refactor: move tests and examples to crate/binding level
