+++ tests/r/testthat/test-fastlowess.R
@@ -1,262 +1,0 @@
-#' @srrstats {G5.0} Tests use standard data patterns (sin, linear, constant).
-#' @srrstats {G5.1} Test datasets are reproducible via set.seed().
-#' @srrstats {G5.3} Tests verify no NA/NaN in outputs.
-#' @srrstats {G5.4, G5.4a} Correctness tests against expected behavior.
-#' @srrstats {G5.5} Fixed random seeds in all tests.
-#' @srrstats {G5.6, G5.6a} Parameter recovery within tolerance.
-#' @srrstats {G5.7} Performance tests (parallel vs serial).
-#' @srrstats {G5.8c, G5.8d} Edge cases: minimum points, constant values.
-#' @srrstats {G5.9a} Noise susceptibility via rnorm() variations.
-#' @srrstats {RE4.10} Residuals tested in Lowess residuals test.
-#' @srrstats {RE4.11} Goodness-of-fit tested in diagnostics test.
-#' @srrstats {RE5.0} Confidence/prediction intervals tested.
-#' @srrstats {RE6.0} Cross-validation tested.
-test_that("Lowess basic functionality works", {
-    x <- c(1, 2, 3, 4, 5)
-    y <- c(2, 4, 6, 8, 10)
-
-    result <- Lowess(fraction = 0.67)$fit(as.double(x), as.double(y))
-
-    expect_type(result, "list")
-    expect_named(result, c("x", "y", "fraction_used"))
-    expect_length(result$x, length(x))
-    expect_length(result$y, length(y))
-    expect_type(result$x, "double")
-    expect_type(result$y, "double")
-})
-
-test_that("Lowess handles different fractions", {
-    set.seed(42)
-    x <- seq(0, 10, length.out = 50)
-    y <- sin(x) + rnorm(50, sd = 0.1)
-
-    result_low <- Lowess(fraction = 0.2)$fit(as.double(x), as.double(y))
-    result_high <- Lowess(fraction = 0.8)$fit(as.double(x), as.double(y))
-
-    expect_length(result_low$y, length(y))
-    expect_length(result_high$y, length(y))
-
-    # Higher fraction should produce smoother results
-    expect_lt(sd(diff(result_high$y)), sd(diff(result_low$y)))
-})
-
-test_that("Lowess robustness iterations work", {
-    set.seed(123)
-    x <- seq(0, 10, length.out = 50)
-    y <- sin(x) + rnorm(50, sd = 0.1)
-
-    # Add outliers
-    y[c(10, 25, 40)] <- y[c(10, 25, 40)] + 5
-
-    model_no_robust <- Lowess(fraction = 0.3, iterations = 0)
-    result_no_robust <- model_no_robust$fit(as.double(x), as.double(y))
-    model_robust <- Lowess(fraction = 0.3, iterations = 5)
-    result_robust <- model_robust$fit(as.double(x), as.double(y))
-
-    expect_length(result_no_robust$y, length(y))
-    expect_length(result_robust$y, length(y))
-})
-
-test_that("Lowess confidence intervals work", {
-    set.seed(42)
-    x <- seq(0, 10, length.out = 50)
-    y <- sin(x) + rnorm(50, sd = 0.2)
-
-    model <- Lowess(fraction = 0.5, confidence_intervals = 0.95)
-    result <- model$fit(as.double(x), as.double(y))
-
-    expect_true("confidence_lower" %in% names(result))
-    expect_true("confidence_upper" %in% names(result))
-    expect_length(result$confidence_lower, length(y))
-    expect_length(result$confidence_upper, length(y))
-
-    # CI bounds should bracket the fitted values
[…]
