### 2026-08-16 ed02724
refactor: move Makevars.check to r bidning dir
