### 2026-08-16 c6ac8ba
refactor: inline prepare_cran.sh into Makefile
