### 2026-08-16 acabad2
build: inline fix_rd_style script, replace style_pkg with air
