### 2026-08-16 849c7b8
refactor: inline format_julia into Makefile
