+++ .github/workflows/ci.yml
@@ -93,9 +93,9 @@ jobs:
 
       - name: "Rust: Lowess Tests"
         run: |
-          cargo test -q -p lowess-project-tests --test lowess
+          cargo test -q -p lowess --test lowess
           if [ "${{ matrix.os }}" = "ubuntu-latest" ]; then
-            cargo test -q -p lowess-project-tests --test lowess --no-default-features
+            cargo test -q -p lowess --test lowess --no-default-features
           fi
         shell: bash
 
@@ -122,9 +122,9 @@ jobs:
 
              test -n "$VK_ICD_FILENAMES"
              vulkaninfo --summary
-             cargo test -q -p lowess-project-tests --test fastLowess --features gpu
+             cargo test -q -p fastLowess --test fastLowess --features gpu
           else
-             cargo test -q -p lowess-project-tests --test fastLowess --features cpu
+             cargo test -q -p fastLowess --test fastLowess --features cpu
           fi
         shell: bash
 
@@ -138,10 +138,15 @@ jobs:
 
       - name: "Rust: Run Examples"
         run: |
-          for example in examples/lowess/*.rs examples/fastLowess/*.rs; do
+          for example in crates/lowess/examples/*.rs; do
             name=$(basename "$example" .rs)
             echo "Running example: $name"
-            cargo run -p examples --example "$name" --features cpu
+            cargo run -p lowess --example "$name"
+          done
+          for example in crates/fastLowess/examples/*.rs; do
+            name=$(basename "$example" .rs)
+            echo "Running example: $name"
+            cargo run -p fastLowess --example "$name" --features cpu
           done
         shell: bash
 
@@ -158,19 +163,19 @@ jobs:
       - name: "Python: Lint (Ruff)"
         working-directory: bindings/python
         run: |
-          ruff format --check python/fastlowess/ ../../tests/python/
-          ruff check python/fastlowess/ ../../tests/python/
+          ruff format --check python/fastlowess/ tests/
+          ruff check python/fastlowess/ tests/
 
       - name: "Python: Tests"
         working-directory: bindings/python
-        run: pytest ../../tests/python/ -v
+        run: pytest tests/ -v
 
       - name: "Python: Examples"
         working-directory: bindings/python
         run: |
-          python ../../examples/python/batch_smoothing.py
-          python ../../examples/python/streaming_smoothing.py
-          python ../../examples/python/online_smoothing.py
+          python examples/batch_smoothing.py
+          python examples/streaming_smoothing.py
+          python examples/online_smoothing.py
 
       # ========================================================================
       # JULIA
@@ -208,7 +213,7 @@ jobs:
           else
             export FASTLOWESS_LIB="target/release/libfastlowess_jl.so"
           fi
-          julia -e 'using Pkg; Pkg.activate("bindings/julia/julia"); Pkg.instantiate(); include("tests/julia/test_FastLOWESS.jl")'
+          julia -e 'using Pkg; Pkg.activate("bindings/julia/julia"); Pkg.instantiate(); include("bindings/julia/tests/test_FastLOWESS.jl")'
 
       - name: "Julia: Examples"
         shell: bash
@@ -220,9 +225,9 @@ jobs:
           else
             export FASTLOWESS_LIB="target/release/libfastlowess_jl.so"
           fi
-          julia --project=bindings/julia/julia examples/julia/batch_smoothing.jl
-          julia --project=bindings/julia/julia examples/julia/streaming_smoothing.jl
[…]
