### 2026-08-17 a023a4d
fix: minor compilation errors
