+++ bindings/nodejs/examples/online_smoothing.js
@@ -1,4 +1,4 @@
-const fastlowess = require('../../bindings/nodejs');
+const fastlowess = require('..');
 
 /**
  * fastlowess Online Smoothing Example
