+++ bindings/r/R/methods.R
@@ -64,10 +64,15 @@ print.LowessResult <- function(x, ...) {
 plot.LowessResult <- function(x, main = "LOWESS Fit", ...) {
     # Plot the smoothed curve
     plot(
-        x$x, x$y,
-        type = "l", col = "blue", lwd = 2,
-        xlab = "x", ylab = "Fitted",
-        main = main, ...
+        x$x,
+        x$y,
+        type = "l",
+        col = "blue",
+        lwd = 2,
+        xlab = "x",
+        ylab = "Fitted",
+        main = main,
+        ...
     )
 
     # If confidence intervals exist, plot them
