+++ bindings/cpp/examples/online_smoothing.cpp
@@ -13,7 +13,7 @@
 #include <random>
 #include <vector>
 
-#include "../../bindings/cpp/include/fastlowess.hpp"
+#include "../include/fastlowess.hpp"
 
 namespace {
 
