+++ bindings/nodejs/examples/batch_smoothing.js
@@ -1,4 +1,4 @@
-const fastlowess = require('../../bindings/nodejs');
+const fastlowess = require('..');
 
 /**
  * fastlowess Batch Smoothing Example
