### 2026-08-17 0e401c1
fix: remove r binding level tests from gitignore
