+++ .github/workflows/R-CMD-check.yaml
@@ -87,7 +87,7 @@ jobs:
           {
             echo "Authors and Copyright Holders for Rust Dependencies:"
             echo ""
-            (cd ../.. && cargo metadata --locked --format-version 1) | \
+            (cd ../.. && cargo metadata --format-version 1) | \
               jq -r '. as $root | [.packages[] | select(.id as $id | ($root.workspace_members | index($id)) == null)] | unique_by(.name)[] | "Package: \(.name) (\(.version))\nAuthors: \(.authors | join(", "))\nLicense: \(.license // "Unknown")\n" + "-" * 40'
           } > inst/AUTHORS
           test -f tests/testthat.R

[…]
