### 2026-08-17 40a8a10
fix: workflow bugs, contributing docs
