+++ .github/workflows/R-CMD-check.yaml
@@ -97,7 +97,7 @@ jobs:
         with:
           upload-snapshots: true
           build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
-          args: 'c("--no-manual", "--as-cran", "--no-check-urls")'
+          args: 'c("--no-manual", "--as-cran")'
           check-dir: '"check"'
           error-on: '"error"'
           upload-results: false
