### 2026-08-17 e3df5de
ci: suppress third party compilation warnings in SRR-check workflow
