### 2026-08-17 4879c59
fix: remaining R CMD check warnings
