+++ bindings/r/CONTRIBUTING.md
@@ -1,3 +1,3 @@
 # Contributing to rfastlowess
 
-For contribution guidelines, please see the [Contributing Guide](https://lowess.readthedocs.io/contributing/) in the docs.
+For contribution guidelines, please see the [Contributing Guide](https://lowess.readthedocs.io/en/latest/contributing/) in the docs.
