### 2026-08-17 31a19b0
refactor: added Depends: R (>=4.2) to Description for R
