+++ .github/workflows/R-CMD-check.yaml
@@ -17,13 +17,18 @@ jobs:
   # R CMD CHECK (multi-OS matrix)
   # ============================================================================
   r-cmd-check:
-    name: R-CMD-check (${{ matrix.os }})
+    name: R-CMD-check (${{ matrix.os }}, R ${{ matrix.r }})
     runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
+        r: [release]
         rust: [stable]
+        include:
+          - os: ubuntu-latest
+            r: "4.2"
+            rust: stable
 
     steps:
       - name: Checkout Repository
@@ -42,12 +47,12 @@ jobs:
       - name: Cache dependencies
         uses: Swatinem/rust-cache@v2
         with:
-          key: r-cmd-check-${{ matrix.os }}-cargo
+          key: r-cmd-check-${{ matrix.os }}-r${{ matrix.r }}-cargo
 
       - name: Setup R
         uses: r-lib/actions/setup-r@v2
         with:
-          r-version: release
+          r-version: ${{ matrix.r }}
           use-public-rspm: true
 
       - name: Setup Pandoc
