### 2026-08-17 689da56
refactor: change list-closure API in R binding to S3 generics, fix docs
