### 2026-08-17 947de60
docs: add cross-reference links in all docs/api/ files, fix Rust API option values
