+++ docs/api/julia.md
@@ -212,6 +212,8 @@ Returned by `add_point()` once the window has enough points (`nothing` until the
 
 ### weight_function
 
+*See: [Weight Functions](../user-guide/kernels.md)*
+
 * `"tricube"` (default)
 * `"epanechnikov"`
 * `"gaussian"`
@@ -222,11 +224,15 @@ Returned by `add_point()` once the window has enough points (`nothing` until the
 
 ### robustness_method
 
+*See: [Robustness](../user-guide/robustness.md)*
+
 * `"bisquare"` (default; alias: `"biweight"`)
 * `"huber"`
 * `"talwar"`
 
 ### boundary_policy
+
+*See: [Boundary Handling](../user-guide/boundary.md)*
 
 * `"extend"` (default; alias: `"pad"`)
 * `"reflect"` (alias: `"mirror"`)
@@ -235,11 +241,15 @@ Returned by `add_point()` once the window has enough points (`nothing` until the
 
 ### scaling_method
 
+*See: [Scaling Methods](../user-guide/scaling.md)*
+
 * `"mad"` (default; alias: `"median_absolute_deviation"`)
 * `"mar"` (alias: `"median_absolute_residual"`)
 * `"mean"` (alias: `"mean_absolute_residual"`)
 
 ### zero_weight_fallback
+
+*See: [Parameters](../user-guide/parameters.md)*
 
 * `"use_local_mean"` (default; aliases: `"local_mean"`, `"mean"`)
 * `"return_original"` (alias: `"original"`)
@@ -247,12 +257,16 @@ Returned by `add_point()` once the window has enough points (`nothing` until the
 
 ### merge_strategy
 
+*See: [Merge Strategies](../user-guide/merge.md)*
+
 * `"weighted_average"` (default; alias: `"weighted"`)
 * `"average"` (alias: `"mean"`)
 * `"take_first"` (alias: `"first"`)
 * `"take_last"` (alias: `"last"`)
 
 ### update_mode
+
+*See: [Execution Modes](../user-guide/adapters.md)*
 
 * `"full"` (default; alias: `"resmooth"`)
 * `"incremental"` (alias: `"single"`)

[…]
