### 2026-08-18 f66c979
docs: separate streaming/online docs from batch in docs/api/
