+++ docs/api/rust.md
@@ -2,6 +2,8 @@
 
 The Rust bindings provide the core implementation and high-performance extensions. The API uses a Builder pattern consistent across both the `lowess` (pure Rust) and `fastLowess` (accelerated) crates.
 
+> **StreamingLowess** and **OnlineLowess** are documented separately: [rust-streaming.md](rust-streaming.md), [rust-online.md](rust-online.md)
+
 ## Structs & Usage
 
 The `fastLowess` crate exposes three dedicated wrapper structs — `Lowess`, `StreamingLowess`, and `OnlineLowess` — that mirror the distinct classes available in other language bindings. Each struct wraps a `LowessBuilder<f64>` and its `build()` method delegates to the corresponding parallel adapter.
@@ -43,105 +45,9 @@ fn main() -> Result<(), LowessError> {
 * Fits the model to the provided `x` and `y` arrays.
 * Returns `Result<LowessResult<T>, LowessError>`.
 
-### `StreamingLowess`
+See [rust-streaming.md](rust-streaming.md) for `StreamingLowess`.
 
-Streaming mode for large datasets.
-
-**Constructor:**
-
-```rust
-use fastLowess::prelude::*;
-
-fn main() -> Result<(), LowessError> {
-    let mut processor = StreamingLowess::new();
-
-    Ok(())
-}
-```
-
-**Methods:**
-
-```rust
-use fastLowess::prelude::*;
-use std::f64::consts::TAU;
-
-fn main() -> Result<(), LowessError> {
-    let n = 100usize;
-    let x: Vec<f64> = (0..n).map(|i| i as f64 * TAU / (n - 1) as f64).collect();
-    let y: Vec<f64> = x.iter().map(|&xi| xi.sin() + 0.1).collect();
-
-    let mut processor = StreamingLowess::new().build()?;
-    let result = processor.process_chunk(&x, &y)?;
-
-    Ok(())
-}
-```
-
-* Processes a chunk of data. Returns `LowessResult<T>` with partial results.
-
-```rust
-use fastLowess::prelude::*;
-use std::f64::consts::TAU;
-
-fn main() -> Result<(), LowessError> {
-    let n = 100usize;
-    let x: Vec<f64> = (0..n).map(|i| i as f64 * TAU / (n - 1) as f64).collect();
-    let y: Vec<f64> = x.iter().map(|&xi| xi.sin() + 0.1).collect();
-
-    let mut processor = StreamingLowess::new().build()?;
-    processor.process_chunk(&x, &y)?;
-    let final_result = processor.finalize()?;
-
-    Ok(())
-}
-```
-
-* Finalizes processing and returns remaining buffered results.
-
-### `OnlineLowess`
-
-Online mode for real-time data.
-
-**Constructor:**
-
-```rust
-use fastLowess::prelude::*;
-
-fn main() -> Result<(), LowessError> {
-    let mut processor = OnlineLowess::new();
-
-    Ok(())
-}
-```
-
-**Methods:**
-
-```rust
-use fastLowess::prelude::*;
-
-fn main() -> Result<(), LowessError> {
-    let mut processor = OnlineLowess::new().build()?;
-    let output = processor.add_point(1.0f64, 2.0f64)?;
-
-    Ok(())
-}
-```
-
-* Adds a single point `(x, y)` to the window.
-* Returns `Result<Option<OnlineOutput<T>>, LowessError>`.
-
-```rust
-use fastLowess::prelude::*;
-
-fn main() -> Result<(), LowessError> {
-    let mut processor = OnlineLowess::new().build()?;
-    processor.reset();
-
-    Ok(())
-}
-```
-
-* Clears the internal window buffer. **Rust-only** — this method is not exposed in other language bindings, where creating a new instance is the idiomatic alternative.
[…]
