### 2026-08-18 c1d66a8
docs: expand R binding internal docs
