+++ bindings/r/R/StreamingLowess.R
@@ -8,11 +8,13 @@
 #'
 #' @inheritParams Lowess
 #' @param ... Not used; forces all subsequent arguments to be named.
-#' @param chunk_size Points per chunk.
-#' @param overlap Overlap between chunks.
+#' @param chunk_size Number of data points per processing chunk.
+#' @param overlap Number of overlapping points between consecutive chunks.
 #' @param merge_strategy Strategy for reconciling overlapping chunk regions:
-#'   \code{"weighted_average"} (default), \code{"average"},
-#'   \code{"take_first"}, or \code{"take_last"}.
+#'   \code{"weighted_average"} (default; alias: \code{"weighted"}),
+#'   \code{"average"} (alias: \code{"mean"}),
+#'   \code{"take_first"} (alias: \code{"first"}), or
+#'   \code{"take_last"} (alias: \code{"last"}).
 #'
 #' @return A StreamingLowess object.
 #' @examples
