+++ bindings/r/R/Lowess.R
@@ -15,29 +15,54 @@
 #' @srrstats {RE5.0} O(n) scaling documented in README.
 #'
 #' @param ... Not used; forces all subsequent arguments to be named.
-#' @param fraction Smoothing fraction (0 to 1). Default: 0.67.
-#' @param iterations Robustness iterations. Default: 3.
-#' @param delta Interpolation threshold. NULL = auto.
-#' @param weight_function Kernel name. Default: "tricube".
-#' @param robustness_method Method: "bisquare", "huber", "talwar".
-#' @param scaling_method Scale estimation: "mad", "mar", "mean".
-#' @param boundary_policy Edge handling: "extend", "reflect", "zero",
-#'   "noboundary".
-#' @param confidence_intervals Confidence level (e.g., 0.95). NULL disables.
-#' @param prediction_intervals Prediction level (e.g., 0.95). NULL disables.
-#' @param return_diagnostics Return fit metrics. Default: FALSE.
-#' @param return_residuals Return residuals. Default: FALSE.
-#' @param return_robustness_weights Return weights. Default: FALSE.
-#' @param zero_weight_fallback Fallback: "use_local_mean", "return_original",
-#'   "return_none".
-#' @param auto_converge Convergence tolerance. NULL disables.
-#' @param cv_fractions Fractions for cross-validation. NULL disables.
-#' @param cv_method CV method: "kfold", "loocv".
-#' @param cv_k Folds for k-fold CV. Default: 5.
-#' @param parallel Enable parallel processing. Default: TRUE.
+#' @param fraction Smoothing fraction (between 0 and 1). Default: 0.67.
+#' @param iterations Number of robustness iterations (non-negative integer).
+#'   Default: 3.
+#' @param delta Interpolation distance threshold; points within \code{delta}
+#'   of each other on x share the same local fit. \code{NULL} (default) sets
+#'   it automatically to 1/100th of the x range.
+#' @param weight_function Kernel weight function. One of \code{"tricube"}
+#'   (default), \code{"gaussian"}, \code{"uniform"} (alias: \code{"boxcar"}),
+#'   \code{"cosine"}, \code{"epanechnikov"},
+#'   \code{"biweight"} (alias: \code{"bisquare"}), or
+#'   \code{"triangle"} (alias: \code{"triangular"}).
+#' @param robustness_method Outlier downweighting method: \code{"bisquare"}
+#'   (default; alias: \code{"biweight"}), \code{"huber"}, or \code{"talwar"}.
+#' @param scaling_method Residual scale estimation for robustness weights:
+#'   \code{"mad"} (default; alias: \code{"median_absolute_deviation"}),
+#'   \code{"mar"} (alias: \code{"median_absolute_residual"}), or
+#'   \code{"mean"} (alias: \code{"mean_absolute_residual"}).
+#' @param boundary_policy Boundary handling strategy: \code{"extend"}
+#'   (default; alias: \code{"pad"}), \code{"reflect"} (alias:
+#'   \code{"mirror"}), \code{"zero"}, or
+#'   \code{"noboundary"} (alias: \code{"none"}).
+#' @param confidence_intervals Confidence level for confidence intervals
+#'   (e.g., 0.95). \code{NULL} (default) disables confidence intervals.
+#' @param prediction_intervals Confidence level for prediction intervals
+#'   (e.g., 0.95). \code{NULL} (default) disables prediction intervals.
+#' @param return_diagnostics Logical; if \code{TRUE}, return fit-quality
+#'   metrics (RMSE, MAE, R-squared, AIC, etc.). Default: \code{FALSE}.
+#' @param return_residuals Logical; if \code{TRUE}, return residuals in the
+#'   result. Default: \code{FALSE}.
[…]
