+++ bindings/r/man/OnlineLowess.Rd
@@ -5,65 +5,91 @@
 \title{LOWESS Online Smoothing}
 \usage{
 OnlineLowess(
-    fraction = 0.67,
-    window_capacity = 1000L,
-    min_points = 3L,
-    iterations = 3L,
-    delta = NULL,
-    weight_function = "tricube",
-    robustness_method = "bisquare",
-    scaling_method = "mad",
-    boundary_policy = "extend",
-    zero_weight_fallback = "use_local_mean",
-    update_mode = "full",
-    auto_converge = NULL,
-    return_robustness_weights = FALSE,
-    return_diagnostics = FALSE,
-    return_residuals = FALSE,
-    parallel = FALSE,
-    confidence_intervals = NULL,
-    prediction_intervals = NULL
+  fraction = 0.67,
+  window_capacity = 1000L,
+  min_points = 3L,
+  ...,
+  iterations = 3L,
+  delta = NULL,
+  weight_function = "tricube",
+  robustness_method = "bisquare",
+  scaling_method = "mad",
+  boundary_policy = "extend",
+  zero_weight_fallback = "use_local_mean",
+  update_mode = "full",
+  auto_converge = NULL,
+  return_robustness_weights = FALSE,
+  return_diagnostics = FALSE,
+  return_residuals = FALSE,
+  parallel = FALSE,
+  confidence_intervals = NULL,
+  prediction_intervals = NULL
 )
 }
 \arguments{
-\item{fraction}{Smoothing fraction (0 to 1). Default: 0.67.}
+\item{fraction}{Smoothing fraction (between 0 and 1). Default: 0.67.}
 
-\item{window_capacity}{Max points in sliding window. Default: 1000.}
+\item{window_capacity}{Maximum number of points kept in the sliding window.}
 
-\item{min_points}{Minimum points before smoothing. Default: 3.}
+\item{min_points}{Minimum number of points required before smoothing begins.}
 
-\item{iterations}{Robustness iterations. Default: 3.}
+\item{...}{Not used; forces all subsequent arguments to be named.}
 
-\item{delta}{Interpolation threshold. NULL = auto.}
+\item{iterations}{Number of robustness iterations (non-negative integer).
+Default: 3.}
 
-\item{weight_function}{Kernel name. Default: "tricube".}
+\item{delta}{Interpolation distance threshold; points within \code{delta}
+of each other on x share the same local fit. \code{NULL} (default) sets
+it automatically to 1/100th of the x range.}
 
-\item{robustness_method}{Method: "bisquare", "huber", "talwar".}
+\item{weight_function}{Kernel weight function. One of \code{"tricube"}
+(default), \code{"gaussian"}, \code{"uniform"} (alias: \code{"boxcar"}),
+\code{"cosine"}, \code{"epanechnikov"},
+\code{"biweight"} (alias: \code{"bisquare"}), or
+\code{"triangle"} (alias: \code{"triangular"}).}
 
-\item{scaling_method}{Scale estimation: "mad", "mar", "mean".}
+\item{robustness_method}{Outlier downweighting method: \code{"bisquare"}
+(default; alias: \code{"biweight"}), \code{"huber"}, or \code{"talwar"}.}
 
-\item{boundary_policy}{Edge handling: "extend", "reflect", "zero",
-"noboundary".}
+\item{scaling_method}{Residual scale estimation for robustness weights:
+\code{"mad"} (default; alias: \code{"median_absolute_deviation"}),
+\code{"mar"} (alias: \code{"median_absolute_residual"}), or
+\code{"mean"} (alias: \code{"mean_absolute_residual"}).}
 
[…]
