+++ bindings/r/man/add_point.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/methods.R
+\name{add_point}
+\alias{add_point}
+\title{Add a single point to an online LOWESS model}
+\usage{
+add_point(model, ...)
+}
+\arguments{
+\item{model}{An \code{OnlineLowess} object.}
+
+\item{...}{Not used.}
+
+\item{x}{A single numeric x value.}
+
+\item{y}{A single numeric y value.}
+}
+\value{
+An online result list, or \code{NULL} if fewer than
+\code{min_points} have been added.
+}
+\description{
+Add a single point to an online LOWESS model
+}
+\examples{
+model <- OnlineLowess(fraction = 0.2, window_capacity = 20L)
+result <- add_point(model, 1.0, 0.5)
+}

[…]
