+++ bindings/r/man/Lowess.Rd
@@ -113,3 +113,6 @@ result <- fit(model, x, y)
 plot(x, y)
 lines(x, result$y, col = "red")
 }
+\seealso{
+\url{https://lowess.readthedocs.io/} for full documentation.
+}
