+++ bindings/r/man/StreamingLowess.Rd
@@ -106,3 +106,6 @@ res1 <- process_chunk(model, x[1:50], y[1:50])
 res2 <- process_chunk(model, x[51:100], y[51:100])
 final <- finalize(model)
 }
+\seealso{
+\url{https://lowess.readthedocs.io/} for full documentation.
+}
