### 2026-08-18 081be53
docs: removed examples
