+++ bindings/python/examples/online_smoothing.py
@@ -1,118 +1,0 @@
-#!/usr/bin/env python3
-"""
-fastlowess Online Smoothing Example
-
-This example demonstrates online LOWESS smoothing for real-time data:
-- Basic incremental processing with streaming data
-- Real-time sensor data smoothing
-- Different update modes (Full vs Incremental)
-- Memory-bounded processing with sliding window
-
-The OnlineLowess class is designed for:
-- Real-time data streams
-- Sensors and monitoring
-- Low-latency applications
-"""
-
-import os
-
-import matplotlib.pyplot as plt
-import numpy as np
-from fastlowess import OnlineLowess
-
-# Get script directory for relative paths
-SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
-PLOTS_DIR = os.path.join(SCRIPT_DIR, "plots")
-
-
-def main():
-    """Run the online smoothing example and save the generated plots."""
-    print("=== fastlowess Online Smoothing Example ===")
-
-    # 1. Simulate a real-time signal
-    # A sine wave with changing frequency and random noise
-    n_points = 1000
-    x = np.arange(n_points, dtype=float)
-    y_true = 20.0 + 5.0 * np.sin(x * 0.1) + 2.0 * np.sin(x * 0.02)
-    y = y_true + np.random.normal(0, 1.2, n_points)
-
-    # Add some sudden spikes (sensor glitches)
-    y[200:205] += 15.0
-    y[600:610] -= 10.0
-
-    print(f"Simulating {n_points} real-time data points...")
-
-    # 2. Sequential Online Processing
-    def online_smooth(model, x_arr, y_arr):
-        """Process all points one at a time, falling back to original y when
-        not enough points have been accumulated yet."""
-        result = np.empty_like(y_arr)
-        for i, (xi, yi) in enumerate(zip(x_arr, y_arr)):
-            r = model.add_point(float(xi), float(yi))
-            result[i] = r.smoothed if r is not None else yi
-        return result
-
-    # Full Update Mode (higher accuracy)
-    print("Processing with 'full' update mode...")
-    model_full = OnlineLowess(
-        fraction=0.3, window_capacity=50, iterations=3, update_mode="full"
-    )
-    res_full = online_smooth(model_full, x, y)
-
-    # Incremental Update Mode (faster for large windows)
-    print("Processing with 'incremental' update mode...")
-    model_inc = OnlineLowess(
-        fraction=0.3, window_capacity=50, iterations=3, update_mode="incremental"
-    )
-    res_inc = online_smooth(model_inc, x, y)
-
-    # Plotting
-    os.makedirs(PLOTS_DIR, exist_ok=True)
-
-    fig1 = plt.figure(figsize=(12, 7))
-
-    # Original Data
-    plt.scatter(x, y, s=5, alpha=0.3, color="gray", label="Raw Sensor Stream")
-    plt.plot(x, y_true, "k--", alpha=0.6, label="True Signal")
-
-    # Online Results
-    plt.plot(x, res_full, "r-", linewidth=2, label="Online LOWESS (Full)")
-    plt.plot(
-        x,
-        res_inc,
-        "b-",
-        linewidth=1.5,
-        alpha=0.7,
-        label="Online LOWESS (Incremental)",
-    )
-
-    # Highlight a zoom area to show the windowing effect
-    plt.axvspan(400, 500, color="yellow", alpha=0.1, label="Zoom Area")
-
-    plt.title("fastlowess: Real-time Online Smoothing (Sliding Window)")
[…]
