+++ crates/fastLowess/examples/fast_batch_smoothing.rs
@@ -1,292 +1,0 @@
-//! fastLowess Parallel Smoothing Examples
-//!
-//! This example demonstrates features specific to `fastLowess`:
-//! - Parallel execution using `rayon`
-//! - Sequential fallback
-//! - `ndarray` integration
-//! - Cross-validation for automatic parameter selection
-//! - Performance comparison (simulated)
-
-#[cfg(feature = "cpu")]
-use fastLowess::prelude::*;
-
-#[cfg(feature = "cpu")]
-use ndarray::Array1;
-#[cfg(feature = "cpu")]
-use std::time::Instant;
-
-#[cfg(feature = "cpu")]
-fn main() -> Result<(), LowessError> {
-    #[cfg(feature = "cpu")]
-    {
-        println!("{}", "=".repeat(80));
-        println!("fastLowess Parallel Smoothing Examples");
-        println!("{}", "=".repeat(80));
-        println!();
-
-        example_1_parallel_execution()?;
-        example_2_sequential_fallback()?;
-        example_3_ndarray_integration()?;
-        example_4_robust_parallel()?;
-        example_5_cross_validation()?;
-        example_6_custom_weights()?;
-    }
-    Ok(())
-}
-
-#[cfg(not(feature = "cpu"))]
-fn main() {}
-
-#[cfg(feature = "cpu")]
-/// Example 1: Parallel Execution
-/// Demonstrates the default parallel execution mode
-fn example_1_parallel_execution() -> Result<(), LowessError> {
-    println!("Example 1: Parallel Execution");
-    println!("{}", "-".repeat(80));
-
-    // Generate a larger synthetic dataset
-    let n = 10_000;
-    let x: Vec<f64> = (0..n).map(|i| i as f64).collect();
-    let y: Vec<f64> = x
-        .iter()
-        .map(|&xi| (xi * 0.1).sin() + (xi * 0.01).cos())
-        .collect();
-
-    let start = Instant::now();
-    let model = Lowess::new()
-        .fraction(0.5) // Use 50% of data for each local fit
-        .iterations(3) // 3 robustness iterations
-        .parallel(true)
-        .build()?;
-
-    let result = model.fit(&x, &y)?;
-    let duration = start.elapsed();
-
-    println!("Processed {} points in {:?}", n, duration);
-    println!("Execution mode: Parallel");
-    println!("Result summary:\n{}", result);
-
-    println!();
-    Ok(())
-}
-
-#[cfg(feature = "cpu")]
-/// Example 2: Sequential Fallback
-/// Demonstrates explicitly disabling parallelism
-fn example_2_sequential_fallback() -> Result<(), LowessError> {
-    println!("Example 2: Sequential Fallback");
-    println!("{}", "-".repeat(80));
-
-    let n = 10_000;
-    let x: Vec<f64> = (0..n).map(|i| i as f64).collect();
-    let y: Vec<f64> = x
-        .iter()
-        .map(|&xi| (xi * 0.1).sin() + (xi * 0.01).cos())
-        .collect();
-
-    let start = Instant::now();
-    let model = Lowess::new()
-        .parallel(false)
-        .build()?;
-
-    let _result = model.fit(&x, &y)?;
-    let duration = start.elapsed();
-
-    println!("Processed {} points in {:?}", n, duration);
-    println!("Execution mode: Sequential");
-
-    println!();
-    Ok(())
-}
-
-#[cfg(feature = "cpu")]
-/// Example 3: NdArray Integration
-/// Demonstrates direct usage with ndarray types
-fn example_3_ndarray_integration() -> Result<(), LowessError> {
-    println!("Example 3: NdArray Integration");
[…]
