### 2026-08-19 e363671
docs: add streaming and online results structures
