### 2026-08-19 f385537
docs: print outputs, sync snippets
