+++ docs/api/wasm-streaming.md
@@ -26,10 +26,11 @@ const { StreamingLowess } = require('fastlowess-wasm');
 
 const n = 100;
 const x = Float64Array.from({ length: n }, (_, i) => i * 2 * Math.PI / (n - 1));
-const y = Float64Array.from(x, (xi, i) => Math.sin(xi) + (((i * 7 + 3) % 17) / 17 - 0.5) * 0.6);
+const y = Float64Array.from(x, xi => Math.sin(xi) + 0.1);
 
-const stream = new StreamingLowess({ fraction: 0.3 }, { chunk_size: 50, overlap: 10 });
+const stream = new StreamingLowess({ fraction: 0.5 }, { chunk_size: 50, overlap: 10 });
 const partialResult = stream.process_chunk(x.slice(0, 50), y.slice(0, 50));
+console.log(partialResult.fraction_used);  // 0.5
 ```
 
 * Processes a chunk of data. Returns partial results.
@@ -39,11 +40,13 @@ const { StreamingLowess } = require('fastlowess-wasm');
 
 const n = 100;
 const x = Float64Array.from({ length: n }, (_, i) => i * 2 * Math.PI / (n - 1));
-const y = Float64Array.from(x, (xi, i) => Math.sin(xi) + (((i * 7 + 3) % 17) / 17 - 0.5) * 0.6);
+const y = Float64Array.from(x, xi => Math.sin(xi) + 0.1);
 
 const stream = new StreamingLowess({ fraction: 0.3 }, { chunk_size: 50, overlap: 10 });
-stream.process_chunk(x, y);
+stream.process_chunk(x.slice(0, 50), y.slice(0, 50));
+stream.process_chunk(x.slice(50), y.slice(50));
 const finalResult = stream.finalize();
+console.log(finalResult.fraction_used);  // 0.3
 ```
 
 * Finalizes the smoothing process and returns any remaining buffered results.
