### 2026-08-19 05c6db9
chore: verify rfastloess-intro.Rmd snippets
