+++ dev/verify_snippets.py
@@ -35,9 +35,9 @@ import textwrap
 import threading
 import time
 from collections import defaultdict
+from collections.abc import Iterator
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Iterator, List, Optional, Tuple
 
 # ---------------------------------------------------------------------------
 # Paths
@@ -45,6 +45,9 @@ from typing import Iterator, List, Optional, Tuple
 
 REPO_ROOT = Path(__file__).resolve().parent.parent
 DOCS_DIR = REPO_ROOT / "docs"
+VIGNETTES_DIRS: list[Path] = [
+    REPO_ROOT / "bindings" / "r" / "vignettes",
+]
 
 # ---------------------------------------------------------------------------
 # Terminal colours (disabled on non-TTY or Windows without colour support)
@@ -98,12 +101,13 @@ def _find_python_with_fastlowess() -> str:
                 r = subprocess.run(
                     [str(c), "-c", "import fastlowess"],
                     capture_output=True,
+                    check=False,
                     timeout=10,
                 )
                 if r.returncode == 0:
                     return str(c)
-            except Exception:
-                pass
+            except OSError:
+                continue
     return sys.executable
 
 
@@ -186,11 +190,11 @@ class Snippet:
     file: Path
     line: int  # 1-based line number of the opening fence
     lang_tag: str  # code-block language tag (e.g. "python")
-    tab: Optional[str]  # nearest === "Tab" label, or None
+    tab: str | None  # nearest === "Tab" label, or None
     code: str
 
     @property
-    def runner(self) -> Optional[str]:
+    def runner(self) -> str | None:
         """Return which runner handles this snippet, or None to skip."""
         for runner, tags in _LANG_TAGS.items():
             if self.lang_tag.lower() in tags:
@@ -226,13 +230,26 @@ _TAB_RE = re.compile(r'^[ \t]*===\s+"([^"]+)"', re.MULTILINE)
 _FENCE_RE = re.compile(r"^([ \t]*)```(\w+)", re.MULTILINE)
 
 
-def extract_snippets(md_file: Path) -> List[Snippet]:
-    """Extract all fenced code blocks from a markdown file."""
+def _rmd_chunk_is_runnable(opts_str: str) -> bool:
+    """Return False if an Rmd chunk header has eval=FALSE or include=FALSE."""
+    opts = opts_str.lower().replace(" ", "")
+    return "eval=false" not in opts and "include=false" not in opts
+
+
+def extract_snippets(md_file: Path) -> list[Snippet]:
+    """Extract all fenced code blocks from a markdown file (also handles .Rmd).
+
+    For .Rmd files all runnable R chunks are combined into one snippet, since
+    Rmd chunks share state (later chunks depend on variables set by earlier ones).
+    """
     text = md_file.read_text(encoding="utf-8")
     lines = text.splitlines()
-    result: List[Snippet] = []
+    result: list[Snippet] = []
+    is_rmd = md_file.suffix.lower() == ".rmd"
 
-    current_tab: Optional[str] = None
+    current_tab: str | None = None
+    # Accumulate Rmd chunks to combine at the end
+    rmd_chunks: list[tuple[int, str, str]] = []  # (start_line, chunk_name, code)
     i = 0
     while i < len(lines):
         line = lines[i]
@@ -244,13 +261,36 @@ def extract_snippets(md_file: Path) -> List[Snippet]:
             i += 1
             continue
 
+        # Rmd-style fence: ```{r} or ```{r chunk-name} or ```{r chunk-name, opts}
+        m_rmd = re.match(r"^([ \t]*)```\{r([^}]*)\}\s*$", line)
+        if m_rmd:
+            fence_indent = m_rmd.group(1)
[…]
