### 2026-08-19 535dc39
docs: expanded R vignettes
