+++ bindings/r/vignettes/rfastlowess-intro.Rmd
@@ -33,6 +33,18 @@ knitr::opts_chunk$set(
 `rfastlowess` is a high-performance R package for LOWESS (Locally Weighted
 Scatterplot Smoothing) built on a Rust backend.
 
+LOWESS fits a smooth curve through a scatter plot by performing a sequence of
+*local polynomial regressions*. For each point $x_i$, the algorithm selects
+the nearest `fraction` of observations, assigns distance-based weights (points
+farther away receive lower weight), fits a weighted least-squares polynomial,
+and reads off the predicted value at $x_i$. Because no global model is
+assumed, LOWESS adapts naturally to non-linear and non-stationary data.
+
+Optional robustness iterations guard against outliers: after each pass,
+residuals are used to recompute per-observation weights so that high-leverage
+outliers are progressively downweighted. This makes `iterations > 0` the right
+choice whenever the data may contain measurement errors or anomalies.
+
 **Full documentation**: <https://lowess.readthedocs.io/>
 
 ## Installation
@@ -46,9 +58,12 @@ install.packages("rfastlowess", repos = "https://thisisamirv.r-universe.dev")
 
 ### Basic Smoothing
 
-The package uses S3 classes to provide a user-friendly interface.
-Models and results can be inspected using `print()`.
-Results can also be visualized using `plot()`.
+`Lowess()` accepts all smoothing options at construction time and returns a
+reusable model object. Calling `fit(model, x, y)` runs the algorithm and
+returns a `LowessResult` list. Key result fields are `$x` (input x values),
+`$y` (smoothed fitted values), `$fraction_used`, and `$iterations_used`. The
+`fraction` parameter controls the bandwidth: a larger value produces a smoother
+curve at the cost of following local features less closely.
 
 ```{r basic_example}
 library(rfastlowess)
@@ -76,6 +91,18 @@ lines(result$x, result$y, col = "red", lwd = 2)
 
 ### Robust Smoothing with Intervals
 
+Real-world data often contains a small fraction of gross errors. Setting
+`iterations > 0` activates robustness re-weighting: after the initial fit,
+large residuals identify likely outliers and their contribution is downweighted
+for the next iteration. The bisquare robustness function (the default) reduces
+the weight of any point whose residual exceeds about six times the median
+absolute deviation to zero.
+
+Setting `confidence_intervals = 0.95` attaches pointwise 95% confidence bands
+to the smoothed curve. These reflect uncertainty in the mean response, not
+individual observations; `prediction_intervals` provides the wider bands
+covering future observations.
+
 ```{r robust_example}
 # Add outliers
 y_outliers <- y
@@ -95,13 +122,109 @@ lines(result_robust$x, result_robust$confidence_lower, col = "blue", lty = 2)
 lines(result_robust$x, result_robust$confidence_upper, col = "blue", lty = 2)
 ```
 
+### Diagnostics
+
+When `return_diagnostics = TRUE`, the result includes goodness-of-fit metrics
+that are useful for comparing different `fraction` values or kernel choices.
+`return_residuals = TRUE` attaches the raw residuals, and
+`return_robustness_weights = TRUE` exposes the final per-point robustness
+weights so you can see which observations were downweighted.
+
+```{r diagnostics_example}
+result_diag <- fit(Lowess(
+    fraction = 0.3,
+    iterations = 3,
+    return_diagnostics = TRUE,
+    return_residuals = TRUE,
+    return_robustness_weights = TRUE
+), x, y)
+
+cat("R-squared:  ", result_diag$r_squared, "\n")
[…]
