### 2026-08-19 1fd586f
docs: fix minor inconsistencies
