### 2026-08-19 d7ee2c2
fix: add missing functions to _pkgdown.yml in R
