+++ bindings/r/_pkgdown.yml
@@ -17,6 +17,10 @@ reference:
   - title: Results and Utilities
     desc: Objects returned by fit methods and helper functions.
     contents:
+      - fit.Lowess
+      - process_chunk.StreamingLowess
+      - finalize.StreamingLowess
+      - add_point.OnlineLowess
       - plot.LowessResult
       - print.Lowess
       - print.LowessResult
@@ -29,7 +33,7 @@ reference:
 
 navbar:
   structure:
-    left:  [intro, reference, articles, news]
+    left: [intro, reference, articles, news]
     right: [search, github]
   components:
     github:
