### 2026-08-19 8be4805
fix: julia iterations_used now returns nothing instead of -1 when robustness iterations were not applicable
