### 2026-08-19 d3542a1
fix: wasm add_point returns null instead of undefined like nodejs
