### 2026-08-19 ee8f21a
fix: _pkgdown.yml topic names
