### 2026-08-19 43307d0
fix: r binding minor warnings, add GPU benchmarks
