+++ benchmarks/gpu_transfer.py
@@ -1,0 +1,349 @@
+"""
+benchmarks/gpu_transfer.py — GPU data-transfer overhead benchmark.
+
+Isolates host↔device buffer transfer time from kernel execution time by
+reproducing the exact buffer footprint of one fastLowess GPU fit:
+  - Upload:   2 × n f32 arrays  (x, y)
+  - Download: 3 × n f32 arrays  (y_smooth, robustness_weights, residuals)
+
+The script uses wgpu-py (same wgpu version as fastLowess) to run the transfers
+without any compute, then cross-references against an end-to-end GPU benchmark
+JSON (if present) to compute the transfer fraction of total GPU time.
+
+Install:
+    pip install wgpu numpy
+Run:
+    python benchmarks/gpu_transfer.py
+    python benchmarks/gpu_transfer.py --sizes 1000 10000 100000 --iters 30
+    python benchmarks/gpu_transfer.py --output benchmarks/output/gpu_transfer.json
+"""
+
+from __future__ import annotations
+
+import argparse
+import json
+import math
+import statistics
+import sys
+import time
+from pathlib import Path
+
+# ── defaults ──────────────────────────────────────────────────────────────────
+
+# Sizes that span the interesting crossover region (where transfer dominates
+# at small n and kernel dominates at large n).
+DEFAULT_SIZES = [1_000, 5_000, 10_000, 50_000, 100_000, 500_000, 1_000_000]
+DEFAULT_WARMUP = 5
+DEFAULT_ITERS = 20
+F32_BYTES = 4
+
+# fastLowess GPU buffer layout (without optional SE / interval buffers):
+#   upload arrays  : x, y                        → 2 × n f32
+#   download arrays: y_smooth, weights, residuals → 3 × n f32
+UPLOAD_ARRAYS = 2
+DOWNLOAD_ARRAYS = 3
+
+
+# ── wgpu helpers ──────────────────────────────────────────────────────────────
+
+
+def _require_wgpu():
+    try:
+        import wgpu
+
+        return wgpu
+    except ImportError:
+        print(
+            "ERROR: wgpu not installed.\n"
+            "       Run:  pip install wgpu\n"
+            "       Docs: https://wgpu-py.readthedocs.io/",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+
+def _require_numpy():
+    try:
+        import numpy as np
+
+        return np
+    except ImportError:
+        print("ERROR: numpy not installed. Run: pip install numpy", file=sys.stderr)
+        sys.exit(1)
+
+
+def _init_device(wgpu):
+    """Request a high-performance adapter and return (adapter, device)."""
+    # wgpu-py exposes the GPU singleton at wgpu.gpu, not on the module directly.
+    adapter = wgpu.gpu.request_adapter_sync(power_preference="high-performance")
+    if adapter is None:
+        print("ERROR: No GPU adapter found.", file=sys.stderr)
+        sys.exit(1)
+    device = adapter.request_device_sync()
+    return adapter, device
+
+
+# ── core measurement ──────────────────────────────────────────────────────────
+
+
+def _sync_read(device, wgpu, sync_buf: object) -> None:
+    """Force CPU/GPU synchronisation by reading a tiny 4-byte buffer."""
+    device.queue.read_buffer(sync_buf, 0, size=4)
+
+
+def bench_transfer(wgpu, np, device, sync_buf, n: int, iters: int, warmup: int) -> dict:
+    """
+    Measure upload and download latency for one fastLowess-sized dataset of n
+    points.  Returns a dict with per-phase and round-trip timing statistics.
+    """
[…]
