+++ bindings/r/R/methods.R
@@ -134,6 +134,7 @@ print.OnlineLowess <- function(x, ...) {
 #' @export
 fit <- function(model, ...) UseMethod("fit")
 
+#' @rdname fit
 #' @export
 fit.Lowess <- function(model, x, y, custom_weights = NULL, ...) {
     validated_args <- validate_common_args(
@@ -160,6 +161,7 @@ fit.Lowess <- function(model, x, y, custom_weights = NULL, ...) {
 #' @export
 process_chunk <- function(model, ...) UseMethod("process_chunk")
 
+#' @rdname process_chunk
 #' @export
 process_chunk.StreamingLowess <- function(model, x, y, ...) {
     args <- validate_common_args(
@@ -204,6 +206,7 @@ finalize.StreamingLowess <- function(model, ...) {
 #' @export
 add_point <- function(model, ...) UseMethod("add_point")
 
+#' @rdname add_point
 #' @export
 add_point.OnlineLowess <- function(model, x, y, ...) {
     model$handle$add_point(as.double(x), as.double(y))
