### 2026-08-19 6f1ff80
bench: add gpu vs cpu crossover bench
