### 2026-08-19 1b906d5
feat: expose gpu backend to bindings
