+++ bindings/wasm/src/lib.rs
@@ -384,6 +384,7 @@ fn options_to_builder(opts: Option<SmoothOptions>) -> Result<LowessBuilder<f64>,
                 confidence_intervals: opts.confidence_intervals,
                 prediction_intervals: opts.prediction_intervals,
                 parallel: opts.parallel,
+                backend: None,
                 chunk_size: None,
                 overlap: None,
                 merge_strategy: None,
