+++ bindings/cpp/include/fastlowess.hpp
@@ -123,6 +123,9 @@ struct LowessOptions {
   bool return_robustness_weights = false;
   bool return_se = false; ///< Return standard errors
   bool parallel = true;
+  /// Execution backend: "cpu" (default) or "gpu" (requires the library to be
+  /// built with the `gpu` Cargo feature and a Vulkan/Metal/DX12-capable GPU).
+  std::string backend = "cpu";
 
   // Cross-validation options
   std::vector<double> cv_fractions;
@@ -423,7 +426,7 @@ public:
         options.cv_fractions.empty() ? nullptr : options.cv_fractions.data(),
         static_cast<unsigned long>(options.cv_fractions.size()),
         options.cv_method.c_str(), options.cv_k, options.parallel ? 1 : 0,
-        options.return_se ? 1 : 0);
+        options.return_se ? 1 : 0, options.backend.c_str());
     if (options.cv_seed > 0) {
       cpp_lowess_set_cv_seed(ptr_, static_cast<unsigned long>(options.cv_seed));
     }

[…]
