### 2026-08-19 a559751
feat: add available-on-demand gpu backend
