### 2026-08-19 4a6086e
fix: R positional arg rejection, CI stale code, lint R
