+++ bindings/r/R/Lowess.R
@@ -103,14 +103,13 @@ Lowess <- function(
     return_se = FALSE,
     backend = "cpu"
 ) {
-    if (...length() > 0) {
-        stop("All arguments after 'fraction' must be named.", call. = FALSE)
-    }
+    reject_extra_positional_args(sys.call(), "fraction")
     if (identical(backend, "gpu") && !gpu_available()) {
         stop(
-            "GPU backend not installed in this build. Run `install_gpu()` once to ",
-            "download and install a GPU-enabled build, then restart R. See ",
-            "https://lowess.readthedocs.io/api/r/#gpu-acceleration for details.",
+            "GPU backend not installed in this build. Run `install_gpu()` ",
+            "once to download and install a GPU-enabled build, then ",
+            "restart R. See https://lowess.readthedocs.io/api/r/",
+            "#gpu-acceleration for details.",
             call. = FALSE
         )
     }
