+++ .github/workflows/gpu-backend.yml
@@ -109,7 +109,8 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install maturin numpy
-          maturin develop --release --features gpu
+          maturin build --release --features gpu -o dist
+          pip install --force-reinstall dist/*.whl
 
       - name: "Python: GPU smoke test"
         if: matrix.language == 'python'
