### 2026-08-19 0f0ef9a
fix: lint python, add missing R _pkgdown sections, fix gpu test workflow
