+++ .github/workflows/gpu-backend.yml
@@ -148,7 +148,7 @@ jobs:
         working-directory: bindings/nodejs
         run: |
           cat > /tmp/gpu_smoke.js <<'JS'
-          const fastlowess = require('.');
+          const fastlowess = require(process.cwd());
 
           if (!fastlowess.gpuAvailable()) {
               throw new Error('GPU backend not detected in Node.js build');
@@ -176,6 +176,18 @@ jobs:
         if: matrix.language == 'julia'
         working-directory: bindings/julia
         run: cargo build --release --features gpu
+
+      - name: "Julia: Relax JLL dependency for CI"
+        if: matrix.language == 'julia'
+        shell: bash
+        run: |
+          PROJECT_TOML="bindings/julia/julia/Project.toml"
+          LATEST=$(julia -e 'using Pkg; Pkg.activate(temp=true); try Pkg.add("fastlowess_jll"); print(Pkg.dependencies()[Base.UUID("7381d214-3001-5784-8d9a-8eb8e8688076")].version) catch; print("1.0.0") end' | cut -d'+' -f1)
+          CURRENT=$(grep "^version =" "$PROJECT_TOML" | cut -d"\"" -f2)
+          echo "Latest Registered JLL: $LATEST"
+          echo "Current Local Version: $CURRENT"
+          julia -e "using TOML; path = \"$PROJECT_TOML\"; p = TOML.parsefile(path); p[\"compat\"][\"fastlowess_jll\"] = \"$LATEST, $CURRENT\"; open(path, \"w\") do io; TOML.print(io, p); end"
+          echo "Modified $PROJECT_TOML (fastlowess_jll = \"$LATEST, $CURRENT\") using TOML parser."
 
       - name: "Julia: GPU smoke test"
         if: matrix.language == 'julia'
