+++ bindings/r/_pkgdown.yml
@@ -30,6 +30,11 @@ reference:
     desc: Helper types for documentation.
     contents:
       - Nullable
+  - title: GPU Backend
+    desc: Check for and install the optional GPU-accelerated backend.
+    contents:
+      - gpu_available
+      - install_gpu
 
 navbar:
   structure:
