### 2026-08-19 fbfe9f4
ci: improve ci and reduce redunancy
