+++ .github/workflows/R-CMD-check.yaml
@@ -1,131 +1,0 @@
-name: R-CMD-check
-
-on:
-  push:
-    branches: [main, master, develop]
-  pull_request:
-    branches: [main, master, develop]
-
-env:
-  CARGO_TERM_COLOR: always
-  RUST_BACKTRACE: 1
-  GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-  FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true
-
-jobs:
-  # ============================================================================
-  # R CMD CHECK (multi-OS matrix)
-  # ============================================================================
-  r-cmd-check:
-    name: R-CMD-check (${{ matrix.os }}, R ${{ matrix.r }})
-    runs-on: ${{ matrix.os }}
-    strategy:
-      fail-fast: false
-      matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
-        r: [release]
-        rust: [stable]
-        include:
-          - os: ubuntu-latest
-            r: "4.2"
-            rust: stable
-
-    steps:
-      - name: Checkout Repository
-        uses: actions/checkout@v6
-
-      - name: Install Rust Toolchain
-        uses: dtolnay/rust-toolchain@stable
-        with:
-          toolchain: ${{ matrix.rust }}
-          components: rustfmt, clippy
-
-      - name: Install Rust GNU (Windows)
-        if: matrix.os == 'windows-latest'
-        run: rustup toolchain install stable-x86_64-pc-windows-gnu
-
-      - name: Cache dependencies
-        uses: Swatinem/rust-cache@v2
-        with:
-          key: r-cmd-check-${{ matrix.os }}-r${{ matrix.r }}-cargo
-
-      - name: Setup R
-        uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: ${{ matrix.r }}
-          use-public-rspm: true
-
-      - name: Setup Pandoc
-        uses: r-lib/actions/setup-pandoc@v2
-
-      - name: Bootstrap pak (Windows)
-        if: runner.os == 'Windows'
-        shell: Rscript {0}
-        run: |
-          install.packages(
-            "pak",
-            repos = sprintf(
-              "https://r-lib.github.io/p/pak/stable/%s/%s/%s",
-              .Platform$pkgType, R.Version()$os, R.Version()$arch
-            )
-          )
-
-      - name: "R: Setup Dependencies"
-        uses: r-lib/actions/setup-r-dependencies@v2
-        with:
-          working-directory: bindings/r
-          extra-packages: |
-            any::rcmdcheck
-            any::BiocCheck
-            any::XML
-            any::devtools
-            any::lintr
-            any::BiocManager
-          needs: check
-
-      - name: "R: Prepare Package Sources"
-        working-directory: bindings/r
-        shell: bash
-        run: |
-          chmod +x configure cleanup
-          mkdir -p inst
-          {
-            echo "Authors and Copyright Holders for Rust Dependencies:"
-            echo ""
-            (cd ../.. && cargo metadata --format-version 1) | \
-              jq -r '. as $root | [.packages[] | select(.id as $id | ($root.workspace_members | index($id)) == null)] | unique_by(.name)[] | "Package: \(.name) (\(.version))\nAuthors: \(.authors | join(", "))\nLicense: \(.license // "Unknown")\n" + "-" * 40'
-          } > inst/AUTHORS
-          test -f tests/testthat.R
-
-      - name: "R: CMD Check"
-        uses: r-lib/actions/check-r-package@v2
-        with:
-          upload-snapshots: true
-          build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
[…]
