### 2026-08-19 4b463ad
fix: add Downloads to julia deps, fix ci bugs
