+++ .github/workflows/ci.yml
@@ -419,10 +419,6 @@ jobs:
         os: [ubuntu-latest, macos-latest, windows-latest]
         r: [release]
         rust: [stable]
-        include:
-          - os: ubuntu-latest
-            r: "4.2"
-            rust: stable
 
     steps:
       - name: Checkout Repository
