+++ bindings/julia/julia/Project.toml
@@ -14,10 +14,12 @@ keywords = ["lowess", "smoothing", "statistics", "regression", "bioinformatics"]
 categories = ["science", "algorithms"]
 
 [deps]
+Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 fastlowess_jll = "7381d214-3001-5784-8d9a-8eb8e8688076"
 
 [compat]
 julia = "1.6"
+Downloads = "1"
 fastlowess_jll = "2.0.0"
 Test = "1"
 Random = "1"
