### 2026-08-19 66895db
refactor: change std_error and smoothed to standard_error and y in online adapter, fix r gpu linker
