+++ bindings/r/R/Lowess.R
@@ -104,15 +104,7 @@ Lowess <- function(
     backend = "cpu"
 ) {
     reject_extra_positional_args(sys.call(), "fraction")
-    if (identical(backend, "gpu") && !gpu_available()) {
-        stop(
-            "GPU backend not installed in this build. Run `install_gpu()` ",
-            "once to download and install a GPU-enabled build, then ",
-            "restart R. See https://lowess.readthedocs.io/api/r/",
-            "#gpu-acceleration for details.",
-            call. = FALSE
-        )
-    }
+    check_gpu_backend(backend)
     validate_params(fraction = fraction, iterations = iterations)
     handle <- do.call(RLowess$new, env_args(lowess_params))
 
