### 2026-08-19 9b8b80c
fix: docs snippets verification, minor R quality check notes, make clean, readmes
