### 2026-08-19 c69d71b
docs: add rOpenSci review status badge
