### 2026-08-20 3cf04cf
fix: r vignettes special charactars, make clean
