+++ Makefile
@@ -221,6 +221,7 @@ docs-serve:
 
 docs-clean:
 	@echo "Cleaning documentation build..."
+	@$(PYTHON) dev/kill_locked_venv.py $(DOCS_VENV)
 	@git clean -fdX -- site $(DOCS_VENV)
 	@echo "Documentation clean complete!"
 
@@ -244,6 +245,7 @@ all-coverage: lowess-coverage fastLowess-coverage python-coverage r-coverage
 all-clean: r-clean lowess-clean fastLowess-clean python-clean julia-clean nodejs-clean wasm-clean cpp-clean
 	@echo "Cleaning project root..."
 	@cargo clean
+	@$(PYTHON) dev/kill_locked_venv.py $(PY_VENV) $(DOCS_VENV)
 	@git clean -fdX .
 	@echo "All clean completed!"
 
