### 2026-08-20 8d8e9a8
build: bump lowess to 3.0.0
