### 2026-08-20 51c043e
build: bump fastLowess to 3.0.0
