### 2026-08-20 4855462
build: bump bindings to v3.0.0
