+++ bindings/r/tests/testthat/test-gpu.R
@@ -1,0 +1,214 @@
+#' @srrstats {G5.3} No NA/NaN in validated outputs.
+#' @srrstats {G5.8, G5.8a} Edge condition tests for GPU backend helpers.
+# Tests targeting uncovered lines in gpu.R:
+#   gpu_available / gpu_enabled (extendr-wrappers.R:110)
+#   check_gpu_backend (both branches)
+#   gpu_asset_info (all three platform branches, via mocking)
+#   gpu_confirm_download (both branches)
+#   gpu_download_to (success and failure, via mocking)
+#   install_gpu (already-active and needs-confirmation branches)
+
+check_gpu_backend <- getFromNamespace("check_gpu_backend", "rfastlowess")
+gpu_asset_info <- getFromNamespace("gpu_asset_info", "rfastlowess")
+gpu_confirm_download <- getFromNamespace("gpu_confirm_download", "rfastlowess")
+gpu_download_to <- getFromNamespace("gpu_download_to", "rfastlowess")
+
+# ── gpu_available / gpu_enabled ──────────────────────────────────────────────
+
+test_that("gpu_available returns a single logical", {
+    result <- gpu_available()
+    expect_type(result, "logical")
+    expect_length(result, 1)
+})
+
+# ── check_gpu_backend ────────────────────────────────────────────────────────
+
+test_that("check_gpu_backend allows any non-gpu backend", {
+    expect_null(check_gpu_backend("cpu"))
+    expect_null(check_gpu_backend(NULL))
+    expect_null(check_gpu_backend("nonsense"))
+})
+
+test_that("check_gpu_backend errors for gpu backend when unavailable", {
+    skip_if(gpu_available(), "GPU backend is active in this build")
+    expect_error(
+        check_gpu_backend("gpu"),
+        "GPU backend not installed in this build"
+    )
+})
+
+test_that("check_gpu_backend allows gpu backend when available", {
+    testthat::local_mocked_bindings(gpu_available = function() TRUE)
+    expect_null(check_gpu_backend("gpu"))
+})
+
+# ── gpu_asset_info ───────────────────────────────────────────────────────────
+
+test_that("gpu_asset_info builds a well-formed asset name and URL", {
+    info <- gpu_asset_info("1.2.3")
+    expect_true(grepl("^librfastlowess-gpu-v1\\.2\\.3-", info$asset))
+    expect_true(endsWith(info$asset, info$ext))
+    expect_identical(info$repo, "thisisamirv/lowess-project")
+    expect_identical(
+        info$url,
+        sprintf(
+            "https://github.com/%s/releases/download/v1.2.3/%s",
+            info$repo,
+            info$asset
+        )
+    )
+    expect_true(info$ext %in% c(".dll", ".so"))
+})
+
+test_that("gpu_asset_info handles Windows platform", {
+    testthat::local_mocked_bindings(
+        `Sys.info` = function() c(sysname = "Windows", machine = "x86-64"),
+        .package = "base"
+    )
+    info <- gpu_asset_info("1.0.0")
+    expect_identical(info$ext, ".dll")
+    expect_true(grepl("windows-x86_64\\.dll$", info$asset))
+})
+
+test_that("gpu_asset_info handles macOS arm64 platform", {
+    testthat::local_mocked_bindings(
+        `Sys.info` = function() c(sysname = "Darwin", machine = "arm64"),
+        .package = "base"
+    )
+    info <- gpu_asset_info("1.0.0")
+    expect_identical(info$ext, ".so")
+    expect_true(grepl("macos-aarch64\\.so$", info$asset))
+})
+
+test_that("gpu_asset_info handles Linux platform", {
+    testthat::local_mocked_bindings(
+        `Sys.info` = function() c(sysname = "Linux", machine = "x86_64"),
+        .package = "base"
+    )
[…]
