+++ bindings/nodejs/index.d.ts
@@ -2,176 +2,166 @@
 /* eslint-disable */
 /** Batch LOWESS smoothing. */
 export declare class Lowess {
-    /** Create a new batch LOWESS smoother. */
-    constructor(options?: SmoothOptions | undefined | null)
-    /** Fit the model. */
-    fit(x: Float64Array, y: Float64Array, customWeights?: Float64Array | undefined | null): LowessResult
-    /** Fit the model asynchronously. */
-    fit_async(x: Float64Array, y: Float64Array, customWeights?: Float64Array | undefined | null): Promise<unknown>
+  /** Create a new batch LOWESS smoother. */
+  constructor(options?: SmoothOptions | undefined | null)
+  /** Fit the model. */
+  fit(x: Float64Array, y: Float64Array, customWeights?: Float64Array | undefined | null): LowessResult
+  /** Fit the model asynchronously. */
+  fit_async(x: Float64Array, y: Float64Array, customWeights?: Float64Array | undefined | null): Promise<unknown>
 }
 
 /** Result of a LOWESS fit. */
 export declare class LowessResult {
-    /** Get the sorted x values. */
-    get x(): Float64Array
-    /** Get the smoothed y values. */
-    get y(): Float64Array
-    /** Get residuals (if requested). */
-    get residuals(): Float64Array | null
-    /** Get standard errors (if requested/computed). */
-    get standard_errors(): Float64Array | null
-    /** Get lower confidence bounds (if requested). */
-    get confidence_lower(): Float64Array | null
-    /** Get upper confidence bounds (if requested). */
-    get confidence_upper(): Float64Array | null
-    /** Get lower prediction bounds (if requested). */
-    get prediction_lower(): Float64Array | null
-    /** Get upper prediction bounds (if requested). */
-    get prediction_upper(): Float64Array | null
-    /** Get robustness weights (if requested). */
-    get robustness_weights(): Float64Array | null
-    /** Get diagnostics (if requested). */
-    get diagnostics(): Diagnostics | null
-    /** Get cross-validation scores (if CV was performed). */
-    get cv_scores(): Float64Array | null
-    /** Get the fraction used for smoothing. */
-    get fraction_used(): number
-    /** Get the number of iterations performed. */
-    get iterations_used(): number | null
+  /** Get the sorted x values. */
+  get x(): Float64Array
+  /** Get the smoothed y values. */
+  get y(): Float64Array
+  /** Get residuals (if requested). */
+  get residuals(): Float64Array | null
+  /** Get standard errors (if requested/computed). */
+  get standard_errors(): Float64Array | null
+  /** Get lower confidence bounds (if requested). */
+  get confidence_lower(): Float64Array | null
+  /** Get upper confidence bounds (if requested). */
+  get confidence_upper(): Float64Array | null
+  /** Get lower prediction bounds (if requested). */
+  get prediction_lower(): Float64Array | null
+  /** Get upper prediction bounds (if requested). */
+  get prediction_upper(): Float64Array | null
+  /** Get robustness weights (if requested). */
+  get robustness_weights(): Float64Array | null
+  /** Get diagnostics (if requested). */
+  get diagnostics(): Diagnostics | null
+  /** Get cross-validation scores (if CV was performed). */
+  get cv_scores(): Float64Array | null
+  /** Get the fraction used for smoothing. */
+  get fraction_used(): number
+  /** Get the number of iterations performed. */
+  get iterations_used(): number | null
 }
 
 /** Online LOWESS smoother for real-time data. */
 export declare class OnlineLowess {
-    /** Create a new online LOWESS smoother. */
[…]
