+++ bindings/python/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fastLowess-py"
-version = "2.0.0"
+version = "3.0.0"
 authors = ["Amir Valizadeh <thisisamirv@gmail.com>"]
 edition = "2024"
 license = "MIT OR Apache-2.0"
@@ -14,7 +14,10 @@ keywords = ["lowess", "smoothing", "statistics", "regression", "bioinformatics"]
 categories = ["science", "algorithms"]
 
 [dependencies]
-fastLowess = { version = "2.0", features = ["dev", "cpu"] }
+fastLowess = { path = "../../crates/fastLowess", version = "3.0", features = [
+    "dev",
+    "cpu",
+] }
 pyo3 = { version = "0.29", features = ["abi3-py38"] }
 numpy = "0.29"
 

[…]
