### 2026-08-20 8b41256
fix: release-gpu workflow_dispatch with tag input, pypi abi3, napi target, BiocGenerics
