+++ .github/workflows/release-pypi.yml
@@ -22,7 +22,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python: ["3.8", "3.9", "3.10", "3.11", "3.12", "3.13", "3.14"]
+        python: ["3.8"]
     steps:
       - uses: actions/checkout@v7
 
@@ -31,7 +31,7 @@ jobs:
         with:
           working-directory: bindings/python
           target: x86_64
-          args: --release --out dist --interpreter python${{ matrix.python }}
+          args: --release --out dist
           manylinux: auto
 
       - name: Upload wheels
