+++ .github/workflows/release-gpu.yml
@@ -3,7 +3,11 @@ name: Release GPU
 on:
   release:
     types: [published]
-  workflow_dispatch: # Allow manual trigger for testing
+  workflow_dispatch:
+    inputs:
+      tag:
+        description: "Release tag to upload assets to (e.g. v3.0.0)"
+        required: true
 
 permissions:
   contents: read
@@ -16,11 +20,8 @@ jobs:
   # BUILD GPU WHEELS - Linux x86_64
   # ============================================================================
   linux-x86_64:
-    name: Build Linux x86_64 GPU (py${{ matrix.python }})
+    name: Build Linux x86_64 GPU
     runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python: ["3.8", "3.9", "3.10", "3.11", "3.12", "3.13", "3.14"]
     steps:
       - uses: actions/checkout@v7
 
@@ -29,13 +30,13 @@ jobs:
         with:
           working-directory: bindings/python
           target: x86_64
-          args: --release --out dist --interpreter python${{ matrix.python }} --features gpu
+          args: --release --out dist --features gpu
           manylinux: auto
 
       - name: Upload wheels
         uses: actions/upload-artifact@v7
         with:
-          name: gpu-wheels-linux-x86_64-py${{ matrix.python }}
+          name: gpu-wheels-linux-x86_64
           path: bindings/python/dist
 
   # ============================================================================
@@ -342,7 +343,7 @@ jobs:
 
       - name: Build GPU binary
         working-directory: bindings/nodejs
-        run: npx napi build --release --target x86_64-unknown-linux-gnu --features gpu
+        run: npx napi build --release --features gpu
 
       - name: Rename asset
         working-directory: bindings/nodejs
@@ -376,7 +377,7 @@ jobs:
 
       - name: Build GPU binary
         working-directory: bindings/nodejs
-        run: npx napi build --release --target x86_64-apple-darwin --features gpu
+        run: npx napi build --release --features gpu
 
       - name: Rename asset
         working-directory: bindings/nodejs
@@ -410,7 +411,7 @@ jobs:
 
       - name: Build GPU binary
         working-directory: bindings/nodejs
-        run: npx napi build --release --target aarch64-apple-darwin --features gpu
+        run: npx napi build --release --features gpu
 
       - name: Rename asset
         working-directory: bindings/nodejs
@@ -444,7 +445,7 @@ jobs:
 
       - name: Build GPU binary
         working-directory: bindings/nodejs
-        run: npx napi build --release --target x86_64-pc-windows-msvc --features gpu
+        run: npx napi build --release --features gpu
 
       - name: Rename asset
         working-directory: bindings/nodejs
@@ -469,6 +470,9 @@ jobs:
       - uses: actions/checkout@v7
 
       - uses: r-lib/actions/setup-r@v2
+
+      - name: Install R dependencies
+        run: Rscript -e 'options(repos = c(CRAN = "https://cloud.r-project.org")); if (!requireNamespace("BiocManager", quietly = TRUE)) install.packages("BiocManager"); BiocManager::install("BiocGenerics", ask = FALSE, update = FALSE)'
 
       - name: Build and install with GPU feature
         working-directory: benchmarks
@@ -497,6 +501,9 @@ jobs:
 
       - uses: r-lib/actions/setup-r@v2
 
+      - name: Install R dependencies
[…]
