### 2026-08-20 c63821a
fix: napi build without --target outputs fastlowess.node, not platform-suffixed name
