### 2026-08-20 ab74c3b
fix: R on Windows installs DLL to libs/x64/, not libs/
