### 2026-08-20 ecf3066
fix: checkout repo in attach-to-release so gh can find the git context
