### 2026-08-20 #769 @thisisamirv
Dear @etiennebacher. Thank you for the thorough review again. I just released a major revision (version 3.0.0) that addresses all your comments and also adds some new features to the package. Below is a point-by-point response to each comment.

---

## Setup

**1. `make r` failing with "python: not found"**

The R build no longer requires Python at all. All logic previously spread across `dev/prepare_cargo.py`, `dev/patch_vendor_crates.py`, `dev/clean_checksums.py`, and `dev/prepare_cran.sh` has been inlined directly into `bindings/r/Makefile`. There is no Python invocation remaining in the R build path.

**2. `ModuleNotFoundError: No module named 'tomli_w'`**

Resolved by the same change above — because Python scripts have been removed entirely from the R build, this module is no longer needed or referenced.

---

## Documentation

**R reference page: three functions on one page**

The `StreamingLowess` and `OnlineLowess` content has been split out of the main R API reference page into dedicated `r-streaming.md` and `r-online.md` files. Each of the three functions now has its own page. The same layout was applied to every other binding.

**Text `...: Arguments corresponding to LowessOptions fields.` in wrong location**

`...` has been added to `Lowess()`, `StreamingLowess()`, and `OnlineLowess()` to force named arguments for optional parameters (see Code section below), and roxygen2 `@param` documentation has been expanded throughout. The text now appears in the correct position alongside the `...` argument it documents.

**Documentation drift between website and `?Lowess`**

The roxygen2 `@param` docs have been expanded and a `See Also` section linking to <https://lowess.readthedocs.io/> has been added to every function. Furthermore, CI uses `dev/verify_snippets.py` which runs the code snippets in docs published to the website and those in the vignettes to ensure no inaccuracies exist. I should add that I prefer to keep the offline existing docs in the package minimal, as the actual guide for the package and every parameter would be too large for offline documentation in the package. Instead, I aim to provide minimal guide offline for basic usage, while redirecting advanced users to the website for an in-depth guide.

**Tutorials/examples: little explanation, code not evaluated, vignette not on website**

- `rfastlowess-intro.Rmd` has been expanded with additional prose and context.
- The `tutorials/` pages have been moved into a new `user-guide/use-cases/` section and
  grouped more logically.
- The `examples` pages have been partially integrated into `user-guide` pages where appropriate and the redundant sections have been removed.
- All `docs/api/` code examples have been standardized with expected output comments so
  readers can see results without running code.

**`robustness_method` and other non-trivial arguments lack explanation**

All `@param` entries have been expanded with descriptions of accepted values and their
effects. A `See Also` section linking to the online documentation has been added to direct
users to fuller explanations where argument interaction is complex.

**Broken/wrongly formatted links (e.g. `contributing/` in `CONTRIBUTING.md`)**

Incorrect URLs in the R binding docs have been fixed.

---

## Code

**Positional arguments risk: recommend `...` to force named arguments**

`...` has been added to `Lowess()`, `StreamingLowess()`, and `OnlineLowess()`, and a `reject_extra_positional_args()` helper has been introduced to error clearly if any unnamed arguments are passed. Optional parameters must now be specified by name.

**List-closure API (`$add_point()`) unfamiliar to R users**

The list-closure API was actually a direct port from the Rust builder API. However, I agree that such an API is not idiomatic in R. As a result, the API has been replaced with S3 generics: `fit()`, `process_chunk()`, `finalize()`, and `add_point()` are now standalone functions dispatched on the object class. This is the idiomatic R pattern and also makes the reference documentation easier to navigate.

---

## Other

**`Depends: R (>= ...)` missing from `DESCRIPTION`**

`Depends: R (>= 4.6)` has been added to `DESCRIPTION` and a matching entry has been added
to the CI test matrix.

---

## Additional Changes in v3.0.0

Beyond the above, v3.0.0 includes the following new features, breaking changes, and fixes.

### New Feature: GPU Backend

A `backend` option (`"cpu"` default, `"gpu"`) has been exposed on `Lowess` across all bindings (R, Python, Julia, Node.js, C++). The GPU backend is significantly faster than the CPU backend for large datasets. This feature is opt-in and not enabled in CRAN/Bioconductor/PyPI/npm releases. Users can opt in by calling `install_gpu()`. If you are interested, you can read more about it here: <https://lowess.readthedocs.io/en/latest/user-guide/gpu-backend/>

### Breaking Changes

**`OnlineOutput` field renames (all bindings):**
`OnlineOutput`'s `smoothed` and `std_error` fields have been renamed to `y` and `standard_error` across every binding (R, Python, Julia, Node.js, WASM, C++), matching the field names used in `LowessResult`.

**Monorepo restructure:**

- The monolithic root `Makefile` has been split into per-crate/binding sub-Makefiles (e.g. `crates/lowess/Makefile`, `bindings/r/Makefile`), each invokable directly. The root `Makefile` now only aggregates top-level targets.
- Rust and binding tests have been moved into their respective crate/binding directories (e.g. `tests/lowess/` → `crates/lowess/tests/lowess/`). This makes each binding/crate self-contained.
- Several now-redundant dev scripts have been removed: `dev/isolate_cargo.py`, `dev/check_root_cargo.py`, `dev/fix_doc_snippets.py`, and `check_js_licenses.js`.

### Bug Fixes

- **Docs:** `docs/api/rust.md` was showing Rust enum variant names instead of the string
  option values that the API actually accepts. Fixed.
