### 2026-08-20 #769 @thisisamirv
> Submitting Author Name: Amir Valizadeh
> Submitting Author Github Handle: [@thisisamirv](https://github.com/thisisamirv)
> Other Package Authors Github handles:
> Repository: https://github.com/thisisamirv/lowess-project
> Sub-directory: bindings/r
> Version submitted: 1.3.0
> Submission type: Stats Badge grade: gold
> Editor: [@etiennebacher](https://github.com/etiennebacher)
> Reviewers: [@JosiahParry](https://github.com/JosiahParry)
>
> Due date for [@JosiahParry](https://github.com/JosiahParry): 2026-09-02
> Archive: TBD
> Version accepted: TBD
> Language: en
>
> .
> .
> .
> ## Code of conduct
> * [x]  I agree to abide by [rOpenSci's Code of Conduct](https://ropensci.org/code-of-conduct/) during the review process and in maintaining my package should it be accepted.

By the way, should I change the version and other updated info here, or leave this be?
