### 2026-08-20 d398930
build: change R version dependency to >= 4.2.0
