+++ bindings/r/src/Cargo.toml
@@ -32,10 +32,6 @@ dev = []
 # Not used for CRAN/Bioconductor submission builds.
 gpu = ["fastLowess/gpu"]
 
-
-
-
-
 [workspace]
 
 [patch.crates-io]
