### 2026-08-20 7569deb
fix(r): strip bcryptprimitives.dll raw-dylib stub from archive before lld 19 linking on aarch64-pc-windows-gnullvm
