### 2026-08-20 80e68c3
fix(r): strip all raw-dylib *.dll stubs from archive on gnullvm to avoid lld 19 .idata errors
