+++ docs/changelog.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 **R:**
 
-- Strip bcryptprimitives.dll raw-dylib stub from archive before lld 19 linking on aarch64-pc-windows-gnullvm. It is required for Windows arm64 builds on R-Universe.
+- Strip all raw-dylib `*.dll` import stubs from archive before lld 19 linking on `aarch64-pc-windows-gnullvm`. Required for Windows arm64 builds on R-Universe.
 
 ## 3.0.0
 
