### 2026-08-20 1b3107c
fix(r): add -lbcryptprimitives -lsynchronization on gnullvm to restore symbols lost when stripping raw-dylib stubs
