### 2026-08-20 60e3db4
fix(r): patch .idata out of raw-dylib stubs with llvm-objcopy instead of deleting them
