### 2026-08-21 d9fa11b
fix(r): always re-insert ar member after objcopy regardless of exit code
