### 2026-08-21 d85fe44
fix(r): use ar x (no member arg) to correctly extract long-name archive members
