+++ docs/changelog.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 **R:**
 
-- Used `llvm-objcopy --remove-section=.idata$4` to zero out invalid relocations in `.idata$4` section headers inside raw-dylib import stubs in `librfastlowess.a`, which lld 19 rejects on `aarch64-pc-windows-gnullvm`. Required for Windows arm64 builds on R-Universe.
+- Used `ar x` (no member argument) to extract all members from `librfastlowess.a` into a temp directory, then `objcopy --remove-section=.idata$4` on each `.dll` stub, then `ar r` to re-insert. `ar x` without a member name resolves long-name archive entries automatically; named extraction silently fails for entries stored as `/<offset>` (long names >16 chars). Required for Windows arm64 builds on R-Universe, where lld 19 rejects `.idata$4` relocations referencing special section 0 in rustc-generated raw-dylib stubs.
 
 ## 3.0.0
 
