### 2026-08-21 a98ed73
fix(r): reference rtools by bare filename instead of a hardcoded drive path
