### 2026-08-21 62fbc1a
fix(r): extended the existing archive-patching step on windows
