### 2026-08-21 bf83499
fix(r): drop --whole-archive on gnullvm to avoid duplicate raw-dylib symbol crashes
