### 2026-08-21 7c0c378
docs: minor comment fixes in R
