+++ docs/changelog.md
@@ -18,9 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 **R:**
 
-- Used `ar x` (no member argument) to extract all members from `librfastlowess.a` into a temp directory, then `objcopy --remove-section=.idata$4` on each `.dll` stub, then `ar r` to re-insert. `ar x` without a member name resolves long-name archive entries automatically; named extraction silently fails for entries stored as `/<offset>` (long names >16 chars). Required for Windows arm64 builds on R-Universe, where lld 19 rejects `.idata$4` relocations referencing special section 0 in rustc-generated raw-dylib stubs.
-- Fixed `ld.lld` crashing (`Exception Code: 0xC0000005`) or dropping symbols entirely (`undefined symbol: WakeByAddressSingle`/`WaitOnAddress`) while linking `rfastlowess.dll` on Windows arm64/R-Universe. `--whole-archive` forces every crate's private raw-dylib stub object for a given DLL into the link, even when different crates' stubs cover different, non-overlapping symbols of that DLL (e.g. std's stub provides `WaitOnAddress`/`WakeByAddressAll`, `once_cell`'s provides `WakeByAddressSingle`); `--allow-multiple-definition` tolerates the resulting duplicates on x86_64 but crashes `ld.lld`'s arm64pe backend, and naively deduping same-named stub archive members to a single copy silently discarded symbols that only one of the duplicates provided. Fixed by dropping `--whole-archive`/`--no-whole-archive` entirely on the `gnullvm` target, letting normal single-definition-per-symbol archive resolution apply instead (nothing needed from the archive is lost, since `entrypoint.c` already references the extendr init symbol directly).
-- Fixed CRAN Windows build failure (`error: linker ... not found`): `cargo-config.toml` hardcoded the MinGW linker and `ar` as absolute `c:/rtools45/...` paths, which don't exist on machines where Rtools is installed elsewhere (e.g. CRAN's Windows build server uses `d:/rtools45`). Fixed by referencing the tools by name only, resolved via `PATH`.
+- Fixed Windows arm64 (R-Universe) build: `ar x` without a member name correctly resolves long-name archive entries (>16 chars stored as `/<offset>`); named extraction silently fails for such entries. Used `objcopy --remove-section=.idata$4` on each extracted `.dll` stub to strip the invalid relocations that lld 19 rejects, then `ar r` to re-insert.
+- Fixed `ld.lld` crashing or dropping symbols (`WakeByAddressSingle`, `WaitOnAddress`) on Windows arm64: `--whole-archive` pulls every crate's raw-dylib stub for a given DLL into the link, but different crates' stubs cover different, non-overlapping symbols of that DLL — `--allow-multiple-definition` works on x86_64 but crashes lld's arm64pe backend. Fixed by dropping `--whole-archive` on `gnullvm`; normal archive resolution applies and nothing is lost since `entrypoint.c` already references the extendr init symbol directly.
+- Fixed CRAN Windows build (`error: linker not found`): `cargo-config.toml` hardcoded linker/ar as `c:/rtools45/...` absolute paths, which break when Rtools is installed on a different drive. Fixed by using bare tool names resolved via `PATH`.
 
 ## 3.0.0
 
