### 2026-08-21 2c03437
build: split make docs-test and CI to language level
