### 2026-08-22 f2b996c
ci: add nodejs/wasm lint
