+++ .github/workflows/ci-wasm.yml
@@ -55,6 +55,7 @@ jobs:
         run: |
           cargo fmt --all -- --check
           cargo clippy --all-targets -- -D warnings
+          npm run lint
 
       - name: make wasm
         run: make wasm
