+++ .github/workflows/ci-nodejs.yml
@@ -55,6 +55,7 @@ jobs:
         run: |
           cargo fmt --all -- --check
           cargo clippy --all-targets -- -D warnings
+          npm run lint
 
       - name: make nodejs
         run: make nodejs
