### 2026-08-22 2a19727
fix(cpp): always target msvc on windows
