### 2026-08-22 a98582f
fix(cpp): makefile bugs, snippet verification
