### 2026-08-22 8ac77a2
fix: cran windows build error, eslint was replaced by oxlint
