### 2026-08-22 43627a1
ci: add npm install to wasm ci
