### 2026-08-22 7811a8a
fix: verify_snippet msvc linker
