### 2026-08-22 ca8d8e2
fix(cpp): re-derive cl.exe from vcvarsall PATH to prevent MSVC compiler/headers version mismatch
