+++ .github/workflows/ci-cpp.yml
@@ -33,9 +33,9 @@ jobs:
         if: matrix.os == 'windows-latest'
         run: rustup toolchain install stable-x86_64-pc-windows-gnu
 
-      - name: Install GNU Make (Windows)
+      - name: Install GNU Make and cppcheck (Windows)
         if: matrix.os == 'windows-latest'
-        run: choco install make -y
+        run: choco install make cppcheck -y
 
       - uses: Swatinem/rust-cache@v2
         with:
