### 2026-08-22 19c49fd
build: split makefiles into default and dev
