+++ .github/workflows/ci-r.yml
@@ -101,9 +101,9 @@ jobs:
           R CMD build .
           Rscript -e "BiocCheck::BiocCheck(list.files(pattern='[.]tar[.]gz$')[1], new_package=FALSE, no_check_vignettes=TRUE, quit_with_status=TRUE)"
 
-      - name: make r
+      - name: make r-dev
         if: matrix.os == 'ubuntu-latest'
-        run: make r
+        run: make r-dev
 
   srr:
     name: Software Review Requirements
