+++ Makefile
@@ -1,7 +1,7 @@
 # ==============================================================================
 # Configuration
 # ==============================================================================
-FEATURE_SET   ?= all
+FEATURE_SET ?= all
 RUN_GPU_TESTS ?= auto
 
 # Make shell commands fail on error
@@ -29,9 +29,9 @@ else
 	NPX := npx
 endif
 
-PYTHON      ?= python3
+PYTHON ?= python3
 PYO3_PYTHON ?= $(PYTHON)
-NODE        ?= node
+NODE ?= node
 
 TEMP ?= /tmp
 ifeq ($(OS),Windows_NT)
@@ -49,19 +49,19 @@ else
 endif
 
 # lowess crate
-LOWESS_PKG      := lowess
-LOWESS_DIR      := crates/lowess
+LOWESS_PKG := lowess
+LOWESS_DIR := crates/lowess
 LOWESS_FEATURES := std dev
 
 # fastLowess crate
-FASTLOWESS_PKG      := fastLowess
-FASTLOWESS_DIR      := crates/fastLowess
+FASTLOWESS_PKG := fastLowess
+FASTLOWESS_DIR := crates/fastLowess
 FASTLOWESS_FEATURES := cpu gpu dev
 
 # Python bindings
-PY_PKG  := fastLowess-py
-PY_DIR      := bindings/python
-PY_VENV     := .venv
+PY_PKG := fastLowess-py
+PY_DIR := bindings/python
+PY_VENV := .venv
 ifeq ($(OS),Windows_NT)
 	PY_ACTIVATE    := $(PY_VENV)/Scripts/activate
 	PY_VENV_PYTHON := $(PY_VENV)/Scripts/python.exe
@@ -72,8 +72,8 @@ endif
 
 # R bindings
 R_PKG_NAME := rfastlowess
-R_DIR      := bindings/r
-R_LIB_DIR  := $(R_DIR)/.r-lib
+R_DIR := bindings/r
+R_LIB_DIR := $(R_DIR)/.r-lib
 
 # Julia bindings
 JL_DIR := bindings/julia
@@ -87,15 +87,15 @@ else
 endif
 
 # Node.js bindings
-NODE_DIR      := bindings/nodejs
+NODE_DIR := bindings/nodejs
 
 # WebAssembly bindings
-WASM_DIR      := bindings/wasm
+WASM_DIR := bindings/wasm
 
 # C++ bindings
-CPP_DIR           := bindings/cpp
+CPP_DIR := bindings/cpp
 CPP_CARGO_PROFILE := --profile release-c
-CPP_LIBRARY_DIR   := target/release-c
+CPP_LIBRARY_DIR := target/release-c
 
 ifeq ($(OS),Windows_NT)
 	_CPP_GCC_MACHINE := $(shell gcc -dumpmachine 2>/dev/null)
@@ -115,6 +115,9 @@ DOCS_VENV := docs-venv
 lowess:
 	@"$(MAKE)" -f crates/lowess/Makefile FEATURE_SET="$(FEATURE_SET)"
 
+lowess-dev:
+	@"$(MAKE)" -f crates/lowess/Makefile dev FEATURE_SET="$(FEATURE_SET)"
+
 lowess-coverage:
 	@"$(MAKE)" -f crates/lowess/Makefile coverage
 
@@ -132,6 +135,11 @@ fastLowess:
 		FEATURE_SET="$(FEATURE_SET)" \
 		RUN_GPU_TESTS="$(RUN_GPU_TESTS)"
 
+fastLowess-dev:
+	@"$(MAKE)" -f crates/fastLowess/Makefile dev \
[…]
