### 2026-08-22 249bc13
fix: fall back to local libgcc_eh.a stub when gcc lib dir is not writable
