+++ docs/changelog.md
@@ -77,6 +77,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed `ld.lld` crashing or dropping symbols (`WakeByAddressSingle`, `WaitOnAddress`) on Windows arm64: `--whole-archive` pulls every crate's raw-dylib stub for a given DLL into the link, but different crates' stubs cover different, non-overlapping symbols of that DLL — `--allow-multiple-definition` works on x86_64 but crashes lld's arm64pe backend. Fixed by dropping `--whole-archive` on `gnullvm`; normal archive resolution applies and nothing is lost since `entrypoint.c` already references the extendr init symbol directly.
 - Fixed CRAN Windows build (`error: linker not found`): `cargo-config.toml` hardcoded linker/ar as `c:/rtools45/...` absolute paths, which break when Rtools is installed on a different drive. Fixed by using bare tool names resolved via `PATH`.
 - Fixed CRAN Windows build (`cannot find -lgcc_eh`): the Rtools45 posix-static toolchain does not ship `libgcc_eh.a`, causing `proc-macro2` and other build-script crates to fail to link. Fixed in `Makevars.win` by locating the GCC lib directory via `gcc -print-file-name=libgcc.a` and creating a dummy empty archive there before the `cargo build` step, so the link succeeds regardless of which drive Rtools is installed on.
+- Fixed CRAN Windows build (`libgcc_eh.a: Permission denied`): on CRAN's Windows server the Rtools gcc lib directory is not writable, so the existing `ar rcs` fallback failed with a permission error. `Makevars.win` now falls back to creating the stub in `$(TARGET_DIR)` instead and injects `rustflags = ["-C", "symbol-mangling-version=v0", "-C", "link-arg=-L<path>"]` into the `[target.x86_64-pc-windows-gnu]` section of `.cargo/config.toml` via `awk`, so the linker finds the local stub.
 
 ## 3.0.0
 
