### 2026-08-22 3ed89aa
fix: correct oxlint path in wasm package.json
