### 2026-08-22 140178c
docs: document ... parameter in R API constructor sections
