+++ docs/api/r.md
@@ -24,6 +24,8 @@ print(model)
 #>   Parallel:          TRUE
 ```
 
+* `...`: Not used. All parameters after `fraction` must be passed by name; passing extra positional arguments raises an error.
+
 **Methods:**
 
 ```r
@@ -55,6 +57,7 @@ See [r-online.md](r-online.md) for the `OnlineLowess` class.
 | Field | Type | Default | Description |
 | --- | --- | --- | --- |
 | `fraction` | `numeric` | `0.67` | Smoothing fraction (bandwidth) |
+| `...` | — | — | Not used. Forces all parameters after `fraction` to be passed by name. |
 | `iterations` | `integer` | `3` | Number of robustifying iterations |
 | `delta` | `numeric` | `NULL` | Interpolation distance (NULL for auto) |
 | `weight_function` | `character` | `"tricube"` | Weight function name |
