### 2026-08-22 e080516
refactor(dev): split verify_snippets into per-language runner modules
