### 2026-08-22 cd07813
fix(r): use absolute path in LIBRARY_PATH for libgcc_mock
