### 2026-08-23 0c120ac
ci: add R >= 4.2.0
