+++ bindings/r/R/utils.R
@@ -92,7 +92,9 @@ reject_extra_positional_args <- function(call, boundary_name) {
         arg_names <- rep("", length(call) - 1)
     }
     unnamed <- which(arg_names == "")
-    if (length(unnamed) > 1L || (length(unnamed) == 1L && unnamed[[1L]] != 1L)) {
+    if (
+        length(unnamed) > 1L || (length(unnamed) == 1L && unnamed[[1L]] != 1L)
+    ) {
         stop(
             sprintf("All arguments after '%s' must be named.", boundary_name),
             call. = FALSE
