+++ .github/workflows/ci-r.yml
@@ -26,7 +26,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        r: [release]
+        r: [release, '4.2']
     steps:
       - uses: actions/checkout@v7
 
