### 2026-08-23 6d62b52
docs: create gh-pages workflow for julia and r
