+++ bindings/julia/julia/docs/src/api.md
@@ -1,0 +1,5 @@
+# API Reference
+
+```@autodocs
+Modules = [FastLOWESS]
+```
