+++ bindings/julia/julia/docs/Project.toml
@@ -1,0 +1,7 @@
+[deps]
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+FastLOWESS = "5d760c97-4851-4f57-8eaf-1b3e829bbf83"
+
+[compat]
+Documenter = "1"
+julia = "1.6"
