+++ bindings/r/_pkgdown.yml
@@ -1,4 +1,4 @@
-url: https://thisisamirv.github.io/lowess-project/
+url: https://thisisamirv.github.io/lowess-project/r/
 template:
   bootstrap: 5
   bootswatch: flatly
