### 2026-08-23 bfc8333
ci: fix R 4.2.0 ci
