+++ r/CONTRIBUTING.md
@@ -1,0 +1,5 @@
+# Contributing to rfastlowess
+
+For contribution guidelines, please see the [Contributing
+Guide](https://lowess.readthedocs.io/en/latest/contributing/) in the
+docs.
