+++ r/articles/rfastlowess-intro.md
@@ -1,0 +1,290 @@
+# Introduction to rfastlowess
+
+## Overview
+
+`rfastlowess` is a high-performance R package for LOWESS (Locally
+Weighted Scatterplot Smoothing) built on a Rust backend.
+
+LOWESS fits a smooth curve through a scatter plot by performing a
+sequence of *local polynomial regressions*. For each point $`x_i`$, the
+algorithm selects the nearest `fraction` of observations, assigns
+distance-based weights (points farther away receive lower weight), fits
+a weighted least-squares polynomial, and reads off the predicted value
+at $`x_i`$. Because no global model is assumed, LOWESS adapts naturally
+to non-linear and non-stationary data.
+
+Optional robustness iterations guard against outliers: after each pass,
+residuals are used to recompute per-observation weights so that
+high-leverage outliers are progressively downweighted. This makes
+`iterations > 0` the right choice whenever the data may contain
+measurement errors or anomalies.
+
+**Full documentation**: <https://lowess.readthedocs.io/>
+
+## Installation
+
+``` r
+
+# From R-universe (pre-built binaries, no Rust required)
+install.packages("rfastlowess", repos = "https://thisisamirv.r-universe.dev")
+```
+
+## Quick Start
+
+### Basic Smoothing
+
+[`Lowess()`](https://thisisamirv.github.io/lowess-project/r/reference/Lowess.md)
+accepts all smoothing options at construction time and returns a
+reusable model object. Calling `fit(model, x, y)` runs the algorithm and
+returns a `LowessResult` list. Key result fields are `$x` (input x
+values), `$y` (smoothed fitted values), `$fraction_used`, and
+`$iterations_used`. The `fraction` parameter controls the bandwidth: a
+larger value produces a smoother curve at the cost of following local
+features less closely.
+
+``` r
+
+library(rfastlowess)
+
+# Generate example data
+set.seed(42)
+x <- seq(0, 10, length.out = 100)
+y <- sin(x) + rnorm(100, sd = 0.3)
+
+# Initialize model
+model <- Lowess(fraction = 0.3)
+print(model)
+#> <Lowess Model>
+#>   Fraction:          0.3 
+#>   Iterations:        3 
+#>   Weight Function:   tricube 
+#>   Parallel:          TRUE
+
+# Fit model
+result <- fit(model, x, y)
+print(result)
+#> <LowessResult>
+#>   Points:            100 
+#>   Fraction Used:     0.3
+
+# Quick visualization using the S3 plot method
+plot(result, main = "Auto-plot of LowessResult")
+```
+
+![](rfastlowess-intro_files/figure-html/basic_example-1.png)
+
+``` r
+
+
+# For custom plotting, access components directly
+plot(x, y, pch = 16, col = "gray", main = "Manual Overlay")
+lines(result$x, result$y, col = "red", lwd = 2)
+```
+
+![](rfastlowess-intro_files/figure-html/basic_example-2.png)
+
+### Robust Smoothing with Intervals
+
+Real-world data often contains a small fraction of gross errors. Setting
+`iterations > 0` activates robustness re-weighting: after the initial
+fit, large residuals identify likely outliers and their contribution is
+downweighted for the next iteration. The bisquare robustness function
+(the default) reduces the weight of any point whose residual exceeds
+about six times the median absolute deviation to zero.
+
[…]

[…]
