### 2026-08-23 4c12357
docs: remove now stale CI badge
