### 2026-08-23 40626b5
ci: fix lib path for R 4.2.0 in CI
