### 2026-08-23 4be3e9f
ci: change R 4.2 dep installation
