### 2026-08-23 27b2c69
ci(r): slim R 4.2 deps to check-only packages, skip devtools/BiocStyle
