### 2026-08-23 9fcb3ee
build: bump minimum R to 4.4, drop 4.2 BiocManager workaround
