+++ .github/workflows/ci-r.yml
@@ -26,7 +26,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        r: [release, "4.2"]
+        r: [release, "4.4"]
     steps:
       - uses: actions/checkout@v7
 
@@ -50,7 +50,7 @@ jobs:
       - uses: r-lib/actions/setup-pandoc@v2
 
       - name: Bootstrap pak (Windows)
-        if: runner.os == 'Windows' && matrix.r == 'release'
+        if: runner.os == 'Windows'
         shell: Rscript {0}
         run: |
           install.packages(
@@ -61,25 +61,7 @@ jobs:
             )
           )
 
-      # R 4.2: Bioc 3.15/16 source tarballs give 404 — pak cannot resolve them; use BiocManager directly
-      - name: Install R dependencies (R 4.2)
-        if: matrix.r == '4.2'
-        shell: bash
-        run: |
-          Rscript --no-save -e '
-            lib <- Sys.getenv("R_LIBS_USER")
-            dir.create(lib, recursive = TRUE, showWarnings = FALSE)
-            .libPaths(c(lib, .libPaths()))
-            repos <- c(RSPM = Sys.getenv("RSPM", "https://cloud.r-project.org"), CRAN = "https://cloud.r-project.org")
-            install.packages(c("rcmdcheck", "testthat", "knitr", "rmarkdown"), lib = lib, repos = repos)
-            install.packages("BiocManager", lib = lib, repos = repos)
-            BiocManager::install("BiocGenerics", lib = lib, ask = FALSE, update = FALSE)
-            if (!dir.exists(file.path(lib, "rcmdcheck"))) stop("rcmdcheck not installed")
-            if (!dir.exists(file.path(lib, "BiocGenerics"))) stop("BiocGenerics not installed")
-          '
-
       - uses: r-lib/actions/setup-r-dependencies@v2
-        if: matrix.r == 'release'
         with:
           working-directory: bindings/r
           extra-packages: |
@@ -105,7 +87,6 @@ jobs:
           working-directory: bindings/r
 
       - name: Install Package
-        if: matrix.r == 'release'
         working-directory: bindings/r
         run: Rscript -e "devtools::install(quiet = TRUE)"
 
