+++ r/pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.2.1
 pkgdown_sha: ~
 articles:
   rfastlowess-intro: rfastlowess-intro.html
-last_built: 2026-08-23T01:03Z
+last_built: 2026-08-23T02:00Z
 urls:
   reference: https://thisisamirv.github.io/lowess-project/r/reference
   article: https://thisisamirv.github.io/lowess-project/r/articles
