### 2026-08-23 4dd17d1
docs: move r docs to github pages made with pkgdown
