+++ bindings/r/vignettes/rfastlowess-intro.Rmd
@@ -1,244 +1,0 @@
----
-title: "Introduction to rfastlowess"
-pagetitle: "Introduction to rfastlowess"
-author: "Amir Valizadeh"
-date: "`r Sys.Date()`"
-output: BiocStyle::html_document
-vignette: >
-  %\VignetteIndexEntry{Introduction to rfastlowess}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r setup, include = FALSE}
-knitr::opts_chunk$set(
-    collapse = TRUE,
-    comment = "#>",
-    fig.width = 7,
-    fig.height = 5
-)
-```
-
-```{r srr-tags, eval = FALSE, echo = FALSE}
-#' @srrstats {G1.0} Package introduction with academic context.
-#' @srrstats {G1.4} Examples use roxygen2-documented functions.
-#' @srrstats {RE1.2} Documents expected numeric vector input format.
-#' @srrstats {RE1.4} Documents LOWESS assumptions and use cases.
-#' @srrstats {RE4.17} Demonstrates print() method for model objects.
-#' @srrstats {RE6.0, RE6.2} Demonstrates plot() method with CI visualization.
-```
-
-## Overview
-
-`rfastlowess` is a high-performance R package for LOWESS (Locally Weighted
-Scatterplot Smoothing) built on a Rust backend.
-
-LOWESS fits a smooth curve through a scatter plot by performing a sequence of
-*local polynomial regressions*. For each point $x_i$, the algorithm selects
-the nearest `fraction` of observations, assigns distance-based weights (points
-farther away receive lower weight), fits a weighted least-squares polynomial,
-and reads off the predicted value at $x_i$. Because no global model is
-assumed, LOWESS adapts naturally to non-linear and non-stationary data.
-
-Optional robustness iterations guard against outliers: after each pass,
-residuals are used to recompute per-observation weights so that high-leverage
-outliers are progressively downweighted. This makes `iterations > 0` the right
-choice whenever the data may contain measurement errors or anomalies.
-
-**Full documentation**: <https://lowess.readthedocs.io/>
-
-## Installation
-
-```{r installation, eval=FALSE}
-# From R-universe (pre-built binaries, no Rust required)
-install.packages("rfastlowess", repos = "https://thisisamirv.r-universe.dev")
-```
-
-## Quick Start
-
-### Basic Smoothing
-
-`Lowess()` accepts all smoothing options at construction time and returns a
-reusable model object. Calling `fit(model, x, y)` runs the algorithm and
-returns a `LowessResult` list. Key result fields are `$x` (input x values),
-`$y` (smoothed fitted values), `$fraction_used`, and `$iterations_used`. The
-`fraction` parameter controls the bandwidth: a larger value produces a smoother
-curve at the cost of following local features less closely.
-
-```{r basic_example}
-library(rfastlowess)
-
-# Generate example data
-set.seed(42)
-x <- seq(0, 10, length.out = 100)
-y <- sin(x) + rnorm(100, sd = 0.3)
-
-# Initialize model
-model <- Lowess(fraction = 0.3)
-print(model)
-
-# Fit model
-result <- fit(model, x, y)
-print(result)
-
-# Quick visualization using the S3 plot method
-plot(result, main = "Auto-plot of LowessResult")
-
-# For custom plotting, access components directly
-plot(x, y, pch = 16, col = "gray", main = "Manual Overlay")
-lines(result$x, result$y, col = "red", lwd = 2)
-```
-
-### Robust Smoothing with Intervals
-
-Real-world data often contains a small fraction of gross errors. Setting
[…]
