### 2026-08-23 a498025
build: skip BiocCheck on R 4.4
