+++ .github/workflows/ci-nodejs.yml
@@ -2,9 +2,9 @@ name: CI — Node.js
 
 on:
   push:
-    branches: [main, master, develop]
+    branches: [main, master]
   pull_request:
-    branches: [main, master, develop]
+    branches: [main, master]
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.head_ref || github.ref }}
