+++ .github/workflows/ci-python.yml
@@ -2,9 +2,9 @@ name: CI — Python
 
 on:
   push:
-    branches: [main, master, develop]
+    branches: [main, master]
   pull_request:
-    branches: [main, master, develop]
+    branches: [main, master]
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.head_ref || github.ref }}
