+++ .github/workflows/ci-cpp.yml
@@ -2,9 +2,9 @@ name: CI — C++
 
 on:
   push:
-    branches: [main, master, develop]
+    branches: [main, master]
   pull_request:
-    branches: [main, master, develop]
+    branches: [main, master]
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.head_ref || github.ref }}
