### 2026-08-23 b6f7aa8
feat: add length gaurd for R extra arguments, fix minor R docs issues
