+++ r/llms.txt
@@ -601,7 +601,7 @@ Check for and install the optional GPU-accelerated backend.
 
 # Articles
 
-### Getting Started
+### Introduction
 
 - [Installation](https://thisisamirv.github.io/lowess-project/r/articles/installation.md):
 - [Quick
@@ -647,6 +647,6 @@ Check for and install the optional GPU-accelerated backend.
 - [Use Case: Real-Time
   Processing](https://thisisamirv.github.io/lowess-project/r/articles/use-case-real-time.md):
 
-### Reference
+### Performance
 
 - [Benchmarks](https://thisisamirv.github.io/lowess-project/r/articles/benchmarks.md):
