+++ r/pkgdown.yml
@@ -20,7 +20,7 @@ articles:
   use-case-genomics: use-case-genomics.html
   use-case-real-time: use-case-real-time.html
   use-case-time-series: use-case-time-series.html
-last_built: 2026-08-23T02:50Z
+last_built: 2026-08-23T03:14Z
 urls:
   reference: https://thisisamirv.github.io/lowess-project/r/reference
   article: https://thisisamirv.github.io/lowess-project/r/articles
