### 2026-08-23 ce01b7f
docs: move changelog and contributing to repo root
