+++ bindings/r/NEWS.md
@@ -1,4 +1,4 @@
 # rfastlowess 0.0.0
 
 * For the changelog, see:
-<https://lowess.readthedocs.io/en/latest/changelog/>
+<https://github.com/thisisamirv/lowess-project/blob/main/CHANGELOG.md>
