+++ bindings/r/_pkgdown.yml
@@ -1,5 +1,6 @@
 url: https://thisisamirv.github.io/lowess-project/r/
 template:
+  package: srr
   bootstrap: 5
   bootswatch: flatly
 
