+++ docs/mkdocs.yml
@@ -157,5 +157,3 @@ nav:
       - WebAssembly: api/wasm.md
       - C++: api/cpp.md
   - Benchmarks: benchmarks.md
-  - Changelog: changelog.md
-  - Contributing: contributing.md
