### 2026-08-23 bdf98ca
docs: fix vignette drifts from docs
