+++ bindings/r/vignettes/intervals.Rmd
@@ -109,11 +109,11 @@ legend("topright",
 
 Pass any value between 0 and 1 (exclusive). Common choices:
 
-| Level | Interpretation |
-| --- | --- |
-| 0.90 | 90% of the true curve / future points fall inside |
-| 0.95 | Standard choice (95%) |
-| 0.99 | Conservative (wider bands) |
+| Level | z-value | Interpretation |
+| --- | --- | --- |
+| 0.90 | 1.645 | 90% of intervals contain true value |
+| 0.95 | 1.960 | 95% of intervals contain true value |
+| 0.99 | 2.576 | 99% of intervals contain true value |
 
 ```{r}
 # 99% confidence intervals

[…]
