### 2026-08-23 88a68bd
docs: in vignettes break adapters to four pages
