+++ bindings/r/vignettes/adapters.Rmd
@@ -1,173 +1,0 @@
----
-
-title: "Execution Modes"
-output: BiocStyle::html_document
-vignette: >
-  %\VignetteIndexEntry{Execution Modes}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-<!-- markdownlint-disable MD024 -->
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(collapse = TRUE, comment = "#>", eval = FALSE)
-```
-
-## Overview
-
-![Execution mode comparison](../man/figures/adapter_comparison.svg)
-
-`rfastlowess` provides three execution modes for different data sizes and processing needs:
-
-| Mode | Use Case | Memory | Key Features |
-| --- | --- | --- | --- |
-| **Batch** | Complete datasets in memory | Entire dataset | All features: CI, PI, CV, GPU |
-| **Streaming** | Large files (>100K points) | One chunk at a time | Chunked processing with overlap |
-| **Online** | Real-time / live data | Fixed sliding window | Point-by-point incremental updates |
-
----
-
-## Batch Mode
-
-The default mode. Processes the entire dataset at once and supports all features.
-
-### When to Use
-
-- Dataset fits comfortably in memory
-- Need confidence/prediction intervals
-- Need cross-validation
-- One-shot analysis
-
-### Parameters
-
-| Parameter | Default | Description |
-| --- | --- | --- |
-| `fraction` | 0.67 | Neighbourhood size |
-| `iterations` | 3 | Robustness iterations |
-| `confidence_intervals` | `NULL` | CI coverage (e.g. 0.95) |
-| `prediction_intervals` | `NULL` | PI coverage (e.g. 0.95) |
-| `parallel` | `FALSE` | Enable CPU parallelism |
-| `backend` | `"cpu"` | `"cpu"` or `"gpu"` |
-
-### Example
-
-```{r}
-library(rfastlowess)
-set.seed(42)
-x <- seq(0, 2 * pi, length.out = 100)
-y <- sin(x) + rnorm(100, sd = 0.3)
-
-model <- Lowess(
-    fraction = 0.5,
-    iterations = 3,
-    confidence_intervals = 0.95,
-    prediction_intervals = 0.95,
-    return_diagnostics = TRUE
-)
-result <- fit(model, x, y)
-
-print(result$confidence_lower)
-print(result$confidence_upper)
-print(result$diagnostics$r_squared)
-```
-
-![Gap handling](../man/figures/gap_handling.svg)
-
----
-
-## Streaming Mode
-
-Processes data in fixed-size chunks with configurable overlap. Results for each chunk are returned after calling `process_chunk()`. Call `finalize()` after the last chunk to flush remaining buffered points.
-
-### When to Use
-
-- Dataset is too large to fit in memory
-- Processing data from a file or stream
-- Batch pipeline with memory constraints
-
-### Parameters
-
-| Parameter | Default | Description |
-| --- | --- | --- |
-| `chunk_size` | 5000 | Points per chunk |
-| `overlap` | 500 | Overlap between chunks |
-| `merge_strategy` | `"weighted_average"` | How to merge overlapping values |
-
-### Merge Strategies
-
-| Strategy | Behavior |
-| --- | --- |
-| `"average"` | Arithmetic mean of both estimates |
-| `"weighted_average"` | Distance-weighted blend (recommended) |
-| `"take_first"` | Keep left-chunk estimate |
-| `"take_last"` | Keep right-chunk estimate |
-
-### Example
-
-```{r}
-library(rfastlowess)
-set.seed(42)
-x <- seq(0, 2 * pi, length.out = 100)
[…]
