+++ bindings/r/_pkgdown.yml
@@ -46,7 +46,10 @@ articles:
   - title: User Guide
     contents:
       - parameters
-      - adapters
+      - adapter-choice
+      - batch
+      - streaming
+      - online
       - intervals
       - cross-validation
   - title: Weight & Robustness

[…]
