### 2026-08-23 dcdddec
docs: fix the position of figures in vignettes
