+++ r/llms.txt
@@ -512,13 +512,14 @@ result.cv_scores()
 
 ## Contributing
 
-Contributions are welcome! Please see the [Contributing
-Guide](https://lowess.readthedocs.io/en/latest/contributing/) for more
-information.
+Contributions are welcome! Please see
+[CONTRIBUTING.md](https://github.com/thisisamirv/lowess-project/blob/main/CONTRIBUTING.md)
+for more information.
 
 ## Changelog
 
-See the [Changelog](https://lowess.readthedocs.io/en/latest/changelog/)
+See
+[CHANGELOG.md](https://github.com/thisisamirv/lowess-project/blob/main/CHANGELOG.md)
 for a history of changes.
 
 ## License
@@ -613,8 +614,14 @@ Check for and install the optional GPU-accelerated backend.
 
 - [Parameters
   Reference](https://thisisamirv.github.io/lowess-project/r/articles/parameters.md):
-- [Execution
-  Modes](https://thisisamirv.github.io/lowess-project/r/articles/adapters.md):
+- [Choosing an Execution
+  Mode](https://thisisamirv.github.io/lowess-project/r/articles/adapter-choice.md):
+- [Batch Mode
+  (Lowess)](https://thisisamirv.github.io/lowess-project/r/articles/batch.md):
+- [Streaming Mode
+  (StreamingLowess)](https://thisisamirv.github.io/lowess-project/r/articles/streaming.md):
+- [Online Mode
+  (OnlineLowess)](https://thisisamirv.github.io/lowess-project/r/articles/online.md):
 - [Confidence and Prediction
   Intervals](https://thisisamirv.github.io/lowess-project/r/articles/intervals.md):
 - [Cross-Validation](https://thisisamirv.github.io/lowess-project/r/articles/cross-validation.md):
