### 2026-08-23 e2b440e
docs: move julia docs from docs/ to github pages
