+++ bindings/r/R/StreamingLowess.R
@@ -15,7 +15,6 @@
 #'   \code{"take_first"} (alias: \code{"first"}), or
 #'   \code{"take_last"} (alias: \code{"last"}).
 #'
-#' @seealso \url{https://lowess.readthedocs.io/} for full documentation.
 #' @return A StreamingLowess object.
 #' @examples
 #' x <- seq(0, 10, length.out = 100)
