+++ julia/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"category":"section","location":"api/#API-Reference","page":"API Reference","text":"","title":"API Reference"},{"category":"module","location":"api/#FastLOWESS.FastLOWESS","page":"API Reference","text":"FastLOWESS\n\nHigh-performance LOWESS (Locally Weighted Scatterplot Smoothing) for Julia.\n\nProvides bindings to the fastlowess Rust library for fast, robust LOWESS smoothing.\n\nMain API\n\nLowess(; kwargs...): Configure batch LOWESS\nfit(model, x, y): Fit and return results\nStreamingLowess(; kwargs...) / process_chunk / finalize: Streaming mode\nOnlineLowess(; kwargs...) / add_point: Online sliding-window mode\n\nExample\n\nusing FastLOWESS\n\nx = collect(1.0:0.1:10.0)\ny = sin.(x) .+ 0.1 .* randn(length(x))\n\nresult = fit(Lowess(fraction=0.3), x, y)\nprintln(\"Smoothed values: \", result.y)\n\n\n\n\n\n","title":"FastLOWESS.FastLOWESS"},{"category":"type","location":"api/#FastLOWESS.Diagnostics","page":"API Reference","text":"Diagnostics\n\nDiagnostic statistics for LOWESS fit quality.\n\nFields\n\nrmse::Float64: Root Mean Squared Error\nmae::Float64: Mean Absolute Error\nr_squared::Float64: R-squared (coefficient of determination)\naic::Float64: Akaike Information Criterion (NaN if not computed)\naicc::Float64: Corrected AIC (NaN if not computed)\neffective_df::Float64: Effective degrees of freedom (NaN if not computed)\nresidual_sd::Float64: Residual standard deviation\n\n\n\n\n\n","title":"FastLOWESS.Diagnostics"},{"category":"type","location":"api/#FastLOWESS.Lowess","page":"API Reference","text":"Lowess(; kwargs...)\n\nStateful batch LOWESS smoother.\n\nKeyword Arguments\n\nfraction::Float64 = 0.67: Smoothing fraction (proportion of data used for each fit)\niterations::Int = 3: Number of robustness iterations\ndelta::Float64 = NaN: Interpolation optimization threshold (NaN for auto)\nweight_function::String = \"tricube\": Kernel function\nrobustness_method::String = \"bisquare\": Robustness method\nscaling_method::String = \"mad\": Scaling method for robustness\nboundary_policy::String = \"extend\": Handling of edge effects\nconfidence_intervals::Float64 = NaN: Confidence level (e.g., 0.95), NaN to disable\nprediction_intervals::Float64 = NaN: Prediction interval level, NaN to disable\nreturn_diagnostics::Bool = false: Whether to compute RMSE, MAE, R², etc.\nreturn_residuals::Bool = false: Whether to include residuals\nreturn_robustness_weights::Bool = false: Whether to include robustness weights\nzero_weight_fallback::String = \"use_local_mean\": Fallback when all weights are zero\nauto_converge::Float64 = NaN: Tolerance for auto-convergence, NaN to disable\ncv_fractions::Vector{Float64} = Float64[]: Fractions for cross-validation\ncv_method::String = \"kfold\": CV method (\"loocv\" or \"kfold\")\ncv_k::Int = 5: Number of folds for k-fold CV\nparallel::Bool = true: Enable parallel execution\ncv_seed::Union{Int,Nothing} = nothing: Seed for reproducible K-fold splits\nreturn_se::Bool = false: Return standard errors\nbackend::String = \"cpu\": Execution backend (\"cpu\" or \"gpu\"). GPU requires the library to be built with the gpu Cargo feature and a Vulkan/Metal/DX12- capable GPU driver.\n\nExample\
[…]
