+++ r/reference/StreamingLowess.md
@@ -132,10 +132,6 @@ StreamingLowess(
 
 A StreamingLowess object.
 
-## See also
-
-<https://lowess.readthedocs.io/> for full documentation.
-
 ## Examples
 
 ``` r
