+++ r/pkgdown.yml
@@ -23,7 +23,7 @@ articles:
   use-case-genomics: use-case-genomics.html
   use-case-real-time: use-case-real-time.html
   use-case-time-series: use-case-time-series.html
-last_built: 2026-08-23T04:05Z
+last_built: 2026-08-23T04:32Z
 urls:
   reference: https://thisisamirv.github.io/lowess-project/r/reference
   article: https://thisisamirv.github.io/lowess-project/r/articles
