### 2026-08-23 d13d2a3
docs: fix julia links
