+++ bindings/julia/julia/docs/src/use-case-time-series.md
@@ -171,6 +171,6 @@ println("R²: ", result.diagnostics.r_squared)
 ## See Also
 
 - [Real-Time Processing](real-time.md) — For streaming time series
-- [Cross-Validation](../user-guide/cross-validation.md) — Optimal fraction selection
-- [Boundary Handling](../user-guide/boundary.md) — Edge bias in trend extraction
-- [Parameters](../user-guide/parameters.md) — Full parameter reference
+- [Cross-Validation](cross-validation.md) — Optimal fraction selection
+- [Boundary Handling](boundary.md) — Edge bias in trend extraction
+- [Parameters](parameters.md) — Full parameter reference
