+++ bindings/julia/julia/docs/src/adapter-choice.md
@@ -40,4 +40,4 @@ graph LR
 ## Next Steps
 
 - [Parameters](parameters.md) — All configuration options
-- [Tutorials](../tutorials/real-time.md) — Real-time processing guide
+- [Tutorials](use-case-real-time.md) — Real-time processing guide
