### 2026-08-23 1b78f8c
docs: add figures to julia docs
