+++ bindings/julia/julia/docs/src/intervals.md
@@ -4,7 +4,7 @@ Confidence and prediction intervals for uncertainty quantification.
 
 ## Overview
 
-![Confidence and Prediction Intervals](../assets/diagrams/intervals_comparison.svg)
+![Confidence and Prediction Intervals](assets/intervals_comparison.svg)
 
 !!! note "Adapter support"
     Confidence and prediction intervals are available in **Batch** mode only. Streaming and Online modes do not support intervals.
