+++ bindings/julia/julia/docs/src/kernels.md
@@ -6,7 +6,7 @@ Kernel functions for distance weighting.
 
 Weight functions (kernels) determine how neighboring points contribute to each local fit. Points closer to the target receive higher weights.
 
-![Weight Functions](../assets/diagrams/kernel_comparison.svg)
+![Weight Functions](assets/kernel_comparison.svg)
 
 ---
 
