+++ bindings/julia/julia/docs/src/adapter-choice.md
@@ -19,7 +19,7 @@ graph LR
 | **Streaming** | Large files (>100K) | Chunked | Residuals, robustness |
 | **Online** | Real-time sensors | Fixed window | Incremental updates |
 
-![Adapter Comparison](../assets/diagrams/adapter_comparison.svg)
+![Adapter Comparison](assets/adapter_comparison.svg)
 
 ---
 
