+++ bindings/julia/julia/docs/src/merge.md
@@ -22,7 +22,7 @@ Overlap:            [=====]
 | `"take_last"` | Right-chunk estimate only | Low | Fastest |
 | `"weighted_average"` | Distance-weighted mean | High | Moderate |
 
-![Merge Strategies](../assets/diagrams/merge_comparison.svg)
+![Merge Strategies](assets/merge_comparison.svg)
 
 ---
 
