+++ bindings/julia/julia/docs/src/online.md
@@ -8,7 +8,7 @@ Incremental updates with a sliding window for real-time data.
 - Need real-time smoothed values
 - Fixed memory budget
 
-![Online Adapter](../assets/diagrams/online_comparison.svg)
+![Online Adapter](assets/online_comparison.svg)
 
 ## Parameters
 
