### 2026-08-23 b1dcdc4
docs: migrate rust docs to docs.rs
