+++ crates/lowess/src/lib.rs
@@ -232,6 +232,9 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
+#[cfg(doc)]
+pub mod doc;
+
 #[cfg(not(feature = "std"))]
 #[macro_use]
 extern crate alloc;
