+++ dev/runners/base.py
@@ -10,6 +10,11 @@ REPO_ROOT = Path(__file__).resolve().parent.parent.parent
 DOCS_DIR = REPO_ROOT / "docs"
 JULIA_DOCS_DIR = REPO_ROOT / "bindings" / "julia" / "julia" / "docs" / "src"
 VIGNETTES_DIRS: list[Path] = [REPO_ROOT / "bindings" / "r" / "vignettes"]
+RUST_CRATE_DOCS_DIRS: list[Path] = [
+    REPO_ROOT / "crates" / "fastLowess" / "docs",
+    REPO_ROOT / "crates" / "lowess" / "docs",
+]
+NODEJS_BINDING_DOCS_DIR = REPO_ROOT / "bindings" / "nodejs" / "docs"
 
 _TAB_ALIASES: dict[str, set[str]] = {
     "python": {"Python"},
