+++ docs/getting-started/installation.md
@@ -26,23 +26,6 @@ Install the LOWESS library for your preferred language.
     pip install maturin
     maturin develop --release
     ```
-=== "Node.js"
-
-    === "From NPM (recommended)"
-
-    ```bash
-    npm install fastlowess
-    ```
-
-    === "From Source"
-
-    ```bash
-    git clone https://github.com/thisisamirv/lowess-project
-    cd lowess-project/bindings/nodejs
-    npm install
-    npm run build
-    ```
-
 === "WebAssembly"
 
     === "From NPM (recommended)"
@@ -140,19 +123,6 @@ Install the LOWESS library for your preferred language.
     result = model.fit(x, y)
     print("Installed successfully!")
     ```
-=== "Node.js"
-
-    ```javascript
-    const fl = require('fastlowess');
-    
-    const x = new Float64Array([1.0, 2.0, 3.0]);
-    const y = new Float64Array([2.0, 4.0, 6.0]);
-    
-    const model = new fl.Lowess({});
-    const result = model.fit(x, y);
-    console.log("Installed successfully!");
-    ```
-
 === "WebAssembly"
 
     ```javascript
