### 2026-08-23 a92bb03
docs: add workflow for publishing nidejs docs
