+++ docs/getting-started/quickstart.md
@@ -23,21 +23,6 @@ Smooth a noisy sine wave — the kind of signal where LOWESS shines. Each exampl
 
     print(f"First smoothed value: {result.y[0]:.4f}  (true: {np.sin(x[0]):.4f})")
     ```
-=== "WebAssembly"
-
-    ```javascript
-    const { Lowess } = require('./fastlowess_wasm.js');
-
-    const n = 100;
-    const x = Float64Array.from({ length: n }, (_, i) => i * 2 * Math.PI / (n - 1));
-    const y = Float64Array.from(x, (xi, i) => Math.sin(xi) + (((i * 7 + 3) % 17) / 17 - 0.5) * 0.6);
-
-    const model = new Lowess({ fraction: 0.3, iterations: 3 });
-    const result = model.fit(x, y);
-
-    console.log(`First smoothed: ${result.y[0].toFixed(4)}`);
-    ```
-
 === "C++"
 
     ```cpp
@@ -92,30 +77,6 @@ Smooth a noisy sine wave — the kind of signal where LOWESS shines. Each exampl
     print("CI Upper:", result.confidence_upper)
     print("R²:", result.diagnostics.r_squared)
     ```
-=== "WebAssembly"
-
-    ```javascript
-    const { Lowess } = require('./fastlowess_wasm.js');
-
-    const n = 100;
-    const x = Float64Array.from({ length: n }, (_, i) => i * 2 * Math.PI / (n - 1));
-    const y = Float64Array.from(x, (xi, i) => Math.sin(xi) + (((i * 7 + 3) % 17) / 17 - 0.5) * 0.6);
-
-    const model = new Lowess({
-        fraction: 0.5,
-        iterations: 3,
-        confidence_intervals: 0.95,
-        prediction_intervals: 0.95,
-        return_diagnostics: true
-    });
-    const result = model.fit(x, y);
-
-    console.log("Smoothed:", result.y);
-    console.log("CI Lower:", result.confidence_lower);
-    console.log("CI Upper:", result.confidence_upper);
-    console.log("R²:", result.diagnostics.r_squared);
-    ```
-
 === "C++"
 
     ```cpp
@@ -185,31 +146,6 @@ LOWESS can robustly handle outliers through iterative reweighting:
         if w < 0.5:
             print(f"Point {i} is likely an outlier (weight: {w:.3f})")
     ```
-=== "WebAssembly"
-
-    ```javascript
-    const { Lowess } = require('./fastlowess_wasm.js');
-
-    // Data with an outlier at position 3
-    const x = new Float64Array([1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
-    const yWithOutlier = new Float64Array([2.0, 4.0, 6.0, 50.0, 10.0, 12.0]);
-
-    const model = new Lowess({
-        fraction: 0.5,
-        iterations: 5,
-        robustness_method: "bisquare",
-        return_robustness_weights: true
-    });
-    const result = model.fit(x, yWithOutlier);
-
-    // Outliers will have low robustness weights
-    result.robustness_weights.forEach((w, i) => {
-        if (w < 0.5) {
-            console.log(`Point ${i} is likely an outlier (weight: ${w.toFixed(3)})`);
-        }
-    });
-    ```
-
 === "C++"
 
     ```cpp
@@ -274,32 +210,6 @@ For datasets too large to fit in memory, stream them in fixed-size chunks with o
     result = model.finalize()
     print(f"Smoothed {len(result.y)} points in streaming mode")
     ```
-=== "WebAssembly"
-
-    ```javascript
-    const { StreamingLowess } = require('./fastlowess_wasm.js');
-
-    const n = 5000;
[…]
