+++ docs/user-guide/boundary.md
@@ -36,18 +36,6 @@ Pads beyond both endpoints by replicating the first and last observed values. Pr
     model = fl.Lowess(boundary_policy="extend")
     result = model.fit(x, y)
     ```
-=== "WebAssembly"
-    ```javascript
-    const { Lowess } = require('./fastlowess_wasm.js');
-
-    const n = 100;
-    const x = Float64Array.from({ length: n }, (_, i) => i * 2 * Math.PI / (n - 1));
-    const y = Float64Array.from(x, (xi, i) => Math.sin(xi) + (((i * 7 + 3) % 17) / 17 - 0.5) * 0.6);
-
-    const model = new Lowess({ boundary_policy: "extend" });
-    const result = model.fit(x, y);
-    ```
-
 === "C++"
     ```cpp
     #include <fastlowess.hpp>
@@ -90,18 +78,6 @@ Mirrors the data about both endpoints before fitting, then discards the reflecte
     model = fl.Lowess(boundary_policy="reflect")
     result = model.fit(x, y)
     ```
-=== "WebAssembly"
-    ```javascript
-    const { Lowess } = require('./fastlowess_wasm.js');
-
-    const n = 100;
-    const x = Float64Array.from({ length: n }, (_, i) => i * 2 * Math.PI / (n - 1));
-    const y = Float64Array.from(x, (xi, i) => Math.sin(xi) + (((i * 7 + 3) % 17) / 17 - 0.5) * 0.6);
-
-    const model = new Lowess({ boundary_policy: "reflect" });
-    const result = model.fit(x, y);
-    ```
-
 === "C++"
     ```cpp
     #include <fastlowess.hpp>
@@ -144,18 +120,6 @@ Pads with zeros beyond both endpoints. Appropriate when the underlying process i
     model = fl.Lowess(boundary_policy="zero")
     result = model.fit(x, y)
     ```
-=== "WebAssembly"
-    ```javascript
-    const { Lowess } = require('./fastlowess_wasm.js');
-
-    const n = 100;
-    const x = Float64Array.from({ length: n }, (_, i) => i * 2 * Math.PI / (n - 1));
-    const y = Float64Array.from(x, (xi, i) => Math.sin(xi) + (((i * 7 + 3) % 17) / 17 - 0.5) * 0.6);
-
-    const model = new Lowess({ boundary_policy: "zero" });
-    const result = model.fit(x, y);
-    ```
-
 === "C++"
     ```cpp
     #include <fastlowess.hpp>
@@ -201,18 +165,6 @@ Applies no padding. Each local fit uses only the points that are actually availa
     model = fl.Lowess(boundary_policy="noboundary")
     result = model.fit(x, y)
     ```
-=== "WebAssembly"
-    ```javascript
-    const { Lowess } = require('./fastlowess_wasm.js');
-
-    const n = 100;
-    const x = Float64Array.from({ length: n }, (_, i) => i * 2 * Math.PI / (n - 1));
-    const y = Float64Array.from(x, (xi, i) => Math.sin(xi) + (((i * 7 + 3) % 17) / 17 - 0.5) * 0.6);
-
-    const model = new Lowess({ boundary_policy: "noboundary" });
-    const result = model.fit(x, y);
-    ```
-
 === "C++"
     ```cpp
     #include <fastlowess.hpp>
