### 2026-08-23 cf0f9b1
docs: migrate cpp docs
