### 2026-08-23 c6aa106
docs: migrate python docs
