+++ bindings/python/docs/user-guide/use-cases/time-series.md
@@ -1,0 +1,215 @@
+<!-- markdownlint-disable MD024 MD046 MD033 MD037 -->
+# Time Series Analysis
+
+LOWESS for trend extraction and temporal smoothing.
+
+## Overview
+
+Time series data often contains noise, seasonality, and trends. LOWESS provides flexible trend extraction without parametric assumptions.
+
+---
+
+## Basic Trend Extraction
+
+`fraction = 0.1` sizes the neighbourhood as 10% of the data at each evaluation point — narrow enough to follow a slowly varying trend without smearing periodic variation. Three robustness `iterations` down-weight noise spikes so they cannot bias the fitted curve; this is especially important when the signal-to-noise ratio is low or when occasional outliers are expected.
+
+```python
+import fastlowess as fl
+import numpy as np
+import matplotlib.pyplot as plt
+
+# Simulate noisy time series with trend
+np.random.seed(42)
+t = np.linspace(0, 100, 500)
+trend = 10 + 0.5 * t + 3 * np.sin(t / 10)
+noise = np.random.normal(0, 3, len(t))
+y = trend + noise
+
+# Extract trend with LOWESS
+model = fl.Lowess(fraction=0.1, iterations=3)
+result = model.fit(t, y)
+
+# Plot
+plt.figure(figsize=(12, 5))
+plt.plot(t, y, "gray", alpha=0.5, label="Observed")
+plt.plot(t, result.y, "b-", linewidth=2, label="Trend (LOWESS)")
+plt.xlabel("Time")
+plt.ylabel("Value")
+plt.legend()
+plt.title("Trend Extraction")
+plt.show()
+```
+
+---
+
+## Detrending
+
+Remove trend to analyze residual patterns.
+
+Setting `return_residuals = True` stores `observed − smoothed` alongside the smooth. A slightly wider `fraction = 0.3` produces a smoother baseline trend, so short-duration oscillations end up in the residuals rather than being absorbed into the trend component. The residual series is then ready for spectral analysis, seasonality detection, or change-point methods.
+
+```python
+import fastlowess as fl
+import numpy as np
+import matplotlib.pyplot as plt
+
+np.random.seed(42)
+t = np.linspace(0, 100, 500)
+trend_true = 10 + 0.5 * t + 3 * np.sin(t / 10)
+y = trend_true + np.random.normal(0, 3, len(t))
+
+# Smooth to get trend
+model = fl.Lowess(fraction=0.3, iterations=3, return_residuals=True)
+result = model.fit(t, y)
+
+trend = result.y
+detrended = result.residuals
+
+# Analyze residuals for seasonality, etc.
+plt.figure(figsize=(12, 4))
+plt.subplot(1, 2, 1)
+plt.plot(t, trend)
+plt.title("Extracted Trend")
+
+plt.subplot(1, 2, 2)
+plt.plot(t, detrended)
+plt.title("Detrended (Residuals)")
+plt.tight_layout()
+```
+
+---
+
+## Forecasting with Prediction Intervals
+
+Prediction intervals widen the uncertainty band to include both the uncertainty in the fitted curve (confidence interval) and the expected scatter of new observations around it. `fraction = 0.2` offers a balance between local detail and stable interval width — too small a fraction produces jagged interval edges; too large a fraction underestimates local variance near turning points.
+
+```python
+import fastlowess as fl
+import numpy as np
+import matplotlib.pyplot as plt
+
+np.random.seed(42)
[…]
