+++ bindings/cpp/README.md
@@ -45,17 +45,17 @@ The fastest, most robust, and most feature-complete language-agnostic LOWESS (Lo
 
 > [!NOTE]
 >
-> Currently available for R, Python, Rust, Julia, Node.js, WebAssembly, and C++. See the [Installation Guide](https://lowess.readthedocs.io/en/latest/getting-started/installation/) for detailed installation instructions.
+> Currently available for R, Python, Rust, Julia, Node.js, WebAssembly, and C++. See the [Installation Guide](https://thisisamirv.github.io/lowess-project/cpp/) for detailed installation instructions.
 
 ### GPU Backend
 
-In addition to `parallel = true` (multi-core CPU), the batch `Lowess` class in every binding — except WebAssembly — as well as the `fastLowess` Rust crate itself, can run on the GPU via `wgpu` (Vulkan/Metal/DX12). It's opt-in and worth enabling for high-throughput processing of large datasets (roughly 10k+ points); for smaller inputs the CPU backend is typically faster. `StreamingLowess`/`OnlineLowess` remain CPU-only. See the [GPU Backend guide](https://lowess.readthedocs.io/en/latest/user-guide/gpu-backend/) for installation instructions and usage.
+In addition to `parallel = true` (multi-core CPU), the batch `Lowess` class in every binding — except WebAssembly — as well as the `fastLowess` Rust crate itself, can run on the GPU via `wgpu` (Vulkan/Metal/DX12). It's opt-in and worth enabling for high-throughput processing of large datasets (roughly 10k+ points); for smaller inputs the CPU backend is typically faster. `StreamingLowess`/`OnlineLowess` remain CPU-only. See the [GPU Backend guide](https://thisisamirv.github.io/lowess-project/cpp/) for installation instructions and usage.
 
 ## Documentation
 
 > [!NOTE]
 >
-> ### 📚 [View the full documentation](https://lowess.readthedocs.io/)
+> ### 📚 [View the full documentation](https://thisisamirv.github.io/lowess-project/cpp/)
 
 ---
 
@@ -79,7 +79,7 @@ In addition to `parallel = true` (multi-core CPU), the batch `Lowess` class in e
 
 The `lowess` project beats the competition in terms of speed, whether in single-threaded or multi-threaded parallel execution. It is on average **200-327x faster** than Python's `statsmodels.lowess` and **2-3x faster** than R's `lowess`.
 
-For more details on the performance comparison, see the [Benchmarks](https://lowess.readthedocs.io/en/latest/benchmarks/) page.
+For more details on the performance comparison, see the [Benchmarks](https://thisisamirv.github.io/lowess-project/cpp/) page.
 
 ### Robustness
 
