+++ cpp/search/pages_4.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['genomics_0',['use-case-genomics',['../md_docs_2use-case-genomics.html',1,'']]],
+  ['gpu_20backend_1',['gpu-backend',['../md_docs_2gpu-backend.html',1,'']]]
+];
