+++ cpp/search/all_b.js
@@ -1,0 +1,25 @@
+var searchData=
+[
+  ['macos_20x64_0',['Pre-built Binaries (macOS (x64))',['../md_docs_2installation.html#autotoc_md126',1,'']]],
+  ['mad_20—_20median_20absolute_20deviation_20default_1',['MAD — Median Absolute Deviation (Default)',['../md_docs_2scaling.html#autotoc_md266',1,'']]],
+  ['mar_20—_20median_20absolute_20residual_2',['MAR — Median Absolute Residual',['../md_docs_2scaling.html#autotoc_md268',1,'']]],
+  ['mean_20—_20mean_20absolute_20residual_3',['Mean — Mean Absolute Residual',['../md_docs_2scaling.html#autotoc_md270',1,'']]],
+  ['measurement_20uncertainty_4',['Propagate Measurement Uncertainty',['../md_docs_2custom-weights.html#autotoc_md106',1,'']]],
+  ['median_20absolute_20deviation_20default_5',['MAD — Median Absolute Deviation (Default)',['../md_docs_2scaling.html#autotoc_md266',1,'']]],
+  ['median_20absolute_20residual_6',['MAR — Median Absolute Residual',['../md_docs_2scaling.html#autotoc_md268',1,'']]],
+  ['merge_7',['merge',['../md_docs_2merge.html',1,'']]],
+  ['merge_20strategies_8',['merge strategies',['../md_docs_2streaming.html#autotoc_md276',1,'Merge Strategies'],['../md_docs_2merge.html#autotoc_md166',1,'Merge Strategies']]],
+  ['merge_2emd_9',['merge.md',['../merge_8md.html',1,'']]],
+  ['merge_5fstrategy_10',['merge_strategy',['../md_docs_2api-streaming.html#autotoc_md20',1,'merge_strategy'],['../md_docs_2api.html#autotoc_md42',1,'merge_strategy'],['../md_docs_2parameters.html#autotoc_md229',1,'merge_strategy']]],
+  ['method_11',['Choosing a Scaling Method',['../md_docs_2scaling.html#autotoc_md272',1,'']]],
+  ['methods_12',['methods',['../md_docs_2parameters.html#autotoc_md221',1,'CV Methods'],['../md_docs_2robustness.html#autotoc_md249',1,'Robustness Methods'],['../md_docs_2scaling.html#autotoc_md263',1,'Scaling Methods']]],
+  ['methylation_20profile_20smoothing_13',['Methylation Profile Smoothing',['../md_docs_2use-case-genomics.html#autotoc_md282',1,'']]],
+  ['metrics_14',['CV Metrics',['../md_docs_2cross-validation.html#autotoc_md88',1,'']]],
+  ['min_5fpoints_15',['min_points',['../md_docs_2parameters.html#autotoc_md233',1,'']]],
+  ['missing_20data_16',['Handling Missing Data',['../md_docs_2use-case-time-series.html#autotoc_md321',1,'']]],
+  ['mode_17',['mode',['../md_docs_2use-case-real-time.html#autotoc_md306',1,'Online Mode'],['../md_docs_2quickstart.html#autotoc_md243',1,'Streaming Mode'],['../md_docs_2use-case-real-time.html#autotoc_md307',1,'Streaming Mode']]],
+  ['mode_3a_20chunk_20processing_18',['Streaming Mode: Chunk Processing',['../md_docs_2use-case-real-time.html#autotoc_md300',1,'']]],
+  ['mode_3a_20point_20by_20point_19',['Online Mode: Point-by-Point',['../md_docs_2use-case-real-time.html#autotoc_md297',1,'']]],
[…]
