+++ cpp/search/all_18.js
@@ -1,0 +1,7 @@
+var searchData=
+[
+  ['—_20c_20api_20reference_0',['— c api reference',['../md_docs_2api-online.html',1,'OnlineLowess — C++ API Reference'],['../md_docs_2api-streaming.html',1,'StreamingLowess — C++ API Reference']]],
+  ['—_20mean_20absolute_20residual_1',['Mean — Mean Absolute Residual',['../md_docs_2scaling.html#autotoc_md270',1,'']]],
+  ['—_20median_20absolute_20deviation_20default_2',['MAD — Median Absolute Deviation (Default)',['../md_docs_2scaling.html#autotoc_md266',1,'']]],
+  ['—_20median_20absolute_20residual_3',['MAR — Median Absolute Residual',['../md_docs_2scaling.html#autotoc_md268',1,'']]]
+];
