+++ cpp/search/all_5.js
@@ -1,0 +1,26 @@
+var searchData=
+[
+  ['fastlowess_0',['fastLowess',['../index.html',1,'']]],
+  ['fastlowess_20c_20api_20reference_1',['fastLowess C++ API Reference',['../md_docs_2api.html',1,'']]],
+  ['fastlowess_3a_3adiagnostics_20tt_2',['&lt;tt&gt;fastlowess::Diagnostics&lt;/tt&gt;',['../md_docs_2api.html#autotoc_md35',1,'']]],
+  ['fastlowess_3a_3alowess_20tt_3',['&lt;tt&gt;fastlowess::Lowess&lt;/tt&gt;',['../md_docs_2api.html#autotoc_md23',1,'']]],
+  ['fastlowess_3a_3alowessresult_20tt_4',['fastlowess::lowessresult tt',['../md_docs_2api-streaming.html#autotoc_md16',1,'&lt;tt&gt;fastlowess::LowessResult&lt;/tt&gt;'],['../md_docs_2api.html#autotoc_md34',1,'&lt;tt&gt;fastlowess::LowessResult&lt;/tt&gt;']]],
+  ['fastlowess_3a_3aonlinelowess_20tt_5',['&lt;tt&gt;fastlowess::OnlineLowess&lt;/tt&gt;',['../md_docs_2api-online.html#autotoc_md5',1,'']]],
+  ['fastlowess_3a_3aonlineoutput_20tt_6',['&lt;tt&gt;fastlowess::OnlineOutput&lt;/tt&gt;',['../md_docs_2api-online.html#autotoc_md9',1,'']]],
+  ['fastlowess_3a_3astreaminglowess_20tt_7',['&lt;tt&gt;fastlowess::StreamingLowess&lt;/tt&gt;',['../md_docs_2api-streaming.html#autotoc_md14',1,'']]],
+  ['feature_20comparison_8',['feature comparison',['../md_docs_2gpu-backend.html#autotoc_md116',1,'Feature Comparison'],['../md_docs_2api.html#autotoc_md30',1,'Feature Comparison']]],
+  ['features_9',['features',['../md_docs_2gpu-backend.html#autotoc_md115',1,'Supported Features'],['../md_docs_2api.html#autotoc_md29',1,'Supported Features']]],
+  ['file_20processing_10',['Log File Processing',['../md_docs_2use-case-real-time.html#autotoc_md301',1,'']]],
+  ['first_11',['Take First',['../md_docs_2merge.html#autotoc_md171',1,'']]],
+  ['fold_20cross_20validation_12',['K-Fold Cross-Validation',['../md_docs_2cross-validation.html#autotoc_md80',1,'']]],
+  ['for_20genomic_20data_13',['Best Practices for Genomic Data',['../md_docs_2use-case-genomics.html#autotoc_md291',1,'']]],
+  ['for_20time_20series_14',['Choosing Fraction for Time Series',['../md_docs_2use-case-time-series.html#autotoc_md327',1,'']]],
+  ['forecasting_20with_20prediction_20intervals_15',['Forecasting with Prediction Intervals',['../md_docs_2use-case-time-series.html#autotoc_md319',1,'']]],
+  ['forge_16',['From conda-forge',['../md_docs_2installation.html#autotoc_md129',1,'']]],
+  ['fraction_17',['fraction',['../md_docs_2parameters.html#autotoc_md190',1,'']]],
+  ['fraction_20for_20time_20series_18',['Choosing Fraction for Time Series',['../md_docs_2use-case-time-series.html#autotoc_md327',1,'']]],
+  ['fraction_20parameter_19',['The Fraction Parameter',['../index.html#autotoc_md66',1,'']]],
[…]

[…]
