+++ julia/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"category":"section","location":"intervals/#Intervals","page":"Intervals","text":"Confidence and prediction intervals for uncertainty quantification.","title":"Intervals"},{"category":"section","location":"intervals/#Overview","page":"Intervals","text":"(Image: Confidence and Prediction Intervals)\n\nnote: Adapter support\nConfidence and prediction intervals are available in Batch mode only. Streaming and Online modes do not support intervals.\n\nType Represents Width Use\nConfidence Uncertainty in mean curve Narrow Where is the true trend?\nPrediction Uncertainty for new points Wide Where will new data fall?\n\n","title":"Overview"},{"category":"section","location":"intervals/#Confidence-Intervals","page":"Intervals","text":"Estimate uncertainty in the smoothed curve itself.\n\nusing FastLOWESS\nusing Random, Statistics\n\nrng = MersenneTwister(42)\nx = collect(range(0, 2π, length=100))\ny = sin.(x) .+ randn(rng, 100) .* 0.3\n\nusing FastLOWESS\n\nmodel = Lowess(; fraction=0.5, confidence_intervals=0.95)\nresult = fit(model, x, y)\n\nfor i in 1:length(result.y)\n    println(\"x=$(result.x[i]): y=$(result.y[i]) [$(result.confidence_lower[i]), $(result.confidence_upper[i])]\")\nend\n\n","title":"Confidence Intervals"},{"category":"section","location":"intervals/#Prediction-Intervals","page":"Intervals","text":"Estimate where new observations might fall.\n\nusing FastLOWESS\nusing Random, Statistics\n\nrng = MersenneTwister(42)\nx = collect(range(0, 2π, length=100))\ny = sin.(x) .+ randn(rng, 100) .* 0.3\n\nmodel = Lowess(; fraction=0.5, prediction_intervals=0.95)\nresult = fit(model, x, y)\n\nprintln(\"Prediction bounds: [$(result.prediction_lower[1]), $(result.prediction_upper[1])]\")\n\n","title":"Prediction Intervals"},{"category":"section","location":"intervals/#Both-Intervals","page":"Intervals","text":"Request both types simultaneously:\n\nusing FastLOWESS\nusing Random, Statistics\n\nrng = MersenneTwister(42)\nx = collect(range(0, 2π, length=100))\ny = sin.(x) .+ randn(rng, 100) .* 0.3\n\nresult = fit(\n    Lowess(fraction=0.5, confidence_intervals=0.95, prediction_intervals=0.95),\n    x, y\n)\n\n","title":"Both Intervals"},{"category":"section","location":"intervals/#Confidence-Levels","page":"Intervals","text":"Common levels and their z-values:\n\nLevel z-value Interpretation\n0.90 1.645 90% of intervals contain true value\n0.95 1.960 95% of intervals contain true value\n0.99 2.576 99% of intervals contain true value\n\nusing FastLOWESS\nusing Random, Statistics\n\nrng = MersenneTwister(42)\nx = collect(range(0, 2π, length=100))\ny = sin.(x) .+ randn(rng, 100) .* 0.3\n\n# 99% confidence interval\nmodel = Lowess(; confidence_intervals=0.99)\nresult = fit(model, x, y)\n\n","title":"Confidence Levels"},{"category":"section","location":"intervals/#Standard-Errors","page":"Intervals","text":"Access standard errors directly (available when intervals are computed):\n\nusing FastLOWESS\nusing Random, Statistics\n\nrng = MersenneTwister(42)\nx = collect(range(0, 2π, length=100))\ny = sin.(x) .+ randn(rng, 100) .* 0.3\n\nmodel = Lowess(; confidence_interval
[…]
