### 2026-08-23 1a9ee84
ci: merge docs workflows
