+++ .github/workflows/docs-nodejs.yml
@@ -1,42 +1,0 @@
-name: Docs — Node.js (TypeDoc)
-
-on:
-  push:
-    branches: [main]
-    paths:
-      - "bindings/nodejs/**"
-      - ".github/workflows/docs-nodejs.yml"
-  workflow_dispatch:
-
-concurrency:
-  group: docs-nodejs-${{ github.ref }}
-  cancel-in-progress: true
-
-permissions:
-  contents: write
-
-jobs:
-  typedoc:
-    name: Build & deploy TypeDoc site
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v7
-
-      - uses: actions/setup-node@v7
-        with:
-          node-version: "22"
-
-      - name: Install dependencies
-        run: npm install
-        working-directory: bindings/nodejs
-
-      - name: Build docs
-        run: npm run docs
-        working-directory: bindings/nodejs
-
-      - uses: JamesIves/github-pages-deploy-action@v4
-        with:
-          branch: gh-pages
-          folder: bindings/nodejs/typedoc-out
-          target-folder: nodejs
-          clean: false
