+++ .github/workflows/docs-cpp.yml
@@ -1,37 +1,0 @@
-name: Docs — C++ (Doxygen)
-
-on:
-  push:
-    branches: [main]
-    paths:
-      - "bindings/cpp/**"
-      - ".github/workflows/docs-cpp.yml"
-  workflow_dispatch:
-
-concurrency:
-  group: docs-cpp-${{ github.ref }}
-  cancel-in-progress: true
-
-permissions:
-  contents: write
-
-jobs:
-  doxygen:
-    name: Build & deploy Doxygen site
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v7
-
-      - name: Install Doxygen
-        run: sudo apt-get install -y doxygen
-
-      - name: Build docs
-        run: doxygen Doxyfile
-        working-directory: bindings/cpp
-
-      - uses: JamesIves/github-pages-deploy-action@v4
-        with:
-          branch: gh-pages
-          folder: bindings/cpp/html
-          target-folder: cpp
-          clean: false
