+++ .github/workflows/docs-wasm.yml
@@ -1,60 +1,0 @@
-name: Docs — WebAssembly (TypeDoc)
-
-on:
-  push:
-    branches: [main]
-    paths:
-      - "bindings/wasm/**"
-      - ".github/workflows/docs-wasm.yml"
-  workflow_dispatch:
-
-concurrency:
-  group: docs-wasm-${{ github.ref }}
-  cancel-in-progress: true
-
-env:
-  FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true
-
-permissions:
-  contents: write
-
-jobs:
-  typedoc:
-    name: Build & deploy TypeDoc site
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v7
-
-      - uses: dtolnay/rust-toolchain@stable
-        with:
-          targets: wasm32-unknown-unknown
-
-      - uses: Swatinem/rust-cache@v2
-        with:
-          key: docs-wasm
-
-      - name: Install wasm-pack
-        run: curl https://rustwasm.github.io/wasm-pack/installer/init.sh -sSf | sh
-
-      - uses: actions/setup-node@v7
-        with:
-          node-version: "22"
-
-      - name: Build WASM
-        run: wasm-pack build --target bundler --release
-        working-directory: bindings/wasm
-
-      - name: Install dependencies
-        run: npm install
-        working-directory: bindings/wasm
-
-      - name: Build docs
-        run: npm run docs
-        working-directory: bindings/wasm
-
-      - uses: JamesIves/github-pages-deploy-action@v4
-        with:
-          branch: gh-pages
-          folder: bindings/wasm/typedoc-out
-          target-folder: wasm
-          clean: false
