+++ .github/workflows/docs-r.yml
@@ -1,56 +1,0 @@
-name: Docs — R (pkgdown)
-
-on:
-  push:
-    branches: [main]
-    paths:
-      - "bindings/r/**"
-      - ".github/workflows/docs-r.yml"
-  workflow_dispatch:
-
-concurrency:
-  group: docs-r-${{ github.ref }}
-  cancel-in-progress: true
-
-env:
-  GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-
-permissions:
-  contents: write
-
-jobs:
-  pkgdown:
-    name: Build & deploy pkgdown site
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v7
-
-      - uses: dtolnay/rust-toolchain@stable
-
-      - uses: Swatinem/rust-cache@v2
-        with:
-          key: docs-r-pkgdown
-
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          use-public-rspm: true
-
-      - uses: r-lib/actions/setup-pandoc@v2
-
-      - uses: r-lib/actions/setup-r-dependencies@v2
-        with:
-          working-directory: bindings/r
-          extra-packages: any::pkgdown any::devtools
-
-      - uses: ./.github/actions/prepare-r-source
-
-      - name: Build pkgdown site
-        shell: Rscript {0}
-        run: pkgdown::build_site_github_pages(pkg = "bindings/r", new_process = FALSE, install = TRUE)
-
-      - uses: JamesIves/github-pages-deploy-action@v4
-        with:
-          branch: gh-pages
-          folder: bindings/r/docs
-          target-folder: r
-          clean: false

[…]
