### 2026-08-23 33b893b
fix: r_ver and grepv errors
