### 2026-08-23 5f12a8e
fix: rust snippets
