### 2026-08-23 580daa1
fix: typedoc plugin linker, r snippets
