### 2026-08-23 7778b25
fix(python): enforce keyword-only arguments beyond the first positional arg
