### 2026-08-23 947c661
docs: update contributing guide
