### 2026-08-23 7a134ad
docs: fix R chunk labels
