+++ bindings/r/vignettes/batch.Rmd
@@ -37,7 +37,7 @@ The default mode. Processes the entire dataset at once and supports all features
 
 ## Example
 
-```{r}
+```{r batch_1}
 library(rfastlowess)
 set.seed(42)
 x <- seq(0, 2 * pi, length.out = 100)
