+++ bindings/r/vignettes/quickstart.Rmd
@@ -15,7 +15,7 @@ knitr::opts_chunk$set(collapse = TRUE, comment = "#>", eval = FALSE)
 
 Smooth a noisy sine wave. `fraction = 0.3` and `iterations = 3` are good starting values for most signals.
 
-```{r}
+```{r quickstart_1}
 library(rfastlowess)
 
 # 100-point noisy sine wave
@@ -36,7 +36,7 @@ cat(sprintf("First smoothed value: %.4f (true: %.4f)\n",
 
 Set `confidence_intervals` and/or `prediction_intervals` to a coverage level (0–1).
 
-```{r}
+```{r quickstart_2}
 library(rfastlowess)
 set.seed(42)
 x <- seq(0, 2 * pi, length.out = 100)
@@ -62,7 +62,7 @@ print(result$diagnostics$r_squared)
 
 LOWESS can robustly handle outliers through iterative reweighting:
 
-```{r}
+```{r quickstart_3}
 library(rfastlowess)
 
 x_out <- seq(1, 6, length.out = 6)
@@ -89,7 +89,7 @@ for (i in seq_along(result$robustness_weights)) {
 
 For large datasets (>100K points) that may not fit in memory.
 
-```{r}
+```{r quickstart_4}
 library(rfastlowess)
 set.seed(42)
 x <- seq(0, 2 * pi, length.out = 10000)
@@ -118,7 +118,7 @@ final <- finalize(model)
 
 For real-time / point-by-point processing.
 
-```{r}
+```{r quickstart_5}
 library(rfastlowess)
 set.seed(42)
 times <- 1:100
@@ -142,7 +142,7 @@ for (i in seq_along(times)) {
 
 ## Plotting Results
 
-```{r}
+```{r quickstart_6}
 library(rfastlowess)
 set.seed(42)
 x <- seq(0, 2 * pi, length.out = 100)
