+++ bindings/r/vignettes/boundary.Rmd
@@ -32,7 +32,7 @@ Pads beyond both endpoints by replicating the first and last observed values. Pr
 
 **Use when**: No strong prior on boundary behaviour; general-purpose smoothing.
 
-```{r}
+```{r boundary_1}
 library(rfastlowess)
 set.seed(42)
 x <- seq(0, 2 * pi, length.out = 100)
@@ -50,7 +50,7 @@ Mirrors data at both endpoints. Prevents inflection artifacts when the signal is
 
 **Use when**: Signal is periodic (e.g. time-of-day) or the boundary is a known symmetry point.
 
-```{r}
+```{r boundary_2}
 model <- Lowess(boundary_policy = "reflect")
 result <- fit(model, x, y)
 ```
@@ -63,7 +63,7 @@ Pads with zeros beyond the endpoints.
 
 **Use when**: The signal is known to decay to zero at the boundaries (e.g. impulse responses, genomic signals at chromosome ends).
 
-```{r}
+```{r boundary_3}
 model <- Lowess(boundary_policy = "zero")
 result <- fit(model, x, y)
 ```
@@ -78,7 +78,7 @@ Reproduces Cleveland's original behaviour — no padding applied.
 
 > **Note:** Without padding, boundary fits can have higher variance and visible edge artefacts, particularly with small `fraction` values.
 
-```{r}
+```{r boundary_4}
 model <- Lowess(boundary_policy = "noboundary")
 result <- fit(model, x, y)
 ```
@@ -87,7 +87,7 @@ result <- fit(model, x, y)
 
 ## Comparing Policies
 
-```{r}
+```{r boundary_5}
 library(rfastlowess)
 set.seed(42)
 x <- seq(0, 2 * pi, length.out = 100)
