### 2026-08-23 4064c08
docs: added eval=FALSE to R installation snippets
