### 2026-08-23 3a68809
docs: add sessionInfo() to r vignettes
