+++ bindings/r/vignettes/benchmarks.Rmd
@@ -76,3 +76,7 @@ mb <- microbenchmark(
 )
 print(mb)
 ```
+
+```{r session-info}
+sessionInfo()
+```
