### 2026-08-23 7e670bd
docs: eval=FALSE individually to r vignettes snippets where appropriate
