+++ bindings/r/vignettes/batch.Rmd
@@ -8,7 +8,7 @@ vignette: >
 ---
 
 ```{r setup, include=FALSE}
-knitr::opts_chunk$set(collapse = TRUE, comment = "#>", eval = FALSE)
+knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 ```
 
 > **See also:** [Choosing an Execution Mode](adapter-choice.html) for a comparison of all three modes.
