### 2026-08-23 04603f9
docs: wrap long lines in R docs
