+++ bindings/r/vignettes/use-case-genomics.Rmd
@@ -13,13 +13,19 @@ knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 
 ## Overview
 
-LOWESS is well-suited for genomic data such as DNA methylation profiles, ChIP-seq signals, and expression arrays. Sequencing depth variation, PCR artifacts, and biological heterogeneity all create noise that LOWESS handles gracefully without parametric assumptions.
+LOWESS is well-suited for genomic data such as DNA methylation profiles,
+ChIP-seq signals, and expression arrays. Sequencing depth variation, PCR
+artifacts, and biological heterogeneity all create noise that LOWESS handles
+gracefully without parametric assumptions.
 
 ---
 
 ## Methylation Profile Smoothing
 
-A small `fraction = 0.1` lets LOWESS follow fine-scale spatial structure without smearing transitions between methylated and unmethylated regions. `confidence_intervals = 0.95` produces uncertainty bands that naturally widen at positions with sparser CpG coverage.
+A small `fraction = 0.1` lets LOWESS follow fine-scale spatial structure without
+smearing transitions between methylated and unmethylated regions.
+`confidence_intervals = 0.95` produces uncertainty bands that naturally widen at
+positions with sparser CpG coverage.
 
 ```{r use_case_genomics_1}
 library(rfastlowess)
@@ -60,9 +66,15 @@ legend("topright", c("Observed", "Smoothed", "95% CI"),
 
 ## ChIP-seq Signal Smoothing
 
-ChIP-seq experiments produce sparse, noisy coverage data. LOWESS can help identify binding regions.
+ChIP-seq experiments produce sparse, noisy coverage data. LOWESS can help
+identify binding regions.
 
-`fraction = 0.05` provides high spatial resolution — important for resolving narrow binding peaks that would otherwise be smeared into the background. The larger `iterations = 5` is deliberate: Poisson-distributed read counts produce tall, isolated spikes, and extra robustness iterations progressively down-weight them so the estimated background level is not inflated by a handful of extreme counts.
+`fraction = 0.05` provides high spatial resolution — important for resolving
+narrow binding peaks that would otherwise be smeared into the background. The
+larger `iterations = 5` is deliberate: Poisson-distributed read counts produce
+tall, isolated spikes, and extra robustness iterations progressively down-weight
+them so the estimated background level is not inflated by a handful of extreme
+counts.
 
 ```{r use_case_genomics_2}
 library(rfastlowess)
