+++ bindings/r/vignettes/intervals.Rmd
@@ -15,13 +15,14 @@ knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 
 ![Confidence and prediction intervals](../man/figures/intervals_comparison.svg)
 
-> **Note:** Confidence and prediction intervals are available in **Batch** mode only.
+> **Note:** Confidence and prediction intervals are available in **Batch** mode
+> only.
 > Streaming and Online modes do not support intervals.
 
 | Type | Represents | Width | Use |
 | --- | --- | --- | --- |
-| **Confidence** | Uncertainty in mean curve | Narrow | Where is the true trend? |
-| **Prediction** | Uncertainty for new points | Wide | Where will new data fall? |
+| **Confidence** | Mean curve uncertainty | Narrow | Where is the true trend? |
+| **Prediction** | New-point uncertainty | Wide | Where will new data fall? |
 
 ---
 

[…]
