### 2026-08-23 08bb91d
docs: fix indentation in R docs
