+++ bindings/r/vignettes/robustness.Rmd
@@ -106,8 +106,8 @@ plot(x, y, pch = 16, col = "gray", main = "Effect of Robustness Iterations")
 lines(result_0$x, result_0$y, col = "red", lwd = 2, lty = 2)
 lines(result_3$x, result_3$y, col = "blue", lwd = 2)
 legend("topright", c("Data", "iterations=0", "iterations=3"),
-       pch = c(16, NA, NA), lty = c(NA, 2, 1),
-       col = c("gray", "red", "blue"))
+        pch = c(16, NA, NA), lty = c(NA, 2, 1),
+        col = c("gray", "red", "blue"))
 ```
 
 ---
