+++ bindings/r/vignettes/custom-weights.Rmd
@@ -94,7 +94,7 @@ model <- Lowess(fraction = 0.3)
 result <- fit(model, x, y, custom_weights = weights)
 
 plot(x, y, pch = 16, cex = 0.5 + weights / max(weights),
-     col = "gray", main = "Inverse-Variance Weighted LOWESS")
+    col = "gray", main = "Inverse-Variance Weighted LOWESS")
 lines(result$x, result$y, col = "blue", lwd = 2)
 ```
 
