+++ r/articles/custom-weights.md
@@ -11,7 +11,8 @@ The effective weight of observation $`j`$ in a local fit centred at
 $`x_i`$ is:
 
 ``` math
-w_{ij} = \text{custom\_weights}[j] \times K\!\left(\frac{d_{ij}}{h_i}\right) \times r_j
+w_{ij} = \text{custom\_weights}[j] \times K\!\left(\frac{d_{ij}}{h_i}\right)
+\times r_j
 ```
 
 where $`K`$ is the distance kernel, $`h_i`$ is the local bandwidth, and
@@ -39,7 +40,7 @@ $`r_j`$ is the robustness weight from the current iteration.
 | **When known** | Before fitting | Computed from residuals |
 | **Knowledge required** | Prior knowledge of quality | None — data-driven |
 | **Effect** | Fixed throughout fit | Adapts each iteration |
-| **Use case** | Known bad sensors, calibration | Unknown outlier contamination |
+| **Use case** | Bad sensors, calibration | Unknown outlier contamination |
 
 They compose: use both simultaneously. Custom weights suppress known bad
 points; robustness iterations handle residual outliers.
@@ -64,6 +65,8 @@ weights[6] <- 0.0          # exclude the spike
 
 model <- Lowess(fraction = 0.5, iterations = 0L)
 result <- fit(model, x, y, custom_weights = weights)
+print(head(result$y))
+#> [1]  2.572621  4.000000  6.000000  8.000000 10.000000 12.000000
 ```
 
 ------------------------------------------------------------------------
@@ -87,9 +90,11 @@ model <- Lowess(fraction = 0.3)
 result <- fit(model, x, y, custom_weights = weights)
 
 plot(x, y, pch = 16, cex = 0.5 + weights / max(weights),
-     col = "gray", main = "Inverse-Variance Weighted LOWESS")
+    col = "gray", main = "Inverse-Variance Weighted LOWESS")
 lines(result$x, result$y, col = "blue", lwd = 2)
 ```
+
+![](custom-weights_files/figure-html/custom_weights_2-1.png)
 
 ------------------------------------------------------------------------
 
@@ -109,4 +114,46 @@ weights[40:50] <- 0.1
 # Also use robustness for unknown outliers
 model <- Lowess(fraction = 0.3, iterations = 3)
 result <- fit(model, x, y, custom_weights = weights)
+print(head(result$y))
+#> [1] 0.5716412 0.5904751 0.6108481 0.6322355 0.6533279 0.6729328
+```
+
+``` r
+
+sessionInfo()
+#> R version 4.6.1 (2026-06-24)
+#> Platform: x86_64-pc-linux-gnu
+#> Running under: Ubuntu 24.04.4 LTS
+#> 
+#> Matrix products: default
+#> BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3 
+#> LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.26.so;  LAPACK version 3.12.0
+#> 
+#> locale:
+#>  [1] LC_CTYPE=C.UTF-8       LC_NUMERIC=C           LC_TIME=C.UTF-8       
+#>  [4] LC_COLLATE=C.UTF-8     LC_MONETARY=C.UTF-8    LC_MESSAGES=C.UTF-8   
+#>  [7] LC_PAPER=C.UTF-8       LC_NAME=C              LC_ADDRESS=C          
+#> [10] LC_TELEPHONE=C         LC_MEASUREMENT=C.UTF-8 LC_IDENTIFICATION=C   
+#> 
+#> time zone: UTC
[…]
