### 2026-08-23 0521656
docs: make julia docs dynamic to run
