+++ bindings/julia/julia/docs/src/parameters.md
@@ -62,7 +62,7 @@ The proportion of data used for each local fit. **Most important parameter.**
 | 0.5–0.7 | Heavy smoothing | Noisy data |
 | 0.7–1.0 | Very smooth | Trend extraction |
 
-```julia
+```@example parameters
 using FastLOWESS
 using Random, Statistics
 
@@ -72,6 +72,7 @@ y = sin.(x) .+ randn(rng, 100) .* 0.3
 
 model = Lowess(; fraction=0.3)
 result = fit(model, x, y)
+println("First smoothed value (fraction=0.3): ", result.y[1])
 ```
 
 ---
@@ -87,7 +88,7 @@ Number of robustness iterations for outlier resistance.
 | 4–6 | Strong | Contaminated data |
 | 7+ | Very strong | Heavy outliers |
 
-```julia
+```@example parameters
 using FastLOWESS
 using Random, Statistics
 
@@ -97,6 +98,7 @@ y = sin.(x) .+ randn(rng, 100) .* 0.3
 
 model = Lowess(; iterations=5)
 result = fit(model, x, y)
+println("First smoothed value (iterations=5): ", result.y[1])
 ```
 
 ---
@@ -108,7 +110,7 @@ Interpolation optimization threshold. Points within `delta` distance reuse the p
 - **Default**: 1% of x-range (Batch), 0.0 (Streaming/Online)
 - **Effect**: Higher values = faster but less accurate
 
-```julia
+```@example parameters
 using FastLOWESS
 using Random, Statistics
 
@@ -118,6 +120,7 @@ y = sin.(x) .+ randn(rng, 100) .* 0.3
 
 model = Lowess(; delta=0.05)
 result = fit(model, x, y)
+println("First smoothed value (delta=0.05): ", result.y[1])
 ```
 
 ---
@@ -138,7 +141,7 @@ Distance weighting kernel for local fits.
 
 See [Weight Functions](kernels.md) for detailed comparison.
 
-```julia
+```@example parameters
 using FastLOWESS
 using Random, Statistics
 
@@ -148,6 +151,7 @@ y = sin.(x) .+ randn(rng, 100) .* 0.3
 
 model = Lowess(; weight_function="epanechnikov")
 result = fit(model, x, y)
+println("First smoothed value (epanechnikov kernel): ", result.y[1])
 ```
 
 ---
@@ -164,7 +168,7 @@ Method for downweighting outliers during iterative refinement.
 
 See [Robustness](robustness.md) for detailed comparison.
 
-```julia
+```@example parameters
 using FastLOWESS
 using Random, Statistics
 
@@ -174,6 +178,7 @@ y = sin.(x) .+ randn(rng, 100) .* 0.3
 
 model = Lowess(; robustness_method="talwar")
 result = fit(model, x, y)
+println("First smoothed value (talwar robustness): ", result.y[1])
 ```
 
 ---
@@ -193,7 +198,7 @@ Edge handling strategy to reduce boundary bias. See [Boundary Handling](boundary
 
 For example:
 
-```julia
+```@example parameters
 using FastLOWESS
 using Random, Statistics
 
@@ -203,6 +208,7 @@ y = sin.(x) .+ randn(rng, 100) .* 0.3
 
 model = Lowess(; boundary_policy="reflect")
 result = fit(model, x, y)
+println("First smoothed value (reflect boundary): ", result.y[1])
 ```
 
 ---
@@ -221,7 +227,7 @@ Method for estimating residual scale during robustness iterations. See [Scaling 
 
 For example:
 
-```julia
+```@example parameters
 using FastLOWESS
 using Random, Statistics
 
@@ -231,6 +237,7 @@ y = sin.(x) .+ randn(rng, 100) .* 0.3
 
[…]
