### 2026-08-23 8b567ac
docs: add context to R vignettes print statements
