+++ bindings/r/vignettes/merge.Rmd
@@ -61,6 +61,7 @@ model <- StreamingLowess(
     overlap = 500
 )
 result <- process_chunk(model, x_chunk, y_chunk)
+cat("First 6 smoothed values (weighted_average strategy):\n")
 print(head(result$y))
 ```
 
@@ -80,6 +81,7 @@ model <- StreamingLowess(
     overlap = 500
 )
 result <- process_chunk(model, x_chunk, y_chunk)
+cat("First 6 smoothed values (average strategy):\n")
 print(head(result$y))
 ```
 
@@ -99,6 +101,7 @@ model <- StreamingLowess(
     overlap = 500
 )
 result <- process_chunk(model, x_chunk, y_chunk)
+cat("First 6 smoothed values (take_first strategy):\n")
 print(head(result$y))
 ```
 
@@ -118,6 +121,7 @@ model <- StreamingLowess(
     overlap = 500
 )
 result <- process_chunk(model, x_chunk, y_chunk)
+cat("First 6 smoothed values (take_last strategy):\n")
 print(head(result$y))
 ```
 
