+++ bindings/r/vignettes/benchmarks.Rmd
@@ -77,6 +77,7 @@ mb <- microbenchmark(
     },
     times = 50
 )
+cat("Benchmark results:\n")
 print(mb)
 ```
 
