+++ bindings/r/vignettes/cross-validation.Rmd
@@ -94,6 +94,7 @@ result <- fit(model, x, y)
 
 cat("Best fraction:", result$fraction_used, "\n")
 cat("CV scores:\n")
+cat("CV scores by fraction:\n")
 print(data.frame(fraction = fractions, score = result$cv_scores))
 ```
 
