### 2026-08-24 b990f40
docs: add println statements to rust docs
