+++ crates/fastLowess/src/lib.rs
@@ -38,7 +38,7 @@
 //! // Fit the model to the data
 //! let result = model.fit(&x, &y)?;
 //!
-//! println!("{}", result);
+//! println!("LOWESS result:\n{}", result);
 //! # Result::<(), LowessError>::Ok(())
 //! ```
 //!
@@ -86,7 +86,7 @@
 //!     .build()?;
 //!
 //! let result = model.fit(&x, &y)?;
-//! println!("{}", result);
+//! println!("LOWESS result:\n{}", result);
 //! # Result::<(), LowessError>::Ok(())
 //! ```
 //!
