+++ crates/fastLowess/docs/parameters.md
@@ -77,6 +77,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (fraction=0.3): {}", result.y[0]);
     Ok(())
 }
 ```
@@ -108,6 +109,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (5 iterations): {}", result.y[0]);
     Ok(())
 }
 ```
@@ -135,6 +137,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value: {}", result.y[0]);
     Ok(())
 }
 ```
@@ -171,6 +174,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (epanechnikov kernel): {}", result.y[0]);
     Ok(())
 }
 ```
@@ -203,6 +207,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (talwar robustness): {}", result.y[0]);
     Ok(())
 }
 ```
@@ -238,6 +243,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (reflect boundary): {}", result.y[0]);
     Ok(())
 }
 ```
@@ -272,6 +278,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (mad scaling): {}", result.y[0]);
     Ok(())
 }
 ```
@@ -306,6 +313,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (use_local_mean fallback): {}", result.y[0]);
     Ok(())
 }
 ```
@@ -331,6 +339,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (auto-converge): {}", result.y[0]);
     Ok(())
 }
 ```
@@ -365,6 +374,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (custom weights): {}", result.y[0]);
     Ok(())
 }
 ```
@@ -461,6 +471,9 @@ fn main() -> Result<(), LowessError> {
     let result = model.fit(&x, &y)?;
     // Points with weight < 0.5 are likely outliers
 
+    if let Some(w) = &result.robustness_weights {
+        println!("First robustness weight: {}", w[0]);
+    }
     Ok(())
 }
 ```
@@ -516,6 +529,9 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    if let (Some(lo), Some(hi)) = (&result.confidence_lower, &result.confidence_upper) {
+        println!("First point 95% CI: [{}, {}]", lo[0], hi[0]);
+    }
     Ok(())
 }
 ```
@@ -546,6 +562,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value: {}", result.y[0]);
     Ok(())
 }
 ```
[…]
