+++ crates/lowess/docs/scaling.md
@@ -46,6 +46,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (mad scaling): {}", result.y[0]);
     Ok(())
 }
 ```
@@ -75,6 +76,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (mar scaling): {}", result.y[0]);
     Ok(())
 }
 ```
@@ -104,6 +106,7 @@ fn main() -> Result<(), LowessError> {
         .build()?;
     let result = model.fit(&x, &y)?;
 
+    println!("First smoothed value (mean scaling): {}", result.y[0]);
     Ok(())
 }
 ```
