### 2026-08-24 4c2932a
docs: fix R docs indentation issues
