+++ bindings/r/vignettes/adapter-choice.Rmd
@@ -24,10 +24,8 @@ processing needs:
 
 ## Quick Decision Guide
 
-- **Data fits in memory and you need intervals or CV** → [Batch
-  Mode](batch.html)
-- **Data is too large for memory or arrives in file chunks** → [Streaming
-  Mode](streaming.html)
+- **Data fits in memory and you need intervals or CV** → [Batch Mode](batch.html)
+- **Data is too large for memory or arrives in file chunks** → [Streaming Mode](streaming.html)
 - **Data arrives point-by-point in real time** → [Online Mode](online.html)
 
 ```{r session-info}
