### 2026-08-24 ab97aaf
docs: replace typedoc plugin with VitePress and add outputs
