### 2026-08-24 c86596a
docs: add wasm doc outputs
