+++ bindings/nodejs/docs/index.md
@@ -1,13 +1,0 @@
----
-layout: home
-hero:
-  name: fastlowess
-  tagline: High-performance LOWESS smoothing for Node.js
-  actions:
-    - theme: brand
-      text: Get Started
-      link: /installation
-    - theme: alt
-      text: API Reference
-      link: /reference/
----
