### 2026-08-24 1427967
docs: add output to cpp docs
