+++ bindings/cpp/docs/parameters.md
@@ -80,8 +80,13 @@ int main() {
     fastlowess::Lowess model({ .fraction = 0.3});
     auto result = model.fit(x, y).value();
 
+    std::cout << "y[0]: " << result.y_vector()[0] << "\n";
     return 0;
 }
+```
+
+```output
+y[0]: 0.258244
 ```
 
 ---
@@ -114,8 +119,13 @@ int main() {
     fastlowess::Lowess model({ .iterations = 5});
     auto result = model.fit(x, y).value();
 
+    std::cout << "y[0]: " << result.y_vector()[0] << "\n";
     return 0;
 }
+```
+
+```output
+y[0]: 0.382603
 ```
 
 ---
@@ -144,8 +154,13 @@ int main() {
     fastlowess::Lowess model({ .delta = 0.05});
     auto result = model.fit(x, y).value();
 
+    std::cout << "y[0]: " << result.y_vector()[0] << "\n";
     return 0;
 }
+```
+
+```output
+y[0]: 0.382608
 ```
 
 ---
@@ -183,8 +198,13 @@ int main() {
     fastlowess::Lowess model({ .weight_function = "epanechnikov"});
     auto result = model.fit(x, y).value();
 
+    std::cout << "y[0]: " << result.y_vector()[0] << "\n";
     return 0;
 }
+```
+
+```output
+y[0]: 0.406728
 ```
 
 ---
@@ -218,8 +238,13 @@ int main() {
     fastlowess::Lowess model({ .robustness_method = "talwar"});
     auto result = model.fit(x, y).value();
 
+    std::cout << "y[0]: " << result.y_vector()[0] << "\n";
     return 0;
 }
+```
+
+```output
+y[0]: 0.378779
 ```
 
 ---
@@ -256,8 +281,13 @@ int main() {
     fastlowess::Lowess model({ .boundary_policy = "reflect"});
     auto result = model.fit(x, y).value();
 
+    std::cout << "y[0]: " << result.y_vector()[0] << "\n";
     return 0;
 }
+```
+
+```output
+y[0]: 0.712762
 ```
 
 ---
@@ -293,8 +323,13 @@ int main() {
     fastlowess::Lowess model({ .scaling_method = "mad"});
     auto result = model.fit(x, y).value();
 
+    std::cout << "y[0]: " << result.y_vector()[0] << "\n";
     return 0;
 }
+```
+
+```output
+y[0]: 0.382608
 ```
 
 ---
@@ -330,8 +365,13 @@ int main() {
     fastlowess::Lowess model({ .zero_weight_fallback = "use_local_mean"});
     auto result = model.fit(x, y).value();
 
+    std::cout << "y[0]: " << result.y_vector()[0] << "\n";
     return 0;
 }
+```
+
+```output
+y[0]: 0.382608
 ```
 
 ---
@@ -357,8 +397,13 @@ int main() {
     fastlowess::Lowess model({ .iterations = 20, .auto_converge = 1e-6});
     auto result = model.fit(x, y).value();
 
+    std::cout << "y[0]: " << result.y_vector()[0] << "\n";
     return 0;
 }
+```
+
+```output
+y[0]: 0.382602
 ```
 
 ---
@@ -394,8 +439,13 @@ int main() {
 
     auto result = fastlowess::Lowess(opts).fit(x, y, weights).value();
 
+    std::cout << "y[0]: " << result.y_vector()[0] << "\n";
     return 0;
 }
+```
+
+```output
+y[0]: 0.336935
 ```
 
 ---
@@ -424,8 +474,13 @@ int main() {
     auto result = model.fit(x, y).value();
     auto residuals = result.residuals();
 
+    std::cout << "residuals[0]: " << result.residuals()[0] << "\n";
     return 0;
 }
+```
+
[…]
