### 2026-08-24 7f3ce32
docs: migrate python docs to sphinx
