+++ bindings/python/docs/user-guide/kernels.md
@@ -1,4 +1,3 @@
-<!-- markdownlint-disable MD024 MD033 MD046 -->
 # Weight Functions
 
 Kernel functions for distance weighting.
@@ -35,7 +34,7 @@ $$w(u) = (1 - |u|^3)^3$$
 
 **Use when**: Default choice for most applications.
 
-```python
+:::{jupyter-execute}
 import fastlowess as fl
 import numpy as np
 
@@ -45,7 +44,8 @@ y = np.sin(x) + rng.normal(0, 0.3, 100)
 
 model = fl.Lowess(weight_function="tricube")
 result = model.fit(x, y)
-```
+print(f"Smoothed y[0]: {result.y[0]:.4f}")
+:::
 
 ---
 
@@ -57,7 +57,7 @@ $$w(u) = \frac{3}{4}(1 - u^2)$$
 
 **Use when**: Optimal MSE properties desired.
 
-```python
+:::{jupyter-execute}
 import fastlowess as fl
 import numpy as np
 
@@ -67,7 +67,8 @@ y = np.sin(x) + rng.normal(0, 0.3, 100)
 
 model = fl.Lowess(weight_function="epanechnikov")
 result = model.fit(x, y)
-```
+print(f"Smoothed y[0]: {result.y[0]:.4f}")
+:::
 
 ---
 
@@ -79,7 +80,7 @@ $$w(u) = \exp(-u^2/2)$$
 
 **Use when**: Maximum smoothness needed, computational cost acceptable.
 
-```python
+:::{jupyter-execute}
 import fastlowess as fl
 import numpy as np
 
@@ -89,7 +90,8 @@ y = np.sin(x) + rng.normal(0, 0.3, 100)
 
 model = fl.Lowess(weight_function="gaussian")
 result = model.fit(x, y)
-```
+print(f"Smoothed y[0]: {result.y[0]:.4f}")
+:::
 
 ---
 
@@ -101,7 +103,7 @@ $$w(u) = (1 - u^2)^2$$
 
 **Use when**: Alternative to Tricube with slightly different properties.
 
-```python
+:::{jupyter-execute}
 import fastlowess as fl
 import numpy as np
 
@@ -111,7 +113,8 @@ y = np.sin(x) + rng.normal(0, 0.3, 100)
 
 model = fl.Lowess(weight_function="biweight")
 result = model.fit(x, y)
-```
+print(f"Smoothed y[0]: {result.y[0]:.4f}")
+:::
 
 ---
 
@@ -123,7 +126,7 @@ $$w(u) = \cos(\pi u / 2)$$
 
 **Use when**: Want smooth kernel with simple form.
 
-```python
+:::{jupyter-execute}
 import fastlowess as fl
 import numpy as np
 
@@ -133,7 +136,8 @@ y = np.sin(x) + rng.normal(0, 0.3, 100)
 
 model = fl.Lowess(weight_function="cosine")
 result = model.fit(x, y)
-```
+print(f"Smoothed y[0]: {result.y[0]:.4f}")
+:::
 
 ---
 
@@ -145,7 +149,7 @@ $$w(u) = 1 - |u|$$
 
 **Use when**: Simple, interpretable weights.
 
-```python
+:::{jupyter-execute}
 import fastlowess as fl
 import numpy as np
 
@@ -155,7 +159,8 @@ y = np.sin(x) + rng.normal(0, 0.3, 100)
 
 model = fl.Lowess(weight_function="triangle")
 result = model.fit(x, y)
-```
+print(f"Smoothed y[0]: {result.y[0]:.4f}")
+:::
 
 ---
 
@@ -167,7 +172,7 @@ $$w(u) = 1$$
 
[…]
