### 2026-08-24 3c13502
docs: add output to rust docs
