+++ dev/add-rust-outputs.py
@@ -1,0 +1,215 @@
+#!/usr/bin/env python3
+"""Compile and run each ```rust snippet in crates/fastLowess/docs/ (or fastLoess),
+inject ```output blocks -- mirrors dev/add-cpp-outputs.py for Rust.
+
+Usage:
+    python dev/add-rust-outputs.py [DOCS_DIR]
+
+DOCS_DIR defaults to crates/fastLowess/docs.
+"""
+
+from __future__ import annotations
+
+import os
+import re
+import subprocess
+import sys
+import tempfile
+from pathlib import Path
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+DOCS_DIR = (
+    Path(sys.argv[1]).resolve()
+    if len(sys.argv) > 1
+    else REPO_ROOT / "crates" / "fastLowess" / "docs"
+)
+CRATE_DIR = DOCS_DIR.parent  # fastLowess or fastLoess crate
+TIMEOUT = 30
+
+FENCE_RE = re.compile(r"```rust\n([\s\S]*?)```", re.MULTILINE)
+OUTPUT_RE = re.compile(r"\n\n```output\n[\s\S]*?```")
+
+CARGO_TOML = """\
+[package]
+name = "snippet-runner"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+{crate} = {{ path = "{path}" }}
+{extra_deps}"""
+
+
+def preprocess(code: str) -> str:
+    """Strip Rust doctest hidden-line prefixes so code compiles as a plain file.
+
+    Lines starting with '# ' are hidden in rendered docs but included in
+    compilation — the prefix is doctest syntax, not valid plain Rust.
+    """
+    out = []
+    for line in code.splitlines():
+        if line.startswith("# "):
+            out.append(line[2:])
+        elif line == "#":
+            out.append("")
+        else:
+            out.append(line)
+    return "\n".join(out)
+
+
+def build_and_run_all(
+    snippets: list[tuple[Path, int, str]],
+) -> dict[tuple[Path, int], str | None]:
+    """Compile ALL snippets sharing one Cargo project.
+
+    fastLowess is compiled once (first bin build); subsequent bins reuse the
+    cached rlib so each only needs to compile its small main file.
+    Per-snippet failures are isolated — a broken snippet does not block others.
+    """
+    crate_name = CRATE_DIR.name
+    results: dict[tuple[Path, int], str | None] = {}
+
+    with tempfile.TemporaryDirectory(prefix="rust-snippets-") as tmp:
+        tmp_path = Path(tmp)
+        bin_dir = tmp_path / "src" / "bin"
+        bin_dir.mkdir(parents=True)
+
+        for i, (_md, _start, code) in enumerate(snippets):
+            (bin_dir / f"s{i}.rs").write_text(preprocess(code), encoding="utf-8")
+
+        # Include all sibling crates so snippets can reference any crate in the workspace
+        extra_deps = ""
+        for sibling in sorted(CRATE_DIR.parent.iterdir()):
+            cargo_toml = sibling / "Cargo.toml"
+            if sibling == CRATE_DIR or not cargo_toml.exists():
+                continue
+            name_m = re.search(
+                r'^name\s*=\s*"([^"]+)"',
+                cargo_toml.read_text(encoding="utf-8"),
+                re.MULTILINE,
+            )
+            if name_m:
+                sib_name = name_m.group(1)
+                sib_path = str(sibling).replace("\\", "/")
+                extra_deps += f'{sib_name} = {{ path = "{sib_path}" }}\n'
+
+        (tmp_path / "Cargo.toml").write_text(
+            CARGO_TOML.format(
+                crate=crate_name,
[…]
