+++ crates/fastLowess/docs/parameters.md
@@ -82,6 +82,10 @@ fn main() -> Result<(), LowessError> {
 }
 ```
 
+```output
+First smoothed value (fraction=0.3): 0.2582436164551495
+```
+
 ---
 
 ### iterations
@@ -114,6 +118,10 @@ fn main() -> Result<(), LowessError> {
 }
 ```
 
+```output
+First smoothed value (5 iterations): 0.3826026982388202
+```
+
 ---
 
 ### delta
@@ -140,6 +148,10 @@ fn main() -> Result<(), LowessError> {
     println!("First smoothed value: {}", result.y[0]);
     Ok(())
 }
+```
+
+```output
+First smoothed value: 0.38260776436644134
 ```
 
 ---
@@ -179,6 +191,10 @@ fn main() -> Result<(), LowessError> {
 }
 ```
 
+```output
+First smoothed value (epanechnikov kernel): 0.40672777844316566
+```
+
 ---
 
 ### robustness_method
@@ -210,6 +226,10 @@ fn main() -> Result<(), LowessError> {
     println!("First smoothed value (talwar robustness): {}", result.y[0]);
     Ok(())
 }
+```
+
+```output
+First smoothed value (talwar robustness): 0.37877937542419027
 ```
 
 ---
@@ -248,6 +268,10 @@ fn main() -> Result<(), LowessError> {
 }
 ```
 
+```output
+First smoothed value (reflect boundary): 0.7127616908322939
+```
+
 ---
 
 ### scaling_method
@@ -281,6 +305,10 @@ fn main() -> Result<(), LowessError> {
     println!("First smoothed value (mad scaling): {}", result.y[0]);
     Ok(())
 }
+```
+
+```output
+First smoothed value (mad scaling): 0.38260776436644134
 ```
 
 ---
@@ -318,6 +346,10 @@ fn main() -> Result<(), LowessError> {
 }
 ```
 
+```output
+First smoothed value (use_local_mean fallback): 0.38260776436644134
+```
+
 ---
 
 ### auto_converge
@@ -342,6 +374,10 @@ fn main() -> Result<(), LowessError> {
     println!("First smoothed value (auto-converge): {}", result.y[0]);
     Ok(())
 }
+```
+
+```output
+First smoothed value (auto-converge): 0.3826022035717678
 ```
 
 ---
@@ -379,6 +415,10 @@ fn main() -> Result<(), LowessError> {
 }
 ```
 
+```output
+First smoothed value (custom weights): 0.33693529914928816
+```
+
 ---
 
 ## Output Options
@@ -407,6 +447,10 @@ fn main() -> Result<(), LowessError> {
 
     Ok(())
 }
+```
+
+```output
[…]
