### 2026-08-24 34c645f
docs: fix broken snippets
