### 2026-08-24 36a7e20
docs: fix broken snippets, fix verification scripts, fix readme files
