### 2026-08-24 216a005
docs: fix minor inconsistencies
