+++ r/llms.txt
@@ -1,21 +1,11 @@
 # LOWESS Project
 
-[![lowess](https://img.shields.io/badge/lowess-000000?logo=rust&logoColor=white)](https://crates.io/crates/lowess)
-[![fastLowess](https://img.shields.io/badge/fastLowess-000000?logo=rust&logoColor=white)](https://crates.io/crates/fastLowess)
-[![PyPI](https://img.shields.io/badge/PyPI-3775A9?logo=pypi&logoColor=white)](https://pypi.org/project/fastlowess/)
 [![R-universe](https://img.shields.io/badge/R--universe-276DC3?logo=r&logoColor=white)](https://thisisamirv.r-universe.dev/rfastlowess)
-[![npm](https://img.shields.io/badge/npm-CB3837?logo=npm&logoColor=white)](https://www.npmjs.com/package/fastlowess)
-[![Julia](https://img.shields.io/badge/Julia-9558B2?logo=julia&logoColor=white)](https://juliahub.com/ui/Packages/General/FastLOWESS)
-[![WASM](https://img.shields.io/badge/WASM-654FF0?logo=webassembly&logoColor=white)](https://www.npmjs.com/package/fastlowess-wasm)
-[![C++](https://img.shields.io/badge/C++-00599C?logo=cplusplus&logoColor=white)](https://github.com/thisisamirv/lowess-project/releases/latest)  
-[![fastlowess
-(Python)](https://img.shields.io/badge/fastlowess_(Python)-44A833?logo=anaconda&logoColor=white)](https://anaconda.org/conda-forge/fastlowess)
-[![libfastlowess
-(C++)](https://img.shields.io/badge/libfastlowess_(C++)-44A833?logo=anaconda&logoColor=white)](https://anaconda.org/conda-forge/libfastlowess)
 [![rfastlowess
-(R)](https://img.shields.io/badge/rfastlowess_(R)-44A833?logo=anaconda&logoColor=white)](https://anaconda.org/conda-forge/r-rfastlowess)  
+(R)](https://img.shields.io/badge/rfastlowess_(R)-44A833?logo=anaconda&logoColor=white)](https://anaconda.org/conda-forge/r-rfastlowess)
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/769_status.svg)](https://github.com/ropensci/software-review/issues/769)
+[![CI](https://github.com/thisisamirv/lowess-project/actions/workflows/ci-r.yml/badge.svg)](https://github.com/thisisamirv/lowess-project/actions/workflows/ci-r.yml)
 
 ![One LOWESS to Rule Them
 All](https://raw.githubusercontent.com/thisisamirv/lowess-project/main/dev/logo.png)  
@@ -26,27 +16,9 @@ LOWESS (Locally Weighted Scatterplot Smoothing) implementation for
 **Rust**, **Python**, **R**, **Julia**, **JavaScript**, **C++**, and
 **WebAssembly**.
 
-> \[!IMPORTANT\]
->
-> The `lowess-project` contains a complete ecosystem for LOWESS
-> smoothing:
->
-> - **[`lowess`](https://crates.io/crates/lowess)** - Core
->   single-threaded Rust implementation with `no_std` support
-> - **[`fastLowess`](https://crates.io/crates/fastLowess)** - Parallel
->   CPU and GPU-accelerated Rust wrapper with ndarray integration  
-> - **[`R bindings`](https://thisisamirv.r-universe.dev/rfastlowess)** -
->   extendr-based R binding
-> - **[`Python bindings`](https://pypi.org/project/fastlowess/)** -
->   PyO3-based Python binding
[…]
