+++ bindings/wasm/Makefile
@@ -59,7 +59,7 @@ dev:
 	@echo "Linting WASM JS files..."
 	@cd $(WASM_DIR) && $(NPM) install -q
 	@cd $(WASM_DIR) && $(NPM) audit || true
-	@cd $(WASM_DIR) && $(NPX) -y depcheck --ignores="oxlint,typedoc-plugin-pages,typedoc-plugin-markdown"
+	@cd $(WASM_DIR) && $(NPX) -y depcheck --ignores="oxlint,typedoc-plugin-pages,typedoc-plugin-markdown,astro:content"
 	@cd $(WASM_DIR) && ($(NPM) outdated | grep -v "fastlowess-" || true)
 	@cd $(WASM_DIR) && $(NPM) ci --dry-run
 	@$(NPX) -y oxlint $(WASM_DIR)/src/*.js $(WASM_TEST_DIR)/*.js
