+++ nodejs/_astro/ui-core.BvAwiB5P.js
@@ -1,0 +1,2 @@
+import{t as e}from"./Search.astro_astro_type_script_index_0_lang.d6R31PM8.js";var t=Object.defineProperty,n=(e,n)=>{for(var r in n)t(e,r,{get:n[r],enumerable:!0})};function r(){}function i(e){return e()}function a(){return Object.create(null)}function o(e){e.forEach(i)}function s(e){return typeof e==`function`}function c(e,t){return e==e?e!==t||e&&typeof e==`object`||typeof e==`function`:t==t}var l;function u(e,t){return e===t||(l||=document.createElement(`a`),l.href=t,e===l.href)}function d(e){return Object.keys(e).length===0}`WeakMap`in(typeof window<`u`?window:typeof globalThis<`u`?globalThis:global);function f(e,t){e.appendChild(t)}function p(e,t,n){e.insertBefore(t,n||null)}function m(e){e.parentNode&&e.parentNode.removeChild(e)}function h(e,t){for(let n=0;n<e.length;n+=1)e[n]&&e[n].d(t)}function g(e){return document.createElement(e)}function _(e){return document.createElementNS(`http://www.w3.org/2000/svg`,e)}function v(e){return document.createTextNode(e)}function y(){return v(` `)}function b(){return v(``)}function x(e,t,n,r){return e.addEventListener(t,n,r),()=>e.removeEventListener(t,n,r)}function S(e,t,n){n==null?e.removeAttribute(t):e.getAttribute(t)!==n&&e.setAttribute(t,n)}function C(e){return Array.from(e.childNodes)}function w(e,t){t=``+t,e.data!==t&&(e.data=t)}function T(e,t){e.value=t??``}function E(e,t,n){e.classList.toggle(t,!!n)}var D=class{is_svg=!1;e=void 0;n=void 0;t=void 0;a=void 0;constructor(e=!1){this.is_svg=e,this.e=this.n=null}c(e){this.h(e)}m(e,t,n=null){this.e||(this.e=this.is_svg?_(t.nodeName):g(t.nodeType===11?`TEMPLATE`:t.nodeName),this.t=t.tagName===`TEMPLATE`?t.content:t,this.c(e)),this.i(n)}h(e){this.e.innerHTML=e,this.n=Array.from(this.e.nodeName===`TEMPLATE`?this.e.content.childNodes:this.e.childNodes)}i(e){for(let t=0;t<this.n.length;t+=1)p(this.t,this.n[t],e)}p(e){this.d(),this.h(e),this.i(this.a)}d(){this.n.forEach(m)}},O;function k(e){O=e}function A(){if(!O)throw Error(`Function called outside component initialization`);return O}function ee(e){A().$$.on_destroy.push(e)}var j=[],M=[],N=[],P=[],F=Promise.resolve(),I=!1;function L(){I||(I=!0,F.then(V))}function R(e){N.push(e)}function te(e){P.push(e)}var z=new Set,B=
[…]
