### 2026-08-24 cfe41e4
docs: replace indexFileName with entryFileName for TypeDoc
