+++ bindings/nodejs/src/content/docs/reference/index.md
@@ -1,24 +1,0 @@
-**fastlowess**
-
-***
-
-# fastlowess
-
-## Classes
-
-- [Lowess](classes/Lowess.md)
-- [LowessResult](classes/LowessResult.md)
-- [OnlineLowess](classes/OnlineLowess.md)
-- [StreamingLowess](classes/StreamingLowess.md)
-
-## Interfaces
-
-- [Diagnostics](interfaces/Diagnostics.md)
-- [OnlineOptions](interfaces/OnlineOptions.md)
-- [OnlineOutput](interfaces/OnlineOutput.md)
-- [SmoothOptions](interfaces/SmoothOptions.md)
-- [StreamingOptions](interfaces/StreamingOptions.md)
-
-## Functions
-
-- [gpu\_enabled](functions/gpu_enabled.md)
