+++ bindings/nodejs/src/content/docs/reference/functions/gpu_enabled.md
@@ -1,17 +1,0 @@
-[**fastlowess**](../index.md)
-
-***
-
-[fastlowess](../index.md) / gpu\_enabled
-
-# Function: gpu\_enabled()
-
-> **gpu\_enabled**(): `boolean`
-
-Defined in: [index.d.ts:80](https://github.com/thisisamirv/lowess-project/blob/a0a9afcfad40aebb60e25a38abb07b76a65a3861/bindings/nodejs/index.d.ts#L80)
-
-True if this addon was built with the `gpu` Cargo feature enabled.
-
-## Returns
-
-`boolean`

[…]
