+++ cpp/search/all_d.js
@@ -1,20 +1,20 @@
 var searchData=
 [
-  ['one_20out_20loocv_0',['Leave-One-Out (LOOCV)',['../md_docs_2cross-validation.html#autotoc_md82',1,'']]],
+  ['one_20out_20loocv_0',['Leave-One-Out (LOOCV)',['../md_docs_2cross-validation.html#autotoc_md88',1,'']]],
   ['online_20adapter_1',['Online Adapter',['../md_docs_2online.html',1,'']]],
-  ['online_20mode_2',['Online Mode',['../md_docs_2use-case-real-time.html#autotoc_md306',1,'']]],
-  ['online_20mode_3a_20point_20by_20point_3',['Online Mode: Point-by-Point',['../md_docs_2use-case-real-time.html#autotoc_md297',1,'']]],
+  ['online_20mode_2',['Online Mode',['../md_docs_2use-case-real-time.html#autotoc_md312',1,'']]],
+  ['online_20mode_3a_20point_20by_20point_3',['Online Mode: Point-by-Point',['../md_docs_2use-case-real-time.html#autotoc_md303',1,'']]],
   ['online_2emd_4',['online.md',['../online_8md.html',1,'']]],
   ['onlinelowess_20—_20c_20api_20reference_5',['OnlineLowess — C++ API Reference',['../md_docs_2api-online.html',1,'']]],
   ['onlineoptions_20tt_20inherits_20tt_20lowessoptions_20tt_6',['&lt;tt&gt;OnlineOptions&lt;/tt&gt; (inherits &lt;tt&gt;LowessOptions&lt;/tt&gt;)',['../md_docs_2api-online.html#autotoc_md7',1,'']]],
-  ['options_7',['options',['../md_docs_2api-streaming.html#autotoc_md19',1,'Options'],['../md_docs_2api.html#autotoc_md36',1,'Options'],['../md_docs_2api-online.html#autotoc_md10',1,'Options'],['../md_docs_2parameters.html#autotoc_md210',1,'Output Options']]],
+  ['options_7',['options',['../md_docs_2api-streaming.html#autotoc_md19',1,'Options'],['../md_docs_2api.html#autotoc_md36',1,'Options'],['../md_docs_2api-online.html#autotoc_md10',1,'Options'],['../md_docs_2parameters.html#autotoc_md216',1,'Output Options']]],
   ['options_20structure_8',['options structure',['../md_docs_2api-streaming.html#autotoc_md17',1,'Options Structure'],['../md_docs_2api-online.html#autotoc_md6',1,'Options Structure']]],
   ['options_20structures_9',['Options Structures',['../md_docs_2api.html#autotoc_md24',1,'']]],
-  ['options_20summary_10',['Parameter Options Summary',['../md_docs_2parameters.html#autotoc_md187',1,'']]],
-  ['out_20loocv_11',['Leave-One-Out (LOOCV)',['../md_docs_2cross-validation.html#autotoc_md82',1,'']]],
-  ['outlier_12',['Suppress a Known Outlier',['../md_docs_2custom-weights.html#autotoc_md102',1,'']]],
-  ['outliers_13',['outliers',['../md_docs_2robustness.html#autotoc_md258',1,'Detecting Outliers'],['../md_docs_2quickstart.html#autotoc_md241',1,'Handling Outliers']]],
-  ['output_20options_14',['Output Options',['../md_docs_2parameters.html#autotoc_md210',1,'']]],
-  ['overlap_15',['overlap',['../md_docs_2parameters.html#autotoc_md227',1,'']]],
[…]
