+++ cpp/search/all_e.js
@@ -1,21 +1,21 @@
 var searchData=
 [
-  ['parameter_0',['The Fraction Parameter',['../index.html#autotoc_md66',1,'']]],
-  ['parameter_20options_20summary_1',['Parameter Options Summary',['../md_docs_2parameters.html#autotoc_md187',1,'']]],
-  ['parameters_2',['parameters',['../md_docs_2parameters.html#autotoc_md224',1,'Adapter Parameters'],['../md_docs_2use-case-real-time.html#autotoc_md305',1,'Choosing Parameters'],['../md_docs_2parameters.html#autotoc_md189',1,'Core Parameters'],['../md_docs_2parameters.html',1,'parameters'],['../md_docs_2streaming.html#autotoc_md275',1,'Parameters'],['../md_docs_2parameters.html#autotoc_md184',1,'Parameters'],['../md_docs_2online.html#autotoc_md180',1,'Parameters']]],
+  ['parameter_0',['The Fraction Parameter',['../index.html#autotoc_md72',1,'']]],
+  ['parameter_20options_20summary_1',['Parameter Options Summary',['../md_docs_2parameters.html#autotoc_md193',1,'']]],
+  ['parameters_2',['parameters',['../md_docs_2parameters.html#autotoc_md230',1,'Adapter Parameters'],['../md_docs_2use-case-real-time.html#autotoc_md311',1,'Choosing Parameters'],['../md_docs_2parameters.html#autotoc_md195',1,'Core Parameters'],['../md_docs_2parameters.html',1,'parameters'],['../md_docs_2streaming.html#autotoc_md281',1,'Parameters'],['../md_docs_2parameters.html#autotoc_md190',1,'Parameters'],['../md_docs_2online.html#autotoc_md186',1,'Parameters']]],
   ['parameters_2emd_3',['parameters.md',['../parameters_8md.html',1,'']]],
-  ['performance_20considerations_4',['performance considerations',['../md_docs_2use-case-real-time.html#autotoc_md309',1,'Performance Considerations'],['../md_docs_2api.html#autotoc_md32',1,'Performance Considerations']]],
-  ['point_20by_20point_5',['Online Mode: Point-by-Point',['../md_docs_2use-case-real-time.html#autotoc_md297',1,'']]],
-  ['points_6',['Emphasize Important Points',['../md_docs_2custom-weights.html#autotoc_md104',1,'']]],
-  ['policy_7',['Choosing a Policy',['../md_docs_2boundary.html#autotoc_md60',1,'']]],
-  ['practices_8',['Best Practices',['../md_docs_2cross-validation.html#autotoc_md94',1,'']]],
-  ['practices_20for_20genomic_20data_9',['Best Practices for Genomic Data',['../md_docs_2use-case-genomics.html#autotoc_md291',1,'']]],
-  ['pre_20built_20binaries_20linux_20x64_10',['Pre-built Binaries (Linux (x64))',['../md_docs_2installation.html#autotoc_md125',1,'']]],
-  ['pre_20built_20binaries_20macos_20x64_11',['Pre-built Binaries (macOS (x64))',['../md_docs_2installation.html#autotoc_md126',1,'']]],
-  ['pre_20built_20binaries_20windows_20x64_12',['Pre-built Binaries (Windows (x64))',['../md_docs_2installation.html#autotoc_md127',1,'']]],
-  ['prediction_20intervals_13',['prediction intervals',['../index.html#autotoc_md70',1,'Confidence vs Prediction Intervals'],['../md_docs_2use-case-time-series.html#autotoc_md319',1,'Forecasting with Prediction Intervals'],['../md_docs_2intervals.html#autotoc_md137',1,'Prediction Intervals']]],
[…]
