+++ cpp/search/pages_1.js
@@ -2,5 +2,6 @@ var searchData=
 [
   ['backend_0',['gpu-backend',['../md_docs_2gpu-backend.html',1,'']]],
   ['batch_20adapter_1',['Batch Adapter',['../md_docs_2batch.html',1,'']]],
-  ['boundary_2',['boundary',['../md_docs_2boundary.html',1,'']]]
+  ['benchmarks_2',['benchmarks',['../md_docs_2benchmarks.html',1,'']]],
+  ['boundary_3',['boundary',['../md_docs_2boundary.html',1,'']]]
 ];
