+++ cpp/search/all_7.js
@@ -1,11 +1,11 @@
 var searchData=
 [
-  ['handling_0',['Boundary Handling',['../md_docs_2boundary.html#autotoc_md49',1,'']]],
-  ['handling_20missing_20data_1',['Handling Missing Data',['../md_docs_2use-case-time-series.html#autotoc_md321',1,'']]],
-  ['handling_20outliers_2',['Handling Outliers',['../md_docs_2quickstart.html#autotoc_md241',1,'']]],
+  ['handling_0',['Boundary Handling',['../md_docs_2boundary.html#autotoc_md55',1,'']]],
+  ['handling_20missing_20data_1',['Handling Missing Data',['../md_docs_2use-case-time-series.html#autotoc_md327',1,'']]],
+  ['handling_20outliers_2',['Handling Outliers',['../md_docs_2quickstart.html#autotoc_md247',1,'']]],
   ['hardware_20requirements_3',['Hardware Requirements',['../md_docs_2api.html#autotoc_md31',1,'']]],
-  ['how_20custom_20weights_20work_4',['How Custom Weights Work',['../md_docs_2custom-weights.html#autotoc_md96',1,'']]],
-  ['how_20it_20works_5',['How It Works',['../index.html#autotoc_md64',1,'']]],
-  ['how_20robustness_20works_6',['How Robustness Works',['../md_docs_2robustness.html#autotoc_md247',1,'']]],
-  ['huber_7',['Huber',['../md_docs_2robustness.html#autotoc_md252',1,'']]]
+  ['how_20custom_20weights_20work_4',['How Custom Weights Work',['../md_docs_2custom-weights.html#autotoc_md102',1,'']]],
+  ['how_20it_20works_5',['How It Works',['../index.html#autotoc_md70',1,'']]],
+  ['how_20robustness_20works_6',['How Robustness Works',['../md_docs_2robustness.html#autotoc_md253',1,'']]],
+  ['huber_7',['Huber',['../md_docs_2robustness.html#autotoc_md258',1,'']]]
 ];

[…]
