### 2026-08-24 5cf4c1e
fix: R docs line length, C++ CI
