+++ .github/workflows/ci-cpp.yml
@@ -37,6 +37,14 @@ jobs:
         if: matrix.os == 'windows-latest'
         run: choco install make cppcheck -y
 
+      - name: Install cppcheck (Linux)
+        if: matrix.os == 'ubuntu-latest'
+        run: sudo apt-get install -y cppcheck
+
+      - name: Install cppcheck (macOS)
+        if: matrix.os == 'macos-latest'
+        run: brew install cppcheck
+
       - uses: Swatinem/rust-cache@v2
         with:
           key: cpp-${{ matrix.os }}
