### 2026-08-24 e7343ac
fix: ignoreConfig in nodejs
