### 2026-08-24 3a962d0
fix: cpp clang-tidy install, R repack, docs landing page
