### 2026-08-24 6c2a187
ci: fix race condition in pages deployment
