### 2026-08-24 ac96e39
fix: R vignettes indentation issue, macos auto fn issue, cpp and julia docs issues
