+++ bindings/julia/julia/docs/src/use-case-time-series.md
@@ -171,7 +171,7 @@ println("R²: ", result.diagnostics.r_squared)
 
 ## See Also
 
-- [Real-Time Processing](real-time.md) — For streaming time series
+- [Real-Time Processing](use-case-real-time.md) — For streaming time series
 - [Cross-Validation](cross-validation.md) — Optimal fraction selection
 - [Boundary Handling](boundary.md) — Edge bias in trend extraction
 - [Parameters](parameters.md) — Full parameter reference
