### 2026-08-24 ddbff0d
build: drop rust assets/ during R vendoring
