### 2026-08-24 af55e63
build: strip symbols and switch vignettes to rmarkdown::html_vignette during R vendoring
