+++ bindings/r/vignettes/cross-validation.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Cross-Validation"
-output: BiocStyle::html_document
+output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Cross-Validation}
   %\VignetteEngine{knitr::rmarkdown}
