+++ bindings/r/vignettes/online.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Online Mode (OnlineLowess)"
-output: BiocStyle::html_document
+output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Online Mode (OnlineLowess)}
   %\VignetteEngine{knitr::rmarkdown}
