+++ bindings/r/vignettes/quickstart.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Quick Start"
-output: BiocStyle::html_document
+output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Quick Start}
   %\VignetteEngine{knitr::rmarkdown}
