### 2026-08-25 9a8af95
build: made R packaging significantly simpler
