+++ bindings/r/R/rfastlowess-package.R
@@ -31,5 +31,4 @@
 #' @useDynLib rfastlowess, .registration = TRUE
 #' @importFrom stats smooth
 #' @importFrom stats setNames
-#' @importFrom BiocGenerics normalize
 "_PACKAGE"
