### 2026-08-25 1cb62e8
docs: fix broken cross-links
