+++ bindings/wasm/src/content/docs/use-case-real-time.md
@@ -159,7 +159,7 @@ Smoothed: 0.0358
 
 ## See Also
 
-- [Execution Modes](../user-guide/adapters.md) — Detailed mode comparison
-- [Merge Strategies](../user-guide/merge.md) — Chunk reconciliation in depth
-- [Scaling Methods](../user-guide/scaling.md) — Robustness scale estimation
-- [Time Series](time-series.md) — General time series analysis
+- [Execution Modes](adapter-choice.md) — Detailed mode comparison
+- [Merge Strategies](merge.md) — Chunk reconciliation in depth
+- [Scaling Methods](scaling.md) — Robustness scale estimation
+- [Time Series](use-case-time-series.md) — General time series analysis
