### 2026-08-25 00dd31b
docs: fix cpp docs figs
