+++ bindings/cpp/docs/scaling.md
@@ -19,7 +19,7 @@ where $B$ is the bisquare function and $\hat{\sigma}$ is the scale estimate. A l
 | `"mar"` | Median of \|residuals\| | Robust | Fast |
 | `"mean"` | Mean of \|residuals\| | Less robust | Fastest |
 
-![Scaling Methods Comparison](../assets/diagrams/scaling_comparison.svg)
+![Scaling Methods Comparison](scaling_comparison.svg)
 
 ---
 
