+++ bindings/cpp/docs/concepts.md
@@ -18,10 +18,7 @@ Unlike parametric methods (linear regression, polynomial fitting), LOWESS adapts
 
 ## How It Works
 
-<figure markdown="span">
-  ![LOWESS Smoothing Concept](../assets/diagrams/lowess_concept.svg){ width="800" }
-  <figcaption>LOWESS fits local weighted regressions at each point, using a focused local window around each evaluation point</figcaption>
-</figure>
+![LOWESS fits local weighted regressions at each point, using a focused local window around each evaluation point](lowess_concept.svg)
 
 For each point in your data, LOWESS:
 
@@ -37,10 +34,7 @@ For each point in your data, LOWESS:
 
 The `fraction` (also called bandwidth or span) is the most important parameter. It controls what proportion of data is used for each local fit.
 
-<figure markdown="span">
-  ![Fraction Effect](../assets/diagrams/fraction_comparison.svg){ width="1000" }
-  <figcaption>Small fraction vs large fraction — bandwidth controls how closely the fit follows local structure</figcaption>
-</figure>
+![Small fraction vs large fraction — bandwidth controls how closely the fit follows local structure](fraction_comparison.svg)
 
 | Fraction | Effect | When to Use |
 | --- | --- | --- |
@@ -58,10 +52,7 @@ The `fraction` (also called bandwidth or span) is the most important parameter. 
 
 Standard LOWESS is sensitive to outliers. **Robustness iterations** downweight points with large residuals:
 
-<figure markdown="span">
-  ![Robustness Effect](../assets/diagrams/robust_iter_comparison.svg){ width="800" }
-  <figcaption>Non-robust LOWESS (iterations=0) vs robust LOWESS — outlier influence is suppressed through iterative reweighting</figcaption>
-</figure>
+![Non-robust LOWESS (iterations=0) vs robust LOWESS — outlier influence is suppressed through iterative reweighting](robust_iter_comparison.svg)
 
 | Iterations | Effect | When to Use |
 | --- | --- | --- |
@@ -74,10 +65,7 @@ Standard LOWESS is sensitive to outliers. **Robustness iterations** downweight p
 
 ## Confidence vs Prediction Intervals
 
-<figure markdown="span">
-  ![Intervals](../assets/diagrams/intervals_comparison.svg){ width="800" }
-  <figcaption>Confidence intervals (narrow, mean curve uncertainty) vs Prediction intervals (wide, new-point uncertainty)</figcaption>
-</figure>
+![Confidence intervals (narrow, mean curve uncertainty) vs Prediction intervals (wide, new-point uncertainty)](intervals_comparison.svg)
 
 | Interval Type | What It Represents | Width |
 | --- | --- | --- |
