+++ cpp/search/pages_8.js
@@ -1,5 +1,5 @@
 var searchData=
 [
   ['online_20adapter_0',['Online Adapter',['../md_docs_2online.html',1,'']]],
-  ['onlinelowess_20—_20c_20api_20reference_1',['OnlineLowess — C++ API Reference',['../md_docs_2api-online.html',1,'']]]
+  ['onlinelowess_20api_1',['OnlineLowess API',['../md_docs_2api-online.html',1,'']]]
 ];
