+++ cpp/search/pages_b.js
@@ -1,6 +1,5 @@
 var searchData=
 [
   ['real_20time_0',['use-case-real-time',['../md_docs_2use-case-real-time.html',1,'']]],
-  ['reference_1',['reference',['../md_docs_2api.html',1,'fastLowess C++ API Reference'],['../md_docs_2api-online.html',1,'OnlineLowess — C++ API Reference'],['../md_docs_2api-streaming.html',1,'StreamingLowess — C++ API Reference']]],
-  ['robustness_2',['robustness',['../md_docs_2robustness.html',1,'']]]
+  ['robustness_1',['robustness',['../md_docs_2robustness.html',1,'']]]
 ];
