+++ cpp/search/all_2.js
@@ -1,47 +1,46 @@
 var searchData=
 [
-  ['c_20api_20reference_0',['c api reference',['../md_docs_2api.html',1,'fastLowess C++ API Reference'],['../md_docs_2api-online.html',1,'OnlineLowess — C++ API Reference'],['../md_docs_2api-streaming.html',1,'StreamingLowess — C++ API Reference']]],
-  ['case_20genomics_1',['use-case-genomics',['../md_docs_2use-case-genomics.html',1,'']]],
-  ['case_20real_20time_2',['use-case-real-time',['../md_docs_2use-case-real-time.html',1,'']]],
-  ['case_20time_20series_3',['use-case-time-series',['../md_docs_2use-case-time-series.html',1,'']]],
-  ['challenge_4',['The Challenge',['../md_docs_2use-case-genomics.html#autotoc_md289',1,'']]],
-  ['checking_20availability_5',['Checking Availability',['../md_docs_2gpu-backend.html#autotoc_md124',1,'']]],
-  ['chip_20seq_20signal_20smoothing_6',['ChIP-seq Signal Smoothing',['../md_docs_2use-case-genomics.html#autotoc_md292',1,'']]],
-  ['choice_7',['adapter-choice',['../md_docs_2adapter-choice.html',1,'']]],
-  ['choosing_20a_20kernel_8',['Choosing a Kernel',['../md_docs_2kernels.html#autotoc_md171',1,'']]],
-  ['choosing_20a_20policy_9',['Choosing a Policy',['../md_docs_2boundary.html#autotoc_md66',1,'']]],
-  ['choosing_20a_20scaling_20method_10',['Choosing a Scaling Method',['../md_docs_2scaling.html#autotoc_md278',1,'']]],
-  ['choosing_20a_20strategy_11',['Choosing a Strategy',['../md_docs_2merge.html#autotoc_md183',1,'']]],
-  ['choosing_20fraction_20for_20time_20series_12',['Choosing Fraction for Time Series',['../md_docs_2use-case-time-series.html#autotoc_md333',1,'']]],
-  ['choosing_20parameters_13',['Choosing Parameters',['../md_docs_2use-case-real-time.html#autotoc_md311',1,'']]],
-  ['chunk_20processing_14',['Streaming Mode: Chunk Processing',['../md_docs_2use-case-real-time.html#autotoc_md306',1,'']]],
-  ['chunk_5fsize_15',['chunk_size',['../md_docs_2parameters.html#autotoc_md231',1,'']]],
-  ['class_16',['class',['../md_docs_2api-online.html#autotoc_md4',1,'Class'],['../md_docs_2api-streaming.html#autotoc_md13',1,'Class']]],
-  ['classes_17',['Classes',['../md_docs_2api.html#autotoc_md22',1,'']]],
-  ['combined_20with_20robustness_20iterations_18',['Combined with Robustness Iterations',['../md_docs_2custom-weights.html#autotoc_md114',1,'']]],
-  ['comparison_19',['comparison',['../md_docs_2cross-validation.html#autotoc_md92',1,'Comparison'],['../md_docs_2robustness.html#autotoc_md262',1,'Comparison'],['../md_docs_2gpu-backend.html#autotoc_md122',1,'Feature Comparison'],['../md_docs_2api.html#autotoc_md30',1,'Feature Comparison']]],
-  ['concepts_20',['Concepts',['../index.html#autotoc_md67',1,'']]],
-  ['concepts_2emd_21',['concepts.md',['../concepts_8md.html',1,'']]],
-  ['conda_20forge_22',['From conda-forge',['../md_docs_2installation.html#autotoc_md135',1,'']]],
[…]
