### 2026-08-25 3b74aa8
docs: add .markdownlint to cpp docs
