+++ bindings/cpp/docs/use-case-real-time.md
@@ -1,4 +1,3 @@
-<!-- markdownlint-disable MD033 -->
 # Real-Time Processing
 
 Streaming and online LOWESS for live data.
