+++ bindings/cpp/docs/adapter-choice.md
@@ -1,4 +1,3 @@
-<!-- markdownlint-disable MD024 MD033 MD046 -->
 # Execution Modes
 
 Choose the right adapter for your use case.
