### 2026-08-25 09446a4
docs: improve consistency
