+++ bindings/r/vignettes/batch.Rmd
@@ -17,14 +17,13 @@ knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 The default mode. Processes the entire dataset at once and supports all
 features.
 
-![Gap handling](../man/figures/gap_handling.svg)
+![Gap Handling](../man/figures/gap_handling.svg)
 
 ## When to Use
 
-- Dataset fits comfortably in memory
-- Need confidence/prediction intervals
-- Need cross-validation
-- One-shot analysis
+- Dataset fits in memory
+- Need intervals, cross-validation, or diagnostics
+- Processing complete files
 
 ## Parameters
 
@@ -54,10 +53,10 @@ model <- Lowess(
 )
 result <- fit(model, x, y)
 
-cat("Confidence lower bounds:\n")
-print(result$confidence_lower)
-cat("Confidence upper bounds:\n")
-print(result$confidence_upper)
+cat("Confidence lower bounds (first 5):\n")
+print(head(result$confidence_lower, 5))
+cat("Confidence upper bounds (first 5):\n")
+print(head(result$confidence_upper, 5))
 cat("R²:", result$diagnostics$r_squared, "\n")
 ```
 
