+++ bindings/r/vignettes/online.Rmd
@@ -16,13 +16,13 @@ knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 
 Maintains a sliding window and processes each incoming point immediately.
 
-![Online adapter comparison](../man/figures/online_comparison.svg)
+![Online Adapter](../man/figures/online_comparison.svg)
 
 ## When to Use
 
-- Real-time data streams (sensors, logs)
-- Each point must be smoothed as it arrives
-- Memory-bounded processing with a fixed window
+- Data arrives incrementally (sensors, streams)
+- Need real-time smoothed values
+- Fixed memory budget
 
 ## Parameters
 
@@ -50,6 +50,7 @@ for (i in seq_along(times)) {
     result <- add_point(model, times[i], temperatures[i])
     if (!is.null(result))
         cat(sprintf("Time %d: %.2f\n", times[i], result$y))
+    if (i >= 10) break  # print only the first few outputs
 }
 ```
 
