+++ cpp/search/pages_a.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['quickstart_0',['quickstart',['../md_docs_2quickstart.html',1,'']]]
+  ['onlinelowess_20api_0',['OnlineLowess API',['../md_docs_2api-online.html',1,'']]]
 ];
