+++ cpp/search/all_13.js
@@ -2,15 +2,11 @@ var searchData=
 [
   ['uncertainty_0',['Propagate Measurement Uncertainty',['../md_docs_2custom-weights.html#autotoc_md112',1,'']]],
   ['uniform_1',['Uniform',['../md_docs_2kernels.html#autotoc_md169',1,'']]],
-  ['update_20modes_2',['Update Modes',['../md_docs_2online.html#autotoc_md187',1,'']]],
-  ['update_5fmode_3',['update_mode',['../md_docs_2parameters.html#autotoc_md241',1,'update_mode'],['../md_docs_2api-online.html#autotoc_md11',1,'update_mode'],['../md_docs_2api.html#autotoc_md43',1,'update_mode']]],
-  ['usage_4',['usage',['../md_docs_2custom-weights.html#autotoc_md107',1,'Basic Usage'],['../md_docs_2api.html#autotoc_md28',1,'Usage'],['../md_docs_2gpu-backend.html#autotoc_md128',1,'Usage']]],
-  ['use_5',['use',['../md_docs_2batch.html#autotoc_md46',1,'When to Use'],['../md_docs_2streaming.html#autotoc_md280',1,'When to Use'],['../md_docs_2online.html#autotoc_md185',1,'When to Use']]],
-  ['use_20case_20genomics_6',['use-case-genomics',['../md_docs_2use-case-genomics.html',1,'']]],
-  ['use_20case_20real_20time_7',['use-case-real-time',['../md_docs_2use-case-real-time.html',1,'']]],
-  ['use_20case_20time_20series_8',['use-case-time-series',['../md_docs_2use-case-time-series.html',1,'']]],
-  ['use_20custom_20weights_9',['When to Use Custom Weights',['../md_docs_2custom-weights.html#autotoc_md104',1,'']]],
-  ['use_2dcase_2dgenomics_2emd_10',['use-case-genomics.md',['../use-case-genomics_8md.html',1,'']]],
-  ['use_2dcase_2dreal_2dtime_2emd_11',['use-case-real-time.md',['../use-case-real-time_8md.html',1,'']]],
-  ['use_2dcase_2dtime_2dseries_2emd_12',['use-case-time-series.md',['../use-case-time-series_8md.html',1,'']]]
+  ['update_5fmode_2',['update_mode',['../md_docs_2api-online.html#autotoc_md12',1,'update_mode'],['../md_docs_2api.html#autotoc_md47',1,'update_mode'],['../md_docs_2parameters.html#autotoc_md235',1,'update_mode']]],
+  ['usage_3',['usage',['../md_docs_2custom-weights.html#autotoc_md107',1,'Basic Usage'],['../md_docs_2api.html#autotoc_md32',1,'Usage'],['../md_docs_2gpu-backend.html#autotoc_md128',1,'Usage']]],
+  ['use_4',['use',['../md_docs_2api-online.html#autotoc_md4',1,'When to Use'],['../md_docs_2api-streaming.html#autotoc_md14',1,'When to Use'],['../md_docs_2api.html#autotoc_md25',1,'When to Use']]],
+  ['use_20custom_20weights_5',['When to Use Custom Weights',['../md_docs_2custom-weights.html#autotoc_md104',1,'']]],
+  ['use_2dcase_2dgenomics_2emd_6',['use-case-genomics.md',['../use-case-genomics_8md.html',1,'']]],
+  ['use_2dcase_2dreal_2dtime_2emd_7',['use-case-real-time.md',['../use-case-real-time_8md.html',1,'']]],
+  ['use_2dcase_2dtime_2dseries_2emd_8',['use-case-time-series.md',['../use-case-time-series_8md.html',1,'']]]
 ];
