+++ cpp/search/pages_c.js
@@ -1,7 +1,4 @@
 var searchData=
 [
-  ['scaling_0',['scaling',['../md_docs_2scaling.html',1,'']]],
-  ['series_1',['use-case-time-series',['../md_docs_2use-case-time-series.html',1,'']]],
-  ['streaming_20adapter_2',['Streaming Adapter',['../md_docs_2streaming.html',1,'']]],
-  ['streaminglowess_20api_3',['StreamingLowess API',['../md_docs_2api-streaming.html',1,'']]]
+  ['quick_20start_0',['Quick Start',['../md_docs_2quickstart.html',1,'']]]
 ];
