+++ cpp/search/all_6.js
@@ -1,13 +1,12 @@
 var searchData=
 [
   ['gaussian_0',['Gaussian',['../md_docs_2kernels.html#autotoc_md161',1,'']]],
-  ['gene_20expression_20time_20course_1',['Gene Expression Time Course',['../md_docs_2use-case-time-series.html#autotoc_md331',1,'']]],
-  ['genome_20coverage_20streaming_2',['Large Genome Coverage (Streaming)',['../md_docs_2use-case-genomics.html#autotoc_md295',1,'']]],
-  ['genomic_20data_3',['Best Practices for Genomic Data',['../md_docs_2use-case-genomics.html#autotoc_md297',1,'']]],
-  ['genomic_20data_20smoothing_4',['Genomic Data Smoothing',['../md_docs_2use-case-genomics.html#autotoc_md285',1,'']]],
-  ['genomics_5',['use-case-genomics',['../md_docs_2use-case-genomics.html',1,'']]],
-  ['gpu_20acceleration_6',['GPU Acceleration',['../md_docs_2api.html#autotoc_md26',1,'']]],
-  ['gpu_20backend_7',['gpu backend',['../md_docs_2benchmarks.html#autotoc_md52',1,'GPU Backend'],['../md_docs_2gpu-backend.html#autotoc_md119',1,'GPU Backend'],['../md_docs_2gpu-backend.html',1,'gpu-backend']]],
-  ['gpu_20support_8',['gpu support',['../md_docs_2api.html#autotoc_md27',1,'Enabling GPU Support'],['../md_docs_2gpu-backend.html#autotoc_md126',1,'Installing GPU Support']]],
-  ['gpu_2dbackend_2emd_9',['gpu-backend.md',['../gpu-backend_8md.html',1,'']]]
+  ['gene_20expression_20time_20course_1',['Gene Expression Time Course',['../md_docs_2use-case-time-series.html#autotoc_md319',1,'']]],
+  ['genome_20coverage_20streaming_2',['Large Genome Coverage (Streaming)',['../md_docs_2use-case-genomics.html#autotoc_md283',1,'']]],
+  ['genomic_20data_3',['Best Practices for Genomic Data',['../md_docs_2use-case-genomics.html#autotoc_md285',1,'']]],
+  ['genomic_20data_20smoothing_4',['Genomic Data Smoothing',['../md_docs_2use-case-genomics.html',1,'']]],
+  ['gpu_20acceleration_5',['GPU Acceleration',['../md_docs_2api.html#autotoc_md30',1,'']]],
+  ['gpu_20backend_6',['gpu backend',['../md_docs_2gpu-backend.html',1,'GPU Backend'],['../md_docs_2benchmarks.html#autotoc_md52',1,'GPU Backend']]],
+  ['gpu_20support_7',['gpu support',['../md_docs_2api.html#autotoc_md31',1,'Enabling GPU Support'],['../md_docs_2gpu-backend.html#autotoc_md126',1,'Installing GPU Support']]],
+  ['gpu_2dbackend_2emd_8',['gpu-backend.md',['../gpu-backend_8md.html',1,'']]]
 ];
