+++ cpp/search/pages_d.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['time_0',['use-case-real-time',['../md_docs_2use-case-real-time.html',1,'']]],
-  ['time_20series_1',['use-case-time-series',['../md_docs_2use-case-time-series.html',1,'']]]
+  ['real_20time_20processing_0',['Real-Time Processing',['../md_docs_2use-case-real-time.html',1,'']]],
+  ['robustness_1',['Robustness',['../md_docs_2robustness.html',1,'']]]
 ];

[…]
