+++ r/articles/online.md
@@ -7,15 +7,15 @@
 Maintains a sliding window and processes each incoming point
 immediately.
 
-![Online adapter comparison](../reference/figures/online_comparison.svg)
+![Online Adapter](../reference/figures/online_comparison.svg)
 
-Online adapter comparison
+Online Adapter
 
 ## When to Use
 
-- Real-time data streams (sensors, logs)
-- Each point must be smoothed as it arrives
-- Memory-bounded processing with a fixed window
+- Data arrives incrementally (sensors, streams)
+- Need real-time smoothed values
+- Fixed memory budget
 
 ## Parameters
 
@@ -44,6 +44,7 @@ for (i in seq_along(times)) {
     result <- add_point(model, times[i], temperatures[i])
     if (!is.null(result))
         cat(sprintf("Time %d: %.2f\n", times[i], result$y))
+    if (i >= 10) break  # print only the first few outputs
 }
 #> Time 5: 22.80
 #> Time 6: 22.72
@@ -51,96 +52,6 @@ for (i in seq_along(times)) {
 #> Time 8: 23.49
 #> Time 9: 25.94
 #> Time 10: 24.14
-#> Time 11: 25.35
-#> Time 12: 27.00
-#> Time 13: 23.99
-#> Time 14: 24.04
-#> Time 15: 24.64
-#> Time 16: 25.60
-#> Time 17: 25.13
-#> Time 18: 23.01
-#> Time 19: 21.96
-#> Time 20: 24.34
-#> Time 21: 24.42
-#> Time 22: 23.41
-#> Time 23: 23.21
-#> Time 24: 23.86
-#> Time 25: 24.49
-#> Time 26: 23.48
-#> Time 27: 22.44
-#> Time 28: 20.54
-#> Time 29: 20.50
-#> Time 30: 20.03
-#> Time 31: 20.29
-#> Time 32: 20.35
-#> Time 33: 20.27
-#> Time 34: 18.98
-#> Time 35: 18.52
-#> Time 36: 16.82
-#> Time 37: 16.15
-#> Time 38: 15.74
-#> Time 39: 14.60
-#> Time 40: 15.17
-#> Time 41: 15.68
-#> Time 42: 15.58
-#> Time 43: 15.97
-#> Time 44: 15.21
-#> Time 45: 14.17
-#> Time 46: 14.57
-#> Time 47: 14.31
-#> Time 48: 15.51
-#> Time 49: 15.33
-#> Time 50: 15.71
-#> Time 51: 15.76
-#> Time 52: 15.26
-#> Time 53: 16.38
-#> Time 54: 16.85
-#> Time 55: 16.94
-#> Time 56: 17.17
-#> Time 57: 17.66
-#> Time 58: 17.85
-#> Time 59: 16.47
-#> Time 60: 17.56
-#> Time 61: 18.29
-#> Time 62: 19.43
-#> Time 63: 20.61
-#> Time 64: 21.85
-#> Time 65: 21.41
-#> Time 66: 22.33
-#> Time 67: 22.58
-#> Time 68: 23.29
-#> Time 69: 23.88
-#> Time 70: 24.23
-#> Time 71: 23.47
-#> Time 72: 23.59
-#> Time 73: 24.27
-#> Time 74: 24.04
-#> Time 75: 24.19
-#> Time 76: 24.92
-#> Time 77: 25.52
-#> Time 78: 25.75
-#> Time 79: 25.01
-#> Time 80: 24.18
-#> Time 81: 25.10
-#> Time 82: 25.15
-#> Time 83: 25.01
-#> Time 84: 24.54
-#> Time 85: 23.35
-#> Time 86: 23.48
[…]
