### 2026-08-25 6b628a5
docs: fix rust docs
