+++ .github/workflows/docs.yml
@@ -32,7 +32,7 @@ jobs:
       wasm: ${{ steps.f.outputs.wasm }}
     steps:
       - uses: actions/checkout@v7
-      - uses: dorny/paths-filter@v3
+      - uses: dorny/paths-filter@v4
         id: f
         with:
           filters: |
