### 2026-08-25 10828f6
docs: fix rust docs format, add theme to nodejs/wasm sites
