+++ crates/lowess/src/evaluation/cv.rs
@@ -3,12 +3,11 @@
 //! This module provides cross-validation tools for selecting the optimal
 //! smoothing fraction (bandwidth) in LOWESS regression. It implements
 //! generic k-fold and leave-one-out cross-validation strategies.
-//!
-//! ## srrstats Compliance
-//!
-//! @srrstats {RE6.0} Leave-one-out cross-validation (LOOCV) for bandwidth selection.
-//! K-fold cross-validation with reproducible shuffling (optional seed).
-//! RMSE-based score aggregation for model selection.
+// ## srrstats Compliance
+//
+// @srrstats {RE6.0} Leave-one-out cross-validation (LOOCV) for bandwidth selection.
+// K-fold cross-validation with reproducible shuffling (optional seed).
+// RMSE-based score aggregation for model selection.
 
 // External dependencies
 #[cfg(not(feature = "std"))]
