### 2026-08-25 5a23a0a
docs: fix readthedocs requirements, fix rust gpu backend docs, improve add outputs scripts
