+++ bindings/python/docs/requirements.txt
@@ -6,5 +6,6 @@ jupyter-sphinx>=0.5
 sphinx-copybutton>=0.5
 sphinxcontrib-mermaid>=1
 ipykernel>=6
+matplotlib>=3.7
 fastlowess
 
