### 2026-08-25 ee28da6
docs: minor fixes, repack
