+++ bindings/wasm/src/content/docs/benchmarks.md
@@ -57,5 +57,5 @@ console.log(`WASM: ${ms.toFixed(2)} ms`);
 ```
 
 ```output
-WASM: 18.28 ms
+WASM: 16.73 ms
 ```
