+++ .github/workflows/ci-r.yml
@@ -51,6 +51,10 @@ jobs:
 
       - uses: r-lib/actions/setup-tinytex@v2
 
+      - name: Install LaTeX packages for PDF manual
+        shell: Rscript {0}
+        run: tinytex::tlmgr_install(c("inconsolata", "times", "helvetic", "courier", "url"))
+
       - name: Bootstrap pak (Windows)
         if: runner.os == 'Windows'
         shell: Rscript {0}
@@ -117,6 +121,10 @@ jobs:
       - uses: r-lib/actions/setup-pandoc@v2
 
       - uses: r-lib/actions/setup-tinytex@v2
+
+      - name: Install LaTeX packages for PDF manual
+        shell: Rscript {0}
+        run: tinytex::tlmgr_install(c("inconsolata", "times", "helvetic", "courier", "url"))
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
