### 2026-08-25 92a361e
docs: replace mermaid chart, fix rust installation guide, add validation workflow
