+++ CHANGELOG.md
@@ -6,13 +6,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 3.1.0
 
 ### Added
 
 **Monorepo:**
 
 - Added a GitHub Pages landing page at the repository root, built from `README.md` via pandoc and deployed by `docs.yml`.
+- Added a GitHub workflow for running validation scripts.
 
 **Julia:**
 
