+++ cpp/search/all_1.js
@@ -1,25 +1,25 @@
 var searchData=
 [
   ['backend_0',['backend',['../md_docs_2gpu-backend.html',1,'GPU Backend'],['../md_docs_2benchmarks.html#autotoc_md52',1,'GPU Backend']]],
-  ['basic_20smoothing_1',['Basic Smoothing',['../md_docs_2quickstart.html#autotoc_md237',1,'']]],
-  ['basic_20trend_20extraction_2',['Basic Trend Extraction',['../md_docs_2use-case-time-series.html#autotoc_md309',1,'']]],
-  ['basic_20usage_3',['Basic Usage',['../md_docs_2custom-weights.html#autotoc_md107',1,'']]],
+  ['basic_20smoothing_1',['Basic Smoothing',['../md_docs_2quickstart.html#autotoc_md239',1,'']]],
+  ['basic_20trend_20extraction_2',['Basic Trend Extraction',['../md_docs_2use-case-time-series.html#autotoc_md311',1,'']]],
+  ['basic_20usage_3',['Basic Usage',['../md_docs_2custom-weights.html#autotoc_md109',1,'']]],
   ['benchmarks_4',['benchmarks',['../md_docs_2benchmarks.html',1,'Benchmarks'],['../md_docs_2benchmarks.html#autotoc_md50',1,'CPU Benchmarks'],['../md_docs_2benchmarks.html#autotoc_md54',1,'Reproducing Benchmarks']]],
   ['benchmarks_2emd_5',['benchmarks.md',['../benchmarks_8md.html',1,'']]],
-  ['best_20practices_6',['Best Practices',['../md_docs_2cross-validation.html#autotoc_md100',1,'']]],
-  ['best_20practices_20for_20genomic_20data_7',['Best Practices for Genomic Data',['../md_docs_2use-case-genomics.html#autotoc_md285',1,'']]],
-  ['binaries_20linux_20x64_8',['Pre-built Binaries (Linux (x64))',['../md_docs_2installation.html#autotoc_md131',1,'']]],
-  ['binaries_20macos_20x64_9',['Pre-built Binaries (macOS (x64))',['../md_docs_2installation.html#autotoc_md132',1,'']]],
-  ['binaries_20windows_20x64_10',['Pre-built Binaries (Windows (x64))',['../md_docs_2installation.html#autotoc_md133',1,'']]],
-  ['bisquare_20default_11',['Bisquare (Default)',['../md_docs_2robustness.html#autotoc_md250',1,'']]],
-  ['biweight_12',['Biweight',['../md_docs_2kernels.html#autotoc_md163',1,'']]],
-  ['both_20intervals_13',['Both Intervals',['../md_docs_2intervals.html#autotoc_md145',1,'']]],
+  ['best_20practices_6',['Best Practices',['../md_docs_2cross-validation.html#autotoc_md102',1,'']]],
+  ['best_20practices_20for_20genomic_20data_7',['Best Practices for Genomic Data',['../md_docs_2use-case-genomics.html#autotoc_md287',1,'']]],
+  ['binaries_20linux_20x64_8',['Pre-built Binaries (Linux (x64))',['../md_docs_2installation.html#autotoc_md133',1,'']]],
+  ['binaries_20macos_20x64_9',['Pre-built Binaries (macOS (x64))',['../md_docs_2installation.html#autotoc_md134',1,'']]],
+  ['binaries_20windows_20x64_10',['Pre-built Binaries (Windows (x64))',['../md_docs_2installation.html#autotoc_md135',1,'']]],
+  ['bisquare_20default_11',['Bisquare (Default)',['../md_docs_2robustness.html#autotoc_md252',1,'']]],
[…]
