+++ cpp/search/all_0.js
@@ -1,26 +1,26 @@
 var searchData=
 [
-  ['a_20kernel_0',['Choosing a Kernel',['../md_docs_2kernels.html#autotoc_md171',1,'']]],
-  ['a_20known_20outlier_1',['Suppress a Known Outlier',['../md_docs_2custom-weights.html#autotoc_md108',1,'']]],
+  ['a_20kernel_0',['Choosing a Kernel',['../md_docs_2kernels.html#autotoc_md173',1,'']]],
+  ['a_20known_20outlier_1',['Suppress a Known Outlier',['../md_docs_2custom-weights.html#autotoc_md110',1,'']]],
   ['a_20policy_2',['Choosing a Policy',['../md_docs_2boundary.html#autotoc_md66',1,'']]],
-  ['a_20scaling_20method_3',['Choosing a Scaling Method',['../md_docs_2scaling.html#autotoc_md272',1,'']]],
-  ['a_20strategy_4',['Choosing a Strategy',['../md_docs_2merge.html#autotoc_md183',1,'']]],
-  ['absolute_20deviation_20default_5',['MAD — Median Absolute Deviation (Default)',['../md_docs_2scaling.html#autotoc_md266',1,'']]],
-  ['absolute_20residual_6',['absolute residual',['../md_docs_2scaling.html#autotoc_md268',1,'MAR — Median Absolute Residual'],['../md_docs_2scaling.html#autotoc_md270',1,'Mean — Mean Absolute Residual']]],
+  ['a_20scaling_20method_3',['Choosing a Scaling Method',['../md_docs_2scaling.html#autotoc_md274',1,'']]],
+  ['a_20strategy_4',['Choosing a Strategy',['../md_docs_2merge.html#autotoc_md185',1,'']]],
+  ['absolute_20deviation_20default_5',['MAD — Median Absolute Deviation (Default)',['../md_docs_2scaling.html#autotoc_md268',1,'']]],
+  ['absolute_20residual_6',['absolute residual',['../md_docs_2scaling.html#autotoc_md270',1,'MAR — Median Absolute Residual'],['../md_docs_2scaling.html#autotoc_md272',1,'Mean — Mean Absolute Residual']]],
   ['acceleration_7',['GPU Acceleration',['../md_docs_2api.html#autotoc_md30',1,'']]],
-  ['adapter_20parameters_8',['Adapter Parameters',['../md_docs_2parameters.html#autotoc_md224',1,'']]],
+  ['adapter_20parameters_8',['Adapter Parameters',['../md_docs_2parameters.html#autotoc_md226',1,'']]],
   ['adapter_2dchoice_2emd_9',['adapter-choice.md',['../adapter-choice_8md.html',1,'']]],
-  ['advantages_10',['Key Advantages',['../index.html#autotoc_md80',1,'']]],
-  ['also_11',['also',['../md_docs_2use-case-real-time.html#autotoc_md305',1,'See Also'],['../md_docs_2custom-weights.html#autotoc_md118',1,'See Also'],['../md_docs_2use-case-genomics.html#autotoc_md287',1,'See Also'],['../md_docs_2gpu-backend.html#autotoc_md129',1,'See also'],['../md_docs_2use-case-time-series.html#autotoc_md323',1,'See Also']]],
-  ['analysis_12',['analysis',['../md_docs_2use-case-time-series.html#autotoc_md317',1,'Multi-Scale Analysis'],['../md_docs_2use-case-time-series.html',1,'Time Series Analysis']]],
+  ['advantages_10',['Key Advantages',['../index.html#autotoc_md82',1,'']]],
[…]
