+++ cpp/search/all_4.js
@@ -1,13 +1,13 @@
 var searchData=
 [
-  ['emphasize_20important_20points_0',['Emphasize Important Points',['../md_docs_2custom-weights.html#autotoc_md110',1,'']]],
+  ['emphasize_20important_20points_0',['Emphasize Important Points',['../md_docs_2custom-weights.html#autotoc_md112',1,'']]],
   ['enabling_20gpu_20support_1',['Enabling GPU Support',['../md_docs_2api.html#autotoc_md31',1,'']]],
-  ['epanechnikov_2',['Epanechnikov',['../md_docs_2kernels.html#autotoc_md159',1,'']]],
-  ['errors_3',['Standard Errors',['../md_docs_2intervals.html#autotoc_md149',1,'']]],
-  ['estimation_4',['Scale Estimation',['../md_docs_2robustness.html#autotoc_md260',1,'']]],
-  ['example_5',['example',['../md_docs_2api.html#autotoc_md48',1,'Example'],['../md_docs_2use-case-real-time.html#autotoc_md297',1,'Real-Time Dashboard Example'],['../md_docs_2use-case-real-time.html#autotoc_md292',1,'Sensor Data Example']]],
+  ['epanechnikov_2',['Epanechnikov',['../md_docs_2kernels.html#autotoc_md161',1,'']]],
+  ['errors_3',['Standard Errors',['../md_docs_2intervals.html#autotoc_md151',1,'']]],
+  ['estimation_4',['Scale Estimation',['../md_docs_2robustness.html#autotoc_md262',1,'']]],
+  ['example_5',['example',['../md_docs_2api.html#autotoc_md48',1,'Example'],['../md_docs_2use-case-real-time.html#autotoc_md299',1,'Real-Time Dashboard Example'],['../md_docs_2use-case-real-time.html#autotoc_md294',1,'Sensor Data Example']]],
   ['execution_20modes_6',['execution modes',['../md_docs_2adapter-choice.html',1,'Execution Modes'],['../index.html#autotoc_md78',1,'Execution Modes']]],
-  ['expression_20time_20course_7',['Gene Expression Time Course',['../md_docs_2use-case-time-series.html#autotoc_md319',1,'']]],
+  ['expression_20time_20course_7',['Gene Expression Time Course',['../md_docs_2use-case-time-series.html#autotoc_md321',1,'']]],
   ['extend_20default_8',['Extend (Default)',['../md_docs_2boundary.html#autotoc_md58',1,'']]],
-  ['extraction_9',['Basic Trend Extraction',['../md_docs_2use-case-time-series.html#autotoc_md309',1,'']]]
+  ['extraction_9',['Basic Trend Extraction',['../md_docs_2use-case-time-series.html#autotoc_md311',1,'']]]
 ];

[…]
