### 2026-08-25 348972e
build: bump to version 3.1.0
