+++ bindings/cpp/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fastlowess-cpp"
-version = "3.0.0"
+version = "3.1.0"
 authors = ["Amir Valizadeh <thisisamirv@gmail.com>"]
 edition = "2024"
 license = "MIT OR Apache-2.0"
@@ -18,7 +18,7 @@ name = "fastlowess_cpp"
 crate-type = ["cdylib", "staticlib"]
 
 [dependencies]
-fastLowess = { path = "../../crates/fastLowess", version = "3.0", features = [
+fastLowess = { path = "../../crates/fastLowess", version = "3.1", features = [
     "cpu",
     "dev",
 ] }
