+++ crates/lowess/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 # Metadata
 name = "lowess"
-version = "3.0.0"
+version = "3.1.0"
 authors = ["Amir Valizadeh <thisisamirv@gmail.com>"]
 edition = "2024"
 rust-version = "1.89"
