### 2026-08-25 db54c28
ci: relax julia/nodejs dependencies during docs build
