+++ .github/workflows/docs.yml
@@ -121,6 +121,9 @@ jobs:
 
       - uses: julia-actions/cache@v3
 
+      - name: Relax JLL dependency for docs build
+        uses: ./.github/actions/relax-julia-jll
+
       - name: Install docs dependencies
         run: |
           julia --project=bindings/julia/julia/docs -e '
@@ -147,12 +150,22 @@ jobs:
     steps:
       - uses: actions/checkout@v7
 
+      - uses: dtolnay/rust-toolchain@stable
+
+      - uses: Swatinem/rust-cache@v2
+        with:
+          key: docs-nodejs
+
       - uses: actions/setup-node@v7
         with:
           node-version: "22"
 
       - name: Install dependencies
         run: npm install
+        working-directory: bindings/nodejs
+
+      - name: Build native addon
+        run: npm run build
         working-directory: bindings/nodejs
 
       - name: Build docs
