### 2026-08-25 6460071
fix: use expect_identical instead of expect_equal in R tests
