### 2026-08-25 #769 @thisisamirv
Hi @etiennebacher, thank you again for the detailed follow-up. A new v3.1.0 was just released that addresses most of the concerns you raised:

## Overview: per-language documentation and CI

> I think this project tries to do too much by displaying the documentation for so many different languages on the same website ... I think it would be better to keep the homepage of this project but then dispatch users to these various documentation websites.

Each binding's documentation has been moved off the shared ReadTheDocs site and onto its own dedicated, idiomatic tool. For R, it now uses **pkgdown**, at <https://thisisamirv.github.io/lowess-project/r/>.

> CI is another place that should probably be split in various workflows.

CI has likewise been split: the monolithic `ci.yml` is now seven per-language workflows (`ci-rust.yml`, `ci-python.yml`, `ci-julia.yml`, `ci-nodejs.yml`, `ci-wasm.yml`, `ci-cpp.yml`, `ci-r.yml`), each independently reviewable and maintainable.

## Setup

> I still hit some errors about package installs ... Why not simply do `Rscript -e "install.packages(c(...))"`?

A simple `Rscript -e "install.packages(c(...))"` is not possible here, for several reasons:

- Not all the dev packages installed in this step are on CRAN. Some are on r-universe servers.
- `install.packages()` on Linux compiles everything from source. It is slow and also requires system libraries. The previous approach had the same system-library requirements, which is probably the reason you ran into issues installing those packages.
- `install.packages()` doesn't raise an error or set a non-zero exit code just because some packages fail to install.
- Without an explicit `lib = ...`, packages install into the user's default personal R library rather than a project-local one, polluting or depending on whatever's already globally installed.
- The script checks `requireNamespace()` first and only installs what's missing, so re-running `make r-dev` doesn't redownload or reinstall everything every time.

The R dev-dependency installation logic has now been replaced with a single [`pak`](https://pak.r-lib.org/)-based block, which handles binary-vs-source selection and CRAN/r-universe sources uniformly. This addresses all the problems the previous installation logic addressed, plus the system-library issue, which is most likely what you encountered.

> The Makefile in bindings/r/Makefile is extremely hard to read and debug. For example, the step "Installing R development packages".

`Makefile` scripts have changed significantly in this version:

- The root `Makefile`, which previously held all the logic for all dependencies, has been split into seven per-language `Makefile`s, each independently reviewable and maintainable.
- More importantly, `make [language name]`, like `make r` (the default target), used to not only build and install a package, but also run a variety of development tools, including linting, documentation builds, vendoring, and more. It now only runs the build and install steps, which is all an end user actually requires. In the case of the R binding, that means running `R CMD INSTALL` only (Rust compilation is handled by the package's own `configure` script from the committed `vendor.tar.xz`). That full dev workflow (formatting, linting, tests, doc-snippet verification) has moved to a separate `make r-dev` target, so a plain `make r`, which is what an end user trying to build and install the package from source needs, should no longer fail the way it did for you.
- Furthermore, I simplified many parts of the development commands in the `Makefile` for R. Updating the vendor archive has been completely moved to a new script named `vendor-update.sh`, which is now easy to read and understand. The R binding has also been completely decoupled from Bioconductor dependencies, as I have decided to target CRAN deployment only.

## Documentation

### Reference page

> there is no page that lists all functions ... I strongly recommend following the standard layout of documentation of R functions

Since R docs are now built with pkgdown, this is addressed structurally rather than by manually reformatting one page.

### `...` documentation and drift between website and `?Lowess`

`...` has been documented via `@param` on `Lowess()`, `StreamingLowess()`, and `OnlineLowess()`, and — since these are now built by pkgdown directly from the package's own roxygen2 docs — there is no longer a separate, hand-maintained website copy of the R reference to drift out of sync with `?Lowess`.

### `dev/verify_snippets.py` and hardcoded output

> `def run_r()` checks whether there are any errors but doesn't actually check the output ... These are hardcoded, meaning that they might not reflect the output

Fair critique, but the purpose of `verify_snippets.py` (which has also been split by language, making it easier to read) is to evaluate the snippets for syntax and execution errors. The numerical accuracy of the package is evaluated by two other methods:

- Comprehensive tests written for each binding.
- Scripts residing in the `validation/` directory at the repository root. These scripts run different scenarios using both fastLowess and R's built-in `stats::lowess()` function, and compare them for numerical accuracy. A new GitHub workflow has also been added to run these validation scripts on every push, making sure future changes don't regress the package's statistical accuracy.

### Broken links

> I still see some links that redirect to a wrongly formatted page ... See for instance the link to `https://lowess.readthedocs.io/benchmarks/`

URLs in several metadata files were double-checked and fixed.

## Code

### `Lowess(fraction = 0.3, 4)` still succeeding

Fixed. `reject_extra_positional_args()` previously counted unnamed arguments without checking their position, so a single unnamed argument in any non-first slot passed validation. It now rejects any unnamed argument that isn't in position 1.

### `fit()`/`add_point()` naming and namespace conflicts

I don't think I want to change this, for a few reasons:

1. Virtually any word with real semantic meaning for these operations (`fit`, `train`, `update`, `add`, `point`, etc.) is already used by some other CRAN package in some form — picking a "unique" name mostly just means picking a less meaningful one.
2. End users should be able to tell what a function does from its name. Something like `fit_lowess()`/`add_point_lowess()` reads fine in isolation, but names like `fit()` and `add_point()` are exactly what a user expects when looking at the method list of a `Lowess`/`OnlineLowess` object, and dispatching on class (as S3 generics do) is the idiomatic way to avoid an actual conflict in the first place.
3. This API is intentionally kept consistent across every binding (Python, Julia, Node.js, WASM, C++, R). Renaming just the R binding's generics would mean either breaking that cross-binding consistency, or renaming across all bindings/crates — a large breaking change for existing users in every language, for a naming concern that doesn't carry over to those other ecosystems (Python/Julia/etc.). I don't think the disruption is proportionate to the benefit here.

### `Depends: R (>= ...)`

`DESCRIPTION` now declares `Depends: R (>= 4.4.0)` (raised from 4.2, since my experiments proved that 4.2 was not well-developed to support complex Rust-based packages like this, requiring many different workarounds), and `ci-r.yml`'s test matrix now runs against both `r: [release, "4.4"]` across ubuntu/macos/windows. R 4.4.0 was released in April 2024, so it's been out for over two years now — the vast majority of users should already be on 4.4 or newer, and this shouldn't meaningfully narrow the pool of people who can install the package.

## Other

Also, not related to the concerns you raised, but still noteworthy, the following changes were made as well:

- The conda build now supports Windows as well. Previously, it was only available on Linux and macOS, since conda-forge's Windows runners were still limited to the 2021 Rust edition.
- I also modified the `README.md` file in each binding/crate to be specific to that binding/crate. This results in not only shorter but more targeted README files — it also means the Rust crates.io URLs that were previously flagged as invalid are now removed, resulting in one less `NOTE` during `R CMD check`.
- `Makevars.win` has now been modified to support Windows arm64 deployment on R-Universe.
- A pre-submission to CRAN resulted in a successful build on their systems. Previously, due to some limitations in the `Makevars.win` file, and also the fact that the CRAN Windows builder only supported the 2021 Rust edition, submitting this package to CRAN was not feasible. The package is now ready to be published on CRAN.

---

Thank you again for the thoroughness here. I hope these changes address your concerns. Please let me know if you think I can improve the package further.

On a separate note, do you think I can go ahead with publishing the package on CRAN, and implement any future changes from rOpenSci reviews as new versions? Or do you believe I should hold off on that for now?

P.S: Since my `LOESS` package is in many ways similar to this `LOWESS` package, I am also actively updating it based on the reviews here, so hopefully, in a future submission of that package, many of these issues won't be present.
