### 2026-08-28 #769 @etiennebacher
Hello @thisisamirv, thank you for the improvements, see other comments below.

### Docs

Thanks for creating a `pkgdown` website, it is already much clearer to me.

Note that the project's homepage (https://lowess.readthedocs.io/en/latest/) seems to be purely Python but it would be good if there was an easily available list of links to redirect users to the docs for each language.

I still think the docs clarity could be improved (from now on, all the comments are about the `pkgdown` website):

1. the current homepage contains duplicated information and has some content that is not properly rendered because of `pkgdown`:
    - the API reference section should be removed now that there is a proper "Reference" page
    - the various `[!TIP]` and `[!NOTE]` are not properly rendered
    - the info about the GPU probably belongs in the "GPU Backend" article so that it doesn't take useful space on the README
    - we could argue that the article about installation instructions should be on the homepage so that people can quickly get started.

    I think using a custom README for this website only would make it better.

2. some "Articles" items are outdated:

    - "Parameters Reference" is no longer needed now that there are proper reference pages
    - similarly, pages such as https://thisisamirv.github.io/lowess-project/r/articles/online.html are basically duplicates of the reference pages.

3. It would be good to have a proper changelog instead of redirecting to the general one so that users don't have to look for the R-related items in the general changelog.

4. Can you add some benchmarks that take more than a few milliseconds? You could adapt those that exist for the GPU approach for instance.

5. On the reference pages:

    - the docs say "Main R6 classes for LOWESS smoothing." but the package doesn't use R6.
    - it would be useful to not assign the last result in the "Examples" section so that users can see what the output look like without having an R session in parallel (e.g. replace `result <- fit(model, x, y)` by `fit(model, x, y)`).
    - does `Nullable()` need to be exported and documented? The docs don't say when to use it.

### Setup

`make r` now works for me.

### Other

> On a separate note, do you think I can go ahead with publishing the package on CRAN, and implement any future changes from rOpenSci reviews as new versions? Or do you believe I should hold off on that for now?

I think you should wait until the rOpenSci review is complete (especially since R-universe gives access to pre-built binaries so it's easy to install the package even if it's not on CRAN).
