### 2026-08-30 0427ba3
docs: minor doc improvements, make README the homepage for docs
