+++ bindings/nodejs/src/content/docs/index.md
@@ -13,3 +13,198 @@ hero:
       icon: open-book
       variant: minimal
 ---
+
+<!-- markdownlint-disable MD024 MD033 -->
+# LOWESS Project
+
+<p align="center">
+  <a href="https://www.npmjs.com/package/fastlowess"><img src="https://img.shields.io/badge/npm-CB3837?logo=npm&logoColor=white" alt="npm"></a>
+  <a href="https://github.com/thisisamirv/lowess-project/actions/workflows/ci-nodejs.yml"><img src="https://github.com/thisisamirv/lowess-project/actions/workflows/ci-nodejs.yml/badge.svg" alt="CI"></a>
+</p>
+
+<p align="center">
+  <img src="https://raw.githubusercontent.com/thisisamirv/lowess-project/main/dev/logo.png" alt="One LOWESS to Rule Them All" width="400">
+  <br>
+  <em>One LOWESS to Rule Them All</em>
+</p>
+
+The fastest, most robust, and most feature-complete language-agnostic LOWESS (Locally Weighted Scatterplot Smoothing) implementation for **Rust**, **Python**, **R**, **Julia**, **JavaScript**, **C++**, and **WebAssembly**.
+
+The `lowess-project` also offers bindings for Rust, Python, R, Julia, Node.js, WebAssembly, and C++ — see the [full repository](https://github.com/thisisamirv/lowess-project).
+
+---
+
+## Installation & Documentation
+
+> Currently available for R, Python, Rust, Julia, Node.js, WebAssembly, and C++. See the [Installation Guide](https://thisisamirv.github.io/lowess-project/nodejs/installation/) for detailed installation instructions.
+>
+> ### 📚 [View the full documentation](https://thisisamirv.github.io/lowess-project/nodejs/)
+
+### GPU Backend
+
+In addition to `parallel = true` (multi-core CPU), the batch `Lowess` class in every binding — except WebAssembly — as well as the `fastLowess` Rust crate itself, can run on the GPU via `wgpu` (Vulkan/Metal/DX12). It's opt-in and worth enabling for high-throughput processing of large datasets (roughly 10k+ points); for smaller inputs the CPU backend is typically faster. `StreamingLowess`/`OnlineLowess` remain CPU-only. See the [GPU Backend guide](https://thisisamirv.github.io/lowess-project/nodejs/gpu-backend/) for installation instructions and usage.
+
+---
+
+## LOESS vs. LOWESS
+
+| Feature | LOESS | LOWESS (This Crate) |
+| --- | --- | --- |
+| **Polynomial Degree** | Linear, Quadratic, Cubic, Quartic | Linear (Degree 1) |
+| **Dimensions** | Multivariate (n-D support) | Univariate (1-D only) |
+| **Flexibility** | High (Distance metrics) | Standard |
+| **Complexity** | Higher (Matrix inversion) | Lower (Weighted average/slope) |
+
+> [!TIP]
+> **Note:** For a **LOESS** implementation, use [`loess-project`](https://github.com/thisisamirv/loess-project).
+
+---
+
+## Why this package?
+
+### Speed
+
+The `lowess` project beats the competition in terms of speed, whether in single-threaded or multi-threaded parallel execution. It is on average **200-327x faster** than Python's `statsmodels.lowess` and **2-3x faster** than R's `lowess`.
+
+For more details on the performance comparison, see the [Benchmarks](https://thisisamirv.github.io/lowess-project/nodejs/benchmarks/) page.
+
+### Robustness
+
+This implementation is *more robust* than R's `lowess` and Python's `statsmodels` due to two key design choices:
+
+**MAD-Based Scale Estimation:**
+
[…]
