+++ cpp/search/all_17.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['zero_0',['Zero',['../md_docs_2boundary.html#autotoc_md62',1,'']]],
-  ['zero_5fweight_5ffallback_1',['zero_weight_fallback',['../md_docs_2api.html#autotoc_md45',1,'zero_weight_fallback'],['../md_docs_2parameters.html#autotoc_md206',1,'zero_weight_fallback']]]
+  ['zero_0',['Zero',['../md_docs_2boundary.html#autotoc_md81',1,'']]],
+  ['zero_5fweight_5ffallback_1',['zero_weight_fallback',['../md_docs_2api.html#autotoc_md64',1,'zero_weight_fallback'],['../md_docs_2parameters.html#autotoc_md225',1,'zero_weight_fallback']]]
 ];
