+++ cpp/search/all_19.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['📚_20a_20href_20https_3a_20thisisamirv_20github_20io_20lowess_20project_20cpp_20view_20the_20full_20documentation_20a_0',['📚 &lt;a href=&quot;https://thisisamirv.github.io/lowess-project/cpp/&quot; &gt;View the full documentation&lt;/a&gt;',['../index.html#autotoc_md3',1,'']]]
+];
