+++ cpp/search/pages_2.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['concepts_0',['Concepts',['../index.html',1,'']]],
+  ['concepts_0',['Concepts',['../md_docs_2concepts.html',1,'']]],
   ['cross_20validation_1',['Cross-Validation',['../md_docs_2cross-validation.html',1,'']]],
   ['custom_20weights_2',['Custom Weights',['../md_docs_2custom-weights.html',1,'']]]
 ];
