+++ cpp/search/all_a.js
@@ -1,12 +1,16 @@
 var searchData=
 [
-  ['large_20genome_20coverage_20streaming_0',['Large Genome Coverage (Streaming)',['../md_docs_2use-case-genomics.html#autotoc_md285',1,'']]],
-  ['last_1',['Take Last',['../md_docs_2merge.html#autotoc_md181',1,'']]],
-  ['leave_20one_20out_20loocv_2',['Leave-One-Out (LOOCV)',['../md_docs_2cross-validation.html#autotoc_md90',1,'']]],
-  ['levels_3',['Confidence Levels',['../md_docs_2intervals.html#autotoc_md149',1,'']]],
-  ['linux_20x64_4',['Pre-built Binaries (Linux (x64))',['../md_docs_2installation.html#autotoc_md133',1,'']]],
-  ['log_20file_20processing_5',['Log File Processing',['../md_docs_2use-case-real-time.html#autotoc_md297',1,'']]],
-  ['loocv_6',['Leave-One-Out (LOOCV)',['../md_docs_2cross-validation.html#autotoc_md90',1,'']]],
-  ['lowess_7',['What is LOWESS?',['../index.html#autotoc_md68',1,'']]],
-  ['lowessoptions_20tt_8',['lowessoptions tt',['../md_docs_2api.html#autotoc_md29',1,'&lt;tt&gt;LowessOptions&lt;/tt&gt;'],['../md_docs_2api-online.html#autotoc_md8',1,'&lt;tt&gt;OnlineOptions&lt;/tt&gt; (inherits &lt;tt&gt;LowessOptions&lt;/tt&gt;)'],['../md_docs_2api-streaming.html#autotoc_md20',1,'&lt;tt&gt;StreamingOptions&lt;/tt&gt; (inherits &lt;tt&gt;LowessOptions&lt;/tt&gt;)']]]
+  ['large_20genome_20coverage_20streaming_0',['Large Genome Coverage (Streaming)',['../md_docs_2use-case-genomics.html#autotoc_md304',1,'']]],
+  ['last_1',['Take Last',['../md_docs_2merge.html#autotoc_md200',1,'']]],
+  ['leave_20one_20out_20loocv_2',['Leave-One-Out (LOOCV)',['../md_docs_2cross-validation.html#autotoc_md109',1,'']]],
+  ['levels_3',['Confidence Levels',['../md_docs_2intervals.html#autotoc_md168',1,'']]],
+  ['license_4',['License',['../index.html#autotoc_md17',1,'']]],
+  ['linux_20x64_5',['Pre-built Binaries (Linux (x64))',['../md_docs_2installation.html#autotoc_md152',1,'']]],
+  ['loess_20vs_20lowess_6',['LOESS vs. LOWESS',['../index.html#autotoc_md6',1,'']]],
+  ['log_20file_20processing_7',['Log File Processing',['../md_docs_2use-case-real-time.html#autotoc_md316',1,'']]],
+  ['loocv_8',['Leave-One-Out (LOOCV)',['../md_docs_2cross-validation.html#autotoc_md109',1,'']]],
+  ['lowess_9',['lowess',['../index.html#autotoc_md6',1,'LOESS vs. LOWESS'],['../md_docs_2concepts.html#autotoc_md87',1,'What is LOWESS?']]],
+  ['lowess_20project_10',['LOWESS Project',['../index.html#autotoc_md0',1,'']]],
[…]
