+++ cpp/search/searchdata.js
@@ -1,6 +1,6 @@
 var indexSectionsWithContent =
 {
-  0: "abcdefghiklmnopqrstuvwxz—",
+  0: "abcdefghiklmnopqrstuvwxz—📚",
   1: "abcgikmpqrsu",
   2: "abcdefghimopqrstvw"
 };

[…]
