+++ r/index.md
@@ -22,32 +22,21 @@ repository](https://github.com/thisisamirv/lowess-project).
 
 ------------------------------------------------------------------------
 
-## Installation
+## Installation & Documentation
 
-> \[!NOTE\]
->
 > Currently available for R, Python, Rust, Julia, Node.js, WebAssembly,
 > and C++. See the [Installation
 > Guide](https://thisisamirv.github.io/lowess-project/r/articles/installation.html)
 > for detailed installation instructions.
+>
+> ### 📚 [View the full documentation](https://thisisamirv.github.io/lowess-project/r/)
 
 ### GPU Backend
 
-In addition to `parallel = true` (multi-core CPU), the batch `Lowess`
-class in every binding — except WebAssembly — as well as the
-`fastLowess` Rust crate itself, can run on the GPU via `wgpu`
-(Vulkan/Metal/DX12). It’s opt-in and worth enabling for high-throughput
-processing of large datasets (roughly 10k+ points); for smaller inputs
-the CPU backend is typically faster. `StreamingLowess`/`OnlineLowess`
-remain CPU-only. See the [GPU Backend
+GPU acceleration (`wgpu`: Vulkan/Metal/DX12) is also supported for
+high-throughput batch smoothing. See the [GPU Backend
 guide](https://thisisamirv.github.io/lowess-project/r/articles/gpu-backend.html)
-for installation instructions and usage.
-
-## Documentation
-
-> \[!NOTE\]
->
-> ### 📚 [View the full documentation](https://thisisamirv.github.io/lowess-project/r/)
+for details.
 
 ------------------------------------------------------------------------
 
@@ -60,7 +49,7 @@ for installation instructions and usage.
 | **Flexibility** | High (Distance metrics) | Standard |
 | **Complexity** | Higher (Matrix inversion) | Lower (Weighted average/slope) |
 
-> \[!TIP\] **Note:** For a **LOESS** implementation, use
+> **Note:** For a **LOESS** implementation, use
 > [`loess-project`](https://github.com/thisisamirv/loess-project).
 
 ------------------------------------------------------------------------
@@ -153,54 +142,6 @@ All implementations are **numerical twins** of R’s `lowess`:
 | **Accuracy** | ✅ EXACT MATCH | Max diff \< 1e-12 across all scenarios |
 | **Consistency** | ✅ PERFECT | Multiple scenarios pass with strict tolerance |
 | **Robustness** | ✅ VERIFIED | Robust smoothing matches R exactly |
-
-## API Reference
-
-``` r
-library(rfastlowess)
-
-model <- Lowess(
-    fraction = 0.5,
-    iterations = 3L,
-    delta = 0.01,
-    weight_function = "tricube",
-    robustness_method = "bisquare",
-    scaling_method = "mad",
-    zero_weight_fallback = "use_local_mean",
-    boundary_policy = "extend",
-    confidence_intervals = 0.95,
-    prediction_intervals = 0.95,
-    return_diagnostics = TRUE,
-    return_residuals = TRUE,
-    return_robustness_weights = TRUE,
-    return_se = TRUE,
-    cv_fractions = c(0.3, 0.5, 0.7),
-    cv_method = "kfold",
-    cv_k = 5L,
-    cv_seed = 123L,
-    auto_converge = 1e-4,
-    parallel = TRUE
-)
-custom_weights <- rep(1, length(x))
-result <- fit(model, x, y, custom_weights = custom_weights)
-
-# Result structure:
-result$x,
-result$y,
-result$standard_errors,
-result$confidence_lower,
-result$confidence_upper,
-result$prediction_lower,
-result$prediction_upper,
-result$residuals,
-result$robustness_weights,
-result$diagnostics,
[…]
