### 2026-08-30 f319e97
docs: more docs improvements, especially cpp
