+++ bindings/r/R/Nullable.R
@@ -1,15 +1,5 @@
-#' Nullable Value Wrapper
-#'
-#' @description
-#' Wraps a value to be passed to Rust as an Option.
-#'
-#' @param x Value to wrap or NULL.
-#'
-#' @return The value itself. This is a helper for rextendr conversion.
-#' @examples
-#' Nullable(5)
-#' Nullable(NULL)
-#' @export
+#' Wrap a value to be passed to Rust as an Option
+#' @noRd
 Nullable <- function(x) {
     if (is.null(x)) {
         return(NULL)
