+++ bindings/r/R/rfastlowess-package.R
@@ -31,4 +31,5 @@
 #' @useDynLib rfastlowess, .registration = TRUE
 #' @importFrom stats smooth
 #' @importFrom stats setNames
+#' @keywords internal
 "_PACKAGE"
