### 2026-08-30 3c9dfbf
docs: add theme to cpp docs
