+++ cpp/search/all_14.js
@@ -1,10 +1,9 @@
 var searchData=
 [
-  ['validation_0',['validation',['../md_docs_2cross-validation.html',1,'Cross-Validation'],['../md_docs_2cross-validation.html#autotoc_md107',1,'K-Fold Cross-Validation'],['../index.html#autotoc_md12',1,'Validation']]],
-  ['validation_20rules_1',['Validation Rules',['../md_docs_2custom-weights.html#autotoc_md137',1,'']]],
-  ['verify_20installation_2',['Verify Installation',['../md_docs_2installation.html#autotoc_md158',1,'']]],
-  ['view_20the_20full_20documentation_20a_3',['📚 &lt;a href=&quot;https://thisisamirv.github.io/lowess-project/cpp/&quot; &gt;View the full documentation&lt;/a&gt;',['../index.html#autotoc_md3',1,'']]],
-  ['vs_20lowess_4',['LOESS vs. LOWESS',['../index.html#autotoc_md6',1,'']]],
-  ['vs_20prediction_20intervals_5',['Confidence vs Prediction Intervals',['../md_docs_2concepts.html#autotoc_md95',1,'']]],
-  ['vs_20robustness_20iterations_6',['Custom Weights vs. Robustness Iterations',['../md_docs_2custom-weights.html#autotoc_md126',1,'']]]
+  ['validation_0',['validation',['../md_docs_2cross-validation.html',1,'Cross-Validation'],['../md_docs_2cross-validation.html#autotoc_md106',1,'K-Fold Cross-Validation'],['../index.html#autotoc_md11',1,'Validation']]],
+  ['validation_20rules_1',['Validation Rules',['../md_docs_2custom-weights.html#autotoc_md136',1,'']]],
+  ['verify_20installation_2',['Verify Installation',['../md_docs_2installation.html#autotoc_md157',1,'']]],
+  ['vs_20lowess_3',['LOESS vs. LOWESS',['../index.html#autotoc_md5',1,'']]],
+  ['vs_20prediction_20intervals_4',['Confidence vs Prediction Intervals',['../md_docs_2concepts.html#autotoc_md94',1,'']]],
+  ['vs_20robustness_20iterations_5',['Custom Weights vs. Robustness Iterations',['../md_docs_2custom-weights.html#autotoc_md125',1,'']]]
 ];
