+++ cpp/navtreedata.js
@@ -27,283 +27,287 @@ var NAVTREE =
   [ "fastLowess", "index.html", [
     [ "LOWESS Project", "index.html#autotoc_md0", [
       [ "Installation & Documentation", "index.html#autotoc_md2", [
-        [ "📚 <a href=\"https://thisisamirv.github.io/lowess-project/cpp/\" >View the full documentation</a>", "index.html#autotoc_md3", null ],
-        [ "GPU Backend", "index.html#autotoc_md4", null ]
+        [ "GPU Backend", "index.html#autotoc_md3", null ]
       ] ],
-      [ "LOESS vs. LOWESS", "index.html#autotoc_md6", null ],
-      [ "</blockquote>", "index.html#autotoc_md7", null ],
-      [ "Why this package?", "index.html#autotoc_md8", [
-        [ "Speed", "index.html#autotoc_md9", null ],
-        [ "Robustness", "index.html#autotoc_md10", null ],
-        [ "Features", "index.html#autotoc_md11", null ]
+      [ "LOESS vs. LOWESS", "index.html#autotoc_md5", null ],
+      [ "</blockquote>", "index.html#autotoc_md6", null ],
+      [ "Why this package?", "index.html#autotoc_md7", [
+        [ "Speed", "index.html#autotoc_md8", null ],
+        [ "Robustness", "index.html#autotoc_md9", null ],
+        [ "Features", "index.html#autotoc_md10", null ]
       ] ],
-      [ "Validation", "index.html#autotoc_md12", null ],
-      [ "API Reference", "index.html#autotoc_md13", null ],
-      [ "Contributing", "index.html#autotoc_md15", null ],
-      [ "Changelog", "index.html#autotoc_md16", null ],
-      [ "License", "index.html#autotoc_md17", null ],
-      [ "Citation", "index.html#autotoc_md18", null ]
+      [ "Validation", "index.html#autotoc_md11", null ],
+      [ "API Reference", "index.html#autotoc_md12", null ],
+      [ "Contributing", "index.html#autotoc_md14", null ],
+      [ "Changelog", "index.html#autotoc_md15", null ],
+      [ "License", "index.html#autotoc_md16", null ],
+      [ "Citation", "index.html#autotoc_md17", null ]
     ] ],
     [ "Execution Modes", "md_docs_2adapter-choice.html", [
-      [ "Overview", "md_docs_2adapter-choice.html#autotoc_md20", null ]
+      [ "Overview", "md_docs_2adapter-choice.html#autotoc_md19", null ]
     ] ],
     [ "OnlineLowess API", "md_docs_2api-online.html", [
-      [ "When to Use", "md_docs_2api-online.html#autotoc_md23", null ],
-      [ "Class", "md_docs_2api-online.html#autotoc_md24", [
-        [ "<tt>fastlowess::OnlineLowess</tt>", "md_docs_2api-online.html#autotoc_md25", null ]
+      [ "When to Use", "md_docs_2api-online.html#autotoc_md22", null ],
+      [ "Class", "md_docs_2api-online.html#autotoc_md23", [
+        [ "fastlowess::OnlineLowess", "md_docs_2api-online.html#autotoc_md24", null ]
       ] ],
-      [ "Options Structure", "md_docs_2api-online.html#autotoc_md26", [
-        [ "<tt>OnlineOptions</tt> (inherits <tt>LowessOptions</tt>)", "md_docs_2api-online.html#autotoc_md27", null ]
+      [ "Options Structure", "md_docs_2api-online.html#autotoc_md25", [
[…]
