+++ cpp/search/all_18.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['—_20mean_20absolute_20residual_0',['Mean — Mean Absolute Residual',['../md_docs_2scaling.html#autotoc_md291',1,'']]],
-  ['—_20median_20absolute_20deviation_20default_1',['MAD — Median Absolute Deviation (Default)',['../md_docs_2scaling.html#autotoc_md287',1,'']]],
-  ['—_20median_20absolute_20residual_2',['MAR — Median Absolute Residual',['../md_docs_2scaling.html#autotoc_md289',1,'']]]
+  ['—_20mean_20absolute_20residual_0',['Mean — Mean Absolute Residual',['../md_docs_2scaling.html#autotoc_md290',1,'']]],
+  ['—_20median_20absolute_20deviation_20default_1',['MAD — Median Absolute Deviation (Default)',['../md_docs_2scaling.html#autotoc_md286',1,'']]],
+  ['—_20median_20absolute_20residual_2',['MAR — Median Absolute Residual',['../md_docs_2scaling.html#autotoc_md288',1,'']]]
 ];
