+++ cpp/search/all_d.js
@@ -1,18 +1,18 @@
 var searchData=
 [
-  ['one_20out_20loocv_0',['Leave-One-Out (LOOCV)',['../md_docs_2cross-validation.html#autotoc_md109',1,'']]],
-  ['online_20mode_1',['Online Mode',['../md_docs_2use-case-real-time.html#autotoc_md321',1,'']]],
-  ['online_20mode_3a_20point_20by_20point_2',['Online Mode: Point-by-Point',['../md_docs_2use-case-real-time.html#autotoc_md312',1,'']]],
+  ['one_20out_20loocv_0',['Leave-One-Out (LOOCV)',['../md_docs_2cross-validation.html#autotoc_md108',1,'']]],
+  ['online_20mode_1',['Online Mode',['../md_docs_2use-case-real-time.html#autotoc_md320',1,'']]],
+  ['online_20mode_3a_20point_20by_20point_2',['Online Mode: Point-by-Point',['../md_docs_2use-case-real-time.html#autotoc_md311',1,'']]],
   ['onlinelowess_20api_3',['OnlineLowess API',['../md_docs_2api-online.html',1,'']]],
-  ['onlineoptions_20tt_20inherits_20tt_20lowessoptions_20tt_4',['&lt;tt&gt;OnlineOptions&lt;/tt&gt; (inherits &lt;tt&gt;LowessOptions&lt;/tt&gt;)',['../md_docs_2api-online.html#autotoc_md27',1,'']]],
-  ['options_5',['options',['../md_docs_2api-online.html#autotoc_md30',1,'Options'],['../md_docs_2api-streaming.html#autotoc_md40',1,'Options'],['../md_docs_2api.html#autotoc_md59',1,'Options'],['../md_docs_2parameters.html#autotoc_md231',1,'Output Options']]],
-  ['options_20structure_6',['options structure',['../md_docs_2api-online.html#autotoc_md26',1,'Options Structure'],['../md_docs_2api-streaming.html#autotoc_md38',1,'Options Structure']]],
-  ['options_20structures_7',['Options Structures',['../md_docs_2api.html#autotoc_md47',1,'']]],
-  ['options_20summary_8',['Parameter Options Summary',['../md_docs_2parameters.html#autotoc_md208',1,'']]],
-  ['out_20loocv_9',['Leave-One-Out (LOOCV)',['../md_docs_2cross-validation.html#autotoc_md109',1,'']]],
-  ['outlier_10',['Suppress a Known Outlier',['../md_docs_2custom-weights.html#autotoc_md129',1,'']]],
-  ['outliers_11',['outliers',['../md_docs_2robustness.html#autotoc_md279',1,'Detecting Outliers'],['../md_docs_2quickstart.html#autotoc_md262',1,'Handling Outliers']]],
-  ['output_20options_12',['Output Options',['../md_docs_2parameters.html#autotoc_md231',1,'']]],
-  ['overlap_13',['overlap',['../md_docs_2parameters.html#autotoc_md248',1,'']]],
[…]

[…]
