+++ r/llms.txt
@@ -49,6 +49,9 @@ for details.
 | **Flexibility** | High (Distance metrics) | Standard |
 | **Complexity** | Higher (Matrix inversion) | Lower (Weighted average/slope) |
 
+Read more about how LOWESS works in the
+[Concepts](https://thisisamirv.github.io/lowess-project/r/articles/concepts.html).
+
 > **Note:** For a **LOESS** implementation, use
 > [`loess-project`](https://github.com/thisisamirv/loess-project).
 
@@ -184,9 +187,6 @@ file or the BibTeX entry below:
 
 Main R6 classes for LOWESS smoothing.
 
-- [`rfastlowess`](https://thisisamirv.github.io/lowess-project/r/reference/rfastlowess-package.md)
-  [`rfastlowess-package`](https://thisisamirv.github.io/lowess-project/r/reference/rfastlowess-package.md)
-  : rfastlowess: High-performance LOWESS Smoothing for R
 - [`Lowess()`](https://thisisamirv.github.io/lowess-project/r/reference/Lowess.md)
   : LOWESS Batch Smoothing
 - [`StreamingLowess()`](https://thisisamirv.github.io/lowess-project/r/reference/StreamingLowess.md)
@@ -216,13 +216,6 @@ Objects returned by fit methods and helper functions.
   : Print OnlineLowess Model
 - [`print(`*`<StreamingLowess>`*`)`](https://thisisamirv.github.io/lowess-project/r/reference/print.StreamingLowess.md)
   : Print StreamingLowess Model
-
-## Types
-
-Helper types for documentation.
-
-- [`Nullable()`](https://thisisamirv.github.io/lowess-project/r/reference/Nullable.md)
-  : Nullable Value Wrapper
 
 ## GPU Backend
 
