### 2026-08-30 49092c0
ci: consolidate docs.yml into a single gh-pages deploy per run
