+++ .github/workflows/docs.yml
@@ -55,17 +55,23 @@ jobs:
               - 'bindings/wasm/**'
               - '.github/workflows/docs.yml'
 
+  # Each of the jobs below only builds its own docs and uploads the result as a
+  # workflow artifact — nothing is pushed to gh-pages here. This lets them all
+  # run in parallel, and keeps GitHub's own "pages build and deployment" from
+  # firing once per job. The `deploy` job merges every artifact into a single
+  # commit at the end, so the branch (and its auto-triggered Pages build) is
+  # only touched once per workflow run.
+
   root:
     needs: filter
+    if: needs.filter.outputs.root == 'true' || github.event_name == 'workflow_dispatch'
     name: Root — README
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v7
       - name: Install pandoc
-        if: needs.filter.outputs.root == 'true' || github.event_name == 'workflow_dispatch'
         run: sudo apt-get install -y pandoc
       - name: Convert README.md to HTML
-        if: needs.filter.outputs.root == 'true' || github.event_name == 'workflow_dispatch'
         run: |
           mkdir -p _site
           pandoc README.md \
@@ -76,18 +82,15 @@ jobs:
             --css=https://cdnjs.cloudflare.com/ajax/libs/github-markdown-css/5.6.1/github-markdown-light.min.css \
             -o _site/index.html
           sed -i 's|<body>|<body class="markdown-body" style="max-width:980px;margin:0 auto;padding:45px">|' _site/index.html
-      - name: Ensure .nojekyll
-        run: mkdir -p _site && touch _site/.nojekyll
-      - uses: JamesIves/github-pages-deploy-action@v4
+      - uses: actions/upload-artifact@v7
         with:
-          branch: gh-pages
-          folder: _site
-          target-folder: .
-          clean: false
+          name: docs-root
+          path: _site
+          retention-days: 1
 
   cpp:
-    needs: [filter, root]
-    if: always() && (needs.filter.outputs.cpp == 'true' || github.event_name == 'workflow_dispatch')
+    needs: filter
+    if: needs.filter.outputs.cpp == 'true' || github.event_name == 'workflow_dispatch'
     name: C++ — Doxygen
     runs-on: ubuntu-latest
     steps:
@@ -100,16 +103,15 @@ jobs:
         run: doxygen Doxyfile
         working-directory: bindings/cpp
 
-      - uses: JamesIves/github-pages-deploy-action@v4
+      - uses: actions/upload-artifact@v7
         with:
-          branch: gh-pages
-          folder: bindings/cpp/html
-          target-folder: cpp
-          clean: false
+          name: docs-cpp
+          path: bindings/cpp/html
+          retention-days: 1
 
   julia:
-    needs: [filter, cpp]
-    if: always() && (needs.filter.outputs.julia == 'true' || github.event_name == 'workflow_dispatch')
+    needs: filter
+    if: needs.filter.outputs.julia == 'true' || github.event_name == 'workflow_dispatch'
     name: Julia — Documenter.jl
     runs-on: ubuntu-latest
     steps:
@@ -135,16 +137,15 @@ jobs:
       - name: Build docs
         run: julia --project=bindings/julia/julia/docs bindings/julia/julia/docs/make.jl
 
-      - uses: JamesIves/github-pages-deploy-action@v4
+      - uses: actions/upload-artifact@v7
         with:
-          branch: gh-pages
-          folder: bindings/julia/julia/docs/build
-          target-folder: julia
-          clean: false
+          name: docs-julia
+          path: bindings/julia/julia/docs/build
+          retention-days: 1
 
   nodejs:
-    needs: [filter, julia]
[…]
