### 2026-08-30 eb5c80d
docs: add katex to rust crates
