### 2026-08-30 89fac48
docs: add news with specialized changelog to each binding/crate
