+++ dev/update_changelogs.py
@@ -1,0 +1,173 @@
+#!/usr/bin/env python3
+"""Regenerate per-binding/crate changelog files from the root CHANGELOG.md.
+
+The root CHANGELOG.md documents every language/binding together, grouped as:
+
+    ## <version>
+    ### <Added|Changed|Fixed|Removed>
+    **<Language>:**
+    - <bullet>
+
+For each target below, this script extracts only the bullets listed under its
+matching "**Label:**" heading(s) for every version, and writes the result to
+the target's output file using a CRAN-NEWS-style format:
+
+    # <package> <version>
+    ## <Added|Changed|Fixed|Removed>
+    * <bullet>
+
+Versions with no matching entries are skipped entirely.
+"""
+
+import re
+from pathlib import Path
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+CHANGELOG_PATH = REPO_ROOT / "CHANGELOG.md"
+DESCRIPTION_PATH = REPO_ROOT / "bindings" / "r" / "DESCRIPTION"
+REPO_URL = "https://github.com/thisisamirv/lowess-project"
+
+VERSION_RE = re.compile(r"^## (\[Unreleased\]|\S+)")
+SECTION_RE = re.compile(r"^### (\w+)")
+LANG_RE = re.compile(r"^\*\*([^*]+):\*\*$")
+BULLET_RE = re.compile(r"^- (.*)$")
+
+# Each target's `labels` are matched against the "**Label:**" headings in
+# CHANGELOG.md (case-sensitive, exact match). `package` is the display name
+# used in the "# <package> <version>" heading; use None to resolve it from
+# bindings/r/DESCRIPTION's Package field instead.
+TARGETS = [
+    {"labels": ["R"], "package": None, "output": "bindings/r/NEWS.md"},
+    {
+        "labels": ["C++"],
+        "package": "fastlowess (C++)",
+        "output": "bindings/cpp/docs/NEWS.md",
+    },
+    {
+        "labels": ["Julia"],
+        "package": "FastLOWESS.jl",
+        "output": "bindings/julia/julia/docs/src/NEWS.md",
+    },
+    {
+        "labels": ["Node.js"],
+        "package": "fastlowess (Node.js)",
+        "output": "bindings/nodejs/src/content/docs/NEWS.md",
+        "frontmatter_title": "News",
+    },
+    {
+        "labels": ["Python"],
+        "package": "fastlowess (Python)",
+        "output": "bindings/python/docs/NEWS.md",
+    },
+    {
+        "labels": ["WASM", "WebAssembly"],
+        "package": "fastlowess-wasm",
+        "output": "bindings/wasm/src/content/docs/NEWS.md",
+        "frontmatter_title": "News",
+    },
+    {
+        "labels": ["fastLowess", "lowess and fastLowess", "Rust"],
+        "package": "fastLowess",
+        "output": "crates/fastLowess/docs/news.md",
+    },
+    {
+        "labels": ["lowess", "lowess and fastLowess", "Rust"],
+        "package": "lowess",
+        "output": "crates/lowess/docs/news.md",
+    },
+]
+
+
+def read_description_field(name: str) -> str | None:
+    text = DESCRIPTION_PATH.read_text(encoding="utf-8")
+    for line in text.splitlines():
+        if line.startswith(f"{name}:"):
+            return line.split(":", 1)[1].strip()
+    return None
+
+
+def changelog_url() -> str:
+    bug_reports = read_description_field("BugReports")
+    if bug_reports and bug_reports.endswith("/issues"):
+        return bug_reports[: -len("issues")] + "blob/main/CHANGELOG.md"
[…]
