+++ bindings/nodejs/src/content/docs/NEWS.md
@@ -1,0 +1,120 @@
+---
+title: News
+---
+<!-- markdownlint-disable MD024 MD025 -->
+# fastlowess (Node.js) (development version)
+
+## Changed
+
+* Updated `typedoc-plugin-markdown` to v4.13.
+* `make nodejs-dev` now runs `npm update` after `npm install`, so dependencies are kept current.
+
+## Fixed
+
+* Fixed the docs homepage's "Get Started" button jumping straight to the Installation page without ever showing the README content. `README.md` is now embedded on the Starlight homepage below the hero via a new `dev/add-readme-to-docs.js` script, wired into both `npm run docs` and `make nodejs-dev`.
+
+# fastlowess (Node.js) 3.1.0
+
+## Added
+
+* Added `npm run lint` to the `Lint` step in `ci-nodejs.yml`, so JavaScript source and test files are linted via `oxlint` on every CI run.
+
+## Changed
+
+* Moved Node.js documentation from ReadTheDocs to GitHub Pages, served by Starlight at <https://thisisamirv.github.io/lowess-project/nodejs/>. The ReadTheDocs site no longer includes Node.js-specific content. `dev/add-nodejs-outputs.js` runs as part of `make nodejs-dev`, executing each JavaScript code block in the docs and injecting its output back into the Markdown source.
+* `make nodejs` (`default:`) now builds the native addon and links it globally via `npm link`. The full dev workflow moves to `make nodejs-dev`.
+* Updated `oxlint` dependency to 1.80.
+
+# fastlowess (Node.js) 3.0.0
+
+## Added
+
+* Exposed a `backend` option (`"cpu"` default, `"gpu"`) on `Lowess`, gated behind an opt-in `gpu` Cargo feature not enabled in published npm binaries. Run `await fastlowess.installGpu()` to download a prebuilt GPU addon (requires restarting Node.js), or build locally with `napi build --features gpu`.
+
+## Changed
+
+* Renamed `OnlineOutput`'s `smoothed` and `std_error` fields to `y` and `standard_error`. This is a **breaking change**.
+* Updated `@napi-rs/cli` to v3.8 and `oxlint` to v1.79.
+
+# fastlowess (Node.js) 2.0.0
+
+## Added
+
+* Added `OnlineOutput` object to the Node.js binding. `OnlineLowess.add_point()` now returns `OnlineOutput | null` instead of `number | null`, exposing `smoothed`, `std_error`, `residual`, `robustness_weight`, and `iterations_used`.
+* Added `return_se` and `cv_seed` fields to `SmoothOptions`.
+* Added `customWeights` as an optional per-call argument to `fit(x, y, customWeights?)` and `fit_async(x, y, customWeights?)`. Accepts a `Float64Array` of non-negative per-observation weights. Includes pre-flight length-mismatch and non-negative validation. Batch only.
+* Added JavaScript-layer option key validation: unknown keys in `SmoothOptions`, `StreamingOptions`, or `OnlineOptions` now throw a `TypeError` listing all valid keys, via wrapper classes around the native NAPI exports.
+
+## Changed
+
+* Renamed all `Diagnostics`, `SmoothOptions`, `StreamingOptions`, and `OnlineOptions` interface fields to snake_case (`r_squared`, `effective_df`, `residual_sd`, `chunk_size`, `merge_strategy`, `window_capacity`, `min_points`, `update_mode`, and all smoothing option fields).
+* Renamed binding methods to snake_case: `fit_async`, `process_chunk`, `add_points`.
[…]
