### 2026-08-30 a830143
docs: remove changelog link from readme files
