### 2026-08-30 2db5213
docs: cosolidate R api docs into roxygen docs
