+++ bindings/r/R/StreamingLowess.R
@@ -1,14 +1,35 @@
 #' LOWESS Streaming Smoothing
 #'
 #' @description
-#' Create a stateful LOWESS model for streaming data.
+#' Create a stateful LOWESS model for streaming data. Processes data in
+#' fixed-size chunks with configurable overlap: results for each chunk are
+#' returned by \code{\link{process_chunk}}, and \code{\link{finalize}}
+#' flushes any remaining buffered points after the last chunk.
+#'
+#' @details
+#' Best suited for datasets over 100,000 points, memory-constrained
+#' environments, or batch processing pipelines. For smaller datasets that fit
+#' in memory, see \code{\link{Lowess}}; for point-by-point real-time data,
+#' see \code{\link{OnlineLowess}}.
+#'
+#' Overlapping regions between chunks are reconciled via `merge_strategy`:
+#'
+#' | Strategy | Behavior |
+#' | --- | --- |
+#' | `"average"` | Arithmetic mean of both estimates |
+#' | `"weighted_average"` | Distance-weighted blend (recommended, default) |
+#' | `"take_first"` | Keep left-chunk estimate |
+#' | `"take_last"` | Keep right-chunk estimate |
 #'
 #' @srrstats {G2.0} Input validation for fraction, chunk_size.
 #' @srrstats {G1.6} Memory-efficient streaming for large datasets.
 #'
 #' @inheritParams Lowess
-#' @param chunk_size Number of data points per processing chunk.
-#' @param overlap Number of overlapping points between consecutive chunks.
+#' @param chunk_size Number of data points per processing chunk, at least 10.
+#'   Default: 5000.
+#' @param overlap Number of overlapping points between consecutive chunks,
+#'   less than \code{chunk_size}. \code{NULL} (default) uses the backend's
+#'   default of 500.
 #' @param merge_strategy Strategy for reconciling overlapping chunk regions:
 #'   \code{"weighted_average"} (default; alias: \code{"weighted"}),
 #'   \code{"average"} (alias: \code{"mean"}),
@@ -60,5 +81,95 @@ StreamingLowess <- function(
             )
         ),
         class = "StreamingLowess"
+    )
+}
+code{"zero"}, or
+#'   \code{"noboundary"} (alias: \code{"none"}).
+#' @param confidence_intervals Confidence level for confidence intervals,
+#'   greater than 0 and less than 1 (e.g., 0.95). \code{NULL} (default)
+#'   disables confidence intervals.
+#' @param prediction_intervals Confidence level for prediction intervals,
+#'   greater than 0 and less than 1 (e.g., 0.95). \code{NULL} (default)
+#'   disables prediction intervals.
+#' @param return_diagnostics Logical; if \code{TRUE}, return fit-quality
+#'   metrics (RMSE, MAE, R-squared, AIC, etc.). Default: \code{FALSE}.
+#' @param return_residuals Logical; if \code{TRUE}, return residuals in the
+#'   result. Default: \code{FALSE}.
+#' @param return_robustness_weights Logical; if \code{TRUE}, return per-point
+#'   robustness weights. Default: \code{FALSE}.
+#' @param zero_weight_fallback Fallback policy when all robustness weights drop
+#'   to zero: \code{"use_local_mean"} (default; aliases: \code{"local_mean"},
+#'   \code{"mean"}), \code{"return_original"} (alias: \code{"original"}), or
+#'   \code{"return_none"} (alias: \code{"none"}).
+#' @param auto_converge Convergence tolerance for early stopping of robustness
+#'   iterations. \code{NULL} (default) disables early stopping.
+#' @param cv_fractions Numeric vector of candidate fractions for
[…]
