### 2026-08-30 5f4104c
docs: consolidate parameters into api
