+++ bindings/julia/julia/src/FastLOWESS.jl
@@ -34,80 +34,80 @@ using Downloads: download
 
 # Try to import JLL package first
 try
-    using fastlowess_jll
+	using fastlowess_jll
 catch e
-    # JLL not available, will use fallback
+	# JLL not available, will use fallback
 end
 
 # Library name varies by platform
 const LIBNAME =
-    Sys.iswindows() ? "fastlowess_jl.dll" :
-    Sys.isapple() ? "libfastlowess_jl.dylib" : "libfastlowess_jl.so"
+	Sys.iswindows() ? "fastlowess_jl.dll" :
+	Sys.isapple() ? "libfastlowess_jl.dylib" : "libfastlowess_jl.so"
 
 # Try to load from JLL package first, fall back to local build
 function find_library()
-    # Option 1: Check environment variable (PRIORITY)
-    if haskey(ENV, "FASTLOWESS_LIB")
-        lib = ENV["FASTLOWESS_LIB"]
-        @info "Using library from FASTLOWESS_LIB: $lib"
-        return lib
-    end
+	# Option 1: Check environment variable (PRIORITY)
+	if haskey(ENV, "FASTLOWESS_LIB")
+		lib = ENV["FASTLOWESS_LIB"]
+		@info "Using library from FASTLOWESS_LIB: $lib"
+		return lib
+	end
 
-    # Option 2: Use JLL package if available (for registered package)
-    if @isdefined(fastlowess_jll)
-        try
-            if hasproperty(fastlowess_jll, :libfastlowess_jl)
-                lib = fastlowess_jll.libfastlowess_jl
-                @info "Using fastlowess_jll library: $lib"
-                return lib
-            end
-        catch e
-            @warn "Failed to load from fastlowess_jll" exception = e
-        end
-    end
+	# Option 2: Use JLL package if available (for registered package)
+	if @isdefined(fastlowess_jll)
+		try
+			if hasproperty(fastlowess_jll, :libfastlowess_jl)
+				lib = fastlowess_jll.libfastlowess_jl
+				@info "Using fastlowess_jll library: $lib"
+				return lib
+			end
+		catch e
+			@warn "Failed to load from fastlowess_jll" exception = e
+		end
+	end
 
-    # Option 3: Check relative paths (development mode)
-    # Path: julia/src/fastlowess.jl -> julia/ -> bindings/julia/ -> bindings/ -> lowess-project/
-    src_dir = @__DIR__                        # julia/src/
-    julia_dir = dirname(src_dir)              # julia/
-    bindings_julia_dir = dirname(julia_dir)   # bindings/julia/
-    bindings_dir = dirname(bindings_julia_dir)# bindings/
-    workspace_root = dirname(bindings_dir)    # lowess-project/
+	# Option 3: Check relative paths (development mode)
+	# Path: julia/src/fastlowess.jl -> julia/ -> bindings/julia/ -> bindings/ -> lowess-project/
+	src_dir = @__DIR__                        # julia/src/
+	julia_dir = dirname(src_dir)              # julia/
+	bindings_julia_dir = dirname(julia_dir)   # bindings/julia/
+	bindings_dir = dirname(bindings_julia_dir)# bindings/
+	workspace_root = dirname(bindings_dir)    # lowess-project/
 
-    candidates = [
-        # Workspace root target (most common for workspace members)
-        joinpath(workspace_root, "target", "release", LIBNAME),
-        joinpath(workspace_root, "target", "debug", LIBNAME),
-        # Local target (if built standalone)
-        joinpath(bindings_julia_dir, "target", "release", LIBNAME),
[…]
