+++ bindings/julia/julia/docs/src/index.md
@@ -1,0 +1,142 @@
+<!-- markdownlint-disable MD036 -->
+# LOWESS Project
+
+[![Julia](https://img.shields.io/badge/Julia-9558B2?logo=julia&logoColor=white)](https://juliahub.com/ui/Packages/General/FastLOWESS)
+
+[![CI](https://github.com/thisisamirv/lowess-project/actions/workflows/ci-julia.yml/badge.svg)](https://github.com/thisisamirv/lowess-project/actions/workflows/ci-julia.yml)
+
+![One LOWESS to Rule Them All](https://raw.githubusercontent.com/thisisamirv/lowess-project/main/dev/logo.png)
+
+*One LOWESS to Rule Them All*
+
+The fastest, most robust, and most feature-complete language-agnostic LOWESS (Locally Weighted Scatterplot Smoothing) implementation for **Rust**, **Python**, **R**, **Julia**, **JavaScript**, **C++**, and **WebAssembly**.
+
+The `lowess-project` also offers bindings for Rust, Python, R, Julia, Node.js, WebAssembly, and C++ — see the [full repository](https://github.com/thisisamirv/lowess-project).
+
+---
+
+## Installation & Documentation
+
+> Currently available for R, Python, Rust, Julia, Node.js, WebAssembly, and C++. See the [Installation Guide](https://thisisamirv.github.io/lowess-project/julia/installation/) for detailed installation instructions.
+>
+> ### 📚 [View the full documentation](https://thisisamirv.github.io/lowess-project/julia/)
+
+### GPU Backend
+
+GPU acceleration (`wgpu`: Vulkan/Metal/DX12) is also supported for high-throughput batch smoothing. See the [GPU Backend guide](https://thisisamirv.github.io/lowess-project/julia/gpu-backend/) for details.
+
+---
+
+## LOESS vs. LOWESS
+
+| Feature | LOESS | LOWESS (This Crate) |
+| --- | --- | --- |
+| **Polynomial Degree** | Linear, Quadratic, Cubic, Quartic | Linear (Degree 1) |
+| **Dimensions** | Multivariate (n-D support) | Univariate (1-D only) |
+| **Flexibility** | High (Distance metrics) | Standard |
+| **Complexity** | Higher (Matrix inversion) | Lower (Weighted average/slope) |
+
+Read more about how LOWESS works in the [Concepts](https://thisisamirv.github.io/lowess-project/julia/concepts/).
+
+> **Note:** For a **LOESS** implementation, use [`loess-project`](https://github.com/thisisamirv/loess-project).
+
+---
+
+## Why this package?
+
+### Speed
+
+The `lowess` project beats the competition in terms of speed, whether in single-threaded or multi-threaded parallel execution. It is on average **200-327x faster** than Python's `statsmodels.lowess` and **2-3x faster** than R's `lowess`.
+
+For more details on the performance comparison, see the [Benchmarks](https://thisisamirv.github.io/lowess-project/julia/benchmarks/) page.
+
+### Robustness
+
+This implementation is *more robust* than R's `lowess` and Python's `statsmodels` due to two key design choices:
+
+**MAD-Based Scale Estimation:**
+
+For robustness weight calculations, this crate uses *Median Absolute Deviation (MAD)* for scale estimation:
+
+```text
+s = median(|r_i - median(r)|)
+```
+
+In contrast, `statsmodels` and R's `lowess` uses the median of absolute residuals (MAR):
+
+```text
+s = median(|r_i|)
+```
+
+- MAD is a *breakdown-point-optimal* estimator—it remains valid even when up to 50% of data are outliers.
+- The median-centering step removes asymmetric bias from residual distributions.
[…]
