### 2026-08-30 6ea7e7a
docs: minor fixes
