+++ crates/fastLowess/docs/benchmarks.md
@@ -60,3 +60,9 @@ fn main() -> Result<(), LowessError> {
     Ok(())
 }
 ```
+
+```output
+Serial:   299.22 ms
+Parallel: 54.15 ms
+Speedup:  5.53x
+```
