+++ cpp/search/all_c.js
@@ -1,5 +1,14 @@
 var searchData=
 [
-  ['next_20steps_0',['next steps',['../md_docs_2concepts.html#autotoc_md102',1,'Next Steps'],['../md_docs_2quickstart.html#autotoc_md265',1,'Next Steps']]],
-  ['no_20boundary_1',['No Boundary',['../md_docs_2boundary.html#autotoc_md82',1,'']]]
+  ['dashboard_20example_0',['Real-Time Dashboard Example',['../md_docs_2use-case-real-time.html#autotoc_md284',1,'']]],
+  ['data_1',['data',['../md_docs_2use-case-genomics.html#autotoc_md272',1,'Best Practices for Genomic Data'],['../md_docs_2use-case-time-series.html#autotoc_md302',1,'Handling Missing Data']]],
+  ['data_20example_2',['Sensor Data Example',['../md_docs_2use-case-real-time.html#autotoc_md279',1,'']]],
+  ['data_20smoothing_3',['Genomic Data Smoothing',['../md_docs_2use-case-genomics.html',1,'']]],
+  ['decision_20guide_4',['Quick Decision Guide',['../md_docs_2concepts.html#autotoc_md86',1,'']]],
+  ['default_5',['default',['../md_docs_2robustness.html#autotoc_md237',1,'Bisquare (Default)'],['../md_docs_2boundary.html#autotoc_md65',1,'Extend (Default)'],['../md_docs_2scaling.html#autotoc_md253',1,'MAD — Median Absolute Deviation (Default)'],['../md_docs_2kernels.html#autotoc_md163',1,'Tricube (Default)']]],
+  ['detecting_20outliers_6',['Detecting Outliers',['../md_docs_2robustness.html#autotoc_md245',1,'']]],
+  ['detrending_7',['Detrending',['../md_docs_2use-case-time-series.html#autotoc_md298',1,'']]],
+  ['development_20version_8',['fastlowess (C++) (development version)',['../md_docs_2NEWS.html#autotoc_md190',1,'']]],
+  ['deviation_20default_9',['MAD — Median Absolute Deviation (Default)',['../md_docs_2scaling.html#autotoc_md253',1,'']]],
+  ['documentation_10',['Installation &amp; Documentation',['../index.html#autotoc_md2',1,'']]]
 ];
