+++ cpp/search/all_4.js
@@ -1,13 +1,4 @@
 var searchData=
 [
-  ['emphasize_20important_20points_0',['Emphasize Important Points',['../md_docs_2custom-weights.html#autotoc_md130',1,'']]],
-  ['enabling_20gpu_20support_1',['Enabling GPU Support',['../md_docs_2api.html#autotoc_md49',1,'']]],
-  ['epanechnikov_2',['Epanechnikov',['../md_docs_2kernels.html#autotoc_md179',1,'']]],
-  ['errors_3',['Standard Errors',['../md_docs_2intervals.html#autotoc_md169',1,'']]],
-  ['estimation_4',['Scale Estimation',['../md_docs_2robustness.html#autotoc_md280',1,'']]],
-  ['example_5',['example',['../md_docs_2api.html#autotoc_md66',1,'Example'],['../md_docs_2use-case-real-time.html#autotoc_md317',1,'Real-Time Dashboard Example'],['../md_docs_2use-case-real-time.html#autotoc_md312',1,'Sensor Data Example']]],
-  ['execution_20modes_6',['execution modes',['../md_docs_2adapter-choice.html',1,'Execution Modes'],['../md_docs_2concepts.html#autotoc_md96',1,'Execution Modes']]],
-  ['expression_20time_20course_7',['Gene Expression Time Course',['../md_docs_2use-case-time-series.html#autotoc_md339',1,'']]],
-  ['extend_20default_8',['Extend (Default)',['../md_docs_2boundary.html#autotoc_md76',1,'']]],
-  ['extraction_9',['Basic Trend Extraction',['../md_docs_2use-case-time-series.html#autotoc_md329',1,'']]]
+  ['5_0',['fastlowess (C++) 0.99.5',['../md_docs_2NEWS.html#autotoc_md221',1,'']]]
 ];
