+++ cpp/search/all_0.js
@@ -1,28 +1,10 @@
 var searchData=
 [
-  ['a_20kernel_0',['Choosing a Kernel',['../md_docs_2kernels.html#autotoc_md191',1,'']]],
-  ['a_20known_20outlier_1',['Suppress a Known Outlier',['../md_docs_2custom-weights.html#autotoc_md128',1,'']]],
-  ['a_20policy_2',['Choosing a Policy',['../md_docs_2boundary.html#autotoc_md84',1,'']]],
-  ['a_20scaling_20method_3',['Choosing a Scaling Method',['../md_docs_2scaling.html#autotoc_md292',1,'']]],
-  ['a_20strategy_4',['Choosing a Strategy',['../md_docs_2merge.html#autotoc_md203',1,'']]],
-  ['absolute_20deviation_20default_5',['MAD — Median Absolute Deviation (Default)',['../md_docs_2scaling.html#autotoc_md286',1,'']]],
-  ['absolute_20residual_6',['absolute residual',['../md_docs_2scaling.html#autotoc_md288',1,'MAR — Median Absolute Residual'],['../md_docs_2scaling.html#autotoc_md290',1,'Mean — Mean Absolute Residual']]],
-  ['acceleration_7',['GPU Acceleration',['../md_docs_2api.html#autotoc_md48',1,'']]],
-  ['adapter_8',['When to Use Batch Adapter',['../md_docs_2api.html#autotoc_md43',1,'']]],
-  ['adapter_20parameters_9',['Adapter Parameters',['../md_docs_2parameters.html#autotoc_md244',1,'']]],
-  ['adapter_2dchoice_2emd_10',['adapter-choice.md',['../adapter-choice_8md.html',1,'']]],
-  ['advantages_11',['Key Advantages',['../md_docs_2concepts.html#autotoc_md100',1,'']]],
-  ['also_12',['also',['../md_docs_2use-case-real-time.html#autotoc_md325',1,'See Also'],['../md_docs_2custom-weights.html#autotoc_md138',1,'See Also'],['../md_docs_2use-case-genomics.html#autotoc_md307',1,'See Also'],['../md_docs_2gpu-backend.html#autotoc_md149',1,'See also'],['../md_docs_2use-case-time-series.html#autotoc_md343',1,'See Also']]],
-  ['analysis_13',['analysis',['../md_docs_2use-case-time-series.html#autotoc_md337',1,'Multi-Scale Analysis'],['../md_docs_2use-case-time-series.html',1,'Time Series Analysis']]],
-  ['api_14',['api',['../md_docs_2api.html',1,'API'],['../md_docs_2api-online.html',1,'OnlineLowess API'],['../md_docs_2api-streaming.html',1,'StreamingLowess API']]],
-  ['api_20reference_15',['API Reference',['../index.html#autotoc_md12',1,'']]],
-  ['api_2donline_2emd_16',['api-online.md',['../api-online_8md.html',1,'']]],
-  ['api_2dstreaming_2emd_17',['api-streaming.md',['../api-streaming_8md.html',1,'']]],
-  ['api_2emd_18',['api.md',['../api_8md.html',1,'']]],
-  ['application_19',['Application',['../md_docs_2use-case-genomics.html#autotoc_md301',1,'']]],
-  ['auto_20convergence_20',['Auto-Convergence',['../md_docs_2robustness.html#autotoc_md282',1,'']]],
-  ['auto_5fconverge_21',['auto_converge',['../md_docs_2parameters.html#autotoc_md226',1,'']]],
[…]
