+++ cpp/search/all_21.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['—_20mean_20absolute_20residual_0',['Mean — Mean Absolute Residual',['../md_docs_2scaling.html#autotoc_md257',1,'']]],
+  ['—_20median_20absolute_20deviation_20default_1',['MAD — Median Absolute Deviation (Default)',['../md_docs_2scaling.html#autotoc_md253',1,'']]],
+  ['—_20median_20absolute_20residual_2',['MAR — Median Absolute Residual',['../md_docs_2scaling.html#autotoc_md255',1,'']]]
+];
