### 2026-08-30 900c684
fix: minor workflow fixes
