### 2026-08-30 190ba47
feat(docs): deploy GitHub Pages via official actions/deploy-pages
