+++ r/news/index.md
@@ -70,6 +70,16 @@
 - Fixed `docs.yml` triggering GitHub’s “pages build and deployment” once
   per docs job; per-language jobs now upload artifacts, and a single
   final `deploy` job pushes to `gh-pages` once per run.
+- Fixed `docs.yml`’s reliance on GitHub’s legacy branch-based Pages
+  deployment, which auto-triggers an unpinned, GitHub-managed “pages
+  build and deployment” job on every `gh-pages` push (surfacing
+  deprecation warnings, e.g. for Node.js 20, that aren’t fixable from
+  this repo). The former `deploy` job is now `build`, which still pushes
+  the merged `_site` to `gh-pages` as a cache for future incremental
+  runs, but publishing now goes through `actions/upload-pages-artifact`
+  and a new `deploy` job using the official `actions/deploy-pages`,
+  which this repo pins directly. Requires the repository’s Pages source
+  to be switched to “GitHub Actions” in settings.
 
 ## rfastlowess 3.1.0
 
