### 2026-08-30 2401b9b
fix: various docs fixes
