+++ bindings/r/vignettes/use-case-genomics.Rmd
@@ -95,6 +95,7 @@ model <- Lowess(
     fraction = 0.05,
     iterations = 5
 )
+result <- fit(model, positions, signal_noisy)
 threshold <- quantile(result$y, 0.75)
 
 plot(positions, signal_noisy, pch = 16, cex = 0.4, col = "gray",
