+++ bindings/python/docs/getting-started/quickstart.md
@@ -69,7 +69,7 @@ x_out = np.linspace(1, 6, 6)
 y_with_outlier = np.array([2.0, 4.0, 6.0, 50.0, 10.0, 12.0])
 
 model = fl.Lowess(
-    fraction=0.5,
+    fraction=0.7,
     iterations=5,
     robustness_method="bisquare",
     return_robustness_weights=True

[…]
