### 2026-08-30 8a2f49b
feat(docs): add category grouping to cpp docs
