+++ bindings/cpp/docs/customization.md
@@ -1,0 +1,10 @@
+\page customization Customization
+
+# Customization
+
+Weight functions, outlier handling, and residual scaling.
+
+- \subpage kernels
+- \subpage robustness
+- \subpage scaling
+- \subpage custom_weights
