+++ bindings/cpp/docs/user-guide.md
@@ -1,0 +1,9 @@
+\page user_guide User Guide
+
+# User Guide
+
+Choosing adapters, quantifying uncertainty, and selecting parameters.
+
+- \subpage adapter_choice
+- \subpage intervals
+- \subpage cross_validation
