+++ bindings/cpp/docs/use-cases.md
@@ -1,0 +1,9 @@
+\page use_cases Use Cases
+
+# Use Cases
+
+Application-specific guides.
+
+- \subpage use_case_genomics
+- \subpage use_case_time_series
+- \subpage use_case_real_time
