+++ bindings/cpp/docs/getting-started.md
@@ -1,0 +1,9 @@
+\page getting_started Getting Started
+
+# Getting Started
+
+New to LOWESS? Start here.
+
+- \subpage installation
+- \subpage quickstart
+- \subpage concepts
