### 2026-08-30 7ae3a8a
feat: add spack release for cpp binding
