+++ cpp/search/all_d.js
@@ -1,12 +1,10 @@
 var searchData=
 [
-  ['emphasize_20important_20points_0',['Emphasize Important Points',['../md_docs_2custom-weights.html#autotoc_md116',1,'']]],
-  ['epanechnikov_1',['Epanechnikov',['../md_docs_2kernels.html#autotoc_md165',1,'']]],
-  ['errors_2',['Standard Errors',['../md_docs_2intervals.html#autotoc_md155',1,'']]],
-  ['estimation_3',['Scale Estimation',['../md_docs_2robustness.html#autotoc_md247',1,'']]],
-  ['example_4',['example',['../md_docs_2api.html#autotoc_md55',1,'Example'],['../md_docs_2use-case-real-time.html#autotoc_md284',1,'Real-Time Dashboard Example'],['../md_docs_2use-case-real-time.html#autotoc_md279',1,'Sensor Data Example']]],
-  ['execution_20modes_5',['execution modes',['../md_docs_2adapter-choice.html',1,'Execution Modes'],['../md_docs_2concepts.html#autotoc_md84',1,'Execution Modes']]],
-  ['expression_20time_20course_6',['Gene Expression Time Course',['../md_docs_2use-case-time-series.html#autotoc_md306',1,'']]],
-  ['extend_20default_7',['Extend (Default)',['../md_docs_2boundary.html#autotoc_md65',1,'']]],
-  ['extraction_8',['Basic Trend Extraction',['../md_docs_2use-case-time-series.html#autotoc_md296',1,'']]]
+  ['onlinelowess_0',['onlinelowess',['../classfastlowess_1_1OnlineLowess.html',1,'fastlowess::OnlineLowess'],['../classfastlowess_1_1OnlineOutput.html#a5e236589da58bef059203559a019ad80',1,'fastlowess::OnlineOutput::OnlineLowess'],['../classfastlowess_1_1OnlineLowess.html#a24db2d32e3fec45bb3388336c0200dc0',1,'fastlowess::OnlineLowess::OnlineLowess(const OnlineOptions &amp;options={})'],['../classfastlowess_1_1OnlineLowess.html#a6ad68adac17764c91153df5a02ac1ce5',1,'fastlowess::OnlineLowess::OnlineLowess(const OnlineLowess &amp;)=delete'],['../classfastlowess_1_1OnlineLowess.html#a9a99929c5de21a4dc41a1b7182f48c33',1,'fastlowess::OnlineLowess::OnlineLowess(OnlineLowess &amp;&amp;other) noexcept']]],
+  ['onlinelowess_20api_1',['OnlineLowess API',['../api_online.html',1,'api']]],
+  ['onlineoptions_2',['OnlineOptions',['../structfastlowess_1_1OnlineOptions.html',1,'fastlowess']]],
+  ['onlineoutput_3',['OnlineOutput',['../classfastlowess_1_1OnlineOutput.html',1,'fastlowess']]],
+  ['operator_20bool_4',['operator bool',['../classfastlowess_1_1Expected.html#ad361b5ccf1163a38e469847f2b1adf77',1,'fastlowess::Expected']]],
[…]
