+++ cpp/search/pages_13.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['weight_20functions_0',['Weight Functions',['../kernels.html',1,'customization']]],
+  ['weights_1',['Custom Weights',['../custom_weights.html',1,'customization']]]
+];
