+++ cpp/search/functions_0.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['add_5fpoint_0',['add_point',['../classfastlowess_1_1OnlineLowess.html#a6bb43023f9e7f8884320dd29ac803aa9',1,'fastlowess::OnlineLowess']]],
+  ['aic_1',['aic',['../classfastlowess_1_1Diagnostics.html#aa23b246dc8a2ee2581e742c0e75c3afc',1,'fastlowess::Diagnostics']]],
+  ['aicc_2',['aicc',['../classfastlowess_1_1Diagnostics.html#a54514eec30d53ab40080f8e9896e6ef4',1,'fastlowess::Diagnostics']]]
+];
