+++ cpp/search/functions_2.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['diagnostics_0',['diagnostics',['../classfastlowess_1_1Diagnostics.html#a3c55ce2b41e1ccad911cbe36548f1ddf',1,'fastlowess::Diagnostics::Diagnostics()=default'],['../classfastlowess_1_1Diagnostics.html#a583ebfc139a425ba44a613933cbb1264',1,'fastlowess::Diagnostics::Diagnostics(const fastlowess_CppLowessResult &amp;result)'],['../classfastlowess_1_1LowessResult.html#a92eb3b49b1fffe732a26b5d6db1007c4',1,'fastlowess::LowessResult::diagnostics()']]]
+];
