+++ cpp/search/variables_0.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['auto_5fconverge_0',['auto_converge',['../structfastlowess_1_1LowessOptions.html#ab818fecad4f18a4f6ab947ebe75deb06',1,'fastlowess::LowessOptions::auto_converge'],['../structfastlowess_1_1OnlineOptions.html#a6950c44ed0571eed6219ecd34e78b0bd',1,'fastlowess::OnlineOptions::auto_converge']]]
+];
