+++ cpp/search/functions_d.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['valid_0',['valid',['../classfastlowess_1_1LowessResult.html#a2a5eb76cd3ae3890e8c91597c74aba0d',1,'fastlowess::LowessResult']]],
+  ['value_1',['value',['../classfastlowess_1_1Expected.html#aeb88e40ff81b811771e6e2013fad7a6c',1,'fastlowess::Expected::value() &amp;'],['../classfastlowess_1_1Expected.html#ab9f1bddd03e221b49c256931b13f0dec',1,'fastlowess::Expected::value() const &amp;'],['../classfastlowess_1_1Expected.html#aacabf25f31eaf0838faf06f0165a57ae',1,'fastlowess::Expected::value() &amp;&amp;']]]
+];
