+++ cpp/search/all_f.js
@@ -1,13 +1,5 @@
 var searchData=
 [
-  ['gaussian_0',['Gaussian',['../md_docs_2kernels.html#autotoc_md167',1,'']]],
-  ['gene_20expression_20time_20course_1',['Gene Expression Time Course',['../md_docs_2use-case-time-series.html#autotoc_md306',1,'']]],
-  ['genome_20coverage_20streaming_2',['Large Genome Coverage (Streaming)',['../md_docs_2use-case-genomics.html#autotoc_md270',1,'']]],
-  ['genomic_20data_3',['Best Practices for Genomic Data',['../md_docs_2use-case-genomics.html#autotoc_md272',1,'']]],
-  ['genomic_20data_20smoothing_4',['Genomic Data Smoothing',['../md_docs_2use-case-genomics.html',1,'']]],
-  ['gpu_20acceleration_5',['GPU Acceleration',['../md_docs_2api.html#autotoc_md45',1,'']]],
-  ['gpu_20backend_6',['gpu backend',['../md_docs_2gpu-backend.html',1,'GPU Backend'],['../index.html#autotoc_md3',1,'GPU Backend'],['../md_docs_2benchmarks.html#autotoc_md59',1,'GPU Backend']]],
-  ['gpu_20support_7',['Installing GPU Support',['../md_docs_2gpu-backend.html#autotoc_md131',1,'']]],
-  ['gpu_2dbackend_2emd_8',['gpu-backend.md',['../gpu-backend_8md.html',1,'']]],
-  ['guide_9',['Quick Decision Guide',['../md_docs_2concepts.html#autotoc_md86',1,'']]]
+  ['quick_20start_0',['Quick Start',['../quickstart.html',1,'getting_started']]],
+  ['quickstart_2emd_1',['quickstart.md',['../quickstart_8md.html',1,'']]]
 ];

[…]
