### 2026-08-30 b14d815
fix: cargo doc for gpu-backend, harden spack PR creation code
