### 2026-08-30 9d41b56
dev: fix gh repo fork usage in spack_open_pr.sh (drop unsupported --remote, avoid duplicate upstream remote, harden cleanup trap)
