### 2026-08-30 bb741aa
dev: run gh auth setup-git in spack_open_pr.sh so git push/fetch can authenticate via GH_TOKEN
