+++ benchmarks/rfastlowess.R
@@ -140,7 +140,7 @@ benchmark_scalability <- function(n_iter = 10) {
         model <- make_model(0.1, 3)
         results[[paste0("scale_", size)]] <- run_benchmark(
             paste0("scale_", size), size,
-            function() model$fit(d$x, d$y),
+            function() fit(model, d$x, d$y),
             n_iter
         )
     }
@@ -155,7 +155,7 @@ benchmark_fraction <- function(n_iter = 10) {
         model <- make_model(frac, 3)
         run_benchmark(
             paste0("fraction_", frac), size,
-            function() model$fit(d$x, d$y),
+            function() fit(model, d$x, d$y),
             n_iter
         )
     })
@@ -171,7 +171,7 @@ benchmark_iterations <- function(n_iter = 10) {
         model <- make_model(0.2, it)
         run_benchmark(
             paste0("iterations_", it), size,
-            function() model$fit(d$x, d$y),
+            function() fit(model, d$x, d$y),
             n_iter
         )
     })
@@ -186,7 +186,7 @@ benchmark_financial <- function(n_iter = 10) {
         model <- make_model(0.1, 2)
         results[[paste0("financial_", size)]] <- run_benchmark(
             paste0("financial_", size), size,
-            function() model$fit(d$x, d$y),
+            function() fit(model, d$x, d$y),
             n_iter
         )
     }
@@ -200,7 +200,7 @@ benchmark_scientific <- function(n_iter = 10) {
         model <- make_model(0.15, 3)
         results[[paste0("scientific_", size)]] <- run_benchmark(
             paste0("scientific_", size), size,
-            function() model$fit(d$x, d$y),
+            function() fit(model, d$x, d$y),
             n_iter
         )
     }
@@ -215,7 +215,7 @@ benchmark_genomic <- function(n_iter = 10) {
         size_str <- format(size, scientific = FALSE, trim = TRUE)
         results[[paste0("genomic_", size_str)]] <- run_benchmark(
             paste0("genomic_", size_str), size,
-            function() model$fit(d$x, d$y),
+            function() fit(model, d$x, d$y),
             n_iter
         )
     }
@@ -229,26 +229,97 @@ benchmark_pathological <- function(n_iter = 10) {
     d <- generate_clustered_data(size)
     model <- make_model(0.3, 2)
     results$clustered <- run_benchmark(
-        "clustered", size, function() model$fit(d$x, d$y), n_iter
+        "clustered", size, function() fit(model, d$x, d$y), n_iter
     )
 
     d <- generate_high_noise_data(size)
     model <- make_model(0.5, 5)
     results$high_noise <- run_benchmark(
-        "high_noise", size, function() model$fit(d$x, d$y), n_iter
+        "high_noise", size, function() fit(model, d$x, d$y), n_iter
     )
 
     d <- generate_outlier_data(size)
     model <- make_model(0.2, 10)
     results$extreme_outliers <- run_benchmark(
-        "extreme_outliers", size, function() model$fit(d$x, d$y), n_iter
+        "extreme_outliers", size, function() fit(model, d$x, d$y), n_iter
     )
 
     xk <- as.numeric(seq_len(size))
     yk <- rep(5.0, size)
     model <- make_model(0.2, 2)
     results$constant_y <- run_benchmark(
-        "constant_y", size, function() model$fit(xk, yk), n_iter
+        "constant_y", size, function() fit(model, xk, yk), n_iter
+    )
+
+    results
+}
+
+benchmark_large <- function(n_iter = 3) {
+    # delta = 0 matches stats_lowess.R's large benchmark (which
+    # disables stats::lowess's interpolation shortcut); at n = 50000 this
[…]
