### 2026-08-31 0cb2e67
feat: add go binding
