+++ bindings/go/fastlowess/lowess.go
@@ -1,0 +1,218 @@
+package fastlowess
+
+/*
+#include "fastlowess_go.h"
+*/
+import "C"
+
+import (
+	"errors"
+	"runtime"
+)
+
+// Options configures a Lowess, StreamingLowess, or OnlineLowess model.
+// Use DefaultOptions and override only the fields you need.
+type Options struct {
+	// Fraction is the smoothing fraction, in (0, 1]. Default: 0.67.
+	Fraction float64
+	// Iterations is the number of robustness iterations, in [0, 1000]. Default: 3.
+	Iterations int
+	// Delta is the interpolation distance threshold, as a non-negative
+	// fraction of the x range; points within Delta of each other on x share
+	// the same local fit. Nil sets it automatically to 1/100th of the x range.
+	Delta *float64
+
+	// WeightFunction is the kernel weight function: "tricube" (default),
+	// "gaussian", "uniform" (alias "boxcar"), "cosine", "epanechnikov",
+	// "biweight" (alias "bisquare"), or "triangle" (alias "triangular").
+	WeightFunction string
+	// RobustnessMethod is the outlier downweighting method: "bisquare"
+	// (default, alias "biweight"), "huber", or "talwar".
+	RobustnessMethod string
+	// ScalingMethod is the residual scale estimator for robustness weights:
+	// "mad" (default, alias "median_absolute_deviation"), "mar" (alias
+	// "median_absolute_residual"), or "mean" (alias "mean_absolute_residual").
+	ScalingMethod string
+	// BoundaryPolicy is the boundary handling strategy: "extend" (default,
+	// alias "pad"), "reflect" (alias "mirror"), "zero", or "noboundary"
+	// (alias "none").
+	BoundaryPolicy string
+	// ZeroWeightFallback is the fallback policy when all robustness weights
+	// drop to zero: "use_local_mean" (default, aliases "local_mean", "mean"),
+	// "return_original" (alias "original"), or "return_none" (alias "none").
+	ZeroWeightFallback string
+
+	// ConfidenceIntervals is the confidence level for confidence intervals,
+	// in (0, 1) (e.g. 0.95). Nil disables confidence intervals.
+	ConfidenceIntervals *float64
+	// PredictionIntervals is the confidence level for prediction intervals,
+	// in (0, 1) (e.g. 0.95). Nil disables prediction intervals.
+	PredictionIntervals *float64
+	// AutoConverge is the convergence tolerance for early stopping of
+	// robustness iterations. Nil disables early stopping.
+	AutoConverge *float64
+
+	// ReturnDiagnostics requests fit-quality metrics (RMSE, MAE, R-squared, AIC, etc.).
+	ReturnDiagnostics bool
+	// ReturnResiduals requests residuals in the result.
+	ReturnResiduals bool
+	// ReturnRobustnessWeights requests per-point robustness weights in the result.
+	ReturnRobustnessWeights bool
+	// ReturnSE requests hat-matrix statistics (effective degrees of freedom,
+	// leverage, standard errors). Batch model only.
+	ReturnSE bool
+
+	// CVFractions is a set of candidate fractions for cross-validation.
+	// Empty disables CV. Batch model only.
+	CVFractions []float64
+	// CVMethod is the cross-validation method: "kfold" (default) or "loocv".
+	CVMethod string
+	// CVK is the number of folds for k-fold CV. Default: 5.
+	CVK int
[…]
