+++ bindings/go/include/fastlowess_go.h
@@ -1,0 +1,275 @@
+typedef struct fastlowess_GoLowess fastlowess_GoLowess;
+
+typedef struct fastlowess_GoOnlineLowess fastlowess_GoOnlineLowess;
+
+typedef struct fastlowess_GoStreamingLowess fastlowess_GoStreamingLowess;
+
+/**
+ * Result struct that can be passed across FFI boundary.
+ * All arrays are allocated by Rust and must be freed by Rust.
+ */
+typedef struct fastlowess_GoLowessResult {
+  /**
+   * Sorted x values (length = n)
+   */
+  double *x;
+  /**
+   * Smoothed y values (length = n)
+   */
+  double *y;
+  /**
+   * Number of data points
+   */
+  unsigned long n;
+  /**
+   * Standard errors (NULL if not computed)
+   */
+  double *standard_errors;
+  /**
+   * Lower confidence bounds (NULL if not computed)
+   */
+  double *confidence_lower;
+  /**
+   * Upper confidence bounds (NULL if not computed)
+   */
+  double *confidence_upper;
+  /**
+   * Lower prediction bounds (NULL if not computed)
+   */
+  double *prediction_lower;
+  /**
+   * Upper prediction bounds (NULL if not computed)
+   */
+  double *prediction_upper;
+  /**
+   * Residuals (NULL if not computed)
+   */
+  double *residuals;
+  /**
+   * Robustness weights (NULL if not computed)
+   */
+  double *robustness_weights;
+  /**
+   * Cross-validation scores (NULL if not computed, length = cv_scores_len)
+   */
+  double *cv_scores;
+  /**
+   * Number of cross-validation scores
+   */
+  unsigned long cv_scores_len;
+  /**
+   * Fraction used for smoothing
+   */
+  double fraction_used;
+  /**
+   * Number of iterations performed (-1 if not available)
+   */
+  int iterations_used;
+  /**
+   * Diagnostics (NaN if not computed)
+   */
+  double rmse;
+  double mae;
+  double r_squared;
+  double aic;
+  double aicc;
+  double effective_df;
+  double residual_sd;
+  /**
+   * Error message (NULL if no error)
+   */
+  char *error;
+} fastlowess_GoLowessResult;
+
+typedef struct fastlowess_GoOnlineOutput {
+  int has_value;
+  double y;
+  double standard_error;
+  double residual;
+  double robustness_weight;
+  int iterations_used;
+  char *error;
+} fastlowess_GoOnlineOutput;
+
+const char *go_last_error_message(void);
+
+/**
+ * Returns 1 if this library was built with the `gpu` Cargo feature enabled, 0 otherwise.
+ */
+int go_gpu_enabled(void);
+
+/**
+ * Returns the crate version as a static, null-terminated C string.
+ */
+const char *go_version(void);
+
+/**
+ * Go wrapper constructor.
+ *
+ * # Safety
+ * Pointers must be valid null-terminated strings or null. Arrays must be valid.
+ */
+struct fastlowess_GoLowess *go_lowess_new(double fraction,
+                                          int iterations,
+                                          double delta,
+                                          const char *weight_function,
+                                          const char *robustness_method,
+                                          const char *scaling_method,
+                                          const char *boundary_policy,
+                                          double confidence_intervals,
+                                          double prediction_intervals,
+                                          int return_diagnostics,
+                                          int return_residuals,
+                                          int return_robustness_weights,
+                                          const char *zero_weight_fallback,
+                                          double auto_converge,
+                                          const double *cv_fractions,
+                                          unsigned long cv_fractions_len,
+                                          const char *cv_method,
+                                          int cv_k,
[…]
