### 2026-08-31 71b823d
docs: fix go docs
