+++ bindings/go/docs/installation.md
@@ -1,4 +1,7 @@
-# Installation
+---
+title: "Installation"
+weight: 15
+---
 
 ## Requirements
 

[…]
