### 2026-08-31 08a9373
Step 3: Upgrade Java to 25 - Compile: SUCCESS
- pom.xml: maven.compiler.release 17→25
- pom.xml: pin maven-compiler-plugin to 3.13.0 for stable Java 25 support
- StreamingLowess.java: fix pre-existing javadoc/access-modifier corruption
  (processChunk was accidentally package-private due to mangled comment markers)
