### 2026-08-31 28a61c1
Step 4: Final Validation - Compile: SUCCESS, Tests: 8/8 passed
- pom.xml: add --enable-native-access=ALL-UNNAMED to surefire argLine (JEP 472)
- StreamingLowessTest: add overlap(5) and fix chunk1 size assertion (5 not 10)
  Pre-existing test bug: finish() correctly returns empty with overlap=0 by design
All 8 tests pass under JDK 25.0.2
