+++ bindings/java/src/test/java/com/thisisamirv/fastlowess/StreamingLowessTest.java
@@ -9,7 +9,7 @@ class StreamingLowessTest {
 
     @Test
     void processesChunksAndFinalizes() {
-        try (StreamingLowess model = new StreamingLowess(StreamingOptions.builder().chunkSize(10).build())) {
+        try (StreamingLowess model = new StreamingLowess(StreamingOptions.builder().chunkSize(10).overlap(5).build())) {
             double[] x1 = new double[10];
             double[] y1 = new double[10];
             for (int i = 0; i < 10; i++) {
@@ -17,7 +17,7 @@ class StreamingLowessTest {
                 y1[i] = i * 2.0;
             }
             Result chunk1 = model.processChunk(x1, y1);
-            assertEquals(10, chunk1.x().length);
+            assertEquals(5, chunk1.x().length);
 
             double[] x2 = new double[10];
             double[] y2 = new double[10];
