### 2026-08-31 86ebb43
feat: add java binding
