### 2026-08-31 57c4830
ci: modify java release workflow
