+++ dev/runners/java.py
@@ -19,6 +19,25 @@ _MAIN_RE = re.compile(r"public\s+static\s+void\s+main\s*\(")
 _CLASS_RE = re.compile(r"\bclass\s+(\w+)")
 
 
+def _find_java_exe(name: str) -> str | None:
+    """Prefer JAVA_HOME's bin/ over a bare PATH search.
+
+    A PATH search alone is unreliable here: other installed JDKs (e.g. an
+    Eclipse Adoptium installer adding itself to the *Machine*-scope PATH,
+    which always precedes the User-scope PATH regardless of ordering) can
+    shadow the JDK actually used to compile bindings/java/target/classes,
+    causing a "class file has wrong version" mismatch.
+    """
+    java_home = os.environ.get("JAVA_HOME")
+    if java_home:
+        candidate = (
+            Path(java_home) / "bin" / (f"{name}.exe" if os.name == "nt" else name)
+        )
+        if candidate.is_file():
+            return str(candidate)
+    return _find_exe(name)
+
+
 def skip_reason(snippet: Snippet) -> str | None:
     if not _MAIN_RE.search(snippet.code):
         return "fragment — no public static void main (not a standalone Java program)"
@@ -30,8 +49,8 @@ def skip_reason(snippet: Snippet) -> str | None:
 
 
 def run_java(snippet: Snippet, timeout: int) -> RunResult:
-    javac_exe = _find_exe("javac")
-    java_exe = _find_exe("java")
+    javac_exe = _find_java_exe("javac")
+    java_exe = _find_java_exe("java")
     if javac_exe is None or java_exe is None:
         return RunResult(
             snippet=snippet,
