### 2026-09-01 c66ff73
fix: lint go scripts
