### 2026-09-01 ec1123d
docs: wire go and java into root docs
