### 2026-09-01 38e2945
docs: change java docs structure to a tree
