+++ dev/verify_snippets.py
@@ -360,7 +360,7 @@ def iter_md_files(
     if GO_BINDING_DOCS_DIR.exists():
         yield from sorted(GO_BINDING_DOCS_DIR.glob("*.md"))
     if JAVA_BINDING_DOCS_DIR.exists():
-        yield from sorted(JAVA_BINDING_DOCS_DIR.glob("*.md"))
+        yield from sorted(JAVA_BINDING_DOCS_DIR.rglob("*.md"))
 
 
 def main(argv: list[str] | None = None) -> int:
