+++ bindings/java/docs/guide/_index.md
@@ -1,0 +1,5 @@
+---
+title: "User Guide"
+weight: 25
+bookCollapseSection: true
+---
