+++ bindings/java/docs/weighting/_index.md
@@ -1,0 +1,5 @@
+---
+title: "Weight & Robustness"
+weight: 40
+bookCollapseSection: true
+---
