+++ bindings/java/docs/weighting/custom-weights.md
@@ -224,4 +224,4 @@ First smoothed value (custom weights): 1.7983966900445223
 ## See Also
 
 - [Robustness](robustness.md) — adaptive outlier downweighting via IRLS
-- [API Reference](api.md) — full parameter reference
+- [API Reference](../api/api.md) — full parameter reference
