### 2026-09-01 c567464
docs: change go docs structure to a tree
