+++ bindings/go/docs/introduction/_index.md
@@ -1,0 +1,5 @@
+---
+title: "Introduction"
+weight: 10
+bookCollapseSection: true
+---
