+++ bindings/go/docs/_index.md
@@ -1,0 +1,147 @@
+---
+title: "fastlowess (Go)"
+---
+
+<!-- markdownlint-disable MD024 MD033 -->
+<p align="center">
+  <a href="https://pkg.go.dev/github.com/thisisamirv/lowess-project/bindings/go/fastlowess"><img src="https://img.shields.io/badge/Go-00ADD8?logo=go&logoColor=white" alt="Go"></a>
+  <a href="https://github.com/thisisamirv/lowess-project/actions/workflows/ci-go.yml"><img src="https://github.com/thisisamirv/lowess-project/actions/workflows/ci-go.yml/badge.svg" alt="CI"></a>
+</p>
+
+<p align="center">
+  <img src="https://raw.githubusercontent.com/thisisamirv/lowess-project/main/dev/logo.png" alt="One LOWESS to Rule Them All" width="400">
+  <br>
+  <em>One LOWESS to Rule Them All</em>
+</p>
+
+The fastest, most robust, and most feature-complete language-agnostic LOWESS (Locally Weighted Scatterplot Smoothing) implementation for **Rust**, **Python**, **R**, **Julia**, **JavaScript**, **C++**, and **WebAssembly**.
+
+The `lowess-project` also offers bindings for Rust, Python, R, Julia, Node.js, WebAssembly, C++, Go, and Java — see the [full repository](https://github.com/thisisamirv/lowess-project).
+
+---
+
+## Installation & Documentation
+
+> Currently available for R, Python, Rust, Julia, Node.js, WebAssembly, and C++. See the [Installation Guide](https://thisisamirv.github.io/lowess-project/go/installation/) for detailed installation instructions.
+>
+> ### 📚 [View the full documentation](https://thisisamirv.github.io/lowess-project/go/)
+
+### GPU Backend
+
+GPU acceleration (`wgpu`: Vulkan/Metal/DX12) is also supported for high-throughput batch smoothing. See the [GPU Backend guide](https://thisisamirv.github.io/lowess-project/go/gpu-backend/) for details.
+
+---
+
+## LOESS vs. LOWESS
+
+| Feature | LOESS | LOWESS (This Crate) |
+| --- | --- | --- |
+| **Polynomial Degree** | Linear, Quadratic, Cubic, Quartic | Linear (Degree 1) |
+| **Dimensions** | Multivariate (n-D support) | Univariate (1-D only) |
+| **Flexibility** | High (Distance metrics) | Standard |
+| **Complexity** | Higher (Matrix inversion) | Lower (Weighted average/slope) |
+
+Read more about how LOWESS works in the [Concepts](https://thisisamirv.github.io/lowess-project/go/concepts/).
+
+> **Note:** For a **LOESS** implementation, use [`loess-project`](https://github.com/thisisamirv/loess-project).
+
+---
+
+## Why this package?
+
+### Speed
+
+The `lowess` project beats the competition in terms of speed, whether in single-threaded or multi-threaded parallel execution. It is on average **200-327x faster** than Python's `statsmodels.lowess` and **2-3x faster** than R's `lowess`.
+
+For more details on the performance comparison, see the [Benchmarks](https://thisisamirv.github.io/lowess-project/go/benchmarks/) page.
+
+### Robustness
+
+This implementation is *more robust* than R's `lowess` and Python's `statsmodels` due to two key design choices:
+
+**MAD-Based Scale Estimation:**
+
+For robustness weight calculations, this crate uses *Median Absolute Deviation (MAD)* for scale estimation:
+
+```text
+s = median(|r_i - median(r)|)
+```
+
+In contrast, `statsmodels` and R's `lowess` uses the median of absolute residuals (MAR):
+
+```text
+s = median(|r_i|)
+```
+
[…]
