+++ bindings/go/docs/advanced/_index.md
@@ -1,0 +1,5 @@
+---
+title: "Advanced"
+weight: 55
+bookCollapseSection: true
+---
