+++ bindings/go/docs/api/_index.md
@@ -1,0 +1,5 @@
+---
+title: "API Reference"
+weight: 30
+bookCollapseSection: true
+---
