### 2026-09-01 060f5ab
ci: bump docs.yml go-version to 1.26 for hugo v0.165.0 compat
