+++ .github/workflows/docs.yml
@@ -129,7 +129,7 @@ jobs:
 
       - uses: actions/setup-go@v7
         with:
-          go-version: "1.21"
+          go-version: "1.26"
 
       - name: Install Hugo
         run: go install github.com/gohugoio/hugo@v0.165.0
@@ -154,7 +154,7 @@ jobs:
 
       - uses: actions/setup-go@v7
         with:
-          go-version: "1.21"
+          go-version: "1.26"
 
       - name: Install Hugo
         run: go install github.com/gohugoio/hugo@v0.165.0
