+++ go/en.search.min.9a7f0d8cc1821bb1dc4961146644a406b1c55f3a351a0aad9871d8748d4b9707.js
@@ -1,0 +1,1 @@
+(function(){const o="/lowess-project/go/en.search-data.min.10c035c2c7b5d6592af41bfaf777095b87fefe3d03526632b40d40d0872399b4.json",i="/lowess-project/go/fuse.esm.js",a=Object.assign({cache:!0},{includeScore:!0,useExtendedSearch:!0,fieldNormWeight:1.5,threshold:.3,ignoreLocation:!0,keys:[{name:"title",weight:.7},{name:"content",weight:.3}]}),e=document.querySelector("#book-search-input"),t=document.querySelector("#book-search-results");if(!e)return;let n;e.addEventListener("focus",c,{once:!0}),e.addEventListener("input",function(){clearTimeout(n),n=setTimeout(s,250)}),document.addEventListener("keydown",r);function r(t){if(t.key==="k"&&(t.metaKey||t.ctrlKey)){t.preventDefault(),e.focus();return}if(e===document.activeElement)return;if(t.target.value!==0[0])return;t.key==="/"&&(t.preventDefault(),e.focus())}function c(){e.required=!0,Promise.all([import(i),fetch(o).then(e=>e.json())]).then(([{default:e},t])=>{window.bookSearchIndex=new e(t,a)}).then(()=>e.required=!1).then(s)}function s(){for(;t.firstChild;)t.removeChild(t.firstChild);if(!e.value)return;const n=window.bookSearchIndex.search(e.value,{limit:10});n.forEach(function(e){const n=l("<li><a href></a><small></small></li>"),s=n.querySelector("a"),o=n.querySelector("small");s.href=e.item.href,s.textContent=e.item.title,o.textContent=e.item.section,t.appendChild(n)})}function l(e){const t=document.createElement("div");return t.innerHTML=e,t.firstChild}})()
