+++ java/en.search.min.c59b897ca85c0e945c52b4fb4ede14b90531c8cb57b996c19e526304862eec0d.js
@@ -1,0 +1,1 @@
+(function(){const o="/lowess-project/java/en.search-data.min.044e6031f3fd53c9f25772e6832b3853364c1ca7b30b088c5b73ab18ab94d274.json",i="/lowess-project/java/fuse.esm.js",a=Object.assign({cache:!0},{includeScore:!0,useExtendedSearch:!0,fieldNormWeight:1.5,threshold:.3,ignoreLocation:!0,keys:[{name:"title",weight:.7},{name:"content",weight:.3}]}),e=document.querySelector("#book-search-input"),t=document.querySelector("#book-search-results");if(!e)return;let n;e.addEventListener("focus",c,{once:!0}),e.addEventListener("input",function(){clearTimeout(n),n=setTimeout(s,250)}),document.addEventListener("keydown",r);function r(t){if(t.key==="k"&&(t.metaKey||t.ctrlKey)){t.preventDefault(),e.focus();return}if(e===document.activeElement)return;if(t.target.value!==0[0])return;t.key==="/"&&(t.preventDefault(),e.focus())}function c(){e.required=!0,Promise.all([import(i),fetch(o).then(e=>e.json())]).then(([{default:e},t])=>{window.bookSearchIndex=new e(t,a)}).then(()=>e.required=!1).then(s)}function s(){for(;t.firstChild;)t.removeChild(t.firstChild);if(!e.value)return;const n=window.bookSearchIndex.search(e.value,{limit:10});n.forEach(function(e){const n=l("<li><a href></a><small></small></li>"),s=n.querySelector("a"),o=n.querySelector("small");s.href=e.item.href,s.textContent=e.item.title,o.textContent=e.item.section,t.appendChild(n)})}function l(e){const t=document.createElement("div");return t.innerHTML=e,t.firstChild}})()
