### 2026-09-01 9e56f0a
ci: update go and java workflow, add go and java to gpu release workflow
