+++ .github/workflows/release-gpu.yml
@@ -586,6 +586,205 @@ jobs:
           path: dist
 
   # ============================================================================
+  # BUILD GPU LIBRARY - Go (Linux x86_64)
+  # ============================================================================
+  go-linux-x86_64:
+    name: Build Go GPU library (Linux x86_64)
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v7
+
+      - name: Build GPU library
+        run: cargo build --release --features gpu --manifest-path bindings/go/Cargo.toml
+
+      - name: Rename asset
+        shell: bash
+        run: |
+          mkdir -p dist
+          cp target/release/libfastlowess_go.a "dist/libfastlowess_go-gpu-${{ github.event.release.tag_name || 'dev' }}-linux-x86_64.a"
+
+      - name: Upload artifact
+        uses: actions/upload-artifact@v7
+        with:
+          name: gpu-go-linux-x86_64
+          path: dist
+
+  # ============================================================================
+  # BUILD GPU LIBRARY - Go (macOS x86_64)
+  # ============================================================================
+  go-macos-x86_64:
+    name: Build Go GPU library (macOS x86_64)
+    runs-on: macos-15-intel
+    steps:
+      - uses: actions/checkout@v7
+
+      - name: Build GPU library
+        run: cargo build --release --features gpu --manifest-path bindings/go/Cargo.toml
+
+      - name: Rename asset
+        shell: bash
+        run: |
+          mkdir -p dist
+          cp target/release/libfastlowess_go.a "dist/libfastlowess_go-gpu-${{ github.event.release.tag_name || 'dev' }}-macos-x86_64.a"
+
+      - name: Upload artifact
+        uses: actions/upload-artifact@v7
+        with:
+          name: gpu-go-macos-x86_64
+          path: dist
+
+  # ============================================================================
+  # BUILD GPU LIBRARY - Go (macOS ARM64)
+  # ============================================================================
+  go-macos-aarch64:
+    name: Build Go GPU library (macOS ARM64)
+    runs-on: macos-14
+    steps:
+      - uses: actions/checkout@v7
+
+      - name: Build GPU library
+        run: cargo build --release --features gpu --manifest-path bindings/go/Cargo.toml
+
+      - name: Rename asset
+        shell: bash
+        run: |
+          mkdir -p dist
+          cp target/release/libfastlowess_go.a "dist/libfastlowess_go-gpu-${{ github.event.release.tag_name || 'dev' }}-macos-aarch64.a"
+
+      - name: Upload artifact
+        uses: actions/upload-artifact@v7
+        with:
+          name: gpu-go-macos-aarch64
+          path: dist
+
+  # ============================================================================
+  # BUILD GPU LIBRARY - Go (Windows x86_64)
+  # ============================================================================
+  go-windows-x86_64:
+    name: Build Go GPU library (Windows x86_64)
+    runs-on: windows-latest
+    steps:
+      - uses: actions/checkout@v7
+
+      - name: Install Rust GNU target
+        run: |
+          rustup toolchain install stable-x86_64-pc-windows-gnu
+          rustup target add x86_64-pc-windows-gnu
+
+      - name: Build GPU library
+        run: cargo build --release --features gpu --manifest-path bindings/go/Cargo.toml --target x86_64-pc-windows-gnu
+
+      - name: Rename asset
+        shell: pwsh
+        run: |
+          New-Item -ItemType Directory -Force -Path dist | Out-Null
+          $tag = if ("${{ github.event.release.tag_name }}") { "${{ github.event.release.tag_name }}" } else { "dev" }
[…]
