### 2026-09-01 8dcce62
fix: version export in go, java, and nodejs
