+++ bindings/go/fastlowess/version.go
@@ -1,0 +1,10 @@
+package fastlowess
+
+// version is the released version of this Go binding (bindings/go), tracked
+// independently of the underlying fastLowess Rust core's crate version.
+const version = "3.1.0"
+
+// Version returns the version of this Go binding package.
+func Version() string {
+	return version
+}
