### 2026-09-01 8b3a102
chore: bump to version 3.2.0
