+++ bindings/python/python/fastlowess/__version__.py
@@ -1,3 +1,3 @@
 """Package version for the fastlowess Python binding."""
 
-__version__ = "3.1.0"
+__version__ = "3.2.0"
