+++ bindings/go/docs/NEWS.md
@@ -4,11 +4,11 @@ weight: 100
 ---
 
 <!-- markdownlint-disable MD024 MD025 -->
-# fastlowess (Go) (development version)
+# fastlowess (Go) 3.2.0
 
 ## Added
 
-* Added a new Go binding (`bindings/go`), consuming the `fastLowess` Rust core via a dedicated `cgo`-compatible C ABI (`fastlowess-go` crate, mirroring the C++ binding's FFI approach). Provides `Lowess` (batch), `StreamingLowess`, and `OnlineLowess` model types with an idiomatic Go API (`Options` struct, `Result`/`PointResult`/`Diagnostics` types, `Close()`-based resource management with a GC finalizer safety net). Ships with unit tests, a `Makefile` (`go`/`go-dev`/`go-clean` targets, wired into the root `Makefile`), CI (`ci-go.yml`, `release-go.yml`), and docs (`getting-started.md`, `api.md`, `api-streaming.md`, `api-online.md`, `installation.md`, `benchmarks.md`). The native library builds for the GNU ABI on Windows (`x86_64-pc-windows-gnu`), since Go's `cgo` drives a GCC-compatible compiler there rather than MSVC. Doc snippets are verified via a new `dev/runners/go.py` (wired into `dev/verify_snippets.py`, `dev/add-go-outputs.py`), which runs each snippet in a throwaway module that `replace`s the binding with the local checkout.
+* Added a new Go binding (`bindings/go`), consuming the `fastLowess` Rust core via a dedicated `cgo`-compatible C ABI (`fastlowess-go` crate, mirroring the C++ binding's FFI approach).
 
 ## Changed
 
