+++ bindings/python/Makefile
@@ -88,7 +88,7 @@ dev:
 	@echo "Doc snippet tests (python)..."
 	@echo "=============================================================================="
 	@$(PYTHON) dev/verify_snippets.py --lang python --timeout 120
-	@$(PYTHON) dev/update_changelogs.py
+	@$(PYTHON) dev/update_changelogs.py python
 
 ensure-llvm-cov:
 	@cargo llvm-cov --version > /dev/null 2>&1 || (echo "Installing cargo-llvm-cov..." && cargo install cargo-llvm-cov && cargo llvm-cov install-llvm-tools)
