+++ crates/fastLowess/Makefile
@@ -101,7 +101,7 @@ dev:
 	@$(PYTHON) dev/verify_snippets.py --lang rust --rust-docs $(FASTLOWESS_DIR)/docs --timeout 120
 	@echo "Adding Rust doc snippet outputs..."
 	@$(PYTHON) dev/add-rust-outputs.py $(FASTLOWESS_DIR)/docs
-	@$(PYTHON) dev/update_changelogs.py
+	@$(PYTHON) dev/update_changelogs.py fastlowess
 	@echo "All $(FASTLOWESS_PKG) crate checks completed successfully!"
 
 ensure-llvm-cov:
