+++ bindings/wasm/Makefile
@@ -85,7 +85,7 @@ dev:
 	@$(NODE) dev/add-wasm-outputs.js $(WASM_DIR)
 	@$(NODE) dev/add-readme-to-docs.js $(WASM_DIR)
 	@$(PYTHON) dev/verify_snippets.py --lang wasm --timeout 120
-	@$(PYTHON) dev/update_changelogs.py
+	@$(PYTHON) dev/update_changelogs.py wasm
 	@echo "=============================================================================="
 clean:
 	@echo "Cleaning $(WASM_PKG)..."
