### 2026-09-01 e9b43b1
refactor: verify_snippets now also adds outputs
