### 2026-09-01 06202e3
refactor: consolidate add_readme scripts into one
