### 2026-09-01 4956404
refactor: consolidate spack release workflow
