### 2026-09-01 6eda3cd
refactor: improve maven versioning
