### 2026-09-01 5655dd5
build: add more compiler support to cpp
