+++ .github/workflows/ci-cpp.yml
@@ -73,6 +73,120 @@ jobs:
       - name: make cpp-dev
         run: make cpp-dev
 
+  clangcl:
+    name: C++ (windows-latest, clang-cl)
+    runs-on: windows-latest
+    steps:
+      - uses: actions/checkout@v7
+      - uses: dtolnay/rust-toolchain@stable
+      - name: Install GNU Make (Windows)
+        run: choco install make -y
+      - name: Install cppcheck (Windows)
+        run: |
+          winget install --id Cppcheck.Cppcheck --exact --accept-source-agreements --accept-package-agreements
+          $cppcheckExe = Get-ChildItem -Path "C:\Program Files\Cppcheck","$env:LOCALAPPDATA\Programs\Cppcheck" -Filter cppcheck.exe -Recurse -ErrorAction SilentlyContinue | Select-Object -First 1
+          if (-not $cppcheckExe) { throw "cppcheck.exe not found after winget install" }
+          $cppcheckExe.DirectoryName | Out-File -FilePath $env:GITHUB_PATH -Append -Encoding utf8
+      - uses: Swatinem/rust-cache@v2
+        with:
+          key: cpp-windows-clangcl
+      - name: make cpp-dev (clang-cl toolset)
+        run: make cpp-dev CPP_CMAKE_TOOLSET="-T ClangCL"
+
+  clang-linux:
+    name: C++ (ubuntu-latest, clang)
+    runs-on: ubuntu-latest
+    env:
+      CC: clang
+      CXX: clang++
+    steps:
+      - uses: actions/checkout@v7
+      - uses: dtolnay/rust-toolchain@stable
+        with:
+          components: rustfmt, clippy
+      - name: Install clang and cppcheck
+        run: sudo apt-get install -y --no-install-recommends clang cppcheck
+      - uses: Swatinem/rust-cache@v2
+        with:
+          key: cpp-clang-linux
+      - name: make cpp-dev
+        run: make cpp-dev
+
+  # Experimental: native MinGW-w64 (not the Rtools cross-compiler pinned in the
+  # workspace .cargo/config.toml for the R binding, which lacks libgcc_eh — see
+  # docs/NEWS.md). Non-blocking until proven stable across CI runs.
+  mingw:
+    name: C++ (windows-latest, MinGW-w64) [experimental]
+    runs-on: windows-latest
+    continue-on-error: true
+    env:
+      CARGO_TARGET_X86_64_PC_WINDOWS_GNU_LINKER: gcc
+      CARGO_TARGET_X86_64_PC_WINDOWS_GNU_AR: gcc-ar
+    steps:
+      - uses: actions/checkout@v7
+      - uses: dtolnay/rust-toolchain@stable
+        with:
+          targets: x86_64-pc-windows-gnu
+      - name: Install native MinGW-w64
+        run: |
+          choco install mingw -y
+          $gcc = Get-ChildItem -Path "C:\ProgramData\chocolatey\lib\mingw","C:\tools\mingw64" -Filter gcc.exe -Recurse -ErrorAction SilentlyContinue | Select-Object -First 1
+          if (-not $gcc) { throw "gcc.exe not found after choco install mingw" }
+          $gcc.DirectoryName | Out-File -FilePath $env:GITHUB_PATH -Append -Encoding utf8
+      - uses: Swatinem/rust-cache@v2
+        with:
+          key: cpp-windows-mingw
+      - name: Build (MinGW target)
+        working-directory: bindings/cpp
+        run: cargo build --profile release-c --target x86_64-pc-windows-gnu
+      - name: Configure & build C++ tests (MinGW)
+        run: |
[…]
