+++ bindings/cpp/CMakeLists.txt
@@ -9,7 +9,7 @@ include(GNUInstallDirs)
 FetchContent_Declare(
     Corrosion
     GIT_REPOSITORY https://github.com/corrosion-rs/corrosion.git
-    GIT_TAG v0.5.1
+    GIT_TAG v0.6.1
 )
 FetchContent_MakeAvailable(Corrosion)
 
