### 2026-09-01 d9e3fe7
fix: unify the defaults across the repo
