+++ bindings/julia/julia/src/FastLOWESS.jl
@@ -754,14 +754,14 @@ Stateful online LOWESS smoother.
 # Keyword Arguments
 - `fraction::Float64 = 0.67`: Smoothing fraction. See `Lowess` for guidance on choosing a value.
 - `window_capacity::Int = 1000`: Maximum points to retain in window, at least 3
-- `min_points::Int = 3`: Minimum points before smoothing starts, between 2 and `window_capacity`
+- `min_points::Int = 2`: Minimum points before smoothing starts, between 2 and `window_capacity`
 - `iterations::Int = 3`: Number of robustness iterations. See `Lowess` for guidance on choosing a value.
 - `delta::Float64 = NaN`: Interpolation threshold (NaN auto-sets it to 0.0)
 - `weight_function::String = "tricube"`: Kernel function
 - `robustness_method::String = "bisquare"`: Robustness method
 - `scaling_method::String = "mad"`: Scaling method
 - `boundary_policy::String = "extend"`: Boundary handling
-- `update_mode::String = "full"`: Update strategy ("full" or "incremental")
+- `update_mode::String = "incremental"`: Update strategy ("full" or "incremental")
 - `auto_converge::Float64 = NaN`: Auto-convergence tolerance
 - `return_robustness_weights::Bool = false`: Include weights
 - `return_diagnostics::Bool = false`: Compute diagnostics (RMSE, MAE, R², etc.)
@@ -776,14 +776,14 @@ mutable struct OnlineLowess
     function OnlineLowess(;
         fraction::Float64 = 0.67,
         window_capacity::Int = 1000,
-        min_points::Int = 3,
+        min_points::Int = 2,
         iterations::Int = 3,
         delta::Float64 = NaN,
         weight_function::String = "tricube",
         robustness_method::String = "bisquare",
         scaling_method::String = "mad",
         boundary_policy::String = "extend",
-        update_mode::String = "full",
+        update_mode::String = "incremental",
         auto_converge::Float64 = NaN,
         return_robustness_weights::Bool = false,
         return_diagnostics::Bool = false,
