+++ crates/fastLowess/docs/adapter-choice.md
@@ -33,6 +33,7 @@ Choose the first row below whose condition applies:
 | Diagnostics | ✓ | ✓ | ✗ |
 | Residuals | ✓ | ✓ | ✓ |
 | Robustness weights | ✓ | ✓ | ✓ |
+| Parallel execution | ✓ | ✓ | ✗ |
 
 ---
 

[…]
