### 2026-09-01 38f5632
docs: standardize docs across bindings/crates
