+++ bindings/go/docs/guide/cross-validation.md
@@ -182,7 +182,7 @@ Selected fraction (CV): 0.7
 Cross-validation uses MSE (Mean Squared Error) by default:
 
 ```text
-MSE = mean((y_true - y_pred)²)
+MSE = mean((y_true - y_pred)^2)
 ```
 
 Lower MSE indicates better fit on held-out data.

[…]
