### 2026-09-01 9a1c0cb
docs: restructure to a standard format
