### 2026-09-01 2a3ff69
docs: migrate java docs to antora
