+++ dev/verify_snippets.py
@@ -125,6 +125,38 @@ def _find_python_with_fastlowess() -> str:
 # ---------------------------------------------------------------------------
 
 _TAB_RE = re.compile(r'^[ \t]*===\s+"([^"]+)"', re.MULTILINE)
+_ADOC_SOURCE_RE = re.compile(r"^\[source,(\w+)\]\s*$")
+
+
+def _extract_adoc_snippets(adoc_file: Path) -> list[Snippet]:
+    """Extract `[source,LANG]` / `----` ... `----` blocks from an AsciiDoc file."""
+    lines = adoc_file.read_text(encoding="utf-8").splitlines()
+    result: list[Snippet] = []
+    i = 0
+    while i < len(lines):
+        m = _ADOC_SOURCE_RE.match(lines[i])
+        if m and i + 1 < len(lines) and lines[i + 1].strip() == "----":
+            lang_tag = m.group(1)
+            start_line = i + 1  # 1-based line of the `[source,lang]` attribute
+            i += 2
+            code_lines: list[str] = []
+            while i < len(lines) and lines[i].strip() != "----":
+                code_lines.append(lines[i])
+                i += 1
+            if i < len(lines):
+                i += 1  # consume closing ----
+            result.append(
+                Snippet(
+                    file=adoc_file,
+                    line=start_line,
+                    lang_tag=lang_tag,
+                    tab=None,
+                    code="\n".join(code_lines),
+                )
+            )
+            continue
+        i += 1
+    return result
 
 
 def _rmd_chunk_is_runnable(opts_str: str) -> bool:
@@ -139,6 +171,8 @@ def extract_snippets(md_file: Path) -> list[Snippet]:
     For .Rmd files all runnable R chunks are combined into one snippet, since
     Rmd chunks share state (later chunks depend on variables set by earlier ones).
     """
+    if md_file.suffix.lower() == ".adoc":
+        return _extract_adoc_snippets(md_file)
     text = md_file.read_text(encoding="utf-8")
     lines = text.splitlines()
     result: list[Snippet] = []
@@ -342,6 +376,9 @@ OUTPUT_CAPABLE_RUNNERS = {"cpp", "go", "java", "rust", "nodejs", "wasm"}
 def _splice_output_block(lines: list[str], res: RunResult) -> None:
     """Insert/update/remove the ```output block following one snippet, in place."""
     snippet = res.snippet
+    if snippet.file.suffix.lower() == ".adoc":
+        _splice_adoc_output_block(lines, res)
+        return
     code_line_count = len(snippet.code.split("\n"))
     open_idx = snippet.line - 1  # 0-based index of the opening ``` fence
     close_idx = open_idx + 1 + code_line_count  # 0-based index of the closing ``` fence
@@ -365,6 +402,42 @@ def _splice_output_block(lines: list[str], res: RunResult) -> None:
 
     new_block = (
         ["", "```output", *res.stdout.rstrip().split("\n"), "```"]
+        if res.stdout.strip()
+        else []  # passed with no output -> drop any stale block
+    )
+    if end_of_existing is not None:
+        lines[close_idx + 1 : end_of_existing + 1] = new_block
+    else:
+        lines[close_idx + 1 : close_idx + 1] = new_block
+
+
+def _splice_adoc_output_block(lines: list[str], res: RunResult) -> None:
+    """AsciiDoc analogue of `_splice_output_block`, for `[source,console]` blocks."""
+    snippet = res.snippet
+    code_line_count = len(snippet.code.split("\n"))
+    open_idx = snippet.line - 1  # 0-based index of the `[source,lang]` attribute
[…]
