+++ bindings/java/docs/modules/ROOT/pages/guide/adapter-choice.adoc
@@ -1,0 +1,52 @@
+= Execution Modes
+
+Choose the right adapter for your use case.
+
+== Overview
+
+Choose the first row below whose condition applies:
+
+[cols="1,1",options="header"]
+|===
+|Condition |Adapter
+
+|Data too large to fit in memory |`StreamingLowess`
+|Fits in memory, need real-time/incremental updates |`OnlineLowess`
+|Fits in memory, no real-time requirement |`Lowess`
+|===
+
+[cols="1,1,1,1",options="header"]
+|===
+|Mode |Use Case |Memory |Features
+
+|*Batch* (`Lowess`) |Complete datasets |Full |All features
+|*Streaming* (`StreamingLowess`) |Large files (>100K) |Chunked |Residuals, robustness
+|*Online* (`OnlineLowess`) |Real-time sensors |Fixed window |Incremental updates
+|===
+
+image::adapter_comparison.svg[Adapter Comparison]
+
+'''
+
+== Feature Comparison
+
+[cols="1,1,1,1",options="header"]
+|===
+|Feature |Batch |Streaming |Online
+
+|Confidence intervals |✓ |✗ |✗
+|Prediction intervals |✓ |✗ |✗
+|Cross-validation |✓ |✗ |✗
+|Diagnostics |✓ |✓ |✗
+|Residuals |✓ |✓ |✓
+|Robustness weights |✓ |✓ |✓
+|Parallel execution |✓ |✓ |✗
+|===
+
+'''
+
+== Next Steps
+
+* xref:../api/api.adoc[API Reference] — All configuration options
+* xref:../api/api-streaming.adoc[Streaming API] · xref:../api/api-online.adoc[Online API]
+

[…]
