+++ cpp/search/all_8.js
@@ -1,9 +1,5 @@
 var searchData=
 [
-  ['installation_0',['Installation',['../installation.html',1,'getting_started']]],
-  ['installation_2emd_1',['installation.md',['../installation_8md.html',1,'']]],
-  ['intervals_2',['Intervals',['../intervals.html',1,'user_guide']]],
-  ['intervals_2emd_3',['intervals.md',['../intervals_8md.html',1,'']]],
-  ['iterations_4',['iterations',['../structfastlowess_1_1LowessOptions.html#a3a9b03636d3cb266eb345e3f405ba66b',1,'fastlowess::LowessOptions::iterations'],['../structfastlowess_1_1OnlineOptions.html#a2c8c2dd03559f291a428467d8d231790',1,'fastlowess::OnlineOptions::iterations']]],
-  ['iterations_5fused_5',['iterations_used',['../classfastlowess_1_1OnlineOutput.html#ab895dec754b68f10ca75863ff07c19fc',1,'fastlowess::OnlineOutput::iterations_used()'],['../classfastlowess_1_1LowessResult.html#a8a62026a77efb2bc918a2e4bdde38e09',1,'fastlowess::LowessResult::iterations_used()']]]
+  ['handling_0',['Boundary Handling',['../boundary.html',1,'advanced']]],
+  ['has_5fvalue_1',['has_value',['../classfastlowess_1_1Expected.html#a8515f14a4a841c380b59f8ff702232a5',1,'fastlowess::Expected::has_value()'],['../classfastlowess_1_1OnlineOutput.html#af9ee53d7c3732342c2eb7dce5ec490da',1,'fastlowess::OnlineOutput::has_value()'],['../classfastlowess_1_1Diagnostics.html#a3f6dcd4b9feb2a5760b9892df3b8d200',1,'fastlowess::Diagnostics::has_value()']]]
 ];
