+++ cpp/search/all_5.js
@@ -1,14 +1,8 @@
 var searchData=
 [
-  ['fastlowess_0',['fastlowess',['../namespacefastlowess.html',1,'']]],
-  ['fastlowess_20cpp_1',['fastlowess-cpp',['../index.html',1,'']]],
-  ['fastlowess_2ehpp_2',['fastlowess.hpp',['../fastlowess_8hpp.html',1,'']]],
-  ['fastlowess_3a_3adetail_3',['detail',['../namespacefastlowess_1_1detail.html',1,'fastlowess']]],
-  ['fastlowess_3a_3agpu_4',['gpu',['../namespacefastlowess_1_1gpu.html',1,'fastlowess']]],
-  ['fastlowess_3a_3agpu_3a_3adetail_5',['detail',['../namespacefastlowess_1_1gpu_1_1detail.html',1,'fastlowess::gpu']]],
-  ['finalize_6',['finalize',['../classfastlowess_1_1StreamingLowess.html#a2e00c5df38b02cf9b32e3eba9f75cc09',1,'fastlowess::StreamingLowess']]],
-  ['fit_7',['fit',['../classfastlowess_1_1Lowess.html#a504a722970109b32ab8318352d224a2e',1,'fastlowess::Lowess']]],
-  ['fraction_8',['fraction',['../structfastlowess_1_1LowessOptions.html#ac4c7e581cd561fc11a928f9901bb5ec6',1,'fastlowess::LowessOptions::fraction'],['../structfastlowess_1_1OnlineOptions.html#a6acb9ee334a76cd3701228b0cab4ac62',1,'fastlowess::OnlineOptions::fraction']]],
-  ['fraction_5fused_9',['fraction_used',['../classfastlowess_1_1LowessResult.html#a80b7486011abac87056d6449ed80209b',1,'fastlowess::LowessResult']]],
-  ['functions_10',['Weight Functions',['../kernels.html',1,'customization']]]
+  ['effective_5fdf_0',['effective_df',['../classfastlowess_1_1Diagnostics.html#ac4dafab55a863e978488bc626120787f',1,'fastlowess::Diagnostics']]],
+  ['error_1',['error',['../classfastlowess_1_1Expected.html#a7dcf16472fcc4e069a35aa4e40b50762',1,'fastlowess::Expected::error()'],['../classfastlowess_1_1LowessResult.html#a9baf4ca720c8ae01c38989c1f782af2f',1,'fastlowess::LowessResult::error()']]],
+  ['errortag_2',['ErrorTag',['../structfastlowess_1_1Expected_1_1ErrorTag.html',1,'fastlowess::Expected']]],
+  ['execution_20modes_3',['Execution Modes',['../adapter_choice.html',1,'user_guide']]],
+  ['expected_4',['expected',['../classfastlowess_1_1Expected.html',1,'fastlowess::Expected&lt; T &gt;'],['../classfastlowess_1_1OnlineOutput.html#a815f6718ecaf8321aab0587f5e62d048',1,'fastlowess::OnlineOutput::Expected'],['../classfastlowess_1_1Expected.html#ab76837c64ed44e1c501627a3332be730',1,'fastlowess::Expected::Expected()']]]
 ];
