+++ cpp/search/all_15.js
@@ -1,7 +1,6 @@
 var searchData=
 [
-  ['weight_20functions_0',['Weight Functions',['../kernels.html',1,'customization']]],
-  ['weight_5ffunction_1',['weight_function',['../structfastlowess_1_1LowessOptions.html#a73dba24a8c056492b21bf2215a1c253f',1,'fastlowess::LowessOptions::weight_function'],['../structfastlowess_1_1OnlineOptions.html#a7e39e17c0be1d7c223eb7347b40fcba1',1,'fastlowess::OnlineOptions::weight_function']]],
-  ['weights_2',['Custom Weights',['../custom_weights.html',1,'customization']]],
-  ['window_5fcapacity_3',['window_capacity',['../structfastlowess_1_1OnlineOptions.html#ab1a4b461877e41a676ab9b555995bfab',1,'fastlowess::OnlineOptions']]]
+  ['valid_0',['valid',['../classfastlowess_1_1LowessResult.html#a2a5eb76cd3ae3890e8c91597c74aba0d',1,'fastlowess::LowessResult']]],
+  ['validation_1',['Cross-Validation',['../cross_validation.html',1,'user_guide']]],
+  ['value_2',['value',['../classfastlowess_1_1Expected.html#aeb88e40ff81b811771e6e2013fad7a6c',1,'fastlowess::Expected::value() &amp;'],['../classfastlowess_1_1Expected.html#ab9f1bddd03e221b49c256931b13f0dec',1,'fastlowess::Expected::value() const &amp;'],['../classfastlowess_1_1Expected.html#aacabf25f31eaf0838faf06f0165a57ae',1,'fastlowess::Expected::value() &amp;&amp;']]]
 ];
