+++ cpp/search/all_e.js
@@ -1,9 +1,10 @@
 var searchData=
 [
-  ['parallel_0',['parallel',['../structfastlowess_1_1LowessOptions.html#a689dba738384297e91479aa158dc2e45',1,'fastlowess::LowessOptions::parallel'],['../structfastlowess_1_1OnlineOptions.html#ae627d89402120085520536c4f2737f54',1,'fastlowess::OnlineOptions::parallel']]],
-  ['prediction_5fintervals_1',['prediction_intervals',['../structfastlowess_1_1LowessOptions.html#a48d749c4b025fc0467ef40c62ce6a939',1,'fastlowess::LowessOptions::prediction_intervals'],['../structfastlowess_1_1OnlineOptions.html#aa1e4b6076f7ce0b2835aa1fa57040426',1,'fastlowess::OnlineOptions::prediction_intervals']]],
-  ['prediction_5flower_2',['prediction_lower',['../classfastlowess_1_1LowessResult.html#a42811d23a313b27caf306b4bb70a5d92',1,'fastlowess::LowessResult']]],
-  ['prediction_5fupper_3',['prediction_upper',['../classfastlowess_1_1LowessResult.html#ac0cdc67bf75ce4ca783c0be54ca2b72d',1,'fastlowess::LowessResult']]],
-  ['process_5fchunk_4',['process_chunk',['../classfastlowess_1_1StreamingLowess.html#add63b1329c184527f5f85be31b6710d5',1,'fastlowess::StreamingLowess']]],
-  ['processing_5',['Real-Time Processing',['../use_case_real_time.html',1,'use_cases']]]
+  ['onlinelowess_0',['onlinelowess',['../classfastlowess_1_1OnlineLowess.html',1,'fastlowess::OnlineLowess'],['../classfastlowess_1_1OnlineOutput.html#a5e236589da58bef059203559a019ad80',1,'fastlowess::OnlineOutput::OnlineLowess'],['../classfastlowess_1_1OnlineLowess.html#a24db2d32e3fec45bb3388336c0200dc0',1,'fastlowess::OnlineLowess::OnlineLowess(const OnlineOptions &amp;options={})'],['../classfastlowess_1_1OnlineLowess.html#a6ad68adac17764c91153df5a02ac1ce5',1,'fastlowess::OnlineLowess::OnlineLowess(const OnlineLowess &amp;)=delete'],['../classfastlowess_1_1OnlineLowess.html#a9a99929c5de21a4dc41a1b7182f48c33',1,'fastlowess::OnlineLowess::OnlineLowess(OnlineLowess &amp;&amp;other) noexcept']]],
+  ['onlinelowess_20api_1',['OnlineLowess API',['../api_online.html',1,'api']]],
+  ['onlineoptions_2',['OnlineOptions',['../structfastlowess_1_1OnlineOptions.html',1,'fastlowess']]],
+  ['onlineoutput_3',['OnlineOutput',['../classfastlowess_1_1OnlineOutput.html',1,'fastlowess']]],
+  ['operator_20bool_4',['operator bool',['../classfastlowess_1_1Expected.html#ad361b5ccf1163a38e469847f2b1adf77',1,'fastlowess::Expected']]],
[…]

[…]
