### 2026-09-01 840ed62
feat: add dependabot
