### 2026-09-01 6b55c36
fix: setup-go cache warning
