+++ .github/workflows/ci-cpp.yml
@@ -112,11 +112,10 @@ jobs:
       - name: make cpp-dev
         run: make cpp-dev
 
-  # Experimental: native MinGW-w64 (not the Rtools cross-compiler pinned in the
-  # workspace .cargo/config.toml for the R binding, which lacks libgcc_eh — see
-  # docs/NEWS.md). Non-blocking until proven stable across CI runs.
+  # Native MinGW-w64 (not the Rtools cross-compiler pinned in the workspace
+  # .cargo/config.toml for the R binding, which lacks libgcc_eh — see docs/NEWS.md).
   mingw:
-    name: C++ (windows-latest, MinGW-w64) [experimental]
+    name: C++ (windows-latest, MinGW-w64)
     runs-on: windows-latest
     continue-on-error: true
     env:
@@ -127,35 +126,31 @@ jobs:
       - uses: dtolnay/rust-toolchain@stable
         with:
           targets: x86_64-pc-windows-gnu
+      - name: Install GNU Make (Windows)
+        run: choco install make -y
       - name: Install native MinGW-w64
         run: |
           choco install mingw -y
-          $gcc = Get-ChildItem -Path "C:\ProgramData\chocolatey\lib\mingw","C:\tools\mingw64" -Filter gcc.exe -Recurse -ErrorAction SilentlyContinue | Select-Object -First 1
+          $gcc = Get-ChildItem -Path "C:\ProgramData\mingw64","C:\ProgramData\chocolatey\lib\mingw","C:\tools\mingw64" -Filter gcc.exe -Recurse -ErrorAction SilentlyContinue | Select-Object -First 1
           if (-not $gcc) { throw "gcc.exe not found after choco install mingw" }
           $gcc.DirectoryName | Out-File -FilePath $env:GITHUB_PATH -Append -Encoding utf8
+      - name: Install cppcheck (Windows)
+        run: |
+          winget install --id Cppcheck.Cppcheck --exact --accept-source-agreements --accept-package-agreements
+          $cppcheckExe = Get-ChildItem -Path "C:\Program Files\Cppcheck","$env:LOCALAPPDATA\Programs\Cppcheck" -Filter cppcheck.exe -Recurse -ErrorAction SilentlyContinue | Select-Object -First 1
+          if (-not $cppcheckExe) { throw "cppcheck.exe not found after winget install" }
+          $cppcheckExe.DirectoryName | Out-File -FilePath $env:GITHUB_PATH -Append -Encoding utf8
       - uses: Swatinem/rust-cache@v2
         with:
           key: cpp-windows-mingw
-      - name: Build (MinGW target)
-        working-directory: bindings/cpp
-        run: cargo build --profile release-c --target x86_64-pc-windows-gnu
-      - name: Configure & build C++ tests (MinGW)
-        run: |
-          cmake -S bindings/cpp/tests -B bindings/cpp/tests/build-mingw `
-            -G "MinGW Makefiles" `
-            -DCMAKE_BUILD_TYPE=Release `
-            -DFASTLOWESS_LIB_DIR="$(Resolve-Path target/x86_64-pc-windows-gnu/release-c)"
-          cmake --build bindings/cpp/tests/build-mingw
-      - name: Run C++ tests (MinGW)
-        run: |
-          Copy-Item target/x86_64-pc-windows-gnu/release-c/fastlowess_cpp.dll bindings/cpp/tests/build-mingw/
-          & bindings/cpp/tests/build-mingw/test_fastlowess_suite.exe
+      - name: make cpp-dev
+        run: make cpp-dev CPP_WIN_TOOLCHAIN=mingw
 
-  # Experimental: Intel oneAPI C++ compiler (icpx) as a consumer of the C ABI
[…]
