+++ bindings/cpp/Makefile
@@ -39,10 +39,18 @@ CPP_CARGO_PROFILE := --profile release-c
 CPP_LIBRARY_DIR := target/release-c
 CPP_CARGO_TARGET :=
 
+# Windows toolchain: msvc (default) or mingw, e.g. `CPP_WIN_TOOLCHAIN=mingw` to
+# cross-build/test against the native MinGW-w64 target x86_64-pc-windows-gnu.
+CPP_WIN_TOOLCHAIN ?= msvc
+
 ifeq ($(OS),Windows_NT)
-	_CPP_WIN_TOOLCHAIN := msvc
-	CPP_CARGO_TARGET   := --target x86_64-pc-windows-msvc
-	CPP_LIBRARY_DIR    := target/x86_64-pc-windows-msvc/release-c
+	ifeq ($(CPP_WIN_TOOLCHAIN),mingw)
+		CPP_CARGO_TARGET := --target x86_64-pc-windows-gnu
+		CPP_LIBRARY_DIR  := target/x86_64-pc-windows-gnu/release-c
+	else
+		CPP_CARGO_TARGET := --target x86_64-pc-windows-msvc
+		CPP_LIBRARY_DIR  := target/x86_64-pc-windows-msvc/release-c
+	endif
 endif
 
 # Optional Windows CMake toolset override, e.g. `CPP_CMAKE_TOOLSET="-T ClangCL"`
@@ -50,12 +58,19 @@ endif
 CPP_CMAKE_TOOLSET ?=
 
 ifeq ($(HOST_PLATFORM),windows)
-	CPP_SHARED_LIB      := $(CPP_LIBRARY_DIR)/fastlowess_cpp.dll
-	CPP_EXPORT_SCAN     := objdump -p $(CPP_SHARED_LIB)
-	CPP_TEST_LIB        := $(CPP_LIBRARY_DIR)/fastlowess_cpp.lib
-	CPP_CMAKE_GENERATOR := $(CPP_CMAKE_TOOLSET)
-	CPP_TEST_BUILD      := cmake --build . --config Release
-	CPP_TEST_RUN        := cp -f "$(abspath $(CPP_SHARED_LIB))" Release/ && ./Release/test_fastlowess_suite.exe
+	CPP_SHARED_LIB  := $(CPP_LIBRARY_DIR)/fastlowess_cpp.dll
+	CPP_EXPORT_SCAN := objdump -p $(CPP_SHARED_LIB)
+	ifeq ($(CPP_WIN_TOOLCHAIN),mingw)
+		CPP_TEST_LIB        :=
+		CPP_CMAKE_GENERATOR := -G "MinGW Makefiles"
+		CPP_TEST_BUILD      := cmake --build .
+		CPP_TEST_RUN        := cp -f "$(abspath $(CPP_SHARED_LIB))" . && ./test_fastlowess_suite.exe
+	else
+		CPP_TEST_LIB        := $(CPP_LIBRARY_DIR)/fastlowess_cpp.lib
+		CPP_CMAKE_GENERATOR := $(CPP_CMAKE_TOOLSET)
+		CPP_TEST_BUILD      := cmake --build . --config Release
+		CPP_TEST_RUN        := cp -f "$(abspath $(CPP_SHARED_LIB))" Release/ && ./Release/test_fastlowess_suite.exe
+	endif
 else ifeq ($(HOST_PLATFORM),macos)
 	CPP_SHARED_LIB      := $(CPP_LIBRARY_DIR)/libfastlowess_cpp.dylib
 	CPP_TEST_LIB        := $(CPP_SHARED_LIB)
