### 2026-09-01 7928b69
fix: java docs navigation bar
