+++ bindings/java/docs/antora.yml
@@ -1,4 +1,4 @@
-name: java
+name: ROOT
 title: fastlowess (Java)
 version: ~
 start_page: ROOT:index.adoc
