### 2026-09-02 6144ba2
chore: repack, minor doc fixes, add version check
