+++ bindings/julia/julia/docs/make.jl
@@ -8,37 +8,37 @@ const INDEX_PATH = joinpath(@__DIR__, "src", "index.md")
 # render the raw `<p align="center">` badge/logo HTML blocks at the top of the
 # shared README — convert them to plain Markdown before using it as the homepage.
 function html_center_block_to_markdown(block::AbstractString)
-	lines = String[]
-	for raw_line ∈ eachsplit(strip(block), '\n')
-		line = strip(raw_line)
-		if (
-			m = match(
-				r"^<a href=\"([^\"]+)\"><img src=\"([^\"]+)\" alt=\"([^\"]+)\"></a>$",
-				line,
-			)
-		) !== nothing
-			push!(lines, "[![$(m[3])]($(m[2]))]($(m[1]))")
-		elseif (
-			m = match(
-				r"^<img src=\"([^\"]+)\" alt=\"([^\"]+)\"(?: width=\"[^\"]+\")?>$",
-				line,
-			)
-		) !== nothing
-			push!(lines, "![$(m[2])]($(m[1]))")
-		elseif (m = match(r"^<em>(.+)</em>$", line)) !== nothing
-			push!(lines, "*$(m[1])*")
-		elseif line == "<br>" || isempty(line)
-			continue
-		else
-			push!(lines, line)
-		end
-	end
-	return join(lines, "\n\n")
+    lines = String[]
+    for raw_line ∈ eachsplit(strip(block), '\n')
+        line = strip(raw_line)
+        if (
+            m = match(
+                r"^<a href=\"([^\"]+)\"><img src=\"([^\"]+)\" alt=\"([^\"]+)\"></a>$",
+                line,
+            )
+        ) !== nothing
+            push!(lines, "[![$(m[3])]($(m[2]))]($(m[1]))")
+        elseif (
+            m = match(
+                r"^<img src=\"([^\"]+)\" alt=\"([^\"]+)\"(?: width=\"[^\"]+\")?>$",
+                line,
+            )
+        ) !== nothing
+            push!(lines, "![$(m[2])]($(m[1]))")
+        elseif (m = match(r"^<em>(.+)</em>$", line)) !== nothing
+            push!(lines, "*$(m[1])*")
+        elseif line == "<br>" || isempty(line)
+            continue
+        else
+            push!(lines, line)
+        end
+    end
+    return join(lines, "\n\n")
 end
 
 function convert_center_block(matched::AbstractString)
-	inner = match(r"<p align=\"center\">\n(.*)\n</p>"s, matched).captures[1]
-	return html_center_block_to_markdown(inner)
+    inner = match(r"<p align=\"center\">\n(.*)\n</p>"s, matched).captures[1]
+    return html_center_block_to_markdown(inner)
 end
 
 readme = read(README_PATH, String)
@@ -60,66 +60,66 @@ write(INDEX_PATH, readme)
 const SRC_DIR = joinpath(@__DIR__, "src")
 const LEADING_COMMENT_RE = r"\A<!--.*?-->\n?"s
 
-original_contents = Dict{String, String}()
+original_contents = Dict{String,String}()
 md_paths = String[]
 for (root, _, files) ∈ walkdir(SRC_DIR)
-	for file ∈ files
-		endswith(file, ".md") && push!(md_paths, joinpath(root, file))
-	end
+    for file ∈ files
+        endswith(file, ".md") && push!(md_paths, joinpath(root, file))
+    end
 end
 for path ∈ md_paths
-	content = read(path, String)
-	if occursin(LEADING_COMMENT_RE, content)
-		original_contents[path] = content
-		write(path, replace(content, LEADING_COMMENT_RE => ""))
-	end
+    content = read(path, String)
[…]
