+++ r/llms.txt
@@ -13,8 +13,8 @@ All](https://raw.githubusercontent.com/thisisamirv/lowess-project/main/dev/logo.
 
 The fastest, most robust, and most feature-complete language-agnostic
 LOWESS (Locally Weighted Scatterplot Smoothing) implementation for
-**Rust**, **Python**, **R**, **Julia**, **JavaScript**, **C++**, and
-**WebAssembly**.
+**Rust**, **Python**, **R**, **Julia**, **JavaScript**, **C++**, **Go**,
+**Java**, and **WebAssembly**.
 
 The `lowess-project` also offers bindings for Rust, Python, R, Julia,
 Node.js, WebAssembly, C++, Go, and Java — see the [full
