+++ r/news/index.md
@@ -2,25 +2,32 @@
 
 ## rfastlowess 3.2.0
 
+### Added
+
+- Added `dev/add-readme-to-docs.py`, which auto-detects the Hugo (Go,
+  Java) vs Starlight (Node.js, WASM) docs-site flavor and embeds
+  `README.md` accordingly; wired into the corresponding Makefiles and
+  `package.json` scripts.
+- Added `.github/dependabot.yml`, covering every dependency ecosystem in
+  the repo (`github-actions`, `cargo`, `npm`, `pip`, `maven`, `gomod`).
+  Each directory is grouped so all its updates, including majors, land
+  in a single weekly PR.
+
 ### Changed
 
-- Added a `large` benchmark category to `benchmarks/rfastlowess.R` and
-  `benchmarks/stats_lowess.R` (n = 50000, `delta = 0` to disable
-  [`stats::lowess`](https://rdrr.io/r/stats/lowess.html)’s interpolation
-  shortcut for a fair, exact-computation comparison), since every
-  existing category completed in well under 100ms. Reliably takes
-  ~13-16s for [`stats::lowess`](https://rdrr.io/r/stats/lowess.html) and
-  ~3s for `fastLowess` (serial). Expanded it to 4 scenarios stressing
-  different parameters: `large_delta_0` (the original exact-fit case),
-  `large_delta_0.1` (same workload with `delta` left at its default/auto
-  value, showing the interpolation shortcut’s speedup),
-  `large_high_iter` (10 robustness iterations instead of 3), and
-  `large_high_fraction` (n = 20000, `fraction = 0.67`, since a 0.67
-  fraction at n = 50000 takes over a minute). `benchmarks/compare.py`’s
-  plot grid grew from 5x2 to 7x2 rows to fit the new categories
-  (`large_delta_0`/`large_delta_0.1` share one “large_delta” chart,
-  matching how `fraction_*`/`iterations_*` already group into shared
-  charts).
+- Modified `verify_snippets.py` to verify snippets and also add the
+  output of the snippets to the markdown file.
+- Added a `large` benchmark category (n = 50000) to
+  `benchmarks/rfastlowess.R` and `benchmarks/stats_lowess.R`, since
+  every existing category ran in well under 100ms. Covers 4 scenarios
+  (`large_delta_0`, `large_delta_0.1`, `large_high_iter`,
+  `large_high_fraction`) stressing `delta`, iteration count, and
+  fraction. `benchmarks/compare.py`’s plot grid grew from 5x2 to 7x2 to
+  fit them.
+- Added `.gitattributes`, normalizing all text files to LF line endings
+  (`* text=auto eol=lf`) and marking binary formats (images, archives,
+  compiled libraries, `.rds`/`.RData`, etc.) so Git never treats them as
+  text.
 - Consolidated every crate/binding README: merged the “Installation” and
   “Documentation” sections, replaced GitHub-only alert syntax with plain
   blockquotes, removed the redundant “API Reference” and “Changelog”
@@ -46,28 +53,39 @@
   (mermaid in most bindings/crates, ASCII art in the C++ docs) with an
   equivalent decision table, unifying on a single rendering-agnostic
   format across every binding/crate.
-- Moved the duplicated “GPU Acceleration” section (installation, usage,
-  supported features, feature comparison) out of `api.md` in the C++,
-  Node.js, and Python bindings and the `fastLowess` crate into each
-  one’s dedicated `gpu-backend.md` guide, which also gained the
-  “Hardware Requirements” and “Performance Considerations” subsections
-  previously only in `api.md`; `api.md` now links to `gpu-backend.md`
-  with a short blurb instead. Removed the same section from the `lowess`
[…]
