+++ bindings/java/src/main/java/fastlowess/OnlineOptions.java
@@ -18,6 +18,11 @@ public final class OnlineOptions {
         this.updateMode = b.updateMode;
     }
 
+    /**
+     * Creates a new builder.
+     *
+     * @return a new {@link Builder}
+     */
     public static Builder builder() {
         return new Builder();
     }
@@ -35,76 +40,153 @@ public final class OnlineOptions {
         Builder() {
         }
 
+        /**
+         * @param fraction the fraction of points used to compute each local
+         * regression
+         * @return this builder, for chaining
+         * @see Options.Builder#fraction(double)
+         */
         public Builder fraction(double fraction) {
             common.fraction(fraction);
             return this;
         }
 
+        /**
+         * @param iterations the number of robustifying iterations
+         * @return this builder, for chaining
+         * @see Options.Builder#iterations(int)
+         */
         public Builder iterations(int iterations) {
             common.iterations(iterations);
             return this;
         }
 
+        /**
+         * @param delta the interpolation distance
+         * @return this builder, for chaining
+         * @see Options.Builder#delta(double)
+         */
         public Builder delta(double delta) {
             common.delta(delta);
             return this;
         }
 
+        /**
+         * @param weightFunction the weight function name
+         * @return this builder, for chaining
+         * @see Options.Builder#weightFunction(String)
+         */
         public Builder weightFunction(String weightFunction) {
             common.weightFunction(weightFunction);
             return this;
         }
 
+        /**
+         * @param robustnessMethod the robustness method name
+         * @return this builder, for chaining
+         * @see Options.Builder#robustnessMethod(String)
+         */
         public Builder robustnessMethod(String robustnessMethod) {
             common.robustnessMethod(robustnessMethod);
             return this;
         }
 
+        /**
+         * @param scalingMethod the residual scaling method name
+         * @return this builder, for chaining
+         * @see Options.Builder#scalingMethod(String)
+         */
         public Builder scalingMethod(String scalingMethod) {
             common.scalingMethod(scalingMethod);
             return this;
         }
 
+        /**
+         * @param boundaryPolicy the boundary handling policy name
+         * @return this builder, for chaining
+         * @see Options.Builder#boundaryPolicy(String)
+         */
         public Builder boundaryPolicy(String boundaryPolicy) {
             common.boundaryPolicy(boundaryPolicy);
             return this;
         }
 
+        /**
+         * @param zeroWeightFallback the zero-weight handling strategy name
+         * @return this builder, for chaining
+         * @see Options.Builder#zeroWeightFallback(String)
+         */
         public Builder zeroWeightFallback(String zeroWeightFallback) {
             common.zeroWeightFallback(zeroWeightFallback);
             return this;
         }
 
+        /**
+         * @param autoConverge the auto-convergence tolerance
+         * @return this builder, for chaining
+         * @see Options.Builder#autoConverge(double)
+         */
         public Builder autoConverge(double autoConverge) {
             common.autoConverge(autoConverge);
             return this;
         }
 
+        /**
+         * @param confidenceIntervals the confidence level
+         * @return this builder, for chaining
+         * @see Options.Builder#confidenceIntervals(double)
+         */
         public Builder confidenceIntervals(double confidenceIntervals) {
             common.confidenceIntervals(confidenceIntervals);
             return this;
         }
 
+        /**
+         * @param predictionIntervals the prediction level
+         * @return this builder, for chaining
+         * @see Options.Builder#predictionIntervals(double)
+         */
         public Builder predictionIntervals(double predictionIntervals) {
             common.predictionIntervals(predictionIntervals);
             return this;
         }
 
+        /**
+         * @param returnDiagnostics whether to compute diagnostics
+         * @return this builder, for chaining
[…]
