### 2026-09-02 fb10aa3
ci: suppress pip version-check notice in release-pypi.yml
