+++ .github/workflows/release-pypi.yml
@@ -12,6 +12,7 @@ env:
   # Build wheels for Python 3.8 through 3.14
   PYTHON_VERSIONS: "cp38 cp39 cp310 cp311 cp312 cp313 cp314"
   FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true
+  PIP_DISABLE_PIP_VERSION_CHECK: "1"
 
 jobs:
   # ============================================================================
