### 2026-09-02 4afd4d9
ci: suppress pip version-check notice in release-gpu.yml
