### 2026-09-02 7c2a903
ci: fix spack-release job's detached-HEAD git push failure
