### 2026-09-02 9e5b139
ci: add optional commit input to release workflows' workflow_dispatch trigger
