+++ .github/workflows/release-pypi.yml
@@ -3,7 +3,12 @@ name: Release PyPI
 on:
   release:
     types: [published]
-  workflow_dispatch: # Allow manual trigger for testing
+  workflow_dispatch:
+    inputs:
+      commit:
+        description: "Commit SHA to build from (defaults to the workflow's own ref)"
+        required: false
+        type: string
 
 permissions:
   contents: read
@@ -26,6 +31,8 @@ jobs:
         python: ["3.8"]
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Build wheels
         uses: PyO3/maturin-action@v1
@@ -52,6 +59,8 @@ jobs:
         python: ["3.8"]
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Build wheels
         uses: PyO3/maturin-action@v1
@@ -78,6 +87,8 @@ jobs:
         python: ["3.8"]
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Set up Python
         uses: actions/setup-python@v7
@@ -111,6 +122,8 @@ jobs:
         python: ["3.8"]
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Set up Python
         uses: actions/setup-python@v7
@@ -144,6 +157,8 @@ jobs:
         python: ["3.8"]
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Set up Python
         uses: actions/setup-python@v7
@@ -174,6 +189,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Build sdist
         uses: PyO3/maturin-action@v1
@@ -211,6 +228,8 @@ jobs:
     steps:
       - name: Checkout Repository
         uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Download all artifacts
         uses: actions/download-artifact@v8
