+++ .github/workflows/release-gpu.yml
@@ -8,6 +8,10 @@ on:
       tag:
         description: "Release tag to upload assets to (e.g. v3.0.0)"
         required: true
+      commit:
+        description: "Commit SHA to build from (defaults to the workflow's own ref)"
+        required: false
+        type: string
 
 permissions:
   contents: read
@@ -25,6 +29,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Build GPU wheels
         uses: PyO3/maturin-action@v1
@@ -48,6 +54,8 @@ jobs:
     runs-on: macos-15-intel
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Set up Python
         uses: actions/setup-python@v7
@@ -78,6 +86,8 @@ jobs:
     runs-on: macos-14
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Set up Python
         uses: actions/setup-python@v7
@@ -108,6 +118,8 @@ jobs:
     runs-on: windows-latest
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Set up Python
         uses: actions/setup-python@v7
@@ -138,6 +150,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Build GPU library
         run: cargo build --release --features gpu --manifest-path bindings/julia/Cargo.toml
@@ -162,6 +176,8 @@ jobs:
     runs-on: macos-15-intel
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Build GPU library
         run: cargo build --release --features gpu --manifest-path bindings/julia/Cargo.toml
@@ -186,6 +202,8 @@ jobs:
     runs-on: macos-14
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Build GPU library
         run: cargo build --release --features gpu --manifest-path bindings/julia/Cargo.toml
@@ -210,6 +228,8 @@ jobs:
     runs-on: windows-latest
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Build GPU library
         run: cargo build --release --features gpu --manifest-path bindings/julia/Cargo.toml
@@ -235,6 +255,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Build GPU library
         run: cargo build --release --features gpu --manifest-path bindings/cpp/Cargo.toml
@@ -259,6 +281,8 @@ jobs:
     runs-on: macos-15-intel
     steps:
       - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
 
       - name: Build GPU library
         run: cargo build --release --features gpu --manifest-path bindings/cpp/Cargo.toml
@@ -283,6 +307,8 @@ jobs:
     runs-on: macos-14
     steps:
       - uses: actions/checkout@v7
+        with:
[…]
