### 2026-09-02 cc95b0e
fix: add missing @astrojs/markdown-remark devDependency for astro build
