+++ .github/workflows/release-node.yml
@@ -38,6 +38,14 @@ jobs:
             docker: ghcr.io/napi-rs/napi-rs/nodejs-rust:lts-debian-aarch64
             build_command: npm run build -- --target aarch64-unknown-linux-gnu
             pkg_name: linux-arm64-gnu
+          - host: ubuntu-latest
+            target: aarch64-unknown-linux-musl
+            build_command: npm run build -- --target aarch64-unknown-linux-musl --cross-compile
+            pkg_name: linux-arm64-musl
+          - host: ubuntu-latest
+            target: armv7-unknown-linux-gnueabihf
+            build_command: npm run build -- --target armv7-unknown-linux-gnueabihf
+            pkg_name: linux-arm-gnueabihf
           - host: windows-latest
             target: x86_64-pc-windows-msvc
             build_command: npm run build -- --target x86_64-pc-windows-msvc
@@ -74,9 +82,16 @@ jobs:
           if [[ "${{ matrix.settings.target }}" == "aarch64-unknown-linux-gnu" ]]; then
             sudo apt-get install -y gcc-aarch64-linux-gnu g++-aarch64-linux-gnu
             echo "CARGO_TARGET_AARCH64_UNKNOWN_LINUX_GNU_LINKER=aarch64-linux-gnu-gcc" >> $GITHUB_ENV
+          elif [[ "${{ matrix.settings.target }}" == "armv7-unknown-linux-gnueabihf" ]]; then
+            sudo apt-get install -y gcc-arm-linux-gnueabihf g++-arm-linux-gnueabihf
+            echo "CARGO_TARGET_ARMV7_UNKNOWN_LINUX_GNUEABIHF_LINKER=arm-linux-gnueabihf-gcc" >> $GITHUB_ENV
           elif [[ "${{ matrix.settings.target }}" == "x86_64-unknown-linux-musl" ]]; then
             sudo apt-get install -y musl-tools
           fi
+
+      - name: Install Zig (for musl cross-compile)
+        if: matrix.settings.target == 'aarch64-unknown-linux-musl'
+        uses: mlugg/setup-zig@v2
 
       - name: Install Rust Toolchain
         uses: dtolnay/rust-toolchain@stable
