### 2026-09-02 a2283f4
fix: update release-conda.yml for feedstock's meta.yaml -> recipe.yaml migration
