+++ .github/workflows/release-conda.yml
@@ -97,53 +97,6 @@ jobs:
         env:
           VERSION: ${{ steps.version.outputs.version }}
           SHA256: ${{ steps.sha256.outputs.sha256 }}
-          # Use a multi-line string for the build script
-          BUILD_R_SCRIPT: |
-            #!/bin/bash
-            set -ex
-
-            echo "=== System Info ==="
-            uname -a
-            python3 --version || python --version || true
-
-            # Determine R executable
-            R_EXE=${R:-R}
-            $R_EXE --version
-
-            # Use python provided by conda build
-            PYTHON_EXE=$(which python3 || which python || echo "python")
-            echo "Using Python: $PYTHON_EXE"
-
-            # 1. Extract existing vendor dependencies if present
-            # This contains external crates (extendr-api, etc.)
-            if [ -f bindings/r/src/vendor.tar.xz ]; then
-              echo "Extracting vendor.tar.xz..."
-              (cd bindings/r/src && tar -xf vendor.tar.xz)
-            fi
-
-            # 2. Update local crates in vendor directory
-            # This ensures we build with the latest local changes
-            mkdir -p bindings/r/src/vendor
-            cp -rvL crates/fastLowess bindings/r/src/vendor/
-            cp -rvL crates/lowess bindings/r/src/vendor/
-
-            # Clean up target/locks from vendored sources
-            rm -rf bindings/r/src/vendor/*/target
-            rm -f bindings/r/src/vendor/*/Cargo.lock
-
-            # 3. Patch and Isolate (remove workspace inheritance)
-            echo "Running preparation scripts..."
-            $PYTHON_EXE dev/prepare_cargo.py clean bindings/r/src/Cargo.toml
-            $PYTHON_EXE dev/patch_vendor_crates.py Cargo.toml bindings/r/src/vendor/
-            $PYTHON_EXE dev/prepare_cargo.py isolate bindings/r/src/Cargo.toml
-
-            # Remove main lock file to prevent path conflicts
-            rm -f bindings/r/src/Cargo.lock
-
-            # 4. Final R build
-            echo "Starting R package installation..."
-            cd bindings/r
-            $R_EXE CMD INSTALL --build .
         run: |
           set -x
           cd fastlowess-feedstock
@@ -152,27 +105,16 @@ jobs:
           BRANCH="update-to-v$VERSION"
           git checkout -b "$BRANCH"
 
-          # Update version and SHA256 in meta.yaml
-          sed -i "s/{% set version = \".*\" %}/{% set version = \"$VERSION\" %}/" recipe/meta.yaml
-          sed -i "s/sha256: .*/sha256: $SHA256/" recipe/meta.yaml
-          sed -i "s/number: .*/number: 0/" recipe/meta.yaml
-
-          # Fix R package name in tests
-          sed -i "s/library('fastlowess')/library('rfastlowess')/g" recipe/meta.yaml
-
-          # Add python dependencies to build requirements for R package
-          # Use a single, reliable one-liner for substitution
-          sed -i '/name: r-fastlowess/,/test:/ s/- make  # \[unix\]/- make  # [unix]\n        - python\n        - tomli\n        - tomli-w/' recipe/meta.yaml
-
-          # Save the build script from the env var
-          echo "$BUILD_R_SCRIPT" > recipe/build_r.sh
-          chmod +x recipe/build_r.sh
+          # Update version and SHA256 in recipe.yaml (rattler-build format: the
+          # version lives in a `context.version` block, not the old meta.yaml's
+          # `{% set version = "..." %}` Jinja2 directive).
[…]
