### 2026-09-02 #769 @thisisamirv
Hi @etiennebacher, thank you for taking the time to go through this so carefully — this is very helpful. A new release, `3.2.0`, is now out, and it addresses essentially everything raised below.

## Docs

### lowess.readthedocs.io

The <https://lowess.readthedocs.io/> site is no longer the project's homepage. Every binding's README now points to <https://thisisamirv.github.io/lowess-project/> (the repository root, built from the top-level `README.md`) as the canonical entry point, with each binding's own docs site linked from there.

### pkgdown homepage

The R README (which is the pkgdown homepage) was consolidated along with every other binding's README:

- the "API Reference" section (which duplicated the "Reference" tab) was removed
- the `[!TIP]`/`[!NOTE]` GitHub-only alert syntax was replaced with plain blockquotes, which pkgdown renders correctly
- the GPU section was shortened to a couple of lines with a link to the full [GPU Backend guide](https://thisisamirv.github.io/lowess-project/r/articles/gpu-backend.html), instead of duplicating that content on the homepage
- Installation and Documentation were merged into a single section near the top of the page, right after the intro

### Outdated articles

Also fixed:

- **"Parameters Reference"** — removed. Its content (parameter ranges, defaults, fraction-choice guidance) was merged into the `@param`/`@details` roxygen docs of `Lowess()`, `StreamingLowess()`, and `OnlineLowess()`, so it's now part of the Reference pages instead of a separate article.
- **`batch.Rmd`/`streaming.Rmd`/`online.Rmd`** — same treatment: their unique content (When to Use guidance, merge-strategy comparison) was merged into the same functions' `@description`/`@details`, and the now-redundant vignettes (plus their orphaned `gap_handling.svg`/`online_comparison.svg` diagrams) were removed. Pages like `articles/online.html` no longer exist.

### R-specific changelog

`bindings/r/NEWS.md` is now generated (filtered from the root `CHANGELOG.md`) and lives at the package root, so pkgdown's "Changelog" nav tab now shows only R-relevant entries instead of redirecting to the monorepo-wide changelog.

### Benchmarks

Added a `large` benchmark category (n = 50,000) to `benchmarks/rfastlowess.R`, covering four scenarios (`large_delta_0`, `large_delta_0.1`, `large_high_iter`, `large_high_fraction`) that stress `delta`, iteration count, and fraction. Previously, every category ran in well under 100ms; this one exercises a much heavier workload, similar in spirit to the existing GPU benchmarks. Results (mean of 3 runs):

| Scenario | n | R's `stats::lowess` | rfastlowess (serial) | rfastlowess (parallel) |
| --- | --- | --- | --- | --- |
| `large_delta_0` (exact, `delta = 0`) | 50,000 | 5,805.9 ms | 3,133.1 ms (1.9x) | 1,062.8 ms (5.5x) |
| `large_delta_0.1` (`delta = 0.1`) | 50,000 | 14.5 ms | 8.6 ms (1.7x) | 6.5 ms (2.2x) |
| `large_high_iter` (10 robustness iterations) | 50,000 | 31,694.3 ms | 9,158.9 ms (3.5x) | 3,282.3 ms (9.7x) |
| `large_high_fraction` (`fraction = 0.9`) | 20,000 | 12,627.1 ms | 3,527.8 ms (3.6x) | 724.8 ms (17.4x) |

(speedup vs. R's `stats::lowess` in parentheses)

## Reference pages

- **"Main R6 classes" wording** — corrected to "Main S3 classes for LOWESS smoothing."
- **Not assigning the last result in Examples** — many snippets were previously silent because the result was only assigned (`result <- fit(model, x, y)`) and never printed. Explicit `print()`/output statements were added to the snippets that were quiet before, so their output now actually shows up on the rendered page.
- **`Nullable()` export** — it was an internal helper that had no business being exported/documented, and has been unexported along with removing the `rfastlowess-package` pkgdown topic that duplicated the adapter class list.

## Setup

Glad to hear `make r` works now!

## Other R changes in 3.2.0 (for visibility)

A few more R-focused fixes from this release, not raised above but worth flagging:

- Fixed `OnlineLowess()`'s `min_points` (was `3L`) and `update_mode` (was `"full"`) defaults diverging from the Rust core — they should be `2L`/`"incremental"`; updated the roxygen docs and `man/OnlineLowess.Rd` to match.
- Fixed several roxygen/vignette examples that printed nothing useful, `NULL`, or excessively long output (`OnlineLowess()`, `add_point()`, `robustness.Rmd`, `merge.Rmd`, and the genomics/real-time use-case vignettes).
- Fixed some minor inconsistencies and issues across docs.

## CRAN

Agreed — I'll hold off on submitting to CRAN and wait until the rOpenSci review is complete.

PS: Unrelated to R binding, but in order to complete coverage of all major programming languages, I also added bindings for Go and Java. As a result, you will find the logo of the repository changed slightly to include those two as well.
