### 2026-09-02 10093b3
fix: add missing license files and nested module tag for pkg.go.dev
