### 2026-09-02 bb8c88d
feat: add bump_version, update katex to 0.18.5
