+++ dev/bump_version.py
@@ -1,0 +1,301 @@
+#!/usr/bin/env python3
+"""Bump the project version across every crate, binding, and manifest in the repo.
+
+Covers: Cargo.toml package versions (Rust crates + all bindings), the internal
+fastLowess/lowess path-dependency version requirements (major.minor), each
+binding's own version file (package.json, pyproject-adjacent __version__.py,
+pom.xml, DESCRIPTION, Project.toml, version.go, CMakeLists.txt, FastLowess.java),
+CITATION.cff, and the Spack recipe's example `url`.
+
+Does NOT touch: CHANGELOG.md (write that by hand), generated NEWS.md/docs-site
+content (regenerated via `make <lang>-dev` / dev/update_changelogs.py), or the
+Spack recipe's `version()`/`sha256` block and the conda-forge feedstock -- those
+require a published release tarball to hash, so release-cpp.yml/release-conda.yml
+update them after the fact, not before.
+
+Usage:
+    python dev/bump_version.py --version 3.3.0
+    python dev/bump_version.py --version 3.3.0 --dry-run
+"""
+
+from __future__ import annotations
+
+import argparse
+import re
+import sys
+from pathlib import Path
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+
+VERSION_RE = re.compile(r"^\d+\.\d+\.\d+$")
+
+# Cargo.toml files whose bare `[package]` `version = "X.Y.Z"` line should track
+# the release version 1:1.
+CARGO_PACKAGE_FILES = [
+    "crates/lowess/Cargo.toml",
+    "crates/fastLowess/Cargo.toml",
+    "bindings/cpp/Cargo.toml",
+    "bindings/go/Cargo.toml",
+    "bindings/java/Cargo.toml",
+    "bindings/julia/Cargo.toml",
+    "bindings/nodejs/Cargo.toml",
+    "bindings/python/Cargo.toml",
+    "bindings/r/src/Cargo.toml",
+    "bindings/wasm/Cargo.toml",
+]
+
+# Cargo.toml files with a `fastLowess = { path = "../../crates/fastLowess",
+# version = "X.Y", ... }` internal path-dependency requirement to keep in sync
+# (major.minor only -- these are SemVer caret requirements, not exact pins).
+FASTLOWESS_DEP_FILES = [
+    "bindings/cpp/Cargo.toml",
+    "bindings/go/Cargo.toml",
+    "bindings/java/Cargo.toml",
+    "bindings/julia/Cargo.toml",
+    "bindings/nodejs/Cargo.toml",
+    "bindings/python/Cargo.toml",
+    "bindings/wasm/Cargo.toml",
+]
+
+# Node.js binding platform subpackages (each just has its own bare version).
+NODEJS_NPM_PACKAGES = [
+    "darwin-arm64",
+    "darwin-x64",
+    "linux-arm-gnueabihf",
+    "linux-arm64-gnu",
+    "linux-arm64-musl",
+    "linux-x64-gnu",
+    "linux-x64-musl",
+    "win32-arm64-msvc",
+    "win32-x64-msvc",
+]
+
+
+def _replace(
+    path: Path,
+    pattern: re.Pattern[str],
+    replacement: str,
+    dry_run: bool,
+    count: int = 1,
+) -> bool:
+    """Apply one regex substitution to `path`, reporting what happened."""
+    rel = path.relative_to(REPO_ROOT)
+    if not path.exists():
+        print(f"  SKIP (missing): {rel}")
+        return False
+    text = path.read_text(encoding="utf-8")
+    new_text, n = pattern.subn(replacement, text, count=count)
+    if n == 0:
+        print(f"  WARNING: no match in {rel}")
[…]
