+++ .github/workflows/ci-cpp.yml
@@ -23,22 +23,29 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        os:
+          [
+            ubuntu-latest,
+            ubuntu-24.04-arm,
+            macos-latest,
+            windows-latest,
+            windows-11-arm,
+          ]
     steps:
       - uses: actions/checkout@v7
       - uses: dtolnay/rust-toolchain@stable
         with:
           components: rustfmt, clippy
       - name: Install Rust GNU (Windows)
-        if: matrix.os == 'windows-latest'
+        if: runner.os == 'Windows'
         run: rustup toolchain install stable-x86_64-pc-windows-gnu
 
       - name: Install GNU Make (Windows)
-        if: matrix.os == 'windows-latest'
+        if: runner.os == 'Windows'
         run: choco install make -y
 
       - name: Install cppcheck (Windows)
-        if: matrix.os == 'windows-latest'
+        if: runner.os == 'Windows'
         run: |
           winget install --id Cppcheck.Cppcheck --exact --accept-source-agreements --accept-package-agreements
           $cppcheckExe = Get-ChildItem -Path "C:\Program Files\Cppcheck","$env:LOCALAPPDATA\Programs\Cppcheck" -Filter cppcheck.exe -Recurse -ErrorAction SilentlyContinue | Select-Object -First 1
@@ -46,11 +53,11 @@ jobs:
           $cppcheckExe.DirectoryName | Out-File -FilePath $env:GITHUB_PATH -Append -Encoding utf8
 
       - name: Install cppcheck (Linux)
-        if: matrix.os == 'ubuntu-latest'
+        if: runner.os == 'Linux'
         run: sudo apt-get install -y cppcheck
 
       - name: Install clang-tidy and cppcheck (macOS)
-        if: matrix.os == 'macos-latest'
+        if: runner.os == 'macOS'
         run: |
           brew untap aws/tap || true
           brew install llvm cppcheck
