### 2026-09-03 f6d6aef
fix: julia package version
