+++ bindings/julia/julia/Project.toml
@@ -6,7 +6,7 @@ documentation = "https://thisisamirv.github.io/lowess-project/julia/"
 readme = "README.md"
 repository = "https://github.com/thisisamirv/lowess-project"
 name = "FastLOWESS"
-version = "3.1.0"
+version = "3.2.0"
 description = "High-performance LOWESS (Locally Weighted Scatterplot Smoothing)"
 uuid = "5d760c97-4851-4f57-8eaf-1b3e829bbf83"
 categories = ["science", "algorithms"]
