### 2026-09-03 50b5209
fix: remove non-ASCII characters
