+++ crates/lowess/src/algorithms/interpolation.rs
@@ -6,7 +6,7 @@
 //! often unnecessary.
 // ## srrstats Compliance
 //
-// @srrstats {G1.6} Delta-based interpolation reduces O(n²) to O(n×k) for dense data.
+// @srrstats {G1.6} Delta-based interpolation reduces O(n^2) to O(n×k) for dense data.
 // @srrstats {RE2.2} Linear interpolation between anchor points for non-fitted values.
 
 // External dependencies
