### 2026-09-03 454f384
build: update nodejs/wasm dependencies
