### 2026-09-03 78f9016
feat: add version tracking for MathJax
