### 2026-09-03 2f3acd1
ci: check-versions.yml opens issue on outdated
