### 2026-09-03 cc8adef
chore: bump MathJax to 4.1.3
