### 2026-09-04 342bb97
fix: java defaults and binding support exporting defaults
