+++ CHANGELOG.md
@@ -68,6 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 **Go:**
 
 - Fixed pkg.go.dev showing "License: None detected" and refusing to render documentation, and the "Tagged version"/"Stable version" checks failing: the Go module lives at `bindings/go/fastlowess` but `LICENSE-MIT`/`LICENSE-APACHE` only existed one directory up at `bindings/go`, outside where pkg.go.dev scans for a nested module's license. Copied both license files into `bindings/go/fastlowess`. `release-go.yml` now also pushes a `bindings/go/fastlowess/vX.Y.Z` tag (in addition to the repo-wide `vX.Y.Z` release tag), which Go's nested-module versioning requires to recognize a release as tagged/stable.
+- Fixed `OnlineOptions`'s `MinPoints`/`UpdateMode` defaults (`3`/`"full"`) diverging from every other binding (already fixed in C++/Julia/Python/R) and the Rust core; now `2`/`"incremental"`, matching `docs/api/api-online.md`'s Default column.
 
 **Julia:**
 
@@ -78,6 +79,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed `actions/setup-java@v6`'s deprecation warning for `server-username`/`server-password`/`gpg-passphrase` in `release-java.yml`; migrated to the `-env-var` variants.
 - Fixed Surefire warning that `fastlowess.native.dir` was "configured twice"; moved it from `<argLine>` into `<systemPropertyVariables>`.
 - Fixed ~100 `mvn javadoc:jar` warnings (missing `@param`/`@return` tags, undocumented methods) across the public API. `maven-javadoc-plugin` now sets `failOnWarning`, and `make java-dev` runs it as a dedicated check step.
+- Fixed `OnlineOptions.Builder()` never calling `parallel(false)`, silently defaulting online mode's `parallel` to `true` (inherited from `Options.Builder`) unlike every other binding; fixed the constructor and a matching stale doc sentence in `api-online.adoc`.
 
 **Python:**
 
