+++ .github/workflows/release-pypi.yml
@@ -182,6 +182,45 @@ jobs:
           path: bindings/python/dist
 
   # ============================================================================
+  # BUILD WHEELS - Windows ARM64
+  # ============================================================================
+  windows-arm64:
+    name: Build Windows ARM64 (py${{ matrix.python }})
+    runs-on: windows-11-arm
+    strategy:
+      matrix:
+        # python.org's official Windows installers only ship win-arm64 builds
+        # from 3.11 onward, unlike the other jobs' 3.8 baseline; the resulting
+        # wheel is still abi3-py38-compatible (see Cargo.toml's `abi3-py38`
+        # PyO3 feature), so this doesn't change the wheel's declared support.
+        python: ["3.11"]
+    steps:
+      - uses: actions/checkout@v7
+        with:
+          ref: ${{ github.event.inputs.commit || github.sha }}
+
+      - name: Set up Python
+        uses: actions/setup-python@v7
+        with:
+          python-version: ${{ matrix.python }}
+
+      - name: Upgrade pip
+        run: python -m pip install --upgrade pip
+
+      - name: Build wheels
+        uses: PyO3/maturin-action@v1
+        with:
+          working-directory: bindings/python
+          target: aarch64
+          args: --release --out dist
+
+      - name: Upload wheels
+        uses: actions/upload-artifact@v7
+        with:
+          name: wheels-windows-arm64-py${{ matrix.python }}
+          path: bindings/python/dist
+
+  # ============================================================================
   # BUILD SOURCE DISTRIBUTION
   # ============================================================================
   sdist:
@@ -216,6 +255,7 @@ jobs:
       - macos-x86_64
       - macos-aarch64
       - windows-x86_64
+      - windows-arm64
       - sdist
     runs-on: ubuntu-latest
     # Only publish on release events or manual workflow_dispatch
