+++ go/en.search.min.8db3e3371a2ca493d65a6831ac9d1e6a2c99bb5f17869f99852ce6edc7f9c4d6.js
@@ -1,0 +1,1 @@
+(function(){const o="/lowess-project/go/en.search-data.min.6e89449e76b3d3828aaf3499358355f6f6bc6d2783280754004d685677f7aafc.json",i="/lowess-project/go/fuse.esm.js",a=Object.assign({cache:!0},{includeScore:!0,useExtendedSearch:!0,fieldNormWeight:1.5,threshold:.3,ignoreLocation:!0,keys:[{name:"title",weight:.7},{name:"content",weight:.3}]}),e=document.querySelector("#book-search-input"),t=document.querySelector("#book-search-results");if(!e)return;let n;e.addEventListener("focus",c,{once:!0}),e.addEventListener("input",function(){clearTimeout(n),n=setTimeout(s,250)}),document.addEventListener("keydown",r);function r(t){if(t.key==="k"&&(t.metaKey||t.ctrlKey)){t.preventDefault(),e.focus();return}if(e===document.activeElement)return;if(t.target.value!==0[0])return;t.key==="/"&&(t.preventDefault(),e.focus())}function c(){e.required=!0,Promise.all([import(i),fetch(o).then(e=>e.json())]).then(([{default:e},t])=>{window.bookSearchIndex=new e(t,a)}).then(()=>e.required=!1).then(s)}function s(){for(;t.firstChild;)t.removeChild(t.firstChild);if(!e.value)return;const n=window.bookSearchIndex.search(e.value,{limit:10});n.forEach(function(e){const n=l("<li><a href></a><small></small></li>"),s=n.querySelector("a"),o=n.querySelector("small");s.href=e.item.href,s.textContent=e.item.title,o.textContent=e.item.section,t.appendChild(n)})}function l(e){const t=document.createElement("div");return t.innerHTML=e,t.firstChild}})()
