### 2026-09-04 9a7de7c
fix: bump_version now bumps fastlowess_jll as well
