+++ dev/bump_version.py
@@ -4,8 +4,9 @@
 Covers: Cargo.toml package versions (Rust crates + all bindings), the internal
 fastLowess/lowess path-dependency version requirements (major.minor), each
 binding's own version file (package.json, pyproject-adjacent __version__.py,
-pom.xml, DESCRIPTION, Project.toml, version.go, CMakeLists.txt, FastLowess.java),
-CITATION.cff, and the Spack recipe's example `url`.
+pom.xml, DESCRIPTION, Project.toml (incl. the fastlowess_jll compat floor),
+version.go, CMakeLists.txt, FastLowess.java), CITATION.cff, and the Spack
+recipe's example `url`.
 
 Does NOT touch: CHANGELOG.md (write that by hand), generated NEWS.md/docs-site
 content (regenerated via `make <lang>-dev` / dev/update_changelogs.py), or the
@@ -198,6 +199,17 @@ def build_targets(
             "bindings/julia/julia/Project.toml",
             re.compile(r'^version = "\d+\.\d+\.\d+"$', re.MULTILINE),
             f'version = "{new_version}"',
+            1,
+        )
+    )
+    # `make julia-dev`/CI relax this compat floor to an OR-list of the actual
+    # latest-registered JLL version at test-time (see bindings/julia/Makefile),
+    # so it's safe to bump this to the not-yet-registered version here.
+    targets.append(
+        (
+            "bindings/julia/julia/Project.toml",
+            re.compile(r'^fastlowess_jll = "\d+\.\d+\.\d+"$', re.MULTILINE),
+            f'fastlowess_jll = "{new_version}"',
             1,
         )
     )
