### 2026-09-04 3f92f81
fix: Windows arm64 detection in bindings/go/Makefile under x64 emulation
