+++ .github/workflows/ci-cpp.yml
@@ -37,7 +37,7 @@ jobs:
         with:
           components: rustfmt, clippy
       - name: Install Rust GNU (Windows)
-        if: runner.os == 'Windows'
+        if: runner.os == 'Windows' && runner.arch != 'ARM64'
         run: rustup toolchain install stable-x86_64-pc-windows-gnu
 
       - name: Install GNU Make (Windows)
