### 2026-09-04 eae80c1
fix: pass explicit GO_CARGO_TARGET override for windows-11-arm in ci-go.yml instead of relying on Makefile detection
