### 2026-09-04 e42cb8b
fix: replace winget with choco for cppcheck install in ci-cpp.yml (winget unavailable on windows-11-arm)
