### 2026-09-04 ce9e1ed
fix: fix Windows arm64 target detection in cpp Makefile; override in ci-cpp.yml
