### 2026-09-04 b986ff3
fix: set CC/CXX to aarch64-w64-mingw32-clang for Go cgo build in ci-go.yml
