### 2026-09-04 bc121a6
ci: add PATH fallback after choco install cppcheck in ci-cpp.yml
