### 2026-09-04 fa2cf18
ci: install objdump via choco mingw for windows-11-arm symbol-export check in ci-cpp.yml
