### 2026-09-04 b9345c7
ci: add diagnostics to go ci
