### 2026-09-04 801b4e9
fix: trying more changes for windows arm issues with go and cpp
