+++ bindings/go/fastlowess/ffi.go
@@ -5,7 +5,11 @@ package fastlowess
 #cgo linux LDFLAGS: -L${SRCDIR}/../../../target/release-c -lfastlowess_go -lm -ldl -lpthread
 #cgo darwin LDFLAGS: -L${SRCDIR}/../../../target/release-c -lfastlowess_go
 #cgo windows,amd64 LDFLAGS: -L${SRCDIR}/../../../target/x86_64-pc-windows-gnu/release-c -lfastlowess_go -lws2_32 -luserenv -lbcrypt -lntdll -lpthread
-#cgo windows,arm64 LDFLAGS: -L${SRCDIR}/../../../target/aarch64-pc-windows-gnullvm/release-c -lfastlowess_go -lws2_32 -luserenv -lbcrypt -lntdll -lpthread
+// -static avoids depending on llvm-mingw's runtime DLLs (libunwind/libc++/
+// libwinpthread) being discoverable on PATH at execution time, since they
+// live in a target-specific sysroot subdirectory rather than next to the
+// cross-compiler driver itself.
+#cgo windows,arm64 LDFLAGS: -static -L${SRCDIR}/../../../target/aarch64-pc-windows-gnullvm/release-c -lfastlowess_go -lws2_32 -luserenv -lbcrypt -lntdll -lpthread
 #include <stdlib.h>
 #include "fastlowess_go.h"
 */
