### 2026-09-04 319c9bc
fix: lint cpp, bump to v3.2.1
