+++ bindings/go/src/lib.rs
@@ -127,7 +127,7 @@ pub struct GoOnlineOutput {
 /// All arrays are allocated by Rust and must be freed by Rust.
 #[repr(C)]
 pub struct GoLowessResult {
-    /// Sorted x values (length = n)
+    /// x values, in the same order as the input (length = n)
     pub x: *mut c_double,
     /// Smoothed y values (length = n)
     pub y: *mut c_double,
