+++ dev/runners/rust.py
@@ -2,10 +2,14 @@
 
 from __future__ import annotations
 
+import concurrent.futures
+import json
 import re
+import shutil
 import subprocess
 import textwrap
 import time
+from collections import defaultdict
 
 from .base import REPO_ROOT, RunResult, Snippet, _find_exe
 
@@ -37,88 +41,175 @@ def skip_reason(snippet: Snippet) -> str | None:
 
 
 _RUST_SNIPPET_DIR = REPO_ROOT / "target" / "doc-snippet-runner"
+_RUST_BIN_PREFIX = "snippet_"
 
 
 def _ensure_rust_snippet_project() -> None:
-    """Create the persistent temp Cargo project for Rust snippet execution."""
+    """(Re)write the persistent Cargo project used for batched Rust snippet execution.
+
+    Each snippet becomes its own `src/bin/snippet_NNNN.rs` binary target so that a
+    single `cargo build` compiles (and thus caches shared dependencies for) every
+    snippet at once, instead of paying Cargo's own startup/dependency-resolution
+    overhead once per snippet via repeated `cargo run` invocations.
+    """
     _RUST_SNIPPET_DIR.mkdir(parents=True, exist_ok=True)
-    cargo_toml = _RUST_SNIPPET_DIR / "Cargo.toml"
-    if not cargo_toml.exists():
-        lowess_path = str(REPO_ROOT / "crates" / "lowess").replace("\\", "/")
-        fastlowess_path = str(REPO_ROOT / "crates" / "fastLowess").replace("\\", "/")
-        cargo_toml.write_text(
-            textwrap.dedent(f"""\
-                [workspace]
+    lowess_path = str(REPO_ROOT / "crates" / "lowess").replace("\\", "/")
+    fastlowess_path = str(REPO_ROOT / "crates" / "fastLowess").replace("\\", "/")
+    (_RUST_SNIPPET_DIR / "Cargo.toml").write_text(
+        textwrap.dedent(f"""\
+            [workspace]
 
-                [package]
-                name = "doc-snippet"
-                version = "0.1.0"
-                edition = "2021"
+            [package]
+            name = "doc-snippet"
+            version = "0.1.0"
+            edition = "2021"
 
-                [[bin]]
-                name = "doc-snippet"
-                path = "src/main.rs"
-
-                [dependencies]
-                lowess    = {{ path = "{lowess_path}" }}
-                fastLowess = {{ path = "{fastlowess_path}" }}
-            """),
-            encoding="utf-8",
-        )
-    (_RUST_SNIPPET_DIR / "src").mkdir(exist_ok=True)
+            [dependencies]
+            lowess    = {{ path = "{lowess_path}" }}
+            fastLowess = {{ path = "{fastlowess_path}" }}
+        """),
+        encoding="utf-8",
+    )
+    main_rs = _RUST_SNIPPET_DIR / "src" / "main.rs"
+    if main_rs.exists():
+        main_rs.unlink()
+    bin_dir = _RUST_SNIPPET_DIR / "src" / "bin"
+    if bin_dir.exists():
+        shutil.rmtree(bin_dir)
+    bin_dir.mkdir(parents=True)
 
 
 def run_rust(snippet: Snippet, timeout: int) -> RunResult:
+    return run_rust_batch([snippet], timeout)[0]
+
+
+def run_rust_batch(snippets: list[Snippet], timeout: int) -> list[RunResult]:
+    """Compile every Rust snippet in one `cargo build`, then run each binary.
+
+    This is much faster than one `cargo run` per snippet: shared dependencies and
+    Cargo's own overhead are paid once for the whole batch, `--keep-going` keeps
+    snippet failures independent (a broken snippet doesn't block the others from
[…]
