+++ bindings/r/src/Cargo.toml
@@ -36,11 +36,3 @@ gpu = ["fastLowess/gpu"]
 strip = "symbols"
 
 [workspace]
-
-
-
-
-
-
-
-
