### 2026-09-04 f09225d
cpp: update Spack recipe for v3.2.1
