### 2026-09-04 481b1cc
ci: pin macos-15-intel instead of the deprecated macos-13
