+++ CHANGELOG.md
@@ -6,6 +6,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+**C++:**
+
+- Repinned the macOS x64 job in `release-cpp.yml` from `macos-13` to `macos-15-intel`, GitHub's dedicated Intel-macOS label.
+
+**Go:**
+
+- Repinned the macOS x64 job in `release-go.yml` from `macos-13` to `macos-15-intel`.
+
 ## 3.2.1
 
 ### Added
