+++ .github/workflows/release-go.yml
@@ -32,7 +32,7 @@ jobs:
             asset_name: libfastlowess_go-linux-arm64.a
             cargo_target: ""
             target_dir: target/release
-          - os: macos-13
+          - os: macos-15-intel
             asset_name: libfastlowess_go-macos-x64.a
             cargo_target: ""
             target_dir: target/release
