### 2026-09-04 c0724cd
feat: add return_sorted, improve api docs
