### 2026-09-04 82d1d1d
refactor: drop dead/stale arguments from streaming/online adapters
