+++ bindings/nodejs/index.js
@@ -1,453 +1,707 @@
-/* tslint:disable */
+// prettier-ignore
 /* eslint-disable */
-/* prettier-ignore */
-
+// @ts-nocheck
 /* auto-generated by NAPI-RS */
 
-const { existsSync, readFileSync } = require('fs')
-const { join } = require('path')
-const gpuDownload = require('./gpu-download')
-
-const { version: VERSION } = require('./package.json')
-
-const { platform, arch } = process
-
+const { readFileSync } = require('fs')
 let nativeBinding = null
-let localFileExisted = false
-let loadError = null
+const loadErrors = []
 
-function isMusl() {
-    // For Node 10
-    if (!process.report || typeof process.report.getReport !== 'function') {
-        try {
-            const lddPath = require('child_process').execSync('which ldd').toString().trim()
-            return readFileSync(lddPath, 'utf8').includes('musl')
-        } catch (e) {
-            return true
-        }
-    } else {
-        const { glibcVersionRuntime } = process.report.getReport().header
-        return !glibcVersionRuntime
+const isMusl = () => {
+  let musl = false
+  if (process.platform === 'linux') {
+    musl = isMuslFromFilesystem()
+    if (musl === null) {
+      musl = isMuslFromReport()
     }
+    if (musl === null) {
+      musl = isMuslFromChildProcess()
+    }
+  }
+  return musl
 }
 
-// Check for local build first (prioritize dev/CI environment)
-if (existsSync(join(__dirname, 'fastlowess.node'))) {
-    nativeBinding = require('./fastlowess.node')
-} else {
-    switch (platform) {
-        case 'android':
-            switch (arch) {
-                case 'arm64':
-                    localFileExisted = existsSync(join(__dirname, 'fastlowess.android-arm64.node'))
-                    try {
-                        if (localFileExisted) {
-                            nativeBinding = require('./fastlowess.android-arm64.node')
-                        } else {
-                            nativeBinding = require('fastlowess-android-arm64')
-                        }
-                    } catch (e) {
-                        loadError = e
-                    }
-                    break
-                case 'arm':
-                    localFileExisted = existsSync(join(__dirname, 'fastlowess.android-arm-eabi.node'))
-                    try {
-                        if (localFileExisted) {
-                            nativeBinding = require('./fastlowess.android-arm-eabi.node')
-                        } else {
-                            nativeBinding = require('fastlowess-android-arm-eabi')
-                        }
-                    } catch (e) {
-                        loadError = e
-                    }
-                    break
-                default:
-                    throw new Error(`Unsupported architecture on Android ${arch}`)
-            }
-            break
-        case 'win32':
-            switch (arch) {
-                case 'x64':
-                    localFileExisted = existsSync(join(__dirname, 'fastlowess.win32-x64-msvc.node'))
-                    try {
-                        if (localFileExisted) {
-                            nativeBinding = require('./fastlowess.win32-x64-msvc.node')
-                        } else {
-                            nativeBinding = require('fastlowess-win32-x64-msvc')
-                        }
-                    } catch (e) {
-                        loadError = e
-                    }
-                    break
-                case 'ia32':
-                    localFileExisted = existsSync(join(__dirname, 'fastlowess.win32-ia32-msvc.node'))
-                    try {
-                        if (localFileExisted) {
-                            nativeBinding = require('./fastlowess.win32-ia32-msvc.node')
-                        } else {
-                            nativeBinding = require('fastlowess-win32-ia32-msvc')
-                        }
-                    } catch (e) {
-                        loadError = e
-                    }
-                    break
-                case 'arm64':
[…]
