### 2026-09-05 2d9aece
fix: api inconsistencies
