### 2026-09-05 19bbcc8
docs: api inconsistencies
