### 2026-09-05 3ea57bf
docs: unify docs between all bindings/crates
