+++ bindings/java/docs/modules/ROOT/pages/weighting/robustness.adoc
@@ -213,5 +213,103 @@ public class Example {
 Point 3 is likely an outlier (weight: 0.000)
 ----
 
+'''
+
+== Scale Estimation
+
+Residuals are scaled before computing robustness weights. Two methods:
+
+[cols="1,1,1",options="header"]
+|===
+|Method |Formula |Robustness
+
+|*MAD* |`median(\|r - median(r)\|)` |Very robust (default)
+|*MAR* |`median(\|r\|)` |Robust, uncentered
+|*Mean* |`mean(\|r\|)` |Less robust, fastest
+|===
+
+image::scaling_comparison.svg[Scaling Methods Comparison]
+
+[source,java]
+----
+import fastlowess.Lowess;
+import fastlowess.Options;
+import fastlowess.Result;
+
+public class Example {
+    public static void main(String[] args) {
+        int n = 100;
+        double[] x = new double[n];
+        double[] y = new double[n];
+        for (int i = 0; i < n; i++) {
+            x[i] = i * 2 * Math.PI / (n - 1);
+            y[i] = Math.sin(x[i]) + 0.1;
+        }
+
+        Options options = Options.builder()
+                .iterations(3)
+                .scalingMethod("mad")
+                .build();
+
+        try (Lowess model = new Lowess(options)) {
+            Result result = model.fit(x, y);
+            System.out.println("First smoothed value (mad scaling): " + result.y()[0]);
+        }
+    }
+}
+----
+
+[source,console]
+----
+First smoothed value (mad scaling): 0.38260776436644134
+----
+
+'''
+
+== Auto-Convergence
+
+Stop iterations early when weights stabilize:
+
+[NOTE]
+====
+*Performance:* Auto-convergence can significantly reduce computation when weights stabilize before reaching max iterations.
+====
+
+[source,java]
+----
+import fastlowess.Lowess;
+import fastlowess.Options;
+import fastlowess.Result;
+
+public class Example {
+    public static void main(String[] args) {
+        int n = 100;
+        double[] x = new double[n];
+        double[] y = new double[n];
+        for (int i = 0; i < n; i++) {
+            x[i] = i * 2 * Math.PI / (n - 1);
+            y[i] = Math.sin(x[i]) + 0.1;
+        }
+
+        Options options = Options.builder()
+                .iterations(10)
+                .autoConverge(1e-6)
+                .build();
+
+        try (Lowess model = new Lowess(options)) {
+            Result result = model.fit(x, y);
+            System.out.println("First smoothed value (auto-converge): " + result.y()[0]);
+        }
+    }
+}
+----
+
+[source,console]
+----
+First smoothed value (auto-converge): 0.3826022035717678
+----
+
+'''
+
 All examples above use the same 100-point noisy sine wave (see xref:./kernels.adoc#choosing-a-kernel[Weight Functions] for the generation code), except the outlier-detection example above, which uses a small dataset with a deliberate spike.
 
