+++ bindings/r/vignettes/use-case-genomics.Rmd
@@ -7,34 +7,44 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
+<!-- markdownlint-disable MD024 -->
+
 ```{r setup, include=FALSE}
 knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 ```
 
+LOWESS for methylation profiles, ChIP-seq signals, and other genomic data.
+
 ## Overview
 
-LOWESS is well-suited for genomic data such as DNA methylation profiles,
-ChIP-seq signals, and expression arrays. Sequencing depth variation, PCR
-artifacts, and biological heterogeneity all create noise that LOWESS handles
-gracefully without parametric assumptions.
+Genomic data often contains noise from sequencing depth variation, PCR
+artifacts, or biological heterogeneity. LOWESS smoothing helps reveal
+underlying patterns.
 
 ---
 
 ## Methylation Profile Smoothing
 
-A small `fraction = 0.1` lets LOWESS follow fine-scale spatial structure without
-smearing transitions between methylated and unmethylated regions.
-`confidence_intervals = 0.95` produces uncertainty bands that naturally widen at
-positions with sparser CpG coverage.
+### The Challenge
+
+DNA methylation data (from bisulfite sequencing or arrays) shows
+position-dependent patterns that can be obscured by measurement noise.
+
+### Solution
+
+A small `fraction = 0.1` lets LOWESS follow fine-scale spatial structure
+without smearing the transitions between methylated and unmethylated regions.
+`confidence_intervals = 0.95` produces uncertainty bands that naturally widen
+at positions with sparser CpG coverage, making low-confidence segments
+immediately apparent in the plot.
 
 ```{r use_case_genomics_1}
 library(rfastlowess)
 
-# Deterministic methylation profile along a chromosome
 positions <- seq(0, 99000, by = 1000)
-observed <- (50 + sin(positions / 1000) * 20 + 5) / 100  # Methylation is 0-1
+observed <- 50 + sin(positions / 1000) * 20 + 5
 
-# Smooth
+# positions and observed are your methylation data
 model <- Lowess(
     fraction = 0.1,
     iterations = 3,
@@ -42,7 +52,8 @@ model <- Lowess(
 )
 result <- fit(model, positions, observed)
 
-# Plot
+# Smoothed profile in result$y
+# CI bounds in result$confidence_lower/upper
 plot(positions, observed, pch = ".", col = "gray",
     xlab = "Genomic Position (bp)", ylab = "Methylation Level",
     main = "Methylation Profile Smoothing")
@@ -62,6 +73,8 @@ cat(sprintf(
 
 ## ChIP-seq Signal Smoothing
 
+### Application
+
 ChIP-seq experiments produce sparse, noisy coverage data. LOWESS can help
 identify binding regions.
 
@@ -75,23 +88,17 @@ counts.
 ```{r use_case_genomics_2}
 library(rfastlowess)
 
-# Deterministic ChIP-seq-style coverage profile
 positions <- seq(0, 99000, by = 1000)
 signal <- 50 + sin(positions / 1000) * 20 + 5
 
-# Smooth
 model <- Lowess(
     fraction = 0.05,
     iterations = 5
 )
 result <- fit(model, positions, signal)
-peak_count <- sum(result$y > 65.0)
 
-plot(positions, signal, pch = 16, cex = 0.4, col = "gray",
-    xlab = "Genomic Position", ylab = "Read Count",
-    main = "ChIP-seq Signal Smoothing")
-lines(result$x, result$y, col = "blue", lwd = 2)
[…]

[…]
