### 2026-09-05 7c7479f
ci: add cargo release workflow
