### 2026-09-05 1052cb7
ci: run release-conda on release
