+++ .github/workflows/release-conda.yml
@@ -1,11 +1,13 @@
 name: Release Conda
 
 on:
+  release:
+    types: [published]
   workflow_dispatch:
     inputs:
       version:
         description: "Version to update to (without v prefix, e.g., 0.1.2)"
-        required: true
+        required: false
         type: string
 
 env:
@@ -44,7 +46,12 @@ jobs:
       - name: Determine version
         id: version
         run: |
-          VERSION="${{ github.event.inputs.version }}"
+          if [ -n "${{ github.event.inputs.version }}" ]; then
+            VERSION="${{ github.event.inputs.version }}"
+          else
+            VERSION="${{ github.event.release.tag_name }}"
+            VERSION="${VERSION#v}"
+          fi
           echo "version=$VERSION" >> $GITHUB_OUTPUT
           echo "Updating to version: $VERSION"
 
