### 2026-09-05 c2ee2c0
fix: minor doc inconsistencies, drifted defaults, faster cpp/go snippet check
