+++ dev/runners/go.py
@@ -2,18 +2,23 @@
 
 from __future__ import annotations
 
+import concurrent.futures
 import os
 import re
+import shutil
 import subprocess
-import tempfile
 import time
-from pathlib import Path
+from collections import defaultdict
 
 from .base import REPO_ROOT, RunResult, Snippet, _find_exe
 
 GO_BINDING_DIR = REPO_ROOT / "bindings" / "go"
 GO_MODULE_DIR = GO_BINDING_DIR / "fastlowess"
 MODULE_PATH = "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+
+_GO_SNIPPET_DIR = REPO_ROOT / "target" / "doc-snippet-runner" / "go"
+_GO_BIN_DIR = REPO_ROOT / "target" / "doc-snippet-target" / "go"
+_GO_BIN_PREFIX = "snippet_"
 
 
 def skip_reason(snippet: Snippet) -> str | None:
@@ -24,59 +29,151 @@ def skip_reason(snippet: Snippet) -> str | None:
     return None
 
 
+def _ensure_go_snippet_module(names: list[str], snippets: list[Snippet]) -> None:
+    """(Re)write the persistent module used for batched Go snippet execution.
+
+    Each snippet becomes its own `cmd/snippet_NNNN/main.go` package so a single
+    `go build ./...` compiles the whole batch at once: module resolution and
+    cgo compilation/linking of the shared fastlowess binding are paid once
+    (with the Go toolchain building independent packages concurrently) instead
+    of once per snippet via repeated `go run` invocations, each in its own
+    throwaway module.
+    """
+    cmd_dir = _GO_SNIPPET_DIR / "cmd"
+    if cmd_dir.exists():
+        shutil.rmtree(cmd_dir)
+    cmd_dir.mkdir(parents=True, exist_ok=True)
+    (_GO_SNIPPET_DIR / "go.mod").write_text(
+        "module snippet\n\n"
+        "go 1.23\n\n"
+        f"require {MODULE_PATH} v0.0.0\n\n"
+        f"replace {MODULE_PATH} => {GO_MODULE_DIR.as_posix()}\n",
+        encoding="utf-8",
+    )
+    for name, snippet in zip(names, snippets):
+        pkg_dir = cmd_dir / name
+        pkg_dir.mkdir(parents=True, exist_ok=True)
+        (pkg_dir / "main.go").write_text(snippet.code, encoding="utf-8")
+
+
 def run_go(snippet: Snippet, timeout: int) -> RunResult:
+    return run_go_batch([snippet], timeout)[0]
+
+
+def run_go_batch(snippets: list[Snippet], timeout: int) -> list[RunResult]:
+    """Compile every Go snippet in one `go build ./...`, then run each binary.
+
+    Much faster than one `go run` per snippet: each snippet used to pay its
+    own `go run` module-resolution + cgo compile/link overhead serially in a
+    fresh temp dir; here that cost is paid once for the whole batch, and the
+    already-built binaries are then executed concurrently too.
+    """
+    if not snippets:
+        return []
+
     go_exe = _find_exe("go")
     if go_exe is None:
-        return RunResult(
-            snippet=snippet,
-            runner="go",
-            skipped=True,
-            skip_reason="no 'go' executable found in PATH",
-        )
+        return [
+            RunResult(
+                snippet=s,
+                runner="go",
+                skipped=True,
+                skip_reason="no 'go' executable found in PATH",
+            )
+            for s in snippets
+        ]
 
-    with tempfile.TemporaryDirectory(ignore_cleanup_errors=True) as tmpdir:
-        mod_dir = Path(tmpdir)
-        # A throwaway module that requires the real binding via a local
[…]
