### 2026-09-05 4415711
chore: bump to version 4.0.0
