+++ bindings/java/src/main/java/fastlowess/FastLowess.java
@@ -12,7 +12,7 @@ public final class FastLowess {
      * The released version of this Java binding (bindings/java), tracked
      * independently of the underlying fastLowess Rust core's crate version.
      */
-    public static final String VERSION = "3.2.1";
+    public static final String VERSION = "4.0.0";
 
     /**
      * Returns the version of this Java binding.
