+++ bindings/go/docs/guide/adapter-choice.md
@@ -83,6 +83,11 @@ func main() {
 }
 ```
 
+```output
+95% CI at midpoint: [0.0393, 0.1077]
+R2: 0.9664
+```
+
 ---
 
 ## Streaming Adapter
@@ -156,6 +161,10 @@ func main() {
  }
  fmt.Printf("Smoothed y[0]: %.4f\n", result.Y[0])
 }
+```
+
+```output
+Smoothed y[0]: 0.2578
 ```
 
 ---
@@ -234,6 +243,14 @@ func main() {
   }
  }
 }
+```
+
+```output
+0.3511479871810792
+0.4120334456984871
+0.4716624556603275
+0.5297949120891716
+0.5861967361004687
 ```
 
 ---

[…]
