+++ cpp/search/all_11.js
@@ -6,14 +6,15 @@ var searchData=
   ['residual_3',['residual',['../classfastlowess_1_1OnlineOutput.html#ada2082be0f3d7371969ebca0cfd7b074',1,'fastlowess::OnlineOutput']]],
   ['residual_5fsd_4',['residual_sd',['../classfastlowess_1_1Diagnostics.html#ab0626e0bb903038efeae2785cbcd2f23',1,'fastlowess::Diagnostics']]],
   ['residuals_5',['residuals',['../classfastlowess_1_1LowessResult.html#a2bc1288fa77f608d2018b9a2d8692af5',1,'fastlowess::LowessResult']]],
-  ['return_5fdiagnostics_6',['return_diagnostics',['../structfastlowess_1_1LowessOptions.html#a1afdf08dc8d17bf1820edf63b25fd84e',1,'fastlowess::LowessOptions::return_diagnostics'],['../structfastlowess_1_1OnlineOptions.html#a4f2e9058c0a5e438d95dce1d392f3eba',1,'fastlowess::OnlineOptions::return_diagnostics']]],
-  ['return_5fresiduals_7',['return_residuals',['../structfastlowess_1_1LowessOptions.html#a90450e39a90a6211fdbb7c2812cce131',1,'fastlowess::LowessOptions::return_residuals'],['../structfastlowess_1_1OnlineOptions.html#aa64cc4c93af3542250f730db7e2f5312',1,'fastlowess::OnlineOptions::return_residuals']]],
+  ['return_5fdiagnostics_6',['return_diagnostics',['../structfastlowess_1_1LowessOptions.html#a1afdf08dc8d17bf1820edf63b25fd84e',1,'fastlowess::LowessOptions']]],
+  ['return_5fresiduals_7',['return_residuals',['../structfastlowess_1_1LowessOptions.html#a90450e39a90a6211fdbb7c2812cce131',1,'fastlowess::LowessOptions']]],
   ['return_5frobustness_5fweights_8',['return_robustness_weights',['../structfastlowess_1_1LowessOptions.html#a0e2eb41d4d35c6b037c6a767e0848c2b',1,'fastlowess::LowessOptions::return_robustness_weights'],['../structfastlowess_1_1OnlineOptions.html#a8e6114b58afa71fcbb1228e7a9854aa2',1,'fastlowess::OnlineOptions::return_robustness_weights']]],
   ['return_5fse_9',['return_se',['../structfastlowess_1_1LowessOptions.html#a6b1d86216ff15cd283112a762fc0d302',1,'fastlowess::LowessOptions']]],
-  ['rmse_10',['rmse',['../classfastlowess_1_1Diagnostics.html#a7a7fe5265de253a6d4210ab9635a0bd2',1,'fastlowess::Diagnostics']]],
-  ['robustness_11',['robustness',['../robustness.html',1,'Robustness'],['../weighting.html',1,'Weight &amp; Robustness']]],
-  ['robustness_2emd_12',['robustness.md',['../robustness_8md.html',1,'']]],
-  ['robustness_5fmethod_13',['robustness_method',['../structfastlowess_1_1LowessOptions.html#a503bd268254dd2f6cba9c14d73c3c4a8',1,'fastlowess::LowessOptions::robustness_method'],['../structfastlowess_1_1OnlineOptions.html#a9a7d45c8ecd00f75d3240bb8548be754',1,'fastlowess::OnlineOptions::robustness_method']]],
-  ['robustness_5fweight_14',['robustness_weight',['../classfastlowess_1_1OnlineOutput.html#aa2ed697e5067d27b74def55c301d4443',1,'fastlowess::OnlineOutput']]],
[…]
