+++ cpp/search/variables_2.js
@@ -1,8 +1,8 @@
 var searchData=
 [
   ['chunk_5fsize_0',['chunk_size',['../structfastlowess_1_1StreamingOptions.html#aef557d1b8515c43f5027f0afea85018a',1,'fastlowess::StreamingOptions']]],
-  ['confidence_5fintervals_1',['confidence_intervals',['../structfastlowess_1_1LowessOptions.html#a1d07294a88a01279d0d3668a8a70e23f',1,'fastlowess::LowessOptions::confidence_intervals'],['../structfastlowess_1_1OnlineOptions.html#a43c217d2eb8782d3dd1965c2298f348c',1,'fastlowess::OnlineOptions::confidence_intervals']]],
-  ['custom_5fweights_2',['custom_weights',['../structfastlowess_1_1LowessOptions.html#a1c46edade7f9e883483994d12a49e0da',1,'fastlowess::LowessOptions::custom_weights'],['../structfastlowess_1_1OnlineOptions.html#a4c02d54701abc6d10be96fe54d27f3b4',1,'fastlowess::OnlineOptions::custom_weights']]],
+  ['confidence_5fintervals_1',['confidence_intervals',['../structfastlowess_1_1LowessOptions.html#a1d07294a88a01279d0d3668a8a70e23f',1,'fastlowess::LowessOptions']]],
+  ['custom_5fweights_2',['custom_weights',['../structfastlowess_1_1LowessOptions.html#a1c46edade7f9e883483994d12a49e0da',1,'fastlowess::LowessOptions']]],
   ['cv_5ffractions_3',['cv_fractions',['../structfastlowess_1_1LowessOptions.html#a8df67ab5c18aaea06b523bc3330d3447',1,'fastlowess::LowessOptions']]],
   ['cv_5fk_4',['cv_k',['../structfastlowess_1_1LowessOptions.html#aac93982af4c822840c6264041da36df2',1,'fastlowess::LowessOptions']]],
   ['cv_5fmethod_5',['cv_method',['../structfastlowess_1_1LowessOptions.html#a20d46eb7c38e21cb9c3e7015a1977ab6',1,'fastlowess::LowessOptions']]],
