+++ cpp/search/variables_7.js
@@ -1,5 +1,6 @@
 var searchData=
 [
   ['merge_5fstrategy_0',['merge_strategy',['../structfastlowess_1_1StreamingOptions.html#a5c33b791424c7b5f6aa95e1f8e8a6121',1,'fastlowess::StreamingOptions']]],
-  ['min_5fpoints_1',['min_points',['../structfastlowess_1_1OnlineOptions.html#a749706bb2b58779a0ef79c22d0683d32',1,'fastlowess::OnlineOptions']]]
+  ['min_5fpoints_1',['min_points',['../structfastlowess_1_1OnlineOptions.html#a749706bb2b58779a0ef79c22d0683d32',1,'fastlowess::OnlineOptions']]],
+  ['missing_2',['missing',['../structfastlowess_1_1LowessOptions.html#a3521ff4161c3135d1b6a61ee06bba7cb',1,'fastlowess::LowessOptions::missing'],['../structfastlowess_1_1OnlineOptions.html#aa0d36b5f54febd5315b162d7fabe1b2b',1,'fastlowess::OnlineOptions::missing']]]
 ];

[…]
