### 2026-09-05 4dd1aca
fix(nodejs): add --platform flag to napi build so the platform-suffixed native binding and index.js loader are regenerated correctly
