### 2026-09-05 c65a60e
fix: nodejs gpu installation, removed docs filter
