+++ bindings/nodejs/gpu-installer.js
@@ -1,0 +1,103 @@
+'use strict'
+
+// One-time downloader/installer for the opt-in GPU-enabled fastlowess build.
+//
+// The GPU backend (wgpu) is not included in the native binaries published to
+// npm. This module fetches a prebuilt GPU-enabled `.node` addon from the
+// matching GitHub Release (built by `.github/workflows/release-gpu.yml`) and
+// saves it next to this file, under the platform-specific filename that
+// index.js's loader checks before falling back to the optionalDependency
+// package.
+//
+// napi-rs fully regenerates index.js on every `napi build`, so this logic
+// lives here (and gets re-attached to index.js's exports by
+// scripts/patch-index.js after each build) instead of inside index.js itself.
+
+const path = require('path')
+const { execFileSync } = require('child_process')
+const gpuDownload = require('./gpu-download')
+
+const { version } = require('./package.json')
+
+// Only these 4 platforms are built by release-gpu.yml.
+function currentPlatformSuffix() {
+    if (process.platform === 'win32' && process.arch === 'x64') return 'win32-x64-msvc'
+    if (process.platform === 'darwin' && process.arch === 'x64') return 'darwin-x64'
+    if (process.platform === 'darwin' && process.arch === 'arm64') return 'darwin-arm64'
+    if (process.platform === 'linux' && process.arch === 'x64') return 'linux-x64-gnu'
+    return null
+}
+
+// Runs the availability check in a short-lived child process instead of
+// `require()`-ing index.js in-process: on Windows, a loaded native addon
+// keeps its `.node` file locked for the life of the process, which would
+// make the download below fail with EBUSY when overwriting that same file.
+function gpuAvailable() {
+    try {
+        const out = execFileSync(process.execPath, ['-e', "process.stdout.write(String(require('./index.js').gpu_enabled()))"], {
+            cwd: __dirname,
+            encoding: 'utf8',
+        })
+        return out.trim() === 'true'
+    } catch {
+        return false
+    }
+}
+
+/**
+ * Download and install the GPU-enabled fastlowess native addon for this
+ * platform, then restart Node.js to use it.
+ *
+ * Fetches a prebuilt `.node` addon (built with the `gpu` Cargo feature) from
+ * the matching GitHub Release and saves it as `fastlowess.<platform>.node`
+ * next to this file — the same local-override path index.js's loader
+ * already checks first. A running process cannot swap an already-loaded
+ * native addon, so a restart is required afterwards.
+ *
+ * @param {{ yes?: boolean }} [options] Pass `{ yes: true }` to skip the
+ *   interactive y/N confirmation prompt (required when stdin is not a TTY).
+ */
+async function installGpu(options = {}) {
+    const { yes = false } = options
+    if (gpuAvailable()) {
+        console.log('GPU backend is already active.')
+        return
+    }
+
+    const suffix = currentPlatformSuffix()
+    if (!suffix) {
+        throw new Error(
+            `No prebuilt GPU binary available for ${process.platform}-${process.arch}. ` +
+            'Build from source instead: `npx napi build --release --features gpu`.'
+        )
+    }
+
+    const assetName = `fastlowess-gpu.${suffix}.node`
+    const url = `https://github.com/${gpuDownload.REPO}/releases/download/v${version}/${assetName}`
+    const destPath = path.join(__dirname, `fastlowess.${suffix}.node`)
+
+    if (!yes) {
+        const ok = await gpuDownload.confirm(
+            `Download and install ${assetName} from github.com/${gpuDownload.REPO}? [y/N] `
+        )
+        if (!ok) {
+            console.log('Aborted.')
+            return
+        }
+    }
+
[…]
