### 2026-09-05 bc7dd97
ci: expand gpu target platforms
