### 2026-09-05 1d7f5cd
ci: rewire gpu job to test internal installer
