### 2026-09-05 79bf352
ci: link current julia to doc deployment
