### 2026-09-05 3d67926
feat: add gpu installer to java and go
