+++ bindings/go/fastlowess/gpu_installer.go
@@ -1,0 +1,142 @@
+package fastlowess
+
+// One-time downloader for the opt-in GPU-enabled fastlowess_go static
+// library. Unlike Python/Node.js/R/C++/Julia, cgo links this library
+// statically at build time, so there's no runtime install: this only
+// downloads the library, then instructs the caller to rebuild with
+// CGO_LDFLAGS pointing at it.
+
+import (
+	"bufio"
+	"fmt"
+	"io"
+	"net/http"
+	"os"
+	"path/filepath"
+	"runtime"
+	"strings"
+	"time"
+)
+
+const gpuRepo = "thisisamirv/lowess-project"
+
+// platformArchTag returns the "<platform>-<arch>" suffix release-gpu.yml
+// uses for this binding's GPU static library, or "" if this platform/arch
+// combination isn't built.
+func platformArchTag() string {
+	switch runtime.GOOS {
+	case "linux":
+		switch runtime.GOARCH {
+		case "amd64":
+			return "linux-x86_64"
+		case "arm64":
+			return "linux-arm64"
+		}
+	case "darwin":
+		switch runtime.GOARCH {
+		case "amd64":
+			return "macos-x86_64"
+		case "arm64":
+			return "macos-aarch64"
+		}
+	case "windows":
+		switch runtime.GOARCH {
+		case "amd64":
+			return "windows-x86_64"
+		case "arm64":
+			return "windows-arm64"
+		}
+	}
+	return ""
+}
+
+// InstallGPU downloads a prebuilt GPU-enabled fastlowess_go static library
+// for this platform from the matching GitHub Release, saving it to
+// ~/.fastlowess/gpu/libfastlowess_go.a.
+//
+// Pass yes=true to skip the interactive y/N confirmation prompt; this is
+// required when stdin is not an interactive terminal.
+func InstallGPU(yes bool) error {
+	if GPUEnabled() {
+		fmt.Println("GPU backend is already active.")
+		return nil
+	}
+
+	tag := platformArchTag()
+	if tag == "" {
+		return fmt.Errorf("no prebuilt GPU library available for %s/%s; build from source instead: "+
+			"cargo build -p fastlowess-go --profile release-c --features gpu", runtime.GOOS, runtime.GOARCH)
+	}
+
+	assetName := fmt.Sprintf("libfastlowess_go-gpu-v%s-%s.a", version, tag)
+	url := fmt.Sprintf("https://github.com/%s/releases/download/v%s/%s", gpuRepo, version, assetName)
+
+	if !yes {
+		fmt.Printf("Download and install %s from github.com/%s? [y/N] ", assetName, gpuRepo)
+		reader := bufio.NewReader(os.Stdin)
+		answer, _ := reader.ReadString('\n')
+		answer = strings.TrimSpace(strings.ToLower(answer))
+		if answer != "y" && answer != "yes" {
+			fmt.Println("Aborted.")
+			return nil
+		}
+	}
+
+	home, err := os.UserHomeDir()
+	if err != nil {
+		home = "."
+	}
+	dir := filepath.Join(home, ".fastlowess", "gpu")
+	if err := os.MkdirAll(dir, 0o755); err != nil {
+		return fmt.Errorf("failed to create %s: %w", dir, err)
+	}
+	dest := filepath.Join(dir, "libfastlowess_go.a")
+
+	fmt.Printf("Downloading %s ...\n", url)
+	if err := downloadGPULibrary(url, dest); err != nil {
+		return fmt.Errorf("failed to download %s: %w\n"+
[…]
