### 2026-09-05 e32485e
feat: add local gpu installer option
