+++ r/news/index.md
@@ -13,6 +13,18 @@
   configurable warm-up) to invite contributions.
 - `dev/bump_version.py` now also updates the example crate version in
   `CONTRIBUTING.md`’s “Individual crate Cargo.toml” snippet.
+- Every GPU installer (Python, Node.js, R, Julia, C++, Java, Go) now
+  accepts a local path to an already-built GPU artifact, installing from
+  it directly instead of downloading from a GitHub Release — useful for
+  testing the installer itself or installing an unreleased build.
+  `ci-*.yml`’s `gpu` jobs now build the `gpu` feature locally and
+  exercise this path end-to-end instead of depending on a matching
+  published release existing.
+- `release-gpu.yml` now uploads every GPU artifact (across all languages
+  and all versions) to a single perpetual `gpu-builds` release instead
+  of the release page for the version just published, so version release
+  pages stay uncluttered; each asset’s filename embeds its source
+  version instead. Every installer’s download URL updated to match.
 - Added a `return_sorted` option to
   [`Lowess()`](https://thisisamirv.github.io/lowess-project/r/reference/Lowess.md).
 - Added a `missing` option to
