+++ bindings/java/docs/modules/ROOT/pages/advanced/gpu-backend.adoc
@@ -13,7 +13,7 @@ This is worth enabling for high-throughput processing of large datasets (roughly
 *Batch only.* GPU support applies to the batch `Lowess` type only. `StreamingLowess`/`OnlineLowess` remain CPU-only — the Rust core documents GPU as optimized for static batch data, not incremental chunk/point processing.
 ====
 
-GPU support is *opt-in* and not included in a default build of `fastlowess-java` — it requires building the native library from source with the `gpu` Cargo feature.
+GPU support is *opt-in* and not included in a default build of `fastlowess-java` — download a prebuilt GPU-enabled build via the one-time installer below, or build the native library from source with the `gpu` Cargo feature.
 
 === Supported Features
 
@@ -76,7 +76,7 @@ Building from source with the `gpu` Cargo feature is always available as an alte
 cargo build -p fastlowess-java --release --features gpu
 ----
 
-Since `NativeBridge` loads the shared library by path (via the `fastlowess.native.dir` system property), simply point it at the GPU-enabled build — no application rebuild is required, unlike Go's static-linking approach.
+Since `NativeBridge` loads the shared library by path (via the `fastlowess.native.dir` system property), simply point it at the GPU-enabled build — no application rebuild is required.
 
 '''
 
