### 2026-09-05 e11d3b9
ci: pass WITH_GPU=1 to R gpu job
