### 2026-09-06 0dbaee9
refactor: R gpu installer to satisfy ropensci
