### 2026-09-06 3ce4492
cpp: update Spack recipe for v4.0.0
