### 2026-09-06 2264226
ci: fix gpu release move premission
