### 2026-09-06 #769 @thisisamirv
Hi @etiennebacher, just a quick note to let you know that `v4.0.0` is now out. Nothing here was raised in review, but flagging it for visibility since it touches the R binding's public API:

- Added a `return_sorted` option to `Lowess()`, to return results sorted ascending by `x` instead of input order (default `FALSE`, so existing behavior is unchanged unless you opt in).
- Added a `missing` option to `Lowess()`, `StreamingLowess()`, and `OnlineLowess()`, controlling how non-finite (`NaN`/`Inf`) `x`/`y` values are handled (`"error"` default, or `"drop"`).
- **Breaking:** removed `return_diagnostics`, `return_residuals`, and `parallel` from `OnlineLowess()`'s constructor, and `confidence_intervals`/`prediction_intervals` from both `OnlineLowess()` and `StreamingLowess()` — none of these were ever actually computed for those two modes, so this only removes dead, no-op arguments. `Lowess()` (Batch) is unaffected.
- Fixed a crashing `use-case-real-time.Rmd` example caused by an internal validator's `min_points` being stricter than the Rust core's actual minimum.
- Fixed a rare `install_gpu()` segfault: it used to overwrite the currently-loaded shared library in place, which could corrupt a still memory-mapped file; it now installs atomically via a temp file + rename.
- Expanded docs.

Full details in [CHANGELOG.md](https://github.com/thisisamirv/lowess-project/blob/main/CHANGELOG.md#400).
