### 2026-09-07 225ee0b
feat: add predict
