+++ crates/lowess/tests/lowess/engine_output_tests.rs
@@ -90,6 +90,7 @@ fn test_has_confidence_intervals_true() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: None,
+        fit_state: None,
     };
 
     assert!(
@@ -117,6 +118,7 @@ fn test_has_confidence_intervals_false() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: None,
+        fit_state: None,
     };
 
     assert!(
@@ -144,6 +146,7 @@ fn test_has_prediction_intervals_true() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: None,
+        fit_state: None,
     };
 
     assert!(
@@ -171,6 +174,7 @@ fn test_has_prediction_intervals_false() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: None,
+        fit_state: None,
     };
 
     assert!(
@@ -198,6 +202,7 @@ fn test_has_cv_scores_true() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: Some(vec![0.1, 0.2, 0.3]),
+        fit_state: None,
     };
 
     assert!(lr.has_cv_scores(), "Should have CV scores");
@@ -222,6 +227,7 @@ fn test_has_cv_scores_false() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: None,
+        fit_state: None,
     };
 
     assert!(!lr.has_cv_scores(), "Should not have CV scores");
@@ -251,6 +257,7 @@ fn test_smoothed_accessor() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: None,
+        fit_state: None,
     };
 
     assert_eq!(lr.smoothed(), &y_vals[..]);
@@ -275,6 +282,7 @@ fn test_best_cv_score_present() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: Some(vec![0.3, 0.1, 0.2, 0.5]),
+        fit_state: None,
     };
 
     let best = lr.best_cv_score();
@@ -301,6 +309,7 @@ fn test_best_cv_score_none() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: None,
+        fit_state: None,
     };
 
     assert!(lr.best_cv_score().is_none());
@@ -325,6 +334,7 @@ fn test_best_cv_score_single() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: Some(vec![0.42]),
+        fit_state: None,
     };
 
     let best = lr.best_cv_score();
@@ -355,6 +365,7 @@ fn test_confidence_width() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: None,
+        fit_state: None,
     };
 
     let widths = lr.confidence_width().expect("Should have widths");
@@ -383,6 +394,7 @@ fn test_confidence_width_none() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: None,
+        fit_state: None,
     };
 
     assert!(lr.confidence_width().is_none());
@@ -407,6 +419,7 @@ fn test_prediction_width() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: None,
+        fit_state: None,
     };
 
     let widths = lr.prediction_width().expect("Should have widths");
@@ -435,6 +448,7 @@ fn test_prediction_width_none() {
         iterations_used: None,
         fraction_used: 0.5,
         cv_scores: None,
+        fit_state: None,
     };
 
     assert!(lr.prediction_width().is_none());
@@ -459,6 +473,7 @@ fn test_prediction_wider_than_confidence() {
[…]
