+++ crates/lowess/src/engine/predict.rs
@@ -1,0 +1,346 @@
+//! Out-of-sample prediction for fitted Batch LOWESS models.
+//!
+//! This module holds the fitted-model state retained by `LowessResult::predict()`
+//! (Batch adapter only, opt-in via `.retain_model(true)`) and the logic that
+//! evaluates the local WLS fit at arbitrary query points not in the training set.
+
+// External dependencies
+#[cfg(not(feature = "std"))]
+use alloc::vec;
+#[cfg(not(feature = "std"))]
+use alloc::vec::Vec;
+use num_traits::Float;
+#[cfg(feature = "std")]
+use std::vec;
+#[cfg(feature = "std")]
+use std::vec::Vec;
+
+// Internal dependencies
+use crate::algorithms::regression::{RegressionContext, WLSSolver, ZeroWeightFallback};
+use crate::evaluation::intervals::IntervalMethod;
+use crate::math::kernel::WeightFunction;
+use crate::primitives::errors::LowessError;
+use crate::primitives::window::Window;
+
+// Policy for evaluating query points outside the retained training x-range.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
+pub enum ExtrapolationPolicy {
+    // Clamp to the nearest boundary window (default; matches `predict()`'s original behavior).
+    #[default]
+    Clamp,
+
+    // Linearly extrapolate from the nearest boundary point's local fit and slope.
+    Linear,
+
+    // Fail the whole `predict()` call with `LowessError::PredictOutOfRange` if any query
+    // point falls outside `[min(x_train), max(x_train)]`.
+    Error,
+}
+
+// Options controlling a `LowessResult::predict()` call.
+#[derive(Debug, Clone, Copy)]
+pub struct PredictOptions<T> {
+    // Include standard errors in the output.
+    pub return_se: bool,
+
+    // Confidence interval coverage level (e.g. `Some(0.95)`), or `None` to skip.
+    pub confidence_level: Option<T>,
+
+    // Prediction interval coverage level (e.g. `Some(0.95)`), or `None` to skip.
+    pub prediction_level: Option<T>,
+
+    // Include the local WLS fit's derivative (slope) at each query point.
+    pub return_derivative: bool,
+
+    // Behavior for query points outside the training x-range.
+    pub extrapolation: ExtrapolationPolicy,
+}
+
+impl<T: Float> Default for PredictOptions<T> {
+    fn default() -> Self {
+        Self {
+            return_se: false,
+            confidence_level: None,
+            prediction_level: None,
+            return_derivative: false,
+            extrapolation: ExtrapolationPolicy::default(),
+        }
+    }
+}
+
+// Result of a `LowessResult::predict()` call.
+#[derive(Debug, Clone)]
+pub struct PredictOutput<T> {
+    // Predicted y-values, one per query point in `new_x`.
+    pub y: Vec<T>,
+
+    // Standard errors, if `return_se`/`confidence_level`/`prediction_level` was requested.
+    pub standard_errors: Option<Vec<T>>,
+
+    // Confidence interval bounds for the mean response, if `confidence_level` was set.
+    pub confidence_lower: Option<Vec<T>>,
+    pub confidence_upper: Option<Vec<T>>,
+
+    // Prediction interval bounds for a new observation, if `prediction_level` was set.
+    pub prediction_lower: Option<Vec<T>>,
+    pub prediction_upper: Option<Vec<T>>,
+
+    // Local WLS fit's derivative (slope) at each query point, if `return_derivative` was set.
+    pub derivative: Option<Vec<T>>,
+}
+
+// Per-point predict results before shared confidence/prediction interval math is applied:
[…]
