+++ crates/lowess/src/algorithms/defaults.rs
@@ -38,3 +38,6 @@ pub const DEFAULT_RETURN_ROBUSTNESS_WEIGHTS: bool = false;
 
 // False means return results in the original input order (Batch adapter only).
 pub const DEFAULT_RETURN_SORTED: bool = false;
+
+// False means `fit()` does not retain the fitted model state needed by `predict()` (Batch only).
+pub const DEFAULT_RETAIN_MODEL: bool = false;
