### 2026-09-07 ad6e8be
fix: flatten rust test tree
