+++ crates/fastLowess/tests/fastLowess/main.rs
@@ -1,7 +1,0 @@
-#![cfg(feature = "dev")]
-mod gpu_tests;
-mod integration_tests;
-mod parallel_eval_tests;
-mod parallel_interval_tests;
-mod predict_pass_consistency;
-mod smooth_pass_consistency;
