### 2026-09-07 48bb56f
fix: go versioning
