+++ dev/bump_version.py
@@ -6,7 +6,10 @@ fastLowess/lowess path-dependency version requirements (major.minor), each
 binding's own version file (package.json, pyproject-adjacent __version__.py,
 pom.xml, DESCRIPTION, Project.toml (incl. the fastlowess_jll compat floor),
 version.go, CMakeLists.txt, FastLowess.java), CITATION.cff, and the Spack
-recipe's example `url`.
+recipe's example `url`. Also updates the Go module's `/vN` major-version
+suffix (go.mod files, doc snippets, README/docs badges, the doc-snippet
+runner) whenever a major version bump changes it -- see
+https://go.dev/ref/mod#major-version-suffixes.
 
 Does NOT touch: CHANGELOG.md (write that by hand), generated NEWS.md/docs-site
 content (regenerated via `make <lang>-dev` / dev/update_changelogs.py), or the
@@ -29,6 +32,18 @@ from pathlib import Path
 REPO_ROOT = Path(__file__).resolve().parent.parent
 
 VERSION_RE = re.compile(r"^\d+\.\d+\.\d+$")
+
+# Go requires any module tagged v2.0.0+ to end its module path with /vN, or
+# the Go toolchain (and pkg.go.dev) silently ignores those tags and falls
+# back to commit pseudo-versions. These are the files that reference the Go
+# module's import path and must be kept in sync with its current major.
+GO_MODULE_BASE_PATH = "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+GO_FASTLOWESS_GOMOD = "bindings/go/fastlowess/go.mod"
+GO_TESTS_GOMOD = "bindings/go/tests/go.mod"
+GO_TESTS_MAIN = "bindings/go/tests/lowess_test.go"
+GO_DOC_RUNNER = "dev/runners/go.py"
+GO_README = "bindings/go/README.md"
+GO_DOCS_INDEX = "bindings/go/docs/_index.md"
 
 # Cargo.toml files whose bare `[package]` `version = "X.Y.Z"` line should track
 # the release version 1:1.
@@ -94,6 +109,126 @@ def _replace(
     verb = "Would update" if dry_run else "Updated"
     print(f"  {verb} ({n}x): {rel}")
     return True
+
+
+def _replace_literal_all(
+    path: Path, old: str, new: str, dry_run: bool, required: bool = True
+) -> bool:
+    """Replace every literal occurrence of `old` with `new` in `path`.
+
+    When `required` is False, a file with no occurrences is not an error (used
+    for glob-discovered doc pages that may not contain the pattern at all).
+    """
+    rel = path.relative_to(REPO_ROOT)
+    if not path.exists():
+        print(f"  SKIP (missing): {rel}")
+        return False
+    text = path.read_text(encoding="utf-8")
+    count = text.count(old)
+    if count == 0:
+        if not required:
+            return True
+        print(f"  WARNING: no match in {rel}")
+        return False
+    if not dry_run:
+        path.write_text(text.replace(old, new), encoding="utf-8")
+    verb = "Would update" if dry_run else "Updated"
+    print(f"  {verb} ({count}x): {rel}")
+    return True
+
+
+def _current_go_module_suffix() -> str:
+    """Return the Go module's current major-version suffix, e.g. '/v4', or '' for v0/v1."""
+    gomod = REPO_ROOT / GO_FASTLOWESS_GOMOD
+    text = gomod.read_text(encoding="utf-8")
+    m = re.search(
+        rf"^module {re.escape(GO_MODULE_BASE_PATH)}(/v\d+)?$", text, re.MULTILINE
+    )
+    if m is None:
[…]
