+++ bindings/go/docs/weighting/kernels.md
@@ -45,7 +45,7 @@ import (
  "log"
  "math"
 
- "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+ "github.com/thisisamirv/lowess-project/bindings/go/fastlowess/v4"
 )
 
 func main() {
@@ -95,7 +95,7 @@ import (
  "log"
  "math"
 
- "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+ "github.com/thisisamirv/lowess-project/bindings/go/fastlowess/v4"
 )
 
 func main() {
@@ -145,7 +145,7 @@ import (
  "log"
  "math"
 
- "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+ "github.com/thisisamirv/lowess-project/bindings/go/fastlowess/v4"
 )
 
 func main() {
@@ -195,7 +195,7 @@ import (
  "log"
  "math"
 
- "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+ "github.com/thisisamirv/lowess-project/bindings/go/fastlowess/v4"
 )
 
 func main() {
@@ -245,7 +245,7 @@ import (
  "log"
  "math"
 
- "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+ "github.com/thisisamirv/lowess-project/bindings/go/fastlowess/v4"
 )
 
 func main() {
@@ -295,7 +295,7 @@ import (
  "log"
  "math"
 
- "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+ "github.com/thisisamirv/lowess-project/bindings/go/fastlowess/v4"
 )
 
 func main() {
@@ -345,7 +345,7 @@ import (
  "log"
  "math"
 
- "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+ "github.com/thisisamirv/lowess-project/bindings/go/fastlowess/v4"
 )
 
 func main() {
