+++ bindings/go/docs/use-case/use-case-real-time.md
@@ -27,7 +27,7 @@ import (
  "log"
  "math"
 
- "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+ "github.com/thisisamirv/lowess-project/bindings/go/fastlowess/v4"
 )
 
 func main() {
@@ -92,7 +92,7 @@ import (
  "log"
  "math"
 
- "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+ "github.com/thisisamirv/lowess-project/bindings/go/fastlowess/v4"
 )
 
 func main() {
@@ -152,7 +152,7 @@ import (
  "log"
  "math"
 
- "github.com/thisisamirv/lowess-project/bindings/go/fastlowess"
+ "github.com/thisisamirv/lowess-project/bindings/go/fastlowess/v4"
 )
 
 func main() {
