### 2026-09-07 b5a0218
fix: java badge in readme
