### 2026-09-07 8e7cfe2
fix: nodejs/js naming inconsistency
