+++ bindings/java/docs/modules/ROOT/pages/index.adoc
@@ -15,7 +15,7 @@
 </p>
 ++++
 
-The fastest, most robust, and most feature-complete language-agnostic LOWESS (Locally Weighted Scatterplot Smoothing) implementation for *Rust*, *Python*, *R*, *Julia*, *JavaScript*, *C++*, *Go*, *Java*, and *WebAssembly*.
+The fastest, most robust, and most feature-complete language-agnostic LOWESS (Locally Weighted Scatterplot Smoothing) implementation for *Rust*, *Python*, *R*, *Julia*, *Node.js*, *C++*, *Go*, *Java*, and *WebAssembly*.
 
 The `lowess-project` also offers bindings for Rust, Python, R, Julia, Node.js, WebAssembly, C++, Go, and Java — see the https://github.com/thisisamirv/lowess-project[full repository].
 
