+++ .github/workflows/release-go.yml
@@ -19,8 +19,9 @@ permissions:
 
 jobs:
   build-and-upload:
-    name: Build ${{ matrix.os }}
+    name: Build ${{ matrix.os }}${{ matrix.container && ' (musl)' || '' }}
     runs-on: ${{ matrix.os }}
+    container: ${{ matrix.container }}
     strategy:
       matrix:
         include:
@@ -28,8 +29,18 @@ jobs:
             asset_name: libfastlowess_go-linux-x64.a
             cargo_target: ""
             target_dir: target/release
+          - os: ubuntu-latest
+            container: "alpine:latest"
+            asset_name: libfastlowess_go-linux-x64-musl.a
+            cargo_target: ""
+            target_dir: target/release
           - os: ubuntu-24.04-arm
             asset_name: libfastlowess_go-linux-arm64.a
+            cargo_target: ""
+            target_dir: target/release
+          - os: ubuntu-24.04-arm
+            container: "alpine:latest"
+            asset_name: libfastlowess_go-linux-arm64-musl.a
             cargo_target: ""
             target_dir: target/release
           - os: macos-15-intel
@@ -50,6 +61,10 @@ jobs:
             target_dir: target/aarch64-pc-windows-gnullvm/release
 
     steps:
+      - name: Install Alpine build dependencies
+        if: matrix.container
+        run: apk add --no-cache build-base curl bash git
+
       - uses: actions/checkout@v7
         with:
           ref: ${{ github.event.inputs.commit || github.sha }}
@@ -96,7 +111,7 @@ jobs:
           files: ${{ matrix.asset_name }}
 
       - name: Upload Header to Release
-        if: matrix.os == 'ubuntu-latest'
+        if: matrix.os == 'ubuntu-latest' && !matrix.container
         uses: softprops/action-gh-release@v3
         with:
           tag_name: ${{ github.event.release.tag_name || github.event.inputs.tag }}
