### 2026-09-08 901d05e
feat(java): bundle prebuilt native libraries for all platforms in the published jar
