+++ bindings/java/docs/modules/ROOT/pages/introduction/installation.adoc
@@ -3,7 +3,39 @@
 == Requirements
 
 * Java 17+ (this binding targets Java 25 LTS by default; see the https://github.com/thisisamirv/lowess-project/blob/main/bindings/java/pom.xml[pom.xml] `maven.compiler.release` property)
-* A prebuilt copy of the native `fastlowess_java` library (JNI shared library) on the JVM's native library search path
+
+== As a standalone dependency
+
+Add the Maven coordinate to your `pom.xml`:
+
+[source,xml]
+----
+<dependency>
+    <groupId>io.github.thisisamirv</groupId>
+    <artifactId>fastlowess</artifactId>
+    <version>4.0.0</version>
+</dependency>
+----
+
+The published jar bundles prebuilt native libraries for `linux-x86_64`, `linux-aarch64`, `macos-x86_64`, `macos-aarch64`, `windows-x86_64`, and `windows-aarch64` (see `NativeBridge`'s `osArchDir()`); `NativeBridge` auto-detects your platform, extracts the matching one to a temp file, and loads it — no separate native library installation needed on any of those platforms. (musl/Alpine Linux isn't one of the bundled targets yet, since `NativeBridge` doesn't currently distinguish glibc from musl.)
+
+To override the bundled library (e.g. a custom or GPU-enabled build), either system property takes precedence:
+
+[source,sh]
+----
+java -Dfastlowess.native.path=/path/to/libfastlowess_java.so -cp your-app.jar com.example.Main
+# or point at a directory instead of an exact file:
+java -Dfastlowess.native.dir=/path/to/native/lib -cp your-app.jar com.example.Main
+----
+
+Build the native library yourself from the https://github.com/thisisamirv/lowess-project[`lowess-project`] source:
+
+[source,sh]
+----
+git clone https://github.com/thisisamirv/lowess-project
+cd lowess-project
+cargo build -p fastlowess-java --release
+----
 
 == Within the `lowess-project` monorepo
 
@@ -17,36 +49,7 @@ make java-dev     # full dev checks: fmt, clippy, export verification, mvn test
 
 `NativeBridge` loads the native library from the path given by the `fastlowess.native.dir` system property (or `-Dfastlowess.native.dir=...`), which defaults to `../../target/debug` relative to `bindings/java` — where `cargo build -p fastlowess-java` places the shared library within this repo's layout.
 
-== As a standalone dependency
-
-Outside the monorepo, add the Maven coordinate to your `pom.xml`:
-
-[source,xml]
-----
-<dependency>
-    <groupId>io.github.thisisamirv</groupId>
-    <artifactId>fastlowess</artifactId>
-    <version>3.1.0</version>
-</dependency>
-----
-
-Then point the JVM at a prebuilt native library, either via the `fastlowess.native.dir` system property or by placing it on `java.library.path`:
-
-[source,sh]
-----
-java -Dfastlowess.native.dir=/path/to/native/lib -cp your-app.jar com.example.Main
-----
-
-Build the native library yourself from the https://github.com/thisisamirv/lowess-project[`lowess-project`] source:
-
-[source,sh]
-----
-git clone https://github.com/thisisamirv/lowess-project
-cd lowess-project
-cargo build -p fastlowess-java --release
-----
-
 == GPU backend
 
-GPU acceleration is available via the native library's `gpu` Cargo feature (`wgpu`: Vulkan/Metal/DX12), but is not enabled in the default build. See the C++ binding's https://thisisamirv.github.io/lowess-project/cpp/gpu-backend.html[GPU Backend guide] for the underlying feature; building `fastlowess-java` with `--features gpu` produces a library where `FastLowess.gpuEnabled()` returns `true`.
[…]
